### 2026-02-17 04b86d3
Added a function invokation tracker. Added overloads to the "append" extension to support number/string arrays.
