### 2026-02-17 2e91b22
Added live function invokation update support in the base interpreter.
