+++ AresScript.Tests/Program.cs
@@ -1,10 +1,10 @@
-using System.Diagnostics;
 using Antlr4.Runtime;
 using Ares.Datamodel.Scripting;
 using AresScript.Generated;
 using AresScript.Interpreters;
 using AresScript.ScriptAnalysis;
 using NUnit.Framework;
+using System.Diagnostics;
 
 namespace AresScript.Tests;
 
@@ -41,12 +41,18 @@ public class InterpreterTests
     }
   }
 
-  private static Task RunScriptAsync(string script, CancellationToken cancellationToken = default)
+  private static Task RunScriptAsync(
+    string script,
+    CancellationToken cancellationToken = default,
+    Action<AresFunctionInvocation>? invocationObserver = null)
   {
-    return RunScriptAsync(script, new ScriptExecutionControlToken(cancellationToken));
+    return RunScriptAsync(script, new ScriptExecutionControlToken(cancellationToken), invocationObserver);
   }
 
-  private static async Task RunScriptAsync(string script, ScriptExecutionControlToken executionControlToken)
+  private static async Task RunScriptAsync(
+    string script,
+    ScriptExecutionControlToken executionControlToken,
+    Action<AresFunctionInvocation>? invocationObserver = null)
   {
     var stream = new AntlrInputStream(script);
     var lexer = new AresIndentationLexer(stream);
@@ -60,7 +66,7 @@ public class InterpreterTests
     var env = new AresScriptEnvironment();
     env.AssignSystemFunctions(StandardLibrary.Functions);
     env.AssignExtensionFunctions(StandardLibrary.ExtensionFunctions);
-    var visitor = new AresBaseInterpreter(env, executionControlToken);
+    var visitor = new AresBaseInterpreter(env, executionControlToken, invocationObserver);
 
     await visitor.Visit(programCtx);
   }
@@ -93,6 +99,13 @@ public class InterpreterTests
     var (invocations, diagnostics) = await AresScriptAnalysis.ValidateAndCollectInvocationsAsync(script, env);
     Assert.That(diagnostics, Is.Empty);
     return invocations;
+  }
+
+  private static async Task<AresFunctionInvocation[]> RunAndCollectRuntimeInvocationsAsync(string script)
+  {
+    var invocations = new List<AresFunctionInvocation>();
+    await RunScriptAsync(script, invocationObserver: invocations.Add);
+    return invocations.ToArray();
   }
 
   private static async Task<ScriptSummaryStep[]> BuildScriptSummaryAsync(
@@ -464,13 +477,13 @@ public class InterpreterTests
       """;
 
     var invocations = await ValidateAndCollectInvocationsAsync(script);
-    Assert.That(invocations.Select(i => i.FunctionId), Is.EqualTo(new[] { "range", "id", "print" }));
-    Assert.That(invocations.Select(i => i.Kind), Is.EqualTo(new[]
-    {
+    Assert.That(invocations.Select(i => i.FunctionId), Is.EqualTo(["range", "id", "print"]));
+    Assert.That(invocations.Select(i => i.Kind), Is.EqualTo(
+    [
       AresFunctionInvocationKind.System,
       AresFunctionInvocationKind.User,
       AresFunctionInvocationKind.System
-    }));
+    ]));
   }
 
   [Test]
@@ -506,6 +519,53 @@ public class InterpreterTests
   }
 
   [Test]
+  public async Task Runtime_Collects_Ordered_System_User_And_Extension_Function_Invocations()
+  {
+    var script = """
+      def id(v):
+        return v
+
+      values = range(3)
+      items = []
+      items.append(values)
[…]
