+++ Ares.Device.Serial/AresSerial/SerialDeviceValidationResult.cs
@@ -1,3 +1,0 @@
-﻿namespace Ares.Device.Serial;
-
-public record SerialDeviceValidationResult(bool Success, string Message = "");
