+++ Ares.Core/Device/DeviceDriverRepo.cs
@@ -1,6 +1,5 @@
-using System;
+using Ares.Core.Device.Repos;
 using System.Collections.Concurrent;
-using System.Collections.Generic;
 
 namespace Ares.Core.Device;
 
