+++ AresService/ConfigManagers/StepperControllerConfigManager.cs
@@ -5,7 +5,7 @@ using TicStepperController;
 using TicStepperController.Config;
 
 namespace AresService.ConfigManagers;
-public class StepperControllerConfigManager : DeviceConfigManagerBase<StepperControllerConfig, IStepperController>
+public class StepperControllerConfigManager : DeviceConfigManager<StepperControllerConfig, IStepperController>
 {
   public StepperControllerConfigManager(IDbContextFactory<CoreDatabaseContext> dbContextFactory)
     : base(dbContextFactory)
