+++ Ares.Device.Serial/AresSerial/IAresDeviceConnection.cs
@@ -1,5 +1,0 @@
-﻿namespace Ares.Device.Serial;
-
-public interface IAresDeviceConnection
-{
-}
