+++ Ares.Device.Serial.Tests/SerialPortTests.cs
@@ -1,485 +1,0 @@
-﻿using System.IO.Ports;
-using System.Reactive.Linq;
-using System.Reactive.Threading.Tasks;
-using System.Text;
-using Ares.Device.Serial.Commands;
-using Ares.Device.Serial.Simulation;
-
-namespace Ares.Device.Serial.Tests;
-
-internal class SerialPortTests
-{
-  [Test]
-  [CancelAfter(5000)]
-  public async Task AresSerialPort_Returns_Good_Response_From_Simple_Request(CancellationToken token)
-  {
-    const string stringToTest = "<-Oh Hello->";
-    var port = new TestConnectionWithDelay(new SerialPortConnectionInfo(0, Parity.Even, 0, StopBits.None));
-    var response = await port.Send(new SomeCommandWithResponse(stringToTest), token);
-    // Assert.That(await port.DataBufferState.FirstAsync(), Is.Empty);
-    Assert.That(response, Is.Not.Null);
-    Assert.That(response.Response, Is.EqualTo(stringToTest));
-    // var currentBuffer = await port.DataBufferState.FirstAsync();
-    // Assert.That(currentBuffer, Is.Empty);
-  }
-
-  [Test]
-  [CancelAfter(5000)]
-  public async Task AresSerialPort_Returns_Good_Response_From_Multiple_Data_Adds(CancellationToken token)
-  {
-    const string stringToTest = "<-Oh Hello->";
-    var port = new TestPort2(new SerialPortConnectionInfo(0, Parity.Even, 0, StopBits.None));
-    var response = await port.Send(new SomeCommandWithResponse(stringToTest), token);
-    // Assert.That(await port.DataBufferState.FirstAsync(), Is.Empty);
-    Assert.That(response, Is.Not.Null);
-    Assert.That(response.Response, Is.EqualTo(stringToTest));
-    // var currentBuffer = await port.DataBufferState.FirstAsync();
-    // Assert.That(currentBuffer, Is.Empty);
-  }
-
-  [Test]
-  [CancelAfter(5000)]
-  public async Task AresSerialPort_Returns_Good_Response_From_Multiple_Data_And_Commands(CancellationToken token)
-  {
-    const string stringToTest = "<-Oh Hello->";
-    const string stringToTest2 = "<-Noice->";
-    const string stringToTest3 = "<-This Is A Test->";
-    var port = new TestPort2(new SerialPortConnectionInfo(0, Parity.Even, 0, StopBits.None));
-    var test1 = await port.Send(new SomeCommandWithResponse(stringToTest), token);
-    // Assert.That(await port.DataBufferState.FirstAsync(), Is.Empty);
-    var test2 = await port.Send(new SomeCommandWithResponse(stringToTest2), token);
-    // Assert.That(await port.DataBufferState.FirstAsync(), Is.Empty);
-    var test3 = await port.Send(new SomeCommandWithResponse(stringToTest3), token);
-    // Assert.That(await port.DataBufferState.FirstAsync(), Is.Empty);
-    Assert.Multiple(() =>
-    {
-      Assert.That(test1, Is.Not.Null);
-      Assert.That(test2, Is.Not.Null);
-      Assert.That(test3, Is.Not.Null);
-    });
-
-    Assert.Multiple(() =>
-    {
-      Assert.That(test1.Response, Is.EqualTo(stringToTest));
-      Assert.That(test2.Response, Is.EqualTo(stringToTest2));
-      Assert.That(test3.Response, Is.EqualTo(stringToTest3));
-    });
-
-    // Assert.That(await port.DataBufferState.FirstAsync(), Is.Empty);
-  }
-
-  [Test]
-  [CancelAfter(5000)]
[…]

[…]
