### 2026-02-17 74712e4
Separated the function invocation mapper from the summary mapper class
