+++ AresScript/ScriptAnalysis/AresScriptSummaryMapper.cs
@@ -12,16 +12,7 @@ internal static class AresScriptSummaryMapper
   {
     return invocations
       .Where(invocation => ShouldInclude(invocation.Kind, includeUserFunctions, includeLambdas))
-      .Select((invocation, index) => new ScriptSummaryStep
-      {
-        Order = index + 1,
-        FunctionId = invocation.FunctionId,
-        FunctionName = invocation.FunctionName,
-        Expression = invocation.Expression,
-        Line = invocation.Line,
-        Column = invocation.Column,
-        Kind = MapKind(invocation.Kind)
-      })
+      .Select((invocation, index) => AresFunctionInvocationMapper.ToScriptSummaryStep(invocation, index + 1))
       .ToArray();
   }
 
@@ -37,18 +28,6 @@ internal static class AresScriptSummaryMapper
       AresFunctionInvocationKind.User => includeUserFunctions,
       AresFunctionInvocationKind.Lambda => includeLambdas,
       _ => false
-    };
-  }
-
-  private static FunctionInvocationKind MapKind(AresFunctionInvocationKind kind)
-  {
-    return kind switch
-    {
-      AresFunctionInvocationKind.System => FunctionInvocationKind.System,
-      AresFunctionInvocationKind.Extension => FunctionInvocationKind.Extension,
-      AresFunctionInvocationKind.User => FunctionInvocationKind.User,
-      AresFunctionInvocationKind.Lambda => FunctionInvocationKind.Lambda,
-      _ => FunctionInvocationKind.Unspecified
     };
   }
 }
