### 2026-02-17 7117b1d
Added realtime script event invokation streaming support.
