+++ Ares.Core.Grpc/Services/AresScriptingService.cs
@@ -1,14 +1,16 @@
-using System.Threading.Tasks;
-using Ares.Services;
-using Grpc.Core;
-using System;
-using System.Threading.Channels;
 using Ares.Core.Scripting;
-using Microsoft.Extensions.Logging;
-using Google.Protobuf.WellKnownTypes;
-using AresScript.ScriptAnalysis;
-using Ares.Datamodel.Scripting;
+using Ares.Services;
 using AresScript.Interpreters;
+using AresScript.ScriptAnalysis;
+using Google.Protobuf.WellKnownTypes;
+using Grpc.Core;
+using Microsoft.Extensions.Logging;
+using System;
+using System.Reactive.Disposables;
+using System.Threading.Channels;
+using System.Threading.Tasks;
+using CoreScriptExecutionEvent = Ares.Core.Scripting.ScriptExecutionEvent;
+using GrpcScriptExecutionEvent = Ares.Services.ScriptExecutionEvent;
 
 namespace Ares.Core.Grpc.Services;
 
@@ -23,25 +25,31 @@ public partial class AresScriptingService : Ares.Services.AresScriptingService.A
     _environmentBuilder = environmentBuilder;
 
   }
-  public override async Task ExecuteScript(ScriptExecutionRequest request, IServerStreamWriter<ScriptExecutionOutput> responseStream, ServerCallContext context)
+  public override async Task ExecuteScript(ScriptExecutionRequest request, IServerStreamWriter<GrpcScriptExecutionEvent> responseStream, ServerCallContext context)
   {
-    var channel = Channel.CreateBounded<string>(new BoundedChannelOptions(100));
+    var channel = Channel.CreateBounded<GrpcScriptExecutionEvent>(new BoundedChannelOptions(100));
     var env = _environmentBuilder.Build();
     var runner = new ScriptRunner(env);
-    runner.ScriptOutput.Subscribe(output =>
+    var subscriptions = new CompositeDisposable
     {
-      if(!channel.Writer.TryWrite(output))
-      {
-        _logger.LogWarning("Dropped script output because channel is full. {Output}", output);
-      }
-    });
+      runner.ScriptEvents.Subscribe(executionEvent =>
+        {
+          var grpcEvent = ToGrpcScriptExecutionEvent(executionEvent);
+          Console.WriteLine(grpcEvent);
+          if(!channel.Writer.TryWrite(grpcEvent))
+          {
+            _logger.LogWarning("Dropped script event because channel is full. {Sequence}", grpcEvent.Sequence);
+          }
+        })
+    };
+
     async Task ReadOutputAsync()
     {
       try
       {
         await foreach(var val in channel.Reader.ReadAllAsync(context.CancellationToken))
         {
-          await responseStream.WriteAsync(new ScriptExecutionOutput { Output = val });
+          await responseStream.WriteAsync(val);
         }
       }
       catch(OperationCanceledException) when(context.CancellationToken.IsCancellationRequested)
@@ -66,11 +74,15 @@ public partial class AresScriptingService : Ares.Services.AresScriptingService.A
     }
     catch(Exception e)
     {
-      channel.Writer.TryWrite($"Run failed: {e}");
       channel.Writer.TryComplete(e);
       _logger.LogError("Script runner failed: {Exception}", e);
       throw;
     }
+    finally
+    {
+      subscriptions.Dispose();
+    }
+
     await readTask;
   }
 
@@ -109,5 +121,62 @@ public partial class AresScriptingService : Ares.Services.AresScriptingService.A
     var response = new ValidateScriptResponse();
     response.Diagnostics.AddRange(diagnostics);
     return response;
+  }
+
[…]
