+++ Ares.Core/Scripting/ScriptRunner.cs
@@ -1,5 +1,3 @@
-using System.Reactive.Linq;
-using System.Reactive.Subjects;
 using Antlr4.Runtime;
 using Ares.Datamodel;
 using Ares.Datamodel.Extensions;
@@ -7,25 +5,32 @@ using Ares.Datamodel.Factories;
 using AresScript;
 using AresScript.Generated;
 using AresScript.Interpreters;
+using AresScript.ScriptAnalysis;
+using System.Reactive.Linq;
+using System.Reactive.Subjects;
 
 namespace Ares.Core.Scripting;
 
 public class ScriptRunner
 {
-  private readonly ISubject<string> _outputSubject = new Subject<string>();
+  private readonly Subject<string> _outputSubject = new();
   private readonly ISubject<AresFunctionInvocation> _invocationSubject = Subject.Synchronize(new Subject<AresFunctionInvocation>());
+  private readonly ISubject<ScriptExecutionEvent> _eventSubject = Subject.Synchronize(new Subject<ScriptExecutionEvent>());
   private readonly AresScriptEnvironment _initialEnvironment;
+  private long _eventSequence;
 
   public ScriptRunner(AresScriptEnvironment? initialEnvironment = null)
   {
     ScriptOutput = _outputSubject.AsObservable();
     ScriptInvocations = _invocationSubject.AsObservable();
+    ScriptEvents = _eventSubject.AsObservable();
     Print = new AresSystemFunction("print", "print", (args, _) =>
     {
       var stringy = args.Select(v => v.Stringify());
       foreach(var s in stringy)
       {
         _outputSubject.OnNext(s);
+        PublishEvent(sequence => new ScriptConsoleOutputEvent(sequence, s));
       }
 
       return Task.FromResult(AresValueHelper.CreateUnit());
@@ -57,13 +62,59 @@ public class ScriptRunner
     env.EnterSystemScope("SandboxRunner");
     env.AssignSystemFunctions(Print);
     env.AssignExtensionFunctions(StandardLibrary.ExtensionFunctions);
-    var visitor = new AresBaseInterpreter(env, executionControlToken, invocation => _invocationSubject.OnNext(invocation));
+    var visitor = new AresBaseInterpreter(
+      env,
+      executionControlToken,
+      invocation => _invocationSubject.OnNext(invocation),
+      executionEvent =>
+      {
+        switch(executionEvent.Kind)
+        {
+          case AresFunctionExecutionEventKind.Started:
+            PublishEvent(sequence => new ScriptFunctionStartedEvent(
+              sequence,
+              executionEvent.CallId,
+              executionEvent.ParentCallId,
+              AresFunctionInvocationMapper.ToScriptFunctionInvocation(executionEvent.Invocation, 0)));
+            break;
+          case AresFunctionExecutionEventKind.Completed:
+            PublishEvent(sequence => new ScriptFunctionCompletedEvent(
+              sequence,
+              executionEvent.CallId,
+              executionEvent.Result ?? AresValueHelper.CreateUnit(),
+              executionEvent.Result?.Stringify() ?? string.Empty));
+            break;
+          case AresFunctionExecutionEventKind.Failed:
+            PublishEvent(sequence => new ScriptFunctionFailedEvent(
+              sequence,
+              executionEvent.CallId,
+              executionEvent.Error ?? string.Empty));
+            break;
+        }
+      });
 
-    await visitor.Visit(programCtx);
+    PublishEvent(sequence => new ScriptExecutionStartedEvent(sequence));
+    try
+    {
+      await visitor.Visit(programCtx);
+      PublishEvent(sequence => new ScriptExecutionCompletedEvent(sequence));
[…]
