+++ UI/Features/ScriptPlayground/ScriptPlaygroundViewModel.cs
@@ -1,3 +1,4 @@
+using Ares.Datamodel.Scripting;
 using Ares.Services;
 using Grpc.Core;
 using ReactiveUI;
@@ -11,13 +12,15 @@ public partial class ScriptPlaygroundViewModel : ReactiveObject
 {
   private readonly AresScriptingService.AresScriptingServiceClient _scriptingClient;
   private CancellationTokenSource _cancellationTokenSource = new();
-  private readonly ISubject<string> _scriptOutput = new Subject<string>();
+  private readonly Subject<string> _scriptOutput = new();
+  private readonly Subject<ScriptFunctionInvocation> _scriptInvocations = new();
 
   public ScriptPlaygroundViewModel(
     AresScriptingService.AresScriptingServiceClient scriptingClient)
   {
     _scriptingClient = scriptingClient;
     ScriptOutput = _scriptOutput.AsObservable();
+    ScriptInvocations = _scriptInvocations.AsObservable();
   }
 
   public async Task StartScript(string script)
@@ -31,7 +34,22 @@ public partial class ScriptPlaygroundViewModel : ReactiveObject
     {
       await foreach(var output in something.ResponseStream.ReadAllAsync(token))
       {
-        _scriptOutput.OnNext(output.Output);
+        if(output.FunctionStarted is not null)
+        {
+          _scriptInvocations.OnNext(output.FunctionStarted.Invocation);
+          continue;
+        }
+
+        if(output.ConsoleOutput is not null && !string.IsNullOrEmpty(output.ConsoleOutput.Output))
+        {
+          _scriptOutput.OnNext(output.ConsoleOutput.Output);
+          continue;
+        }
+
+        if(output.ExecutionFailed is not null && !string.IsNullOrEmpty(output.ExecutionFailed.Error))
+        {
+          _scriptOutput.OnNext($"Run failed: {output.ExecutionFailed.Error}");
+        }
       }
     }
     catch(RpcException)
@@ -52,4 +70,5 @@ public partial class ScriptPlaygroundViewModel : ReactiveObject
   public partial bool ScriptRunning { get; private set; }
 
   public IObservable<string> ScriptOutput { get; }
+  public IObservable<ScriptFunctionInvocation> ScriptInvocations { get; }
 }
