### 2026-02-17 2d2f37a
Added flags to prevent script execution to always emit events if there are no subscribers. Should improve performance when scripts are running in the background.
