### 2026-02-18 789348c
Added an ability to ignore function block traversals when building script summaries. They can just get added later once the script actually runs.
