### 2026-02-18 1b5d2ca
RestDevice and SerialRestDevice are now building, but likely broken at runtime and will need more intense reworks later. Low priority for right now.
