+++ Ares.Device/AresDevicePluginBase/AresStateBuilder.cs
@@ -1,197 +1,0 @@
-﻿using Ares.Datamodel;
-using System;
-using System.Collections.Generic;
-
-namespace Ares.Device;
-
-/// <summary>
-/// Fluent builder for constructing <see cref="AresStruct"/> instances.
-/// 
-/// This builder is designed to support immutable-by-convention state updates:
-/// callers create new state objects rather than mutating existing ones.
-/// </summary>
-public class AresStateBuilder
-{
-  private readonly AresStruct _struct = new AresStruct();
-
-  /// <summary>
-  /// Creates a new, empty <see cref="AresStateBuilder"/>.
-  /// </summary>
-  /// <returns>A new builder instance.</returns>
-  public static AresStateBuilder Create() => new AresStateBuilder();
-
-  /// <summary>
-  /// Builds and returns the underlying <see cref="AresStruct"/>.
-  /// 
-  /// After calling this method, the returned struct should be treated as immutable
-  /// by convention and not modified directly.
-  /// </summary>
-  /// <returns>The constructed <see cref="AresStruct"/>.</returns>
-  public AresStruct Build() => _struct;
-
-  /// <summary>
-  /// Adds or replaces a string field in the state.
-  /// 
-  /// If <paramref name="value"/> is <c>null</c>, an empty string is stored.
-  /// </summary>
-  /// <param name="key">The field name.</param>
-  /// <param name="value">The string value.</param>
-  /// <returns>The current builder instance.</returns>
-  public AresStateBuilder Add(string key, string? value)
-  {
-    _struct.Fields[key] = new AresValue { StringValue = value ?? "" };
-    return this;
-  }
-
-  /// <summary>
-  /// Adds or replaces a boolean field in the state.
-  /// </summary>
-  /// <param name="key">The field name.</param>
-  /// <param name="value">The boolean value.</param>
-  /// <returns>The current builder instance.</returns>
-  public AresStateBuilder Add(string key, bool value)
-  {
-    _struct.Fields[key] = new AresValue { BoolValue = value };
-    return this;
-  }
-
-  /// <summary>
-  /// Adds or replaces a numeric (double) field in the state.
-  /// </summary>
-  /// <param name="key">The field name.</param>
-  /// <param name="value">The numeric value.</param>
-  /// <returns>The current builder instance.</returns>
-  public AresStateBuilder Add(string key, double value)
-  {
-    _struct.Fields[key] = new AresValue { NumberValue = value };
-    return this;
-  }
-
-  /// <summary>
-  /// Adds or replaces a numeric field using an integer value.
-  /// 
-  /// The value is stored as a double in the underlying state representation.
-  /// </summary>
-  /// <param name="key">The field name.</param>
-  /// <param name="value">The integer value.</param>
-  /// <returns>The current builder instance.</returns>
-  public AresStateBuilder Add(string key, int value) => Add(key, (double)value);
-
-  /// <summary>
-  /// Adds or replaces a nested struct field.
-  /// 
-  /// The provided <paramref name="builderAction"/> is used to construct the
-  /// nested <see cref="AresStruct"/> via a child builder.
-  /// </summary>
-  /// <param name="key">The field name.</param>
-  /// <param name="builderAction">Action used to populate the nested struct.</param>
-  /// <returns>The current builder instance.</returns>
-  public AresStateBuilder AddStruct(string key, Action<AresStateBuilder> builderAction)
-  {
-    var childBuilder = new AresStateBuilder();
-    builderAction(childBuilder);
-    _struct.Fields[key] = new AresValue { StructValue = childBuilder.Build() };
-    return this;
-  }
-
-  /// <summary>
[…]
