+++ Ares.Device.Tests/Device/TestDeviceCommand.cs
@@ -1,8 +1,0 @@
-﻿namespace Ares.Device.Tests.Device;
-
-public enum TestDeviceCommand
-{
-  Record,
-  Record2,
-  Record3
-}
