+++ Ares.Device.UiKit/Messages/DeviceDeletionMessage.cs
@@ -1,3 +1,0 @@
-﻿namespace Ares.Device.UiKit.Messages;
-
-public record DeviceDeletedMessage(string DeviceId);
