### 2026-02-19 16f3054
Lots of items removed. Started to work on gutting the service and UI of old logic. Working on building the driver sync service and associated logic.
