+++ AresService/Services/Devices/StepperControllerService.cs
@@ -1,266 +1,0 @@
-﻿using System;
-using System.Linq;
-using System.Reactive.Linq;
-using System.Threading.Tasks;
-using Ares.Core.Device;
-using Ares.Core.Device.Repos;
-using AresService.Data;
-using AresService.DeviceManagers;
-using Google.Protobuf.WellKnownTypes;
-using Grpc.Core;
-using Microsoft.EntityFrameworkCore;
-using TicStepperController;
-using TicStepperController.Config;
-using TicStepperController.Messaging;
-using TicStepperController.Proto.Extensions;
-
-namespace AresService.Services.Devices;
-
-public class StepperControllerService : StepperControllerRpc.StepperControllerRpcBase
-{
-  private readonly IDeviceCommandInterpreterRepo _deviceCommandInterpreterRepo;
-  private readonly IDeviceManager<StepperControllerConfig, IStepperController> _stepperManager;
-  private readonly IDbContextFactory<AresDbContext> _dbContextFactory;
-  private readonly IDeviceConfigManager<StepperControllerConfig> _configManager;
-
-  public StepperControllerService(IDeviceCommandInterpreterRepo deviceCommandInterpreterRepo,
-    IDeviceManager<StepperControllerConfig, IStepperController> stepperManager,
-    IDbContextFactory<AresDbContext> dbContextFactory,
-    IDeviceConfigManager<StepperControllerConfig> configManager)
-  {
-    _deviceCommandInterpreterRepo = deviceCommandInterpreterRepo;
-    _stepperManager = stepperManager;
-    _dbContextFactory = dbContextFactory;
-    _configManager = configManager;
-  }
-
-  public override async Task<Empty> AddStepperController(StepperControllerConfig request, ServerCallContext context)
-  {
-    var device = await _stepperManager.Create(request);
-    request = await FillConfig(device, request);
-    await _configManager.Add(device.UniqueId, device.Name, request);
-
-    return new Empty();
-  }
-
-  /// <summary>
-  /// Fills the config automatically based on the values we get directly from the step controller.
-  /// That way the user does not have to manually look up the default/current values when
-  /// first hooking up the controller
-  /// </summary>
-  /// <param name="controller"></param>
-  /// <param name="config"></param>
-  /// <returns></returns>
-  private static async Task<StepperControllerConfig> FillConfig(IStepperController controller, StepperControllerConfig config)
-  {
-    var state = await controller.InternalStateStream.Where(s => s.Valid).Take(1);
-    var newConfig = config.Clone();
-    newConfig.MaxAcceleration = state.MaxAcceleration;
-    newConfig.MaxDeceleration = state.MaxDeceleration;
-    newConfig.MaxSpeed = state.MaxSpeed;
-    newConfig.StartingSpeed = state.StartingSpeed;
-    newConfig.StepMode = state.StepMode;
-
-    return newConfig;
-  }
-
-  public override async Task<Empty> DeEnergize(TicRequest request, ServerCallContext context)
-  {
-    var controller = GetStepperController(request.TicId);
-    await controller.DeEnergize();
-
-    return new Empty();
-  }
-
-  public override async Task<Empty> Energize(TicRequest request, ServerCallContext context)
-  {
-    var controller = GetStepperController(request.TicId);
-    await controller.Energize();
-
-    return new Empty();
-  }
-
-  public override async Task<Empty> EnterSafeStart(TicRequest request, ServerCallContext context)
-  {
-    var controller = GetStepperController(request.TicId);
-    await controller.EnterSafeStart();
-
-    return new Empty();
-  }
-
-  public override async Task<Empty> ExitSafeStart(TicRequest request, ServerCallContext context)
-  {
-    var controller = GetStepperController(request.TicId);
-    await controller.ExitSafeStart();
-
[…]
