### 2026-02-19 37ec3a4
Added type hinting and checking for function contracts.
