+++ AresScript/Interpreters/VariableSchemaCollector.cs
@@ -54,16 +54,10 @@ public sealed class VariableSchemaCollector : AresLangBaseVisitor<object?>
     var decl = context.functionDeclaration();
     if(decl is not null)
     {
-      var ids = decl.ID();
-      var parameters = new List<string>();
-      for(var i = 1; i < ids.Length; i++)
-      {
-        var id = ids[i].GetText();
-        if(!string.IsNullOrWhiteSpace(id))
-        {
-          parameters.Add(id);
-        }
-      }
+      var parameters = decl.parameterList()?.parameter()
+        .Select(p => p.ID().GetText())
+        .Where(id => !string.IsNullOrWhiteSpace(id))
+        .ToList() ?? [];
 
       _pendingFunctionParameters.Push(parameters);
     }
