### 2026-02-19 1f20863
Fixed variables not getting properly assigned with certain expressions during validation. User functions now return a properly types dummy value during validation if types have been hinted.
