### 2026-02-19 26283a5
Change return type hinting to be an arrow.
