### 2026-02-19 fdcc39f
Temporarily gutted the hover provider for script
