### 2026-02-20 5d89c6d
Symbol definition support
