+++ UI/Infrastructure/Monaco/Interops/MonacoHoverProvider.cs
@@ -1,10 +1,6 @@
-using Antlr4.Runtime;
 using Ares.Datamodel;
-using Ares.Datamodel.Extensions;
 using Ares.Datamodel.Scripting;
 using Ares.Services;
-using AresScript;
-using AresScript.Generated;
 using Microsoft.JSInterop;
 using System.Text;
 using UI.Application.Scripting;
@@ -18,63 +14,53 @@ public sealed class MonacoHoverProvider(AresScriptingService.AresScriptingServic
   [JSInvokable]
   public async Task<string?> GetHoverText(string script, int line, int column, string identifier)
   {
-    if(string.IsNullOrWhiteSpace(identifier))
+    var response = await _aresScriptingServiceClient.GetSymbolMetadataAsync(new SymbolMetadataRequest
+    {
+      Script = script,
+      CursorLine = line,
+      CursorColumn = column,
+      Identifier = identifier ?? string.Empty
+    });
+
+    if(response.Metadata is null || !response.Metadata.Found)
     {
       return null;
     }
 
-    var request = new CompletionRequest
-    {
-      CursorColumn = column,
-      CursorLine = line,
-      Script = script
-    };
-
-    var completions = await _aresScriptingServiceClient.GetCompletionsAsync(request);
-    var item = completions.Items.FirstOrDefault(i => string.Equals(i.Label, identifier, StringComparison.Ordinal));
-
-    var hoverMarkdown = BuildHoverMarkdown(item);
-    var markdown = hoverMarkdown?.markdown;
-    var hasSchema = hoverMarkdown?.hasSchema ?? false;
-
-    return string.IsNullOrWhiteSpace(markdown) ? null : markdown;
+    return BuildHoverMarkdown(response.Metadata);
   }
 
-  private static (string markdown, bool hasSchema)? BuildHoverMarkdown(CompletionItem? item)
+  private static string? BuildHoverMarkdown(ScriptSymbolMetadata response)
   {
-    if(item is null)
-    {
-      return null;
-    }
     var sb = new StringBuilder();
-    var schemaFound = false;
-    var label = item.Label ?? string.Empty;
-    var name = string.IsNullOrWhiteSpace(item.ParentIdentifier) ? label : $"{item.ParentIdentifier}.{label}";
+    var displayName = string.IsNullOrWhiteSpace(response.ParentIdentifier)
+      ? response.Identifier
+      : $"{response.ParentIdentifier}.{response.Identifier}";
 
-    if(!string.IsNullOrWhiteSpace(name))
+    if(!string.IsNullOrWhiteSpace(displayName))
     {
-      sb.Append("**").Append(name).Append("**");
+      sb.Append("**").Append(displayName).Append("**");
     }
 
-    AppendDescription(sb, item.Detail, item.Documentation);
-    if(item.InputSchema is not null)
+    AppendDescription(sb, response.Detail, response.Documentation);
+
+    if(response.InputSchema is not null && response.InputSchema.Fields.Count > 0)
     {
-      schemaFound = true;
-      AppendDataSchemaSection(sb, "Inputs", item.InputSchema);
-    }
-    if(item.OutputSchema?.Type != AresDataType.Unit)
-    {
-      schemaFound = true;
-      AppendSchemaEntrySection(sb, "Output", item.OutputSchema);
+      AppendDataSchemaSection(sb, "Inputs", response.InputSchema);
     }
 
-    if(item.Schema is not null)
+    if(response.OutputSchema is not null
+      && response.OutputSchema.Type is not AresDataType.Unit and not AresDataType.UnspecifiedType)
     {
-      schemaFound = true;
-      AppendSchemaEntrySection(sb, "Schema", item.Schema);
[…]

[…]
