### 2026-02-20 75a9ab6
Added script to builder workflow
