+++ AresScript/ScriptBuilding/ScriptNodes.cs
@@ -2,11 +2,6 @@ using System.Text;
 
 namespace AresScript.ScriptBuilding;
 
-internal readonly record struct ScriptBuilderCapabilities(bool AllowReturn, bool AllowLoopControl)
-{
-  public static ScriptBuilderCapabilities Root { get; } = new(AllowReturn: false, AllowLoopControl: false);
-}
-
 internal abstract class ScriptNode
 {
   public abstract void WriteTo(StringBuilder output, int indentLevel, int indentSize);
@@ -19,21 +14,27 @@ internal abstract class ScriptNode
 
 internal sealed class LineNode(string text) : ScriptNode
 {
+  public string Text { get; } = text;
+
   public override void WriteTo(StringBuilder output, int indentLevel, int indentSize)
   {
     WriteIndent(output, indentLevel, indentSize);
-    output.AppendLine(text);
+    output.AppendLine(Text);
   }
 }
 
 internal sealed class BlockNode(string header, IReadOnlyList<ScriptNode> statements) : ScriptNode
 {
+  public string Header { get; } = header;
+
+  public IReadOnlyList<ScriptNode> Statements { get; } = statements;
+
   public override void WriteTo(StringBuilder output, int indentLevel, int indentSize)
   {
     WriteIndent(output, indentLevel, indentSize);
-    output.Append(header);
+    output.Append(Header);
     output.AppendLine(":");
-    foreach(ScriptNode statement in statements)
+    foreach(ScriptNode statement in Statements)
     {
       statement.WriteTo(output, indentLevel + 1, indentSize);
     }
@@ -53,18 +54,25 @@ internal sealed class IfNode(
   IReadOnlyList<ConditionalBranchNode> elifBranches,
   IReadOnlyList<ScriptNode>? elseBranch) : ScriptNode
 {
+  public string Condition { get; } = condition;
+
+  public IReadOnlyList<ScriptNode> ThenStatements { get; } = thenStatements;
+
+  public IReadOnlyList<ConditionalBranchNode> ElifBranches { get; } = elifBranches;
+
+  public IReadOnlyList<ScriptNode>? ElseBranch { get; } = elseBranch;
 
   public override void WriteTo(StringBuilder output, int indentLevel, int indentSize)
   {
-    WriteBranch(output, indentLevel, indentSize, $"if {condition}", thenStatements);
-    foreach(ConditionalBranchNode elifBranch in elifBranches)
+    WriteBranch(output, indentLevel, indentSize, $"if {Condition}", ThenStatements);
+    foreach(ConditionalBranchNode elifBranch in ElifBranches)
     {
       WriteBranch(output, indentLevel, indentSize, elifBranch.Header, elifBranch.Statements);
     }
 
-    if(elseBranch is not null)
+    if(ElseBranch is not null)
     {
-      WriteBranch(output, indentLevel, indentSize, "else", elseBranch);
+      WriteBranch(output, indentLevel, indentSize, "else", ElseBranch);
     }
   }
 
@@ -87,12 +95,13 @@ internal sealed class IfNode(
 
 internal sealed class ParallelNode(IReadOnlyList<string> expressions) : ScriptNode
 {
+  public IReadOnlyList<string> Expressions { get; } = expressions;
 
   public override void WriteTo(StringBuilder output, int indentLevel, int indentSize)
   {
     WriteIndent(output, indentLevel, indentSize);
     output.AppendLine("parallel:");
-    foreach(string expression in expressions)
+    foreach(string expression in Expressions)
     {
       WriteIndent(output, indentLevel + 1, indentSize);
       output.AppendLine(expression);
