+++ AresScript/ScriptBuilding/ScriptImportConverter.cs
@@ -1,0 +1,140 @@
+using AresScript.Generated;
+
+namespace AresScript.ScriptBuilding;
+
+internal static class ScriptImportConverter
+{
+  public static List<ScriptNode> Convert(AresLangParser.ProgramContext program)
+  {
+    return ConvertStatements(program.statement(), ScriptBuilderCapabilities.Root);
+  }
+
+  private static List<ScriptNode> ConvertStatements(
+    IEnumerable<AresLangParser.StatementContext> statements,
+    ScriptBuilderCapabilities capabilities)
+  {
+    return statements.Select(statement => ConvertStatement(statement, capabilities)).ToList();
+  }
+
+  private static ScriptNode ConvertStatement(AresLangParser.StatementContext statement, ScriptBuilderCapabilities capabilities)
+  {
+    return statement switch
+    {
+      AresLangParser.SimpleStmtContext simple => ConvertSimpleStatement(simple.simpleStatement(), capabilities),
+      AresLangParser.IfStmtContext ifStmt => ConvertIfStatement(ifStmt.ifStatement(), capabilities),
+      AresLangParser.WhileStmtContext whileStmt => ConvertWhileStatement(whileStmt.whileStatement(), capabilities),
+      AresLangParser.ForStmtContext forStmt => ConvertForStatement(forStmt.forStatement(), capabilities),
+      AresLangParser.ParallelStmtContext parallel => ConvertParallelStatement(parallel.parallelStatement()),
+      AresLangParser.LoopControlStmtContext loopControl => ConvertLoopControl(loopControl.loopControlStatement(), capabilities),
+      AresLangParser.FuncControlStmtContext funcControl => ConvertFuncControl(funcControl.funcControlStatement(), capabilities),
+      _ => new LineNode(statement.GetText())
+    };
+  }
+
+  private static ScriptNode ConvertSimpleStatement(AresLangParser.SimpleStatementContext statement, ScriptBuilderCapabilities capabilities)
+  {
+    return statement switch
+    {
+      AresLangParser.AssignStmtContext assign =>
+        new LineNode($"{assign.assignment().lvalue().GetText()} = {assign.assignment().expression().GetText()}"),
+      AresLangParser.ExprStmtContext expr =>
+        new LineNode(expr.expression().GetText()),
+      AresLangParser.AssertStmtContext assert =>
+        ConvertAssert(assert.assertStatement()),
+      AresLangParser.FunctionDeclContext function =>
+        ConvertFunctionDeclaration(function.functionDeclaration()),
+      _ => new LineNode(statement.GetText())
+    };
+  }
+
+  private static LineNode ConvertAssert(AresLangParser.AssertStatementContext assert)
+  {
+    var expressions = assert.expression();
+    return expressions.Length == 1
+      ? new LineNode($"assert {expressions[0].GetText()}")
+      : new LineNode($"assert {expressions[0].GetText()}, {expressions[1].GetText()}");
+  }
+
+  private static BlockNode ConvertFunctionDeclaration(AresLangParser.FunctionDeclarationContext declaration)
+  {
+    var ids = declaration.ID();
+    var functionName = ids[0].GetText();
+    var parameters = ids.Skip(1).Select(id => id.GetText()).ToArray();
+    var signature = parameters.Length == 0
+      ? $"def {functionName}()"
+      : $"def {functionName}({string.Join(", ", parameters)})";
+
+    var bodyCapabilities = new ScriptBuilderCapabilities(AllowReturn: true, AllowLoopControl: false);
+    var bodyNodes = ConvertStatements(declaration.funcBlock().statement(), bodyCapabilities);
+    return new BlockNode(signature, bodyNodes);
+  }
+
[…]
