### 2026-02-23 eea8a60
Removed all the old device specific settings lists and things
