+++ UI/Features/Devices/RestDevice/RestDeviceUnitControlViewModel.cs
@@ -1,61 +1,0 @@
-using RestDevice.Services;
-using UI.Application.Devices;
-
-namespace UI.Features.Devices.RestDevice;
-
-public class RestDeviceUnitControlViewModel : DeviceUnitControlViewModel
-{
-  private readonly RestDeviceRpc.RestDeviceRpcClient _client;
-
-  public RestDeviceUnitControlViewModel(string deviceId, string deviceName, RestDeviceRpc.RestDeviceRpcClient client) : base(deviceId, deviceName)
-  {
-    _client = client;
-  }
-
-  public async Task UpdateDeviceCapabilities()
-  {
-    var response = await _client.GetDeviceCapabilitiesAsync(new DeviceRequest() { DeviceId = DeviceId });
-    DeviceMethods.Clear();
-    MethodParameters.Clear();
-
-    foreach(var method in response.DeviceMethods)
-    {
-      DeviceMethods.Add(method.MethodName, method.Parameters.ToList());
-      method.Parameters.ToList().ForEach(p => MethodParameters.Add(p, string.Empty));
-    }
-
-    var blah = DeviceMethods.OrderBy(t => t.Value.Count);
-  }
-
-  public async Task HandleDeviceCommand(string commandName)
-  {
-    var found = DeviceMethods.TryGetValue(commandName, out var parameters);
-
-    if(!found || parameters is null)
-      throw new InvalidOperationException("Unknown method requested!");
-
-    var request = new DeviceMethodRequest() { DeviceId = DeviceId, MethodName = commandName };
-
-    foreach(var param in parameters)
-    {
-      var hasValue = MethodParameters.TryGetValue(param, out var value);
-
-      if(hasValue)
-      {
-        request.ParameterValues.Add(value);
-        request.ParameterNames.Add(param);
-      }
-
-      else
-        throw new InvalidOperationException("Couldn't locate parameter!");
-    }
-
-
-    await _client.CallDeviceMethodAsync(request);
-  }
-
-  public Dictionary<string, List<string>> DeviceMethods { get; set; } = [];
-
-  public Dictionary<string, string> MethodParameters { get; set; } = [];
-}
-
