### 2026-02-23 bd52712
Added symbol interfaces
