+++ AresScript/AresScriptFunction.cs
@@ -9,7 +9,11 @@ public record AresScriptFunction(
   string Name,
   IReadOnlyList<AresScriptParameter> Parameters,
   AresLangParser.FuncBlockContext Body,
-  AresDataType ReturnType = AresDataType.Any)
+  AresDataType ReturnType = AresDataType.Any) : IFunctionSymbol
 {
+  public ScriptSymbolKind Kind => ScriptSymbolKind.Function;
+  public IReadOnlyCollection<ScriptSymbolTag> Tags { get; } = [ScriptSymbolTag.UserDefined];
+  public bool IsUserDefined => true;
+  public bool IsLambda => false;
   public IReadOnlyList<string> ParameterNames { get; } = Parameters.Select(parameter => parameter.Name).ToArray();
 }
