+++ Ares.Core/Scripting/ISymbolProvider.cs
@@ -1,0 +1,6 @@
+namespace Ares.Core.Scripting;
+
+public interface ISymbolProvider
+{
+  
+}
