### 2026-02-23 e3d9abb
Moved all the old device UI stuff into the devices themselves temporarily (this will still be the case later but will be cleaned up). Working towards the UI being buildable again
