### 2026-02-23 fd07daf
More deletions. UI built for the first time in awhile.
