+++ AresService/ConfigManagers/ServoConfigManager.cs
@@ -1,13 +1,0 @@
-﻿using Ares.Core;
-using Ares.Core.Device;
-using HerkulexDRS;
-using HerkulexDRS.Config;
-using Microsoft.EntityFrameworkCore;
-
-namespace AresService.ConfigManagers;
-public class ServoConfigManager : DeviceConfigManager<ServoConfig, IServo>
-{
-  public ServoConfigManager(IDbContextFactory<CoreDatabaseContext> dbContextFactory) : base(dbContextFactory)
-  {
-  }
-}
