+++ AresService/ConfigManagers/FlirCM3ConfigManager.cs
@@ -1,14 +1,0 @@
-﻿using Ares.Core;
-using Ares.Core.Device;
-using FlirCM3;
-using FlirCM3.Config;
-using Microsoft.EntityFrameworkCore;
-
-namespace AresService.ConfigManagers;
-
-public class FlirCM3ConfigManager : DeviceConfigManager<FlirCM3Config, IFlirCM3Camera>
-{
-  public FlirCM3ConfigManager(IDbContextFactory<CoreDatabaseContext> dbContextFactory) : base(dbContextFactory)
-  {
-  }
-}

[…]
