+++ AresScript/Interpreters/AresTypeInferenceInterpreter.cs
@@ -9,7 +9,7 @@ namespace AresScript.Interpreters;
 /// Interpreter specifically to gather the types of symbols so they can be displayed via hover
 /// or otherwise validated when used as inputs/outputs
 /// </summary>
-public sealed class AresTypeInferenceInterpreter : AresLangBaseVisitor<SchemaEntry>
+public sealed class AresTypeInferenceInterpreter : AresLangBaseVisitor<AresValueSchema>
 {
   private readonly AresScriptEnvironment _environment;
 
@@ -18,15 +18,15 @@ public sealed class AresTypeInferenceInterpreter : AresLangBaseVisitor<SchemaEnt
     _environment = environment;
   }
 
-  protected override SchemaEntry DefaultResult => AresSchemaBuilder.Entry(AresDataType.Any).Build();
+  protected override AresValueSchema DefaultResult => AresSchemaBuilder.Entry(AresDataType.Any).Build();
 
-  public override SchemaEntry VisitInt(AresLangParser.IntContext context) => AresSchemaBuilder.Entry(AresDataType.Number).Build();
-  public override SchemaEntry VisitFloat(AresLangParser.FloatContext context) => AresSchemaBuilder.Entry(AresDataType.Number).Build();
-  public override SchemaEntry VisitString(AresLangParser.StringContext context) => AresSchemaBuilder.Entry(AresDataType.String).Build();
-  public override SchemaEntry VisitBool(AresLangParser.BoolContext context) => AresSchemaBuilder.Entry(AresDataType.Boolean).Build();
-  public override SchemaEntry VisitNone(AresLangParser.NoneContext context) => AresSchemaBuilder.Entry(AresDataType.Null).Build();
+  public override AresValueSchema VisitInt(AresLangParser.IntContext context) => AresSchemaBuilder.Entry(AresDataType.Number).Build();
+  public override AresValueSchema VisitFloat(AresLangParser.FloatContext context) => AresSchemaBuilder.Entry(AresDataType.Number).Build();
+  public override AresValueSchema VisitString(AresLangParser.StringContext context) => AresSchemaBuilder.Entry(AresDataType.String).Build();
+  public override AresValueSchema VisitBool(AresLangParser.BoolContext context) => AresSchemaBuilder.Entry(AresDataType.Boolean).Build();
+  public override AresValueSchema VisitNone(AresLangParser.NoneContext context) => AresSchemaBuilder.Entry(AresDataType.Null).Build();
 
-  public override SchemaEntry VisitId(AresLangParser.IdContext context)
+  public override AresValueSchema VisitId(AresLangParser.IdContext context)
   {
     var id = context.ID().GetText();
     if(_environment.TryGetSystemFunction(id, out var sysFunc) && sysFunc?.OutputSchema is not null) 
@@ -44,18 +44,18 @@ public sealed class AresTypeInferenceInterpreter : AresLangBaseVisitor<SchemaEnt
 
     if(_environment.TryGetValue(id, out var envVal))
     {
-      return envVal.ToSchemaEntry();
+      return envVal.ToAresValueSchema();
     }
 
     return AresSchemaBuilder.Entry(AresDataType.Any).Build();
   }
 
-  public override SchemaEntry VisitParens(AresLangParser.ParensContext context)
+  public override AresValueSchema VisitParens(AresLangParser.ParensContext context)
   {
     return Visit(context.expression());
   }
 
-  public override SchemaEntry VisitArray(AresLangParser.ArrayContext context)
+  public override AresValueSchema VisitArray(AresLangParser.ArrayContext context)
   {
     var expressions = context.expression();
     if(expressions.Length == 0)
[…]
