+++ Ares.Core/EntityConfigurations/ParameterMetadataEntityConfiguration.cs
@@ -17,7 +17,7 @@ internal class ParameterMetadataEntityConfiguration : AresEntityTypeBaseConfigur
     builder.Property(parameterMetadata => parameterMetadata.InitialValue).HasAresValue();
 
     builder.Property(parameterMetadata => parameterMetadata.Schema)
-      .HasAresSchemaEntry();
+      .HasAresAresValueSchema();
 
     builder.Navigation(parameterMetadata => parameterMetadata.Constraints)
       .AutoInclude();
