+++ AresScript/AresSystemFunction.cs
@@ -2,8 +2,8 @@ using Ares.Datamodel;
 
 namespace AresScript;
 
-public delegate Task<AresValue> AresFunctionDelegate(List<AresValue> args, ScriptExecutionControlToken token);
+public delegate Task<AresValue> AresFunctionDelegate(IDictionary<string, AresValue> args, ScriptExecutionControlToken token);
 
-public record AresSystemFunction(string Id, string Name, AresFunctionDelegate Body, AresDataSchema InputSchema, SchemaEntry OutputSchema, string Namespace = "", string Description = "")
+public record AresSystemFunction(string Id, string Name, AresFunctionDelegate Body, AresStructSchema InputSchema, AresValueSchema OutputSchema, string Namespace = "", string Description = "")
 {
 }
