### 2026-02-24 78ecf3c
Separated out the script symbols.
