### 2026-02-24 f293236
Fixed script analysis components.
