+++ AresScript/ScriptAnalysis/AresScriptAnalysis.SymbolMetadata.cs
@@ -25,7 +25,6 @@ public static partial class AresScriptAnalysis
     {
       return new ScriptSymbolMetadata
       {
-        Found = false,
         Identifier = string.Empty,
         ParentIdentifier = string.Empty,
         Kind = SymbolKind.Unspecified,
@@ -48,32 +47,25 @@ public static partial class AresScriptAnalysis
 
     if(environment.TryGetSystemFunction(identifier, out var systemFunction))
     {
-      return new ScriptSymbolMetadata
-      {
-        Found = true,
-        Identifier = identifier,
-        ParentIdentifier = string.Empty,
-        Kind = SymbolKind.Function,
-        Detail = systemFunction.Description ?? string.Empty,
-        Documentation = systemFunction.Description ?? string.Empty,
-        InputSchema = systemFunction.InputSchema,
-        OutputSchema = systemFunction.OutputSchema
-      };
+      return BuildFunctionMetadataForSymbolMetadata(
+        identifier,
+        string.Empty,
+        systemFunction.Description ?? string.Empty,
+        systemFunction.Description ?? string.Empty,
+        systemFunction.InputSchema,
+        systemFunction.OutputSchema);
     }
 
     if(environment.TryGetUserFunction(identifier, out var userFunction))
     {
-      return new ScriptSymbolMetadata
-      {
-        Found = true,
-        Identifier = identifier,
-        ParentIdentifier = string.Empty,
-        Kind = SymbolKind.Function,
-        Detail = "User function",
-        Documentation = "User-defined function.",
-        InputSchema = BuildUserFunctionInputSchema(userFunction),
-        OutputSchema = AresSchemaBuilder.Entry(userFunction.ReturnType).Build()
-      };
+      return BuildFunctionMetadataForSymbolMetadata(
+        identifier,
+        string.Empty,
+        "User function",
+        "User-defined function.",
+        BuildUserFunctionInputSchema(userFunction),
+        AresSchemaBuilder.Entry(userFunction.ReturnType).Build(),
+        isUserDefined: true);
     }
 
     if(environment.TryGetUserLambda(identifier, out var lambda))
@@ -84,17 +76,15 @@ public static partial class AresScriptAnalysis
         lambdaSchema.Fields[parameter] = AresSchemaBuilder.Entry(AresDataType.Any).Build();
       }
 
-      return new ScriptSymbolMetadata
-      {
-        Found = true,
-        Identifier = identifier,
-        ParentIdentifier = string.Empty,
-        Kind = SymbolKind.Function,
-        Detail = "Lambda function",
-        Documentation = "User-defined lambda.",
-        InputSchema = lambdaSchema,
-        OutputSchema = AresSchemaBuilder.Entry(AresDataType.Any).Build()
-      };
+      return BuildFunctionMetadataForSymbolMetadata(
+        identifier,
+        string.Empty,
+        "Lambda function",
+        "User-defined lambda.",
+        lambdaSchema,
+        AresSchemaBuilder.Entry(AresDataType.Any).Build(),
+        isUserDefined: true,
+        isLambda: true);
     }
 
     if(environment.TryGetValue(identifier, out var value))
@@ -105,21 +95,17 @@ public static partial class AresScriptAnalysis
     var inferredSchema = TryInferSchema(script, cursorLine, cursorColumn, identifier, environment);
     if(inferredSchema is not null)
     {
-      return new ScriptSymbolMetadata
-      {
-        Found = true,
-        Identifier = identifier,
-        ParentIdentifier = parentIdentifier ?? string.Empty,
[…]
