+++ AresScript/ScriptAnalysis/AresScriptAnalysis.Completions.cs
@@ -1,9 +1,11 @@
 using Antlr4.Runtime;
 using Ares.Datamodel;
 using Ares.Datamodel.Extensions;
+using Ares.Datamodel.Factories;
 using Ares.Datamodel.Scripting;
 using AresScript.Generated;
 using AresScript.Interpreters;
+using AresScript.Symbols;
 using System.Text;
 using System.Text.RegularExpressions;
 
@@ -87,17 +89,29 @@ public static partial class AresScriptAnalysis
       {
         Label = func.Name,
         InsertText = func.Name,
-        Detail = "User function",
-        Kind = SymbolKind.Function
+        Metadata = BuildFunctionMetadata(
+          identifier: func.Name,
+          parentIdentifier: string.Empty,
+          detail: "User function",
+          documentation: "User-defined function.",
+          inputSchema: BuildUserFunctionInputSchemaForCompletions(func),
+          outputSchema: AresSchemaBuilder.Entry(func.ReturnType).Build(),
+          isUserDefined: true)
       }));
 
       items.AddRange(userVariables.Select(variable => new CompletionItem
       {
         Label = variable.Key,
         InsertText = variable.Key,
-        Detail = "User variable",
-        Kind = SymbolKind.Variable,
-        Schema = variable.Value.ToSchemaEntry()
+        Metadata = BuildValueMetadata(
+          identifier: variable.Key,
+          parentIdentifier: string.Empty,
+          detail: "User variable",
+          documentation: "User-defined variable.",
+          kind: variable.Value.KindCase == AresValue.KindOneofCase.StructValue ? SymbolKind.Struct : SymbolKind.Variable,
+          schema: variable.Value.ToSchemaEntry(),
+          value: variable.Value,
+          isUserDefined: true)
       }));
     }
 
@@ -197,12 +211,15 @@ public static partial class AresScriptAnalysis
       {
         Label = extensionFunc.Name,
         InsertText = BuildSnippet(extensionFunc.Name, schemaForCall),
-        Detail = extensionFunc.Description,
-        Documentation = extensionFunc.Description,
-        Kind = SymbolKind.Function,
-        ParentIdentifier = parentIdentifier,
-        InputSchema = schemaForCall,
-        OutputSchema = extensionFunc.OutputSchema
+        InsertTextFormat = InsertTextFormat.Snippet,
+        Metadata = BuildFunctionMetadata(
+          identifier: extensionFunc.Name,
+          parentIdentifier: parentIdentifier,
+          detail: extensionFunc.Description,
+          documentation: extensionFunc.Description,
+          inputSchema: schemaForCall,
+          outputSchema: extensionFunc.OutputSchema,
+          isExtension: true)
       });
     }
   }
@@ -241,12 +258,14 @@ public static partial class AresScriptAnalysis
       {
         Label = label,
         InsertText = BuildSnippet(label, systemFunction.InputSchema),
-        Detail = description,
-        Documentation = description,
-        Kind = SymbolKind.Function,
-        ParentIdentifier = parentIdentifier,
-        InputSchema = systemFunction.InputSchema,
-        OutputSchema = systemFunction.OutputSchema
+        InsertTextFormat = InsertTextFormat.Snippet,
+        Metadata = BuildFunctionMetadata(
+          identifier: label,
+          parentIdentifier: parentIdentifier,
+          detail: description,
+          documentation: description,
+          inputSchema: systemFunction.InputSchema,
+          outputSchema: systemFunction.OutputSchema)
       });
       return;
     }
[…]
