+++ Ares.Core/Scripting/DeviceFunctionProvider.cs
@@ -3,26 +3,29 @@ using Ares.Core.Device;
 using Ares.Datamodel;
 using Ares.Datamodel.Extensions;
 using Ares.Datamodel.Factories;
+using Ares.Datamodel.Scripting;
 using Ares.Datamodel.Templates;
+using AresScript;
 using AresScript.Symbols;
 
 namespace Ares.Core.Scripting;
 
-public class DeviceFunctionProvider(IDeviceCommandInterpreterRepo interpreterRepo) : ISystemFunctionProvider
+public class DeviceFunctionProvider(IDeviceCommandInterpreterRepo interpreterRepo) : ISymbolProvider
 {
   private readonly IDeviceCommandInterpreterRepo _interpreterRepo = interpreterRepo;
 
 
-  public AresSystemFunction[] GetFunctions()
+  public IScriptSymbol[] GetSymbols()
   {
     var interpreters = _interpreterRepo.GetSnapshot();
-    var functions = new List<AresSystemFunction>();
+    var symbols = new List<IScriptSymbol>();
 
     foreach(var interpreter in interpreters)
     {
       var device = interpreter.Device;
       var devicePrefix = SanitizeIdentifier(string.IsNullOrWhiteSpace(device.Name) ? device.UniqueId : device.Name);
       var commandMetadatas = interpreter.CommandsToIndexedMetadatas().ToArray();
+      var deviceFunctionFields = new Dictionary<string, AresSystemValue>(StringComparer.Ordinal);
 
       foreach(var metadata in commandMetadatas)
       {
@@ -36,54 +39,62 @@ public class DeviceFunctionProvider(IDeviceCommandInterpreterRepo interpreterRep
           ? AresSchemaBuilder.Entry(AresDataType.Unit).Build()
           : AresSchemaBuilder.Entry(AresDataType.Struct).WithStructSchema(metadata.OutputMetadata.DataSchema).Build();
 
-        functions.Add(
-          new AresSystemFunction(
-            functionId,
-            commandName,
-            async (args, token) =>
+        var functionSymbol = new AresSystemFunction(
+          functionId,
+          commandName,
+          async (args, token) =>
+          {
+            if(args.Count > parameterMetadatas.Length)
             {
-              if(args.Count > parameterMetadatas.Length)
-              {
-                throw new InvalidOperationException(
-                  $"Function '{functionId}' expected at most {parameterMetadatas.Length} arguments but got {args.Count}.");
-              }
+              throw new InvalidOperationException(
+                $"Function '{functionId}' expected at most {parameterMetadatas.Length} arguments but got {args.Count}.");
+            }
 
-              var template = new CommandTemplate { Metadata = metadata };
-              for(var i = 0; i < args.Count; i++)
-              {
-                var parameterMetadata = parameterMetadatas[i];
-                template.Parameters.Add(
-                  new Parameter
-                  {
-                    Metadata = parameterMetadata,
-                    Value = args[i],
-                    Index = parameterMetadata.Index
-                  });
-              }
+            var template = new CommandTemplate { Metadata = metadata };
+            for(var i = 0; i < args.Count; i++)
+            {
+              var parameterMetadata = parameterMetadatas[i];
+              template.Parameters.Add(
+                new Parameter
+                {
+                  Metadata = parameterMetadata,
+                  Value = args[i],
+                  Index = parameterMetadata.Index
+                });
+            }
 
-              var command = interpreter.TemplateToDeviceCommand(template);
-              var result = await command(token.CancellationToken);
-              if(!result.Success)
-              {
-                throw new InvalidOperationException(
-                  string.IsNullOrWhiteSpace(result.Error)
-                    ? $"Device command '{metadata.Name}' failed."
-                    : result.Error);
[…]
