+++ Ares.Core/Scripting/ISymbolProvider.cs
@@ -1,6 +1,8 @@
+using AresScript.Symbols;
+
 namespace Ares.Core.Scripting;
 
 public interface ISymbolProvider
 {
-  
+  IScriptSymbol[] GetSymbols();
 }
