+++ Ares.Core/Scripting/BaseEnvironmentBuilder.cs
@@ -3,37 +3,39 @@ using AresScript.Symbols;
 
 namespace Ares.Core.Scripting;
 
-public class BaseEnvironmentBuilder(IEnumerable<ISystemFunctionProvider> systemFunctionProviders)
+public class BaseEnvironmentBuilder(IEnumerable<ISymbolProvider> symbolProviders)
 {
-  private readonly IEnumerable<ISystemFunctionProvider> _systemFunctionProvider = systemFunctionProviders;
+  private readonly IEnumerable<ISymbolProvider> _symbolProviders = symbolProviders;
 
   public AresScriptEnvironment Build()
   {
     var env = new AresScriptEnvironment();
-    var functions = _systemFunctionProvider.SelectMany(sfp => sfp.GetFunctions()).ToArray();
-    if(functions.Length == 0)
-    {
-      return env;
-    }
+    var providedSymbols = _symbolProviders.SelectMany(provider => provider.GetSymbols()).ToArray();
+    var allSystemFunctions = StandardLibrary.Functions
+      .Concat(providedSymbols.OfType<AresSystemFunction>())
+      .ToArray();
 
-    // let's stick with global scope for now
-    env.AssignSystemFunctions(StandardLibrary.Functions);
+    env.AssignSystemFunctions(allSystemFunctions);
     env.AssignExtensionFunctions(StandardLibrary.ExtensionFunctions);
-    env.AssignSystemFunctions(functions);
-    env.AssignSystemVariables(BuildSystemValues(env));
+    env.AssignSystemVariables(BuildSystemValues(allSystemFunctions, providedSymbols.OfType<IValueSymbol>()));
 
     return env;
   }
 
-  private static IEnumerable<KeyValuePair<string, AresSystemValue>> BuildSystemValues(AresScriptEnvironment environment)
+  private static IEnumerable<KeyValuePair<string, AresSystemValue>> BuildSystemValues(
+    IEnumerable<AresSystemFunction> systemFunctions,
+    IEnumerable<IValueSymbol> providedValueSymbols)
   {
     var rootVariables = new Dictionary<string, AresSystemValue>(StringComparer.Ordinal);
-    var functions = environment.GetAllSystemFunctions();
 
-    foreach(var func in functions)
+    foreach(var func in systemFunctions)
     {
       var fieldName = string.IsNullOrWhiteSpace(func.Name) ? func.Id : func.Name;
-      if(string.IsNullOrWhiteSpace(func.Namespace))
+      var parentPath = !string.IsNullOrWhiteSpace(func.ParentName)
+        ? func.ParentName
+        : func.Namespace;
+
+      if(string.IsNullOrWhiteSpace(parentPath))
       {
         if(!string.IsNullOrWhiteSpace(fieldName) && !rootVariables.ContainsKey(fieldName))
         {
@@ -47,7 +49,7 @@ public class BaseEnvironmentBuilder(IEnumerable<ISystemFunctionProvider> systemF
         continue;
       }
 
-      var pathSegments = func.Namespace
+      var pathSegments = parentPath
         .Split('.', StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries);
       if(pathSegments.Length == 0)
       {
@@ -57,11 +59,35 @@ public class BaseEnvironmentBuilder(IEnumerable<ISystemFunctionProvider> systemF
       AddFunctionToPath(rootVariables, pathSegments, fieldName, func);
     }
 
+    foreach(var valueSymbol in providedValueSymbols)
+    {
+      var valueName = valueSymbol.Name;
+      if(string.IsNullOrWhiteSpace(valueName))
+      {
+        continue;
+      }
+
+      var systemValue = ToSystemValue(valueSymbol);
[…]
