+++ Ares.Core/Scripting/ISystemFunctionProvider.cs
@@ -1,8 +1,0 @@
-using AresScript.Symbols;
-
-namespace Ares.Core.Scripting;
-
-public interface ISystemFunctionProvider
-{
-  AresSystemFunction[] GetFunctions();
-}

[…]
