### 2026-02-24 ccd47c1
Some small updates to address a few bugs discovered in lab.
- Added a try catch in the Alicat MFC loading logic to test a potential problem point.
- Moved the experiment template copy logic to the core, added additional functionality to the Automation Service to support this with the goal of keeping the UI ignorant to how templates are serialized into JSON.
- Implemented the dispose method of the new syringe pumps to allow for proper device deletion.
