### 2026-02-24 9f6c931
The core ARES libraries are building and running again, preparing for testing the new plugin system
