+++ Ares.Core.Grpc/Services/AresDeviceDriverService.cs
@@ -5,7 +5,6 @@ using Grpc.Core;
 using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace Ares.Core.Grpc.Services;
