+++ Ares.Core/Device/Remote/IRemoteDeviceManager.cs
@@ -10,12 +10,8 @@ namespace Ares.Core.Device.Remote;
 public interface IRemoteDeviceManager
 {
   Task LoadDevices();
-
   Task<RemoteDevice> CreateDevice(string name, string url);
-
   Task<bool> RemoveDevice(string deviceId);
-
   Task UpdateDevice(RemoteDeviceConfig config);
-
   Task UpdateDeviceSettings(DeviceSettings deviceSettings);
 }

[…]
