### 2026-02-24 8cd0d3a
Deleting all the old device specific state messages
