+++ AresMessaging/DeviceStateLogging/TubeFurnace/TubeFurnaceStateEntity.cs
@@ -1,7 +1,0 @@
-﻿using Ares.Datamodel.Device;
-
-namespace Ares.Messages.DeviceStates.TubeFurnace;
-
-public partial class TubeFurnaceStateEntity : IDeviceState
-{
-}
