+++ AresMessaging/DeviceStateLogging/TicStepperController/TicStepperControllerState.cs
@@ -1,7 +1,0 @@
-﻿using Ares.Datamodel.Device;
-
-namespace Ares.Messages.DeviceStates.TicStepperController;
-
-public partial class TicStepperControllerState : IDeviceState
-{
-}
