+++ AresMessaging/DeviceStateLogging/SyringePump/SyringePumpState.cs
@@ -1,6 +1,0 @@
-﻿using Ares.Datamodel.Device;
-
-namespace Ares.Messages.DeviceStates.SyringePump;
-public partial class SyringePumpState : IDeviceState
-{
-}
