+++ AresMessaging/DeviceStateLogging/LaserChiller/ChillerState.cs
@@ -1,7 +1,0 @@
-﻿using Ares.Datamodel.Device;
-
-namespace Ares.Messages.DeviceStates.Chiller;
-
-public partial class ChillerState : IDeviceState
-{
-}
