### 2026-02-24 ce3a4b4
Sorted all the entity configurations into folders
