+++ Ares.Core/EntityConfigurations/Device/Remote/RemoteDeviceConfigEntityConfiguration.cs
@@ -1,0 +1,14 @@
+﻿using Ares.Datamodel.Device;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+
+namespace Ares.Core.EntityConfigurations.Device.Remote;
+internal class RemoteDeviceConfigEntityConfiguration : AresEntityTypeBaseConfiguration<RemoteDeviceConfig>
+{
+  public override void Configure(EntityTypeBuilder<RemoteDeviceConfig> builder)
+  {
+    base.Configure(builder);
+
+    builder.ToTable("RemoteDevices");
+  }
+}
