+++ Ares.Core/EntityConfigurations/Shared/LimitsEntityConfiguration.cs
@@ -1,0 +1,7 @@
+﻿using Ares.Datamodel;
+
+namespace Ares.Core.EntityConfigurations.Shared;
+
+internal class LimitsEntityConfiguration : AresEntityTypeBaseConfiguration<Limits>
+{
+}
