+++ Ares.Core/EntityConfigurations/Analyzer/AnalyzerCapabilitiesEntityConfiguration.cs
@@ -1,0 +1,13 @@
+﻿using Ares.Core.EntityConfigurations.Helpers;
+using Ares.Datamodel.Analyzing;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+
+namespace Ares.Core.EntityConfigurations.Analyzer;
+internal class AnalyzerCapabilitiesEntityConfiguration : AresEntityTypeBaseConfiguration<AnalyzerCapabilities>
+{
+  public override void Configure(EntityTypeBuilder<AnalyzerCapabilities> builder)
+  {
+    base.Configure(builder);
+    builder.Property(p => p.SettingsSchema).HasDataSchema();
+  }
+}
