+++ Ares.Core/EntityConfigurations/Execution/Commands/OutputMetadataEntityConfiguration.cs
@@ -1,0 +1,14 @@
+﻿using Ares.Core.EntityConfigurations.Helpers;
+using Ares.Datamodel.Templates;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+
+namespace Ares.Core.EntityConfigurations.Execution.Commands;
+
+public class OutputMetadataEntityConfiguration : AresEntityTypeBaseConfiguration<OutputMetadata>
+{
+  public override void Configure(EntityTypeBuilder<OutputMetadata> builder)
+  {
+    base.Configure(builder);
+    builder.Property(output => output.DataSchema).HasAresValueSchema();
+  }
+}
