### 2026-02-24 90897d4
Finished off the combined device getter, replaced any remaining references to it
