+++ UI/Features/DeviceStateLogging/DeviceStateFilterViewModelFactory.cs
@@ -1,18 +1,18 @@
 ﻿using Ares.Services;
-using UI.Application.DeviceStateLogging;
+using Ares.Services.Device;
 
 namespace UI.Features.DeviceStateLogging;
 
 public class DeviceStateFilterViewModelFactory
 {
   private readonly AresAutomation.AresAutomationClient _automationClient;
-  readonly ICombinedDeviceGetter _deviceGetter;
+  private readonly AresDevices.AresDevicesClient _devicesClient;
 
-  public DeviceStateFilterViewModelFactory(AresAutomation.AresAutomationClient automationClient, ICombinedDeviceGetter deviceGetter)
+  public DeviceStateFilterViewModelFactory(AresAutomation.AresAutomationClient automationClient, AresDevices.AresDevicesClient devicesClient)
   {
-    _deviceGetter = deviceGetter;
+    _devicesClient = devicesClient;
     _automationClient = automationClient;
   }
 
-  public DeviceStateFilterViewModel Create() => new DeviceStateFilterViewModel(_automationClient, _deviceGetter);
+  public DeviceStateFilterViewModel Create() => new DeviceStateFilterViewModel(_devicesClient, _automationClient);
 }
