+++ UI/Features/DeviceStateLogging/DeviceStateFilterViewModel.cs
@@ -2,26 +2,26 @@
 using Ares.Datamodel.Device;
 using Ares.Messages.DeviceStates;
 using Ares.Services;
+using Ares.Services.Device;
 using Google.Protobuf.WellKnownTypes;
 using ReactiveUI;
 using ReactiveUI.SourceGenerators;
-using UI.Application.DeviceStateLogging;
 
 namespace UI.Features.DeviceStateLogging;
 
 public partial class DeviceStateFilterViewModel : ReactiveObject
 {
+  private readonly AresDevices.AresDevicesClient _deviceClient;
   readonly AresAutomation.AresAutomationClient _automationClient;
 
-  public DeviceStateFilterViewModel(
-    AresAutomation.AresAutomationClient automationClient,
-    ICombinedDeviceGetter deviceGetter)
+  public DeviceStateFilterViewModel(AresDevices.AresDevicesClient devicesClient, AresAutomation.AresAutomationClient automationClient)
   {
     _automationClient = automationClient;
+    _deviceClient = devicesClient;
     _automationClient.GetAvailableCampaignExecutionSummariesAsync(new Empty()).ResponseAsync
       .ContinueWith(task => UpdateCampaigns(task.Result));
-    deviceGetter.GetAvailableDevices()
-      .ContinueWith(task => AvailableDevices = task.Result);
+    _deviceClient.GetAllAvailableDevicesAsync(new Empty()).ResponseAsync
+      .ContinueWith(task => UpdateDevices(task.Result));
 
     var currentTime = DateTime.Now;
     // probably don't need millisecond precision
@@ -35,8 +35,13 @@ public partial class DeviceStateFilterViewModel : ReactiveObject
     Campaigns = response.AvailableCampaignSummaries.Select(result => new CampaignExecutionSummaryMetadata { SummaryId = result.SummaryId, CampaignName = $"result.CampaignName-{result.CompletionTime}", CompletionTime = result.CompletionTime });
   }
 
+  private void UpdateDevices(AvailableDevicesResponse response)
+  {
+    AvailableDevices = response.Devices.ToList();
+  }
+
   [Reactive]
-  public partial IEnumerable<DevicesDescription>? AvailableDevices { get; private set; }
+  public partial IEnumerable<AresDeviceDescription>? AvailableDevices { get; private set; }
 
   public IEnumerable<DevicesDescription>? SelectedDevices { get; set; }
 
