### 2026-02-25 f3b911f
Fixed a few missing bindings and deleted a duplicate file
