### 2026-02-25 0339394
Swapped devices to return AresValue's instead of Structs
