### 2026-02-25 aaa4a86
Updated entity configs to use global conversions when possible. Cleaned up old conversion methods
