+++ AresService.Data/AresDbContext.cs
@@ -1,6 +1,9 @@
-﻿using System.Reflection;
-using Ares.Core;
+﻿using Ares.Core;
+using Ares.Core.EntityConfigurations.Helpers;
+using Ares.Datamodel;
+using Google.Protobuf.WellKnownTypes;
 using Microsoft.EntityFrameworkCore;
+using System.Reflection;
 
 namespace AresService.Data;
 
@@ -19,5 +22,21 @@ public class AresDbContext : CoreDatabaseContext
 
     modelBuilder.ApplyConfigurationsFromAssembly(assembly);
     base.OnModelCreating(modelBuilder);
+  }
+
+  protected override void ConfigureConventions(ModelConfigurationBuilder configurationBuilder)
+  {
+    // Protobuf Data Types
+    configurationBuilder.Properties<AresStruct>().HaveConversion<AresValueConverters>();
+    configurationBuilder.Properties<AresValue>().HaveConversion<AresValueConverter>();
+    
+    // Protobuf Schemas
+    configurationBuilder.Properties<AresValueSchema>().HaveConversion<AresValueSchemaConverter>();
+    configurationBuilder.Properties<AresStructSchema>().HaveConversion<AresStructSchemaConverter>();
+
+    // Utilities
+    configurationBuilder.Properties<Timestamp>().HaveConversion<AresTimestampConverter>();
+
+    base.ConfigureConventions(configurationBuilder);
   }
 }
