+++ Ares.Core/EntityConfigurations/Analyzer/AnalyzerCapabilitiesEntityConfiguration.cs
@@ -1,5 +1,4 @@
-﻿using Ares.Core.EntityConfigurations.Helpers;
-using Ares.Datamodel.Analyzing;
+﻿using Ares.Datamodel.Analyzing;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
 namespace Ares.Core.EntityConfigurations.Analyzer;
@@ -8,6 +7,5 @@ internal class AnalyzerCapabilitiesEntityConfiguration : AresEntityTypeBaseConfi
   public override void Configure(EntityTypeBuilder<AnalyzerCapabilities> builder)
   {
     base.Configure(builder);
-    builder.Property(p => p.SettingsSchema).HasDataSchema();
   }
 }
