+++ Ares.Core/EntityConfigurations/Execution/Parameters/ParameterEntityConfiguration.cs
@@ -1,5 +1,4 @@
-﻿using Ares.Core.EntityConfigurations.Helpers;
-using Ares.Datamodel.Templates;
+﻿using Ares.Datamodel.Templates;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
@@ -25,10 +24,5 @@ internal class ParameterEntityConfiguration : AresEntityTypeBaseConfiguration<Pa
 
     builder.Navigation(parameter => parameter.PlanningMetadata)
       .AutoInclude();
-
-    builder.Navigation(parameter => parameter.Value)
-      .AutoInclude();
-
-    builder.Property(parameter => parameter.Value).HasAresValue();
   }
 }
