+++ Ares.Core/EntityConfigurations/Device/DeviceSettingsEntityConfiguration.cs
@@ -1,5 +1,4 @@
-﻿using Ares.Core.EntityConfigurations.Helpers;
-using Ares.Datamodel.Device;
+﻿using Ares.Datamodel.Device;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
 namespace Ares.Core.EntityConfigurations.Device;
@@ -8,7 +7,5 @@ internal class DeviceSettingsEntityConfiguration : AresEntityTypeBaseConfigurati
   public override void Configure(EntityTypeBuilder<DeviceSettings> builder)
   {
     base.Configure(builder);
-
-    builder.Property(b => b.Settings).HasAresStruct();
   }
 }
