+++ Ares.Core/EntityConfigurations/Device/DeviceCommandDescriptorEntityConfiguration.cs
@@ -1,4 +1,3 @@
-using Ares.Core.EntityConfigurations.Helpers;
 using Ares.Datamodel.Device;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
@@ -9,7 +8,5 @@ public class DeviceCommandDescriptorEntityConfiguration : AresEntityTypeBaseConf
     public override void Configure(EntityTypeBuilder<DeviceCommandDescriptor> builder)
     {
         base.Configure(builder);
-        builder.Property(p => p.InputSchema).HasDataSchema();
-        builder.Property(p => p.OutputSchema).HasAresValueSchema();
     }
 }
