### 2026-02-25 c5557e9
Updated datamodel to remove the only remaining usage of the protobuf Any message. Also removed the Any entity config
