+++ Ares.Core/EntityConfigurations/Shared/AnyTypeEntityConfiguration.cs
@@ -1,19 +1,0 @@
-﻿using Ares.Core.EntityConfigurations.Helpers;
-using Google.Protobuf;
-using Google.Protobuf.WellKnownTypes;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-
-namespace Ares.Core.EntityConfigurations.Shared;
-
-internal class AnyTypeEntityConfiguration : AresEntityTypeBaseConfiguration<Any>
-{
-  public override void Configure(EntityTypeBuilder<Any> builder)
-  {
-    base.Configure(builder);
-
-    builder.Property(any => any.Value)
-      .HasConversion(
-        s => s.ToByteArray(),
-        bytes => ByteString.CopyFrom(bytes), new ByteStringComparer());
-  }
-}
