### 2026-02-25 0d6c4e2
Refactored the hover support a bit. Replaced more items with symbols. Moved files around.
