+++ AresScript/AresExtensionFunction.cs
@@ -6,4 +6,4 @@ namespace AresScript;
 public sealed record AresExtensionFunction(
   AresValue.KindOneofCase ReceiverKind,
   string MemberName,
-  AresSystemFunction Function);
+  AresSystemFunctionSymbol Function);
