### 2026-02-25 b794149
Vibecoded a summary viewer in the script playground :)
