+++ UI/Features/ScriptPlayground/ScriptPlaygroundViewModel.cs
@@ -8,12 +8,27 @@ using System.Reactive.Subjects;
 
 namespace UI.Features.ScriptPlayground;
 
+public abstract record ScriptSummaryEvent;
+
+public sealed record SummaryInitializedEvent(IReadOnlyList<ScriptFunctionInvocation> Steps) : ScriptSummaryEvent;
+
+public sealed record SummaryCallStartedEvent(
+  long Sequence,
+  string CallId,
+  string ParentCallId,
+  ScriptFunctionInvocation Invocation) : ScriptSummaryEvent;
+
+public sealed record SummaryCallCompletedEvent(string CallId, Ares.Datamodel.AresValue Result) : ScriptSummaryEvent;
+
+public sealed record SummaryCallFailedEvent(string CallId, string Error) : ScriptSummaryEvent;
+
 public partial class ScriptPlaygroundViewModel : ReactiveObject
 {
   private readonly AresScriptingService.AresScriptingServiceClient _scriptingClient;
   private CancellationTokenSource _cancellationTokenSource = new();
   private readonly Subject<string> _scriptOutput = new();
   private readonly Subject<ScriptFunctionInvocation> _scriptInvocations = new();
+  private readonly Subject<ScriptSummaryEvent> _scriptSummaryEvents = new();
 
   public ScriptPlaygroundViewModel(
     AresScriptingService.AresScriptingServiceClient scriptingClient)
@@ -21,6 +36,7 @@ public partial class ScriptPlaygroundViewModel : ReactiveObject
     _scriptingClient = scriptingClient;
     ScriptOutput = _scriptOutput.AsObservable();
     ScriptInvocations = _scriptInvocations.AsObservable();
+    ScriptSummaryEvents = _scriptSummaryEvents.AsObservable();
   }
 
   public async Task StartScript(string script)
@@ -28,6 +44,9 @@ public partial class ScriptPlaygroundViewModel : ReactiveObject
     _cancellationTokenSource = new();
     var token = _cancellationTokenSource.Token;
     ScriptRunning = true;
+
+    await PublishSummaryAsync(script, token);
+
     var something = _scriptingClient.ExecuteScript(new ScriptExecutionRequest { Script = script }, new CallOptions(cancellationToken: token));
 
     try
@@ -37,6 +56,27 @@ public partial class ScriptPlaygroundViewModel : ReactiveObject
         if(output.FunctionStarted is not null)
         {
           _scriptInvocations.OnNext(output.FunctionStarted.Invocation);
+          _scriptSummaryEvents.OnNext(new SummaryCallStartedEvent(
+            output.Sequence,
+            output.FunctionStarted.CallId,
+            output.FunctionStarted.ParentCallId,
+            output.FunctionStarted.Invocation));
+          continue;
+        }
+
+        if(output.FunctionCompleted is not null)
+        {
+          _scriptSummaryEvents.OnNext(new SummaryCallCompletedEvent(
+            output.FunctionCompleted.CallId,
+            output.FunctionCompleted.Result));
+          continue;
+        }
+
+        if(output.FunctionFailed is not null)
+        {
+          _scriptSummaryEvents.OnNext(new SummaryCallFailedEvent(
+            output.FunctionFailed.CallId,
+            output.FunctionFailed.Error));
           continue;
         }
 
@@ -61,9 +101,30 @@ public partial class ScriptPlaygroundViewModel : ReactiveObject
     }
   }
 
+  public async Task BuildSummaryAsync(string script)
+  {
+    await PublishSummaryAsync(script, CancellationToken.None);
+  }
+
   public async Task StopScript()
   {
[…]
