### 2026-02-25 9ef8e0d
Fixed the script builder to support new type hinting.
