+++ AresService.Migrations.Sqlite/Migrations/AresDbContextModelSnapshot.cs
@@ -15,7 +15,7 @@ namespace AresService.Migrations.Sqlite.Migrations
         protected override void BuildModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
-            modelBuilder.HasAnnotation("ProductVersion", "10.0.0");
+            modelBuilder.HasAnnotation("ProductVersion", "10.0.1");
 
             modelBuilder.Entity("Ares.Datamodel.AnalysisOverview", b =>
                 {
@@ -435,15 +435,39 @@ namespace AresService.Migrations.Sqlite.Migrations
                     b.Property<string>("DeviceName")
                         .HasColumnType("TEXT");
 
-                    b.Property<string>("DeviceType")
+                    b.Property<string>("DriverId")
                         .HasColumnType("TEXT");
+
+                    b.Property<string>("DriverName")
+                        .HasColumnType("TEXT");
+
+                    b.Property<string>("DriverSettings")
+                        .HasColumnType("TEXT");
+
+                    b.Property<Guid?>("EthernetUniqueId")
+                        .HasColumnType("TEXT");
+
+                    b.Property<bool>("IsSimulated")
+                        .HasColumnType("INTEGER");
 
                     b.Property<DateTime>("LastModified")
                         .ValueGeneratedOnAddOrUpdate()
                         .HasColumnType("TEXT")
                         .HasDefaultValueSql("DATETIME('now')");
 
+                    b.Property<Guid?>("SerialUniqueId")
+                        .HasColumnType("TEXT");
+
+                    b.Property<Guid?>("UsbUniqueId")
+                        .HasColumnType("TEXT");
+
                     b.HasKey("UniqueId");
+
+                    b.HasIndex("EthernetUniqueId");
+
+                    b.HasIndex("SerialUniqueId");
+
+                    b.HasIndex("UsbUniqueId");
 
                     b.ToTable("DeviceConfigs", (string)null);
                 });
@@ -577,6 +601,36 @@ namespace AresService.Migrations.Sqlite.Migrations
                     b.ToTable("DeviceStates");
                 });
 
+            modelBuilder.Entity("Ares.Datamodel.Device.EthernetConnection", b =>
+                {
+                    b.Property<Guid>("UniqueId")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("TEXT");
+
+                    b.Property<DateTime>("CreationTime")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("TEXT")
+                        .HasDefaultValueSql("DATETIME('now')");
+
+                    b.Property<string>("IpAddress")
+                        .HasColumnType("TEXT");
+
+                    b.Property<DateTime>("LastModified")
+                        .ValueGeneratedOnAddOrUpdate()
+                        .HasColumnType("TEXT")
+                        .HasDefaultValueSql("DATETIME('now')");
+
+                    b.Property<int>("Port")
+                        .HasColumnType("INTEGER");
+
+                    b.Property<bool>("UseTls")
+                        .HasColumnType("INTEGER");
+
+                    b.HasKey("UniqueId");
+
+                    b.ToTable("EthernetConnection");
+                });
+
             modelBuilder.Entity("Ares.Datamodel.Device.RemoteDeviceConfig", b =>
                 {
                     b.Property<Guid>("UniqueId")
@@ -602,6 +656,75 @@ namespace AresService.Migrations.Sqlite.Migrations
                     b.HasKey("UniqueId");
 
                     b.ToTable("RemoteDevices", (string)null);
+                });
+
+            modelBuilder.Entity("Ares.Datamodel.Device.SerialConnection", b =>
+                {
+                    b.Property<Guid>("UniqueId")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("TEXT");
+
[…]
