### 2026-02-25 a77ff66
Fixed logging page crashing
