### 2026-02-26 a29b905
Remastered AlicatMFC is building now.
