+++ AlicatMFC/Commands/Requests/BasisCancelValveHoldCommand.cs
@@ -1,15 +1,0 @@
-﻿using AlicatMFC.Commands.Responses;
-using AlicatMFC.Commands.Responses.Parsers;
-using AlicatMFC.Commands.Responses.Streamed;
-
-namespace AlicatMFC.Commands.Requests;
-
-internal class BasisCancelValveHoldCommand : MfcCommandExpectingResponse<LiveDataResponse>
-{
-  public BasisCancelValveHoldCommand(char id, DataFrameFormatEntry[] formatEntries, string firmware) : base(id, new LiveDataParser(formatEntries), firmware)
-  {
-  }
-
-  protected override string SerializeToString()
-    => $"C";
-}
