### 2026-02-26 ad30547
Fixed some type inference not inferencing correctly
