### 2026-02-27 581aa17
Updated alicat csproj to output merged contents. Added yaml manifest
