+++ AlicatMFCRemastered/manifest.yaml
@@ -1,0 +1,30 @@
+# ARES Device Driver Manifest
+# Target: Alicat MFC Remastered
+
+# Basic Metadata
+device_type_name: "Alicat Mass Flow Controller"
+driver_id: "alicat_mfc_remastered_v1"
+connection_type: "Serial"
+
+# The DLL that contains the MassFlowController and ViewModel classes
+assembly_name: "AlicatMFCRemastered.dll"
+
+# Settings required when adding the device through the ARES UI
+settings:
+  - key: "serialId"
+    display_name: "Unit ID"
+    type: "string"
+    description: "The hardware ID of the Alicat unit (e.g., A, B, C)."
+    default: "A"
+
+  - key: "HasValve"
+    display_name: "Has Integrated Valve"
+    type: "boolean"
+    description: "Specify true if this unit is a Controller (MFC), false if it is a Meter (MFM)."
+    default: true
+
+  - key: "IsBasis"
+    display_name: "Is Basis Series"
+    type: "boolean"
+    description: "Set to true if using Alicat Basis/Basis 2 hardware."
+    default: false
