+++ AresService.Data/AresIdentityContext.cs
@@ -1,11 +1,0 @@
-﻿using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore;
-
-namespace AresService.Data;
-
-public class AresIdentityContext : IdentityDbContext<AresUser>
-{
-  public AresIdentityContext(DbContextOptions<AresIdentityContext> options) : base(options)
-  {
-  }
-}
