### 2026-02-27 7c9266c
Moved the experiment result json handler
