### 2026-02-27 c53f8fc
Axed the ARESStarter class. Moved any needed functionality over to the UI. AppSettings was deleted as well.
