+++ UI/Program.cs
@@ -1,7 +1,14 @@
+using Ares.Core;
+using Ares.Core.Grpc;
+using Ares.Services;
+using AresService.Data;
 using CommunityToolkit.Mvvm.Messaging;
+using Microsoft.Data.Sqlite;
 using Microsoft.EntityFrameworkCore;
 using Radzen;
 using Serilog;
+using System.CommandLine;
+using System.Reflection;
 using UI;
 using UI.Infrastructure.Grpc;
 using UI.Application.Notifications;
@@ -10,106 +17,332 @@ using UI.Infrastructure.Persistence;
 using UI.Components.Formatting;
 using UI.Application.Settings;
 
-var builder = WebApplication.CreateBuilder(args);
+#region Command Line Params
 
-// Add services to the container.
+  // 1. Define the command-line interface
+  var migrateOption = new Option<bool>(name: "--migrate")
+  {
+    Description = "Creates and/or updates the database"
+  };
 
-Log.Logger = new LoggerConfiguration()
-  .CreateBootstrapLogger();
+  var checkDbOption = new Option<bool>(name: "--check-database")
+  {
+    Description = "Checks if database exists and/or needs an update. Uses the exit code to return 0 if database is good, 1 if database is good but needs and update, 2 if database does not exist, 3 if other error.",
+  };
 
-builder.Configuration
-  .AddJsonFile("appsettings.ui.json", optional: false, reloadOnChange: true)
-  .AddJsonFile($"appsettings.ui.{builder.Environment.EnvironmentName}.json", optional: true);
+  var rootCommand = new RootCommand("Ares Service")
+  {
+    migrateOption,
+    checkDbOption
+  };
 
-ConfigureDatabaseServices(builder.Services, builder.Configuration);
+  rootCommand.SetAction(async parseResult =>
+  {
+    var shouldMigrate = parseResult.GetValue(migrateOption);
+    if(shouldMigrate)
+    {
+      await RunMigrationsAsync(args);
+      return 0;
+    }
 
-builder.Services.AddDatabaseDeveloperPageExceptionFilter();
+    var checkUpdate = parseResult.GetValue(checkDbOption);
+    if(checkUpdate)
+    {
+      return await CheckDatabase(args);
+    }
 
-builder.Services.AddRazorComponents().AddInteractiveServerComponents();
-builder.Services.AddCascadingAuthenticationState();
-builder.Services.AddRadzenComponents();
-builder.Services.Configure<RemoteServiceSettings>(builder.Configuration.GetSection(nameof(RemoteServiceSettings)));
-builder.Services.Configure<CertificateSettings>(builder.Configuration.GetSection(nameof(CertificateSettings)));
+    await RunWebAppAsync(args);
 
-builder.Services.AddSingleton<IMessenger>(provider => new WeakReferenceMessenger());
-builder.Services.AddScoped<IClientManager, ClientManager>();
-builder.Services.LoadAresModules();
-builder.Services.BindClients();
-builder.Services.AddSingleton<INotificationReceivingService, NotificationReceivingService>();
+    return 0;
+  });
 
-builder.Services.AddOptions();
-builder.Services.AddAntiforgery();
+  var parseResult = rootCommand.Parse(args);
 
-builder.Services.AddHostedService<ServiceStarter>();
+  return await parseResult.InvokeAsync();
 
-builder.Services.AddSerilog((services, lc) => lc
-  .ReadFrom.Configuration(builder.Configuration)
-  .ReadFrom.Services(services)
-  .WriteTo.Console()
-  .Enrich.FromLogContext());
+#endregion
 
-var app = builder.Build();
+#region Database Stuff
+
[…]
