+++ UI/Infrastructure/Grpc/ServiceMapper.cs
@@ -1,0 +1,16 @@
+﻿using Ares.Services;
+using Microsoft.AspNetCore.Authentication;
+using AresScriptingService=Ares.Core.Grpc.Services.AresScriptingService;
+using DeviceStateExportService=Ares.Core.Grpc.Services.DeviceStateExportService;
+
+namespace UI.Infrastructure.Grpc;
+
+public static class ServiceMapper
+{
+  public static void MapAresServices(this IEndpointRouteBuilder routeBuilder)
+  {
+    routeBuilder.MapGrpcService<AuthenticationService>();
+    routeBuilder.MapGrpcService<DeviceStateExportService>();
+    routeBuilder.MapGrpcService<AresScriptingService>();
+  }
+}
