### 2026-02-27 11f721b
Re-added the identity context. Moved all the service collection extension logic from service to the UI as appropriate
