+++ AresService/DbDesignFactories/AresIdentityContextFactory.cs
@@ -1,0 +1,7 @@
+﻿using AresService.Data;
+
+namespace AresService.DbDesignFactories;
+
+public class AresIdentityContextFactory : BaseDesignFactory<AresIdentityContext>
+{
+}
