+++ UI/Program.cs
@@ -19,195 +19,192 @@ using UI.Application.Settings;
 
 #region Command Line Params
 
-  // 1. Define the command-line interface
-  var migrateOption = new Option<bool>(name: "--migrate")
-  {
-    Description = "Creates and/or updates the database"
-  };
+// 1. Define the command-line interface
+var migrateOption = new Option<bool>(name: "--migrate")
+{
+  Description = "Creates and/or updates the database"
+};
 
-  var checkDbOption = new Option<bool>(name: "--check-database")
-  {
-    Description = "Checks if database exists and/or needs an update. Uses the exit code to return 0 if database is good, 1 if database is good but needs and update, 2 if database does not exist, 3 if other error.",
-  };
+var checkDbOption = new Option<bool>(name: "--check-database")
+{
+  Description = "Checks if database exists and/or needs an update. Uses the exit code to return 0 if database is good, 1 if database is good but needs and update, 2 if database does not exist, 3 if other error.",
+};
 
-  var rootCommand = new RootCommand("Ares Service")
+var rootCommand = new RootCommand("Ares Service")
   {
     migrateOption,
     checkDbOption
   };
 
-  rootCommand.SetAction(async parseResult =>
+rootCommand.SetAction(async parseResult =>
+{
+  var shouldMigrate = parseResult.GetValue(migrateOption);
+  if(shouldMigrate)
   {
-    var shouldMigrate = parseResult.GetValue(migrateOption);
-    if(shouldMigrate)
-    {
-      await RunMigrationsAsync(args);
-      return 0;
-    }
+    await RunMigrationsAsync(args);
+    return 0;
+  }
 
-    var checkUpdate = parseResult.GetValue(checkDbOption);
-    if(checkUpdate)
-    {
-      return await CheckDatabase(args);
-    }
+  var checkUpdate = parseResult.GetValue(checkDbOption);
+  if(checkUpdate)
+  {
+    return await CheckDatabase(args);
+  }
 
-    await RunWebAppAsync(args);
+  await RunWebAppAsync(args);
 
-builder.Services.AddSingleton<IMessenger>(provider => new WeakReferenceMessenger());
-builder.Services.AddScoped<IClientManager, ClientManager>();
-builder.Services.LoadAresModules();
-builder.Services.BindClients();
-builder.Services.AddSingleton<INotificationReceivingService, NotificationReceivingService>();
+  return 0;
+});
 
-  var parseResult = rootCommand.Parse(args);
+var parseResult = rootCommand.Parse(args);
 
-  return await parseResult.InvokeAsync();
+return await parseResult.InvokeAsync();
 
 #endregion
 
 #region Database Stuff
 
-  static async Task<int> CheckDatabase(string[] args)
+static async Task<int> CheckDatabase(string[] args)
+{
+  var host = Host.CreateApplicationBuilder(args);
+  host.Configuration
+    .AddJsonFile("appsettings.aresservice.json", optional: false, reloadOnChange: true)
+    .AddJsonFile($"appsettings.aresservice.{host.Environment.EnvironmentName}.json", optional: true);
+
+  try
   {
-    var host = Host.CreateApplicationBuilder(args);
-    host.Configuration
-      .AddJsonFile("appsettings.aresservice.json", optional: false, reloadOnChange: true)
-      .AddJsonFile($"appsettings.aresservice.{host.Environment.EnvironmentName}.json", optional: true);
-    
-    try
+    ConfigureDatabaseServices(host.Services, host.Configuration);
+    var app = host.Build();
+    var settings = host.Configuration.Get<AppSettings>();
+    var provider = settings.DatabaseProvider;
[…]
