### 2026-02-27 c10245f
Fixed rest of Program.cs
