### 2026-02-27 419bccf
Removed AresService project completely :)
