+++ AresService/Program.cs
@@ -1,413 +1,0 @@
-using System;
-using System.CommandLine;
-using System.IO;
-using System.Linq;
-using System.Reflection;
-using System.Security.Cryptography.X509Certificates;
-using System.Text;
-using System.Threading.Tasks;
-using Ares.Core;
-using Ares.Core.Grpc;
-using Ares.Datamodel;
-using Ares.Services;
-using AresService;
-using AresService.Data;
-using AresService.Services.Authentication;
-using Microsoft.AspNetCore.Authentication.Certificate;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.Data.Sqlite;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using Microsoft.IdentityModel.Tokens;
-using Serilog;
-
-public class Program
-{
-  public static async Task<int> Main(string[] args)
-  {
-    // 1. Define the command-line interface
-    var migrateOption = new Option<bool>(name: "--migrate")
-    {
-      Description = "Creates and/or updates the database"
-    };
-
-    var checkDbOption = new Option<bool>(name: "--check-database")
-    {
-      Description = "Checks if database exists and/or needs an update. Uses the exit code to return 0 if database is good, 1 if database is good but needs and update, 2 if database does not exist, 3 if other error.",
-    };
-
-    var rootCommand = new RootCommand("Ares Service")
-    {
-      migrateOption,
-      checkDbOption
-    };
-
-    rootCommand.SetAction(async parseResult =>
-    {
-      var shouldMigrate = parseResult.GetValue(migrateOption);
-      if(shouldMigrate)
-      {
-        await RunMigrationsAsync(args);
-        return 0;
-      }
-
-      var checkUpdate = parseResult.GetValue(checkDbOption);
-      if(checkUpdate)
-      {
-        return await CheckDatabase(args);
-      }
-
-      await RunWebAppAsync(args);
-
-      return 0;
-    });
-
-    var parseResult = rootCommand.Parse(args);
-
-    return await parseResult.InvokeAsync();
-  }
-
-  private static async Task<int> CheckDatabase(string[] args)
-  {
-    var host = Host.CreateApplicationBuilder(args);
-    host.Configuration
-      .AddJsonFile("appsettings.aresservice.json", optional: false, reloadOnChange: true)
-      .AddJsonFile($"appsettings.aresservice.{host.Environment.EnvironmentName}.json", optional: true);
-    
-    try
-    {
-      ConfigureDatabaseServices(host.Services, host.Configuration);
-      var app = host.Build();
-      var settings = host.Configuration.Get<AppSettings>();
-      var provider = settings?.DatabaseProvider;
-      var aresDbContextFactory = app.Services.GetRequiredService<IDbContextFactory<AresDbContext>>();
-      var aresIdentityContextFactory = app.Services.GetRequiredService<IDbContextFactory<AresIdentityContext>>();
-
-      var contextResult = 0;
-      await using(var dbContext = await aresDbContextFactory.CreateDbContextAsync())
-      {
-        contextResult |= await CheckContext(dbContext);
-      }
-
-      await using(var dbContext = await aresIdentityContextFactory.CreateDbContextAsync())
-      {
-        contextResult |= await CheckContext(dbContext);
-      }
-
-      return contextResult;
-    }
-    catch(Exception e)
-    {
-      Console.ForegroundColor = ConsoleColor.Red;
[…]
