### 2026-02-27 11126b6
Removed some stuff from Program into their own classes.
