### 2026-02-27 6c3c698
Replaced the usage of the gRPC services throughout the UI
