+++ UI/Features/CampaignEdit/ViewModels/CampaignListViewModel.cs
@@ -3,34 +3,35 @@ using ReactiveUI;
 using System.Collections.ObjectModel;
 using Ares.Datamodel.Templates;
 using Ares.Services;
+using Ares.Core.Grpc.Services;
 
 namespace UI.Features.CampaignEdit.ViewModels;
 
 public class CampaignListViewModel : ReactiveObject
 {
-  private readonly AresAutomation.AresAutomationClient _automationClient;
+  private readonly AutomationService _automationClient;
   public readonly ObservableCollection<CampaignTemplateSummary> Templates = [];
 
-  public CampaignListViewModel(AresAutomation.AresAutomationClient automationClient)
+  public CampaignListViewModel(AutomationService automationClient)
   {
     _automationClient = automationClient;
   }
 
   public async Task<CampaignTemplate?> GetFullCampaignTemplate(string campaignId)
   {
-    return await _automationClient.GetSingleCampaignAsync(new CampaignRequest { UniqueId = campaignId });
+    return await _automationClient.GetSingleCampaign(new CampaignRequest { UniqueId = campaignId }, null);
   }
 
   public async Task RefreshCampaigns()
   {
-    var campaigns = await _automationClient.GetAllCampaignsAsync(new GetAllCampaignsRequest());
+    var campaigns = await _automationClient.GetAllCampaigns(new GetAllCampaignsRequest(), null);
     Templates.Clear();
     Templates.AddRange(campaigns.Campaigns);
   }
 
   public async Task DeleteCampaign(Guid campaignId)
   {
-    await _automationClient.RemoveCampaignAsync(new CampaignRequest { UniqueId = campaignId.ToString() });
+    await _automationClient.RemoveCampaign(new CampaignRequest { UniqueId = campaignId.ToString() }, null);
     await RefreshCampaigns();
   }
 }
