+++ UI/Features/Execution/ExecutionViewModel.cs
@@ -3,6 +3,7 @@ using Ares.Datamodel.Analyzing;
 using Ares.Datamodel.Planning;
 using Ares.Datamodel.Templates;
 using Ares.Services;
+using Ares.Core.Grpc.Services;
 using DynamicData;
 using Google.Protobuf.WellKnownTypes;
 using ReactiveUI;
@@ -16,15 +17,15 @@ namespace UI.Features.Execution;
 
 public partial class ExecutionViewModel : ReactiveObject, INotifyPropertyChanged
 {
-  private readonly AresAutomation.AresAutomationClient _automationClient;
-  private readonly AresAnalyzerManagementService.AresAnalyzerManagementServiceClient _analyzerService;
+  private readonly AutomationService _automationClient;
+  private readonly AnalyzerService _analyzerService;
   public readonly ObservableCollection<CampaignTemplateSummary> CampaignTemplateSummaries = [];
   private readonly INotificationReceivingService _notificationService;
 
-  public ExecutionViewModel(AresAutomation.AresAutomationClient automationClient,
+  public ExecutionViewModel(AutomationService automationClient,
     IConfiguration configuration,
     INotificationReceivingService notificationService,
-    AresAnalyzerManagementService.AresAnalyzerManagementServiceClient analyzerService)
+    AnalyzerService analyzerService)
   {
     _automationClient = automationClient;
     _notificationService = notificationService;
@@ -40,7 +41,7 @@ public partial class ExecutionViewModel : ReactiveObject, INotifyPropertyChanged
 
   public async Task RefreshCampaigns()
   {
-    var campaigns = await _automationClient.GetAllCampaignsAsync(new GetAllCampaignsRequest());
+    var campaigns = await _automationClient.GetAllCampaigns(new GetAllCampaignsRequest(), null);
     CampaignTemplateSummaries.Clear();
     CampaignTemplateSummaries.AddRange(campaigns.Campaigns);
   }
@@ -50,15 +51,15 @@ public partial class ExecutionViewModel : ReactiveObject, INotifyPropertyChanged
     if(templateSummary is null || templateSummary is not CampaignTemplateSummary campaignTemplateSummary)
       return;
 
-    CampaignTemplate = await _automationClient.GetSingleCampaignAsync(new CampaignRequest { UniqueId = campaignTemplateSummary.UniqueId });
-    await _automationClient.SetCampaignForExecutionAsync(new CampaignRequest { UniqueId = CampaignTemplate.UniqueId });
+    CampaignTemplate = await _automationClient.GetSingleCampaign(new CampaignRequest { UniqueId = campaignTemplateSummary.UniqueId }, null);
+    await _automationClient.SetCampaignForExecution(new CampaignRequest { UniqueId = CampaignTemplate.UniqueId }, null);
     _ = UpdateCurrentTemplate();
     DisplayExecutionSummary = false;
   }
 
   public async Task UpdateCurrentTemplate()
   {
-    var currentTemplateOpt = await _automationClient.GetCurrentlySelectedCampaignAsync(new Empty());
+    var currentTemplateOpt = await _automationClient.GetCurrentlySelectedCampaign(new Empty(), null);
     CampaignTemplate = currentTemplateOpt.Value;
     if(CampaignTemplate is null)
       return;
@@ -74,55 +75,55 @@ public partial class ExecutionViewModel : ReactiveObject, INotifyPropertyChanged
     {
       var request = new AnalyzerInfoRequest();
       request.AnalyzerId = analyzerId;
-      var response = await _analyzerService.GetInfoAsync(request);
+      var response = await _analyzerService.GetInfo(request, null);
       AnalyzerInfo = response.Info;
     }
   }
 
   public async Task SetDesiredAnalysis()
   {
[…]

[…]
