### 2026-03-02 a433b1a
Deleted the device driver sync files as they are no longer needed post merge
