### 2026-03-02 ef453de
Removed the old USB and Serial device connector view models
