+++ UI/ServiceCollectionExtensions.cs
@@ -5,11 +5,8 @@ using Ares.Core;
 using Ares.Core.Grpc.Services;
 using Ares.Core.Grpc.Services.Safety;
 using Ares.Core.Grpc.Services.Notifications;
-using Ares.Services;
-using Ares.Services.Device;
 using AresService.Data;
 using CommunityToolkit.Mvvm.Messaging;
-using Grpc.Health.V1;
 using Microsoft.EntityFrameworkCore;
 using Radzen;
 using Serilog;
