+++ UI/Infrastructure/Devices/SerialDeviceConnectorViewModel.cs
@@ -1,100 +1,0 @@
-using System.Collections.ObjectModel;
-using System.Reactive.Disposables;
-using Ares.Datamodel.Device;
-using Ares.Services.Device;
-using Ares.Core.Grpc.Services;
-using DynamicData;
-using ReactiveUI;
-using UI.Application.Devices;
-
-namespace UI.Infrastructure.Devices;
-
-public abstract class SerialDeviceConnectorViewModel<TDeviceUnitVm> : ReactiveObject, IAsyncDisposable where TDeviceUnitVm : DeviceUnitControlViewModel
-{
-  private readonly ISourceCache<TDeviceUnitVm, string> _connectedSerialDeviceUnitControlVmsSource =
-  new SourceCache<TDeviceUnitVm, string>(vm => vm.DeviceId);
-  private readonly DevicesService _devicesClient;
-  private IDisposable _deviceUpdater = Disposable.Empty;
-  private CancellationTokenSource _deviceUpdaterTokenSource = new CancellationTokenSource();
-
-  public SerialDeviceConnectorViewModel(DevicesService devicesClient)
-  {
-    _devicesClient = devicesClient;
-
-    _connectedSerialDeviceUnitControlVmsSource.Connect().Bind(out var connectedSerialDeviceUnitControlVms).Subscribe();
-    ConnectedSerialDeviceUnitControlVms = connectedSerialDeviceUnitControlVms;
-  }
-
-  protected abstract Task<AresDeviceDescription[]> GetDeviceDescriptions();
-
-  public void Start(TimeSpan interval)
-  {
-    _deviceUpdater = Task.Factory.StartNew(async () =>
-    {
-      while(!_deviceUpdaterTokenSource.Token.IsCancellationRequested)
-      {
-        await UpdateAvailableDevices();
-        await Task.Delay(interval, _deviceUpdaterTokenSource.Token);
-      }
-    }, _deviceUpdaterTokenSource.Token);
-  }
-
-  private async Task UpdateAvailableDevices()
-  {
-    var descriptions = await GetDeviceDescriptions();
-    foreach(var description in descriptions)
-    {
-      var deviceStatusRequest = new DeviceStatusRequest { DeviceId = description.DeviceId };
-      var deviceOperationalStatusResponse = await _devicesClient.GetDeviceStatus(deviceStatusRequest, null);
-
-      if(deviceOperationalStatusResponse.OperationalState == OperationalState.Active)
-      {
-        if(ConnectedSerialDeviceUnitControlVms.Any(vm => vm.DeviceId == description.DeviceId))
-          continue;
-
-        var unitVm = CreateUnitVm(description);
-        _connectedSerialDeviceUnitControlVmsSource.AddOrUpdate(unitVm);
-        continue;
-      }
-
-      if(deviceOperationalStatusResponse.OperationalState == OperationalState.Inactive)
-      {
-        if(ConnectedSerialDeviceUnitControlVms.Any(vm => vm.DeviceId == description.DeviceId))
-          _connectedSerialDeviceUnitControlVmsSource.Remove(description.DeviceId);
-
-        continue;
-      }
-
-      if(deviceOperationalStatusResponse.OperationalState == OperationalState.Error)
-        if(ConnectedSerialDeviceUnitControlVms.Any(vm => vm.DeviceId == description.DeviceId))
-          _connectedSerialDeviceUnitControlVmsSource.Remove(description.DeviceId);
-    }
-  }
-
-  protected abstract TDeviceUnitVm CreateUnitVm(AresDeviceDescription description);
-
-  protected DevicesService DevicesClient => _devicesClient;
-  
[…]
