### 2026-03-02 455fda6
Added a separate store support in the environment for ares schemas so they can live alongsite values. That way we don't have to rely on trying to extract schema out of value
