### 2026-03-02 676140d
Refactored several repositories to follow modern .NET 10 standards.
