### 2026-03-02 e357b87
First time the new device has successfully registered on loading up ARES. Currently unable to add new devices as UI needs to be updated to support it.
