### 2026-03-03 6ad0638
Starting to implement the device plugins UI. To the point where the config editor appears, and you can press save but the logic for creating the devices is yet to be verified.
