+++ UI/Features/Devices/Plugin/Factories/IPluginViewModelFactory.cs
@@ -1,0 +1,6 @@
+﻿namespace UI.Features.Devices.Plugin.Factories;
+
+public interface IPluginViewModelFactory
+{
+  Task<PluginDeviceSettingsListViewModel?> CreateListViewModelAsync(string driverName);
+}
