### 2026-03-03 5b6871d
Added some logging to the driver loader
