### 2026-03-03 6122518
Temporarily removed the ILogger from device constructors. Looks like we're successfully hitting the constructor now, some bugs in the MFC however.
