+++ Ares.Core/Device/Repos/DeviceDriverRepo.cs
@@ -1,3 +1,4 @@
+using Ares.Core.Device.Drivers;
 using DynamicData;
 using System.Collections;
 
