+++ Ares.Core/Device/Repos/IDeviceDriverRepo.cs
@@ -1,3 +1,4 @@
+using Ares.Core.Device.Drivers;
 using Ares.Device;
 using DynamicData;
 
