+++ Ares.Core/Device/Providers/AresDriverProvider.cs
@@ -1,4 +1,5 @@
-﻿using Ares.Core.Device.Repos;
+﻿using Ares.Core.Device.Drivers;
+using Ares.Core.Device.Repos;
 using DynamicData;
 using System.Collections;
 
