+++ Ares.Core/Device/Providers/IAresDriverProvider.cs
@@ -1,4 +1,5 @@
-﻿using DynamicData;
+﻿using Ares.Core.Device.Drivers;
+using DynamicData;
 
 namespace Ares.Core.Device.Providers;
 

[…]
