### 2026-03-03 f17d941
Successfully constructed the first Alicat MFC, though some implementation details were missing so no UI appearance yet.
