### 2026-03-03 ea23972
Enhanced the function input/output contracts by introducing schemas. Instead of just using simple datatypes, functions can take parameters with schema hints, and define an output that's also based on a schema.
