### 2026-03-04 6845838
First time seeing a device successfully add, including a UI appearing on the dashboard. Looks like the config management needs a little work still.
