+++ UI/Features/Devices/Plugin/Factories/IPluginViewModelFactory.cs
@@ -1,6 +1,10 @@
-﻿namespace UI.Features.Devices.Plugin.Factories;
+﻿using Ares.Device;
+using Ares.Toolkit.Device.UI;
+
+namespace UI.Features.Devices.Plugin.Factories;
 
 public interface IPluginViewModelFactory
 {
   Task<PluginDeviceSettingsListViewModel?> CreateListViewModelAsync(string driverName);
+  Task<IDeviceUnitControlViewModel> CreateUnitControlViewModel(IAresDevice device);
 }
