+++ UI/Features/Devices/Remote/RemoteDeviceSettingsViewModel.cs
@@ -141,7 +141,9 @@ public partial class RemoteDeviceSettingsViewModel : ReactiveObject
   private async Task RemoveAsync(Func<Task> onRemoveCallback)
   {
     var request = new RemoveRemoteDeviceRequest() { DeviceId = _deviceInfo.UniqueId };
-    _deviceDeletionMessenger.Send(new DeviceDeletedMessage(_deviceInfo.UniqueId));
+    //TODO: FIX MEEEEEEEEEE
+    throw new NotImplementedException();
+    //_deviceDeletionMessenger.Send(new DeviceDeletedMessage(_deviceInfo.UniqueId));
     await _devicesClient.RemoveRemoteDevice(request, null);
     await onRemoveCallback();
   }

[…]
