### 2026-03-04 1388701
Updated some of the organization in the device section of the ARES core. Added the config provider to make it follow the same "read-only" access patterns as the drivers and devices.
