+++ Ares.Core/Device/Repos/IDeviceDriverRepo.cs
@@ -1,4 +1,4 @@
-using Ares.Core.Device.Drivers;
+using Ares.Core.Device.Plugins.Drivers;
 using Ares.Device;
 using DynamicData;
 
