### 2026-03-04 f68d5dc
Changed name of the driver provider. Updated view model to use the new config provider. Made the method for creating the dashboard view models not async.
