### 2026-03-05 1567259
Updated the data model version. Some changes to the device config
