+++ Ares.Core/Device/IDeviceConfigManager.cs
@@ -4,8 +4,12 @@ namespace Ares.Core.Device;
 
 public interface IDeviceConfigManager
 {
+  /// <summary>
+  /// Initializes the config manager class by adding all known device configs from the database to the local repo 
+  /// </summary>
+  /// <returns></returns>
+  Task LoadConfigs();
   Task Add(string deviceId, string deviceName, DeviceConfig config); 
   Task Remove(string configId);
   Task Update(string configId, DeviceConfig config);
-  Task<DeviceConfig?> GetConfig(string deviceId);
 }
