+++ Ares.Core/EntityConfigurations/Device/Connections/EthernetConnectionEntityConfiguration.cs
@@ -1,7 +1,0 @@
-﻿using Ares.Datamodel.Device;
-
-namespace Ares.Core.EntityConfigurations.Device.Connections;
-
-internal class EthernetConnectionEntityConfiguration : AresEntityTypeBaseConfiguration<EthernetConnection>
-{
-}
