+++ Ares.Core/EntityConfigurations/Device/Connections/UsbConnectionEntityConfiguration.cs
@@ -1,7 +1,0 @@
-﻿using Ares.Datamodel.Device;
-
-namespace Ares.Core.EntityConfigurations.Device.Connections;
-
-internal class UsbConnectionEntityConfiguration : AresEntityTypeBaseConfiguration<UsbConnection>
-{
-}

[…]
