### 2026-03-06 60a0b41
Fixed up one of the entity configurations
