### 2026-03-09 c8ebd61
Updated migration. Added some missing methods to the device service in the datamodel and implemented them. Devices are now loading, and can be updated or removed. Still a bug in the dashboard, not certain if that's the fault of the device or something in the loading process. Also still need to make the driver replacement process more robust.
