+++ AresService.Migrations.Sqlite/Migrations/AresDbContextModelSnapshot.cs
@@ -432,19 +432,16 @@ namespace AresService.Migrations.Sqlite.Migrations
                         .HasColumnType("TEXT")
                         .HasDefaultValueSql("DATETIME('now')");
 
+                    b.Property<string>("DeviceId")
+                        .HasColumnType("TEXT");
+
                     b.Property<string>("DeviceName")
                         .HasColumnType("TEXT");
 
+                    b.Property<string>("DeviceSettings")
+                        .HasColumnType("TEXT");
+
                     b.Property<string>("DriverId")
-                        .HasColumnType("TEXT");
-
-                    b.Property<string>("DriverName")
-                        .HasColumnType("TEXT");
-
-                    b.Property<string>("DriverSettings")
-                        .HasColumnType("TEXT");
-
-                    b.Property<Guid?>("EthernetUniqueId")
                         .HasColumnType("TEXT");
 
                     b.Property<bool>("IsSimulated")
@@ -455,19 +452,12 @@ namespace AresService.Migrations.Sqlite.Migrations
                         .HasColumnType("TEXT")
                         .HasDefaultValueSql("DATETIME('now')");
 
-                    b.Property<Guid?>("SerialUniqueId")
-                        .HasColumnType("TEXT");
-
-                    b.Property<Guid?>("UsbUniqueId")
+                    b.Property<Guid?>("SerialInfoUniqueId")
                         .HasColumnType("TEXT");
 
                     b.HasKey("UniqueId");
 
-                    b.HasIndex("EthernetUniqueId");
-
-                    b.HasIndex("SerialUniqueId");
-
-                    b.HasIndex("UsbUniqueId");
+                    b.HasIndex("SerialInfoUniqueId");
 
                     b.ToTable("DeviceConfigs", (string)null);
                 });
@@ -601,36 +591,6 @@ namespace AresService.Migrations.Sqlite.Migrations
                     b.ToTable("DeviceStates");
                 });
 
-            modelBuilder.Entity("Ares.Datamodel.Device.EthernetConnection", b =>
-                {
-                    b.Property<Guid>("UniqueId")
-                        .ValueGeneratedOnAdd()
-                        .HasColumnType("TEXT");
-
-                    b.Property<DateTime>("CreationTime")
-                        .ValueGeneratedOnAdd()
-                        .HasColumnType("TEXT")
-                        .HasDefaultValueSql("DATETIME('now')");
-
-                    b.Property<string>("IpAddress")
-                        .HasColumnType("TEXT");
-
-                    b.Property<DateTime>("LastModified")
-                        .ValueGeneratedOnAddOrUpdate()
-                        .HasColumnType("TEXT")
-                        .HasDefaultValueSql("DATETIME('now')");
-
-                    b.Property<int>("Port")
-                        .HasColumnType("INTEGER");
-
-                    b.Property<bool>("UseTls")
-                        .HasColumnType("INTEGER");
-
-                    b.HasKey("UniqueId");
-
-                    b.ToTable("EthernetConnection");
-                });
-
             modelBuilder.Entity("Ares.Datamodel.Device.RemoteDeviceConfig", b =>
                 {
                     b.Property<Guid>("UniqueId")
@@ -664,67 +624,22 @@ namespace AresService.Migrations.Sqlite.Migrations
                         .ValueGeneratedOnAdd()
                         .HasColumnType("TEXT");
 
-                    b.Property<int>("BaudRate")
-                        .HasColumnType("INTEGER");
-
                     b.Property<DateTime>("CreationTime")
                         .ValueGeneratedOnAdd()
                         .HasColumnType("TEXT")
                         .HasDefaultValueSql("DATETIME('now')");
-
-                    b.Property<int>("DataBits")
[…]
