### 2026-03-09 480ee6c
Moved device config manager to proper folder
