### 2026-03-09 6b0ad85
Updated database to store an archive of the last known available drivers. This enables us to support smart swapping of drivers around.
