+++ AresService.Migrations.Sqlite/Migrations/20260309165705_AddedDriverEntity_AresDbContext.cs
@@ -1,0 +1,39 @@
+﻿using System;
+using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace AresService.Migrations.Sqlite.Migrations.AresDb
+{
+    /// <inheritdoc />
+    public partial class AddedDriverEntity_AresDbContext : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.CreateTable(
+                name: "DeviceDrivers",
+                columns: table => new
+                {
+                    UniqueId = table.Column<Guid>(type: "TEXT", nullable: false),
+                    DriverId = table.Column<string>(type: "TEXT", nullable: true),
+                    DisplayName = table.Column<string>(type: "TEXT", nullable: true),
+                    Version = table.Column<string>(type: "TEXT", nullable: true),
+                    FileSizeBytes = table.Column<long>(type: "INTEGER", nullable: false),
+                    CreationTime = table.Column<DateTime>(type: "TEXT", nullable: false, defaultValueSql: "DATETIME('now')"),
+                    LastModified = table.Column<DateTime>(type: "TEXT", nullable: false, defaultValueSql: "DATETIME('now')")
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_DeviceDrivers", x => x.UniqueId);
+                });
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropTable(
+                name: "DeviceDrivers");
+        }
+    }
+}

[…]
