### 2026-03-09 5978b4e
Fixed the CSS not loading for device driver plugins.
