+++ UI/WebApplicationExtensions.cs
@@ -1,6 +1,8 @@
 using Ares.Core.Grpc;
 using Ares.Services;
+using Microsoft.Extensions.FileProviders;
 using UI.Components.Formatting;
+using UI.Application.Settings;
 using UI.Infrastructure.Grpc;
 
 namespace UI;
@@ -9,6 +11,16 @@ internal static class WebApplicationExtensions
 {
   public static void UseAresUiPipeline(this WebApplication app)
   {
+    var pluginPath = PluginPathResolver.Resolve(app.Configuration.Get<AppSettings>());
+    if(Directory.Exists(pluginPath))
+    {
+      app.UseStaticFiles(new StaticFileOptions
+      {
+        FileProvider = new PhysicalFileProvider(pluginPath),
+        RequestPath = "/plugins"
+      });
+    }
+
     app.UseStatusCodePagesWithReExecute("/404");
     app.UseHttpsRedirection();
     app.MapStaticAssets();
