### 2026-03-09 d0e9e23
Fixed the MFC driver plugin to be in a decently working condition. Some things are still hardcoded but the proof of concept is working.
