### 2026-03-10 ea2064c
Updated devices and remote devices to use a connection info object to pass nicely typed information needed by all device types.
