### 2026-03-10 58b849f
Updated migrations. Addressed a bug that caused the device configs to not be updated properly. Added additional handling logic that cleans up ghost devices that lack a driver.
