+++ AresService.Migrations.Postgres/Migrations/20260310145633_AddedConnectionInfo_AresDbContext.cs
@@ -1,0 +1,39 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace AresService.Migrations.Postgres.Migrations.AresDb
+{
+    /// <inheritdoc />
+    public partial class AddedConnectionInfo_AresDbContext : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<int>(
+                name: "BaudRate",
+                table: "SerialConnection",
+                type: "integer",
+                nullable: false,
+                defaultValue: 0);
+
+            migrationBuilder.AddColumn<string>(
+                name: "SerialId",
+                table: "SerialConnection",
+                type: "text",
+                nullable: true);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "BaudRate",
+                table: "SerialConnection");
+
+            migrationBuilder.DropColumn(
+                name: "SerialId",
+                table: "SerialConnection");
+        }
+    }
+}
