### 2026-03-10 7272b9f
Fixed the buttons on the MFC dashboard UI. Working on dealing with the special settings that currently appear on the device settings view in the old system
