+++ Ares.Core/Device/Managers/DeviceConfigManager.cs
@@ -1,23 +1,41 @@
-﻿using Ares.Core.Device.Repos;
+﻿using Ares.Core.Device.Plugins.Drivers;
+using Ares.Core.Device.Providers;
+using Ares.Core.Device.Repos;
+using Ares.Core.Notifications;
 using Ares.Datamodel.Device;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
 
 namespace Ares.Core.Device.Managers;
 
 public class DeviceConfigManager : IDeviceConfigManager
 {
   private readonly IDbContextFactory<CoreDatabaseContext> _dbContextFactory;
+  private readonly IDriverDatabaseManager _driverDbManager;
+  private readonly IDeviceDriverProvider _driverProvider;
+  private readonly ILogger<DeviceConfigManager> _logger;
+  private readonly INotificationHandler _notificationHandler;
   private IDeviceConfigRepo _configRepo;
 
-  public DeviceConfigManager(IDbContextFactory<CoreDatabaseContext> dbContextFactory, IDeviceConfigRepo configRepo)
+  public DeviceConfigManager(IDbContextFactory<CoreDatabaseContext> dbContextFactory, 
+    IDriverDatabaseManager driverDbManager, 
+    IDeviceDriverProvider driverProvider, 
+    IDeviceConfigRepo configRepo,
+    ILogger<DeviceConfigManager> logger,
+    INotificationHandler notificationHandler)
   {
     _dbContextFactory = dbContextFactory;
     _configRepo = configRepo;
+    _driverDbManager = driverDbManager;
+    _driverProvider = driverProvider;
+    _notificationHandler = notificationHandler;
+    _logger = logger;
   }
 
   public async Task LoadConfigs()
   {
     await using var context = _dbContextFactory.CreateDbContext();
+    await HandleMissingDriver(context);
     var existingDeviceConfigs = await context.DeviceConfigs.ToListAsync();
     existingDeviceConfigs.ForEach(_configRepo.AddOrUpdate);
   }
@@ -66,5 +84,63 @@ public class DeviceConfigManager : IDeviceConfigManager
 
     await context.SaveChangesAsync();
     _configRepo.AddOrUpdate(existingConfig);
+  }
+
+  private async Task HandleMissingDriver(CoreDatabaseContext context)
+  {
+    var configs = await context.DeviceConfigs.ToListAsync();
+    var archivedDrivers = await _driverDbManager.GetAllDrivers();
+    var currentDrivers = _driverProvider.GetAllDeviceDrivers();
+
+    // Optimization: Pre-compute fast lookups to avoid O(N*M) loop iterations
+    var currentDriverIds = currentDrivers.Select(d => d.UniqueId).ToHashSet();
+    var archivedDriverMap = archivedDrivers.ToDictionary(d => d.DriverId);
+
+    // Track if we made any migrations so we can save the DB context once at the end
+    bool hasUpdates = false;
+
+    foreach(var config in configs)
+    {
+      // Driver found, no action needed
+      if(currentDriverIds.Contains(config.DriverId))
+        continue;
+
+      // Driver missing. Check the archive for a migration path.
+      if(archivedDriverMap.TryGetValue(config.DriverId, out var archivedDriver))
+      {
+        var currentMatch = currentDrivers.FirstOrDefault(cd => cd.Manifest.DeviceTypeName == archivedDriver.DisplayName);
+
+        if(currentMatch is not null)
+        {
+          // Migration successful: map to the new driver and skip deletion
+          config.DriverId = currentMatch.UniqueId;
+          hasUpdates = true;
+          continue;
+        }
+
+        // Migration failed: Archive exists, but no replacement driver found
[…]
