### 2026-03-11 c74d47a
Enforced and implemented a "GetSettings" method for devices that returns the current values for runtime settings.
