### 2026-03-11 7afa06d
Some minor cleanup in the managers
