+++ Ares.Core/Device/Managers/IDeviceManager.cs
@@ -8,15 +8,6 @@ public interface IDeviceManager
   Task<IAresDevice?> Create(DeviceConfig config);
   Task<IAresDevice?> Load(string deviceId, DeviceConfig config);
   Task<IAresDevice?> Update(string deviceId, DeviceConfig config);
-  /// <summary>
-  /// Retrieves all managed devices of a specific type.
-  /// </summary>
-  IReadOnlyCollection<T> GetAll<T>() where T : IAresDevice;
-
-  /// <summary>
-  /// Retrieves a device of a specific type by its ID.
-  /// </summary>
-  T? GetDevice<T>(string id) where T : class, IAresDevice;
   Task Remove(string deviceId);
   void Initialize();
 }
