### 2026-03-11 c6f261c
Added a submit button to the device settings view for updating the settings. Wrapped the update settings login in a try-catch to avoid UI crashes when devices fail to update their information
