### 2026-03-11 f2c2d0d
MFC's are doing more or less the things they should now! Some clean up left to do and still need to update the other devices
