+++ AlicatMFCRemastered/MassFlowController.cs
@@ -1,16 +1,19 @@
-﻿using AlicatMFCRemastered.Commands.Requests;
+﻿using AlicatMFCRemasterd.Commands;
+using AlicatMFCRemastered.Commands.Requests;
 using AlicatMFCRemastered.Commands.Responses;
 using AlicatMFCRemastered.Commands.Responses.Streamed;
+using AlicatMFCRemastered.Commands.Types;
 using AlicatMFCRemastered.Enums;
 using AlicatMFCRemastered.Models;
 using AlicatMFCRemastered.Simulation;
 using Ares.Datamodel;
 using Ares.Datamodel.Device;
 using Ares.Datamodel.Extensions;
+using Ares.Datamodel.Factories;
 using Ares.Device;
 using Ares.Toolkit.Serial;
-using DynamicData;
 using Parsers.AlicatMFCRemastered;
+using ReactiveUI;
 using System.Diagnostics;
 using System.Reactive.Disposables;
 using System.Reactive.Linq;
@@ -28,7 +31,6 @@ public class MassFlowController : AresDevice, IMassFlowController
   private CancellationTokenSource _stateGetterLoopTokenSource = new();
   private CompositeDisposable _stateWatchers = new();
   private Task _stateUpdater = Task.CompletedTask;
-  //readonly ILogger<IAresDevice> _logger;
   private List<GasInfoEntry> _gases = new();
   private List<ManufacturerInfoEntry> _manufacturerInfo = new();
   private List<DataFrameFormatEntry> _dataFrameFormatEntries = new();
@@ -475,6 +477,7 @@ public class MassFlowController : AresDevice, IMassFlowController
     {
       await Initialize();
       activated = true;
+      Status = new DeviceOperationalStatus { OperationalState = OperationalState.Active, Message = $"MFC {Name} is active!" };
     }
     catch(Exception e)
     {
@@ -814,9 +817,117 @@ public class MassFlowController : AresDevice, IMassFlowController
     await StartUpdateLoop(TimeSpan.FromMilliseconds(500));
   }
 
-  public override Task<CommandResult> ExecuteCommand(string command, List<DeviceCommandArgument> arguments, CancellationToken token)
+  
+  public override async Task<CommandResult> ExecuteCommand(string command, List<DeviceCommandArgument> arguments, CancellationToken token)
   {
-    throw new NotImplementedException();
+    if(!Enum.TryParse<MassFlowControllerCommand>(command, out var deviceCommandEnum))
+    {
+      return new CommandResult
+      {
+        Success = false,
+        Error = $"Invalid or unsupported command: '{command}'"
+      };
+    }
+
+    var result = new CommandResult { Success = true };
+
+    // Helper function to safely extract an argument by name (assuming DeviceCommandArgument has Name and Value properties)
+    AresValue? GetArg(MassFlowControllerCommandParameter param) =>
+        arguments.FirstOrDefault(a => a.ArgName == param.ToString())?.ArgValue;
+
+    try
+    {
+      // Route the command
+      switch(deviceCommandEnum)
+      {
+        case MassFlowControllerCommand.PollLiveDataFrame:
+          // TODO: stringify live info
+          break;
+
+        case MassFlowControllerCommand.ManufacturerInfo:
+          // await Device.GetManufacturerDataInfoAsync(token);
+          break;
+
+        case MassFlowControllerCommand.CancelValveHold:
+          await CancelValveHold();
+          break;
+
+        case MassFlowControllerCommand.ChooseDifferentGas:
+          if(GetArg(MassFlowControllerCommandParameter.GasNumber) is not { HasNumberValue: true, NumberValue: var gasNum })
[…]
