### 2026-03-12 f4ee558
Use the schema stringifyer from the datamodel helper.
