### 2026-03-12 0ebb302
Typehinting IsCompatible method added. Compares two QuantitySchema instances for compatibility.
