+++ AresScript/AresScriptTypeHints.cs
@@ -2,6 +2,7 @@ using Ares.Datamodel;
 using Ares.Datamodel.Extensions;
 using Ares.Datamodel.Factories;
 using AresScript.Generated;
+using UnitsNet;
 
 namespace AresScript;
 
@@ -154,9 +155,150 @@ internal static class AresScriptTypeHints
         }
 
         return IsCompatible(expected.ListElementSchema, actual.ListElementSchema);
+      case AresDataType.Quantity:
+        return IsCompatible(expected.QuantitySchema, actual.QuantitySchema);
       default:
         return true;
     }
+  }
+
+  private static bool IsCompatible(QuantitySchema? expected, QuantitySchema? actual)
+  {
+    if(expected is null || actual is null)
+    {
+      return true;
+    }
+
+    if(expected.QuantityType != QuantityType.Unspecified
+      && actual.QuantityType != QuantityType.Unspecified
+      && expected.QuantityType != actual.QuantityType)
+    {
+      return false;
+    }
+
+    if(expected.HasMinScalarValue
+      && actual.HasMinScalarValue
+      && !IsQuantityBoundCompatible(expected, actual, isMinBound: true))
+    {
+      return false;
+    }
+
+    if(expected.HasMaxScalarValue
+      && actual.HasMaxScalarValue
+      && !IsQuantityBoundCompatible(expected, actual, isMinBound: false))
+    {
+      return false;
+    }
+
+    return true;
+  }
+
+  private static bool IsQuantityBoundCompatible(
+    QuantitySchema expectedSchema,
+    QuantitySchema actualSchema,
+    bool isMinBound)
+  {
+    var expectedScalar = isMinBound ? expectedSchema.MinScalarValue : expectedSchema.MaxScalarValue;
+    var actualScalar = isMinBound ? actualSchema.MinScalarValue : actualSchema.MaxScalarValue;
+
+    if(TryConvertToComparableScalars(expectedSchema, actualSchema, isMinBound, out var expectedComparable, out var actualComparable))
+    {
+      return isMinBound
+        ? actualComparable >= expectedComparable
+        : actualComparable <= expectedComparable;
+    }
+
+    return isMinBound
+      ? actualScalar >= expectedScalar
+      : actualScalar <= expectedScalar;
+  }
+
+  private static bool TryConvertToComparableScalars(
+    QuantitySchema expectedSchema,
+    QuantitySchema actualSchema,
+    bool isMinBound,
+    out double expectedComparable,
+    out double actualComparable)
+  {
+    var expectedScalar = isMinBound ? expectedSchema.MinScalarValue : expectedSchema.MaxScalarValue;
+    var actualScalar = isMinBound ? actualSchema.MinScalarValue : actualSchema.MaxScalarValue;
+
+    expectedComparable = expectedScalar;
+    actualComparable = actualScalar;
+
+    if(!TryToUnitsNetQuantity(expectedSchema, actualSchema, expectedScalar, out var expectedQuantity))
+    {
+      return false;
+    }
+
+    if(!TryToUnitsNetQuantity(actualSchema, expectedSchema, actualScalar, out var actualQuantity))
+    {
+      return false;
+    }
+
+    var expectedBaseUnit = expectedQuantity.QuantityInfo.BaseUnitInfo.Value;
+    var actualBaseUnit = actualQuantity.QuantityInfo.BaseUnitInfo.Value;
+    expectedComparable = expectedQuantity.As(expectedBaseUnit);
+    actualComparable = actualQuantity.As(actualBaseUnit);
+    return true;
+  }
+
+  private static bool TryToUnitsNetQuantity(
+    QuantitySchema schema,
[…]
