### 2026-03-12 5cf1e9c
Fixed up the MFC view and corrected some small bugs in the state data. Simulations now seem to be providing good sim data and the views look accurate as to what they were before
