### 2026-03-12 08417c3
Removed the AlicatMFC from the project
