+++ AlicatMFCRemastered/Simulation/IAlicatSim.cs
@@ -1,7 +1,0 @@
-﻿namespace AlicatMFCRemastered.Simulation;
-
-internal interface IAlicatSim : IDisposable
-{
-  char DeviceId { get; }
-  Task SendCommand(byte[] command);
-}
