+++ AlicatMFCRemastered/Commands/Requests/HoldValvesClosedCommand.cs
@@ -1,14 +1,0 @@
-﻿using AlicatMFCRemastered.Commands.Responses;
-using AlicatMFCRemastered.Commands.Responses.Parsers;
-using AlicatMFCRemastered.Commands.Responses.Streamed;
-
-namespace AlicatMFCRemastered.Commands.Requests;
-
-internal class HoldValvesClosedCommand : MfcCommandExpectingResponse<LiveDataResponse>
-{
-  public HoldValvesClosedCommand(char id, DataFrameFormatEntry[] formatEntries, string firmware) : base(id, new LiveDataParser(formatEntries), firmware)
-  { }
-
-  protected override string SerializeToString()
-    => "HC";
-}
