+++ AlicatMFCRemastered/Commands/Responses/GenericLineResponse.cs
@@ -1,8 +1,0 @@
-﻿namespace AlicatMFCRemastered.Commands.Responses;
-
-internal class GenericLineResponse : CommandResponse
-{
-  public GenericLineResponse(char id) : base(id)
-  {
-  }
-}
