+++ AlicatMFCRemastered/Enums/MassFlowControllerCommandParameter.cs
@@ -1,10 +1,0 @@
-﻿namespace AlicatMFCRemastered.Enums;
-
-public enum MassFlowControllerCommandParameter
-{
-  GasNumber,
-  MixNumber,
-  Setpoint,
-  DeviceId,
-  ValvePercent
-}
