+++ AlicatMFCRemastered/Commands/Types/MfcDataTypes.cs
@@ -1,11 +1,0 @@
-﻿using Ares.Datamodel;
-
-namespace AlicatMFCRemastered.Commands.Types;
-
-public static class MfcDataTypes
-{
-  public static readonly KeyValuePair<string, AresDataType> Setpoint = new KeyValuePair<string, AresDataType>(
-    "Setpoint",
-    AresDataType.Number);
-}
-
