+++ HerkulexDRS/Simulation/SimulatedServo.cs
@@ -1,4 +1,4 @@
-﻿using System.Diagnostics;
+using System.Diagnostics;
 
 namespace HerkulexDRS.Simulation;
 public class SimulatedServo
