+++ ValveController/Commands/RelayOne/DisengageRelayOneCommand.cs
@@ -1,4 +1,4 @@
-﻿using Ares.Device.Serial.Commands;
+using Ares.Toolkit.Serial.Commands;
 
 namespace ValveController.Commands.RelayOne;
 public class DisengageRelayOneCommand : SerialCommand
