+++ ValveController/Commands/RelayTwo/DisengageRelayTwoCommand.cs
@@ -1,4 +1,4 @@
-﻿using Ares.Device.Serial.Commands;
+using Ares.Toolkit.Serial.Commands;
 
 namespace ValveController.Commands.RelayTwo;
 public class DisengageRelayTwoCommand : SerialCommand
