+++ ValveController/Commands/Responses/Parsers/RelayStatusResponseParser.cs
@@ -1,45 +1,47 @@
-﻿using Ares.Device.Serial.Commands;
+using Ares.Toolkit.Serial.Commands;
 
 namespace ValveController.Commands.Responses.Parsers;
+
 public class RelayStatusResponseParser : SerialResponseParser<RelayStatusResponse>
 {
   public override bool TryParseResponse(byte[] buffer, out RelayStatusResponse? response, out ArraySegment<byte>? dataToRemove)
   {
-    if (buffer != null && buffer.Length == 1)
+    if (buffer == null || buffer.Length < 1)
     {
-      var responseInteger = (int)buffer[0];
-      ArraySegment<byte> bytes = new ArraySegment<byte>(buffer, 0, 1);
-
-      switch (responseInteger)
-      {
-        case 0:
-          response = new RelayStatusResponse() { RelayOneOn = false, RelayTwoOn = false };
-          dataToRemove = bytes;
-          return true;
-
-        case 1:
-          response = new RelayStatusResponse() { RelayOneOn = true, RelayTwoOn = false };
-          dataToRemove = bytes;
-          return true;
-
-        case 2:
-          response = new RelayStatusResponse() { RelayOneOn = false, RelayTwoOn = true };
-          dataToRemove = bytes;
-          return true;
-
-        case 3:
-          response = new RelayStatusResponse() { RelayOneOn = true, RelayTwoOn = true };
-          dataToRemove = bytes;
-          return true;
-
-        default:
-          break;
-
-      }
+      response = null;
+      dataToRemove = null;
+      return false;
     }
 
-    response = null;
-    dataToRemove = null;
-    throw new Exception("Unexpected Response from Relay Board");
+    var responseInteger = (int)buffer[0];
+    ArraySegment<byte> bytes = new ArraySegment<byte>(buffer, 0, 1);
+
+    switch (responseInteger)
+    {
+      case 0:
+        response = new RelayStatusResponse() { RelayOneOn = false, RelayTwoOn = false };
+        dataToRemove = bytes;
+        return true;
+
+      case 1:
+        response = new RelayStatusResponse() { RelayOneOn = true, RelayTwoOn = false };
+        dataToRemove = bytes;
+        return true;
+
+      case 2:
+        response = new RelayStatusResponse() { RelayOneOn = false, RelayTwoOn = true };
+        dataToRemove = bytes;
+        return true;
+
+      case 3:
+        response = new RelayStatusResponse() { RelayOneOn = true, RelayTwoOn = true };
+        dataToRemove = bytes;
+        return true;
+
+      default:
+        response = null;
+        dataToRemove = new ArraySegment<byte>(buffer, 0, 1); // Remove the invalid byte
+        return false;
+    }
   }
 }
