### 2026-03-12 c11b34a
Quantity symbol provider.
