### 2026-03-12 3e261f7
Updated the V6 Laser
