+++ VerdiV6Laser/LaserConnection.cs
@@ -1,13 +1,17 @@
-﻿using Ares.Device.Serial;
+using Ares.Toolkit.Serial;
 using System.IO.Ports;
 
-namespace VerdiV6Laser
-{
-  public class LaserConnection : AresHardwareConnection, ILaserConnection
-  {
-    public LaserConnection(string portName) : base(new SerialPortConnectionInfo(19200, System.IO.Ports.Parity.None, 8, StopBits.One), portName)
-    {
+namespace VerdiV6Laser;
 
+public class LaserConnection : AresHardwareConnection, ILaserConnection
+{
+  public LaserConnection(string portName) : base(new SerialPortConnectionInfo(19200, Parity.None, 8, StopBits.One), portName,
+    new SerialConnectionOptions
+    {
+      SendBuffer = TimeSpan.FromMilliseconds(50),
+      SendTimeout = TimeSpan.FromSeconds(2)
     }
+  )
+  {
   }
 }
