+++ VerdiV6Laser/IVerdiV6Laser.cs
@@ -1,17 +1,14 @@
-﻿using Ares.Device.Serial;
+namespace VerdiV6Laser;
 
-namespace VerdiV6Laser
+public interface IVerdiV6Laser : IAsyncDisposable
 {
-  public interface IVerdiV6Laser : ISerialDevice<ILaserConnection>, IAsyncDisposable
-  {
-    Task ActivateLaser();
-    Task DeactivateLaser();
-    Task SetLaserPower(double desiredPower);
-    Task SetLaserShutter(bool shutter);
-    Task<bool> GetLaserShutter();
-    Task<double> GetLaserPower();
-    double CurrentPower { get; }
-    double DesiredPower { get; }
-    bool Shutter { get; }
-  }
+  Task ActivateLaser();
+  Task DeactivateLaser();
+  Task SetLaserPower(double desiredPower);
+  Task SetLaserShutter(bool shutter);
+  Task<bool> GetLaserShutter();
+  Task<double> GetLaserPower();
+  double CurrentPower { get; }
+  double DesiredPower { get; }
+  bool Shutter { get; }
 }
