### 2026-03-12 bbc3b4c
Quantity schema support in script. Can now specify Quantity based constraints on function inputs and outputs.
