+++ AresScript/AresScriptTypeHints.cs
@@ -2,6 +2,7 @@ using Ares.Datamodel;
 using Ares.Datamodel.Extensions;
 using Ares.Datamodel.Factories;
 using AresScript.Generated;
+using System.Globalization;
 using UnitsNet;
 
 namespace AresScript;
@@ -14,21 +15,28 @@ internal static class AresScriptTypeHints
     .ToArray();
 
   public static bool TryParseTypeHint(string? typeHint, out AresDataType type)
+    => TryParseTypeHint(typeHint, out type, out _);
+
+  public static bool TryParseTypeHint(string? typeHint, out AresDataType type, out string? error)
   {
+    error = null;
     type = AresDataType.Any;
     if(string.IsNullOrWhiteSpace(typeHint))
     {
+      error = "Type hint is empty.";
       return false;
     }
 
     var leafName = typeHint.Trim().Split('.').Last();
     if(!Enum.TryParse(leafName, true, out AresDataType parsedType))
     {
+      error = $"'{typeHint}' is not a known ARES data type.";
       return false;
     }
 
     if(parsedType == AresDataType.UnspecifiedType)
     {
+      error = "UnspecifiedType is not a valid type hint.";
       return false;
     }
 
@@ -37,7 +45,11 @@ internal static class AresScriptTypeHints
   }
 
   public static bool TryParseTypeHint(AresLangParser.TypeHintContext? typeHint, out SchemaEntry schema)
+    => TryParseTypeHint(typeHint, out schema, out _);
+
+  public static bool TryParseTypeHint(AresLangParser.TypeHintContext? typeHint, out SchemaEntry schema, out string? error)
   {
+    error = null;
     schema = AresSchemaBuilder.Entry(AresDataType.Any).Build();
     if(typeHint is null)
     {
@@ -47,12 +59,10 @@ internal static class AresScriptTypeHints
     switch(typeHint)
     {
       case AresLangParser.NamedTypeRefContext namedTypeHint:
-        if(!TryParseTypeHint(namedTypeHint.namedTypeHint().GetText(), out var resolvedType))
+        if(!TryParseNamedTypeHint(namedTypeHint.namedTypeHint(), namedTypeHint.typeHintConstraints(), out schema, out error))
         {
           return false;
         }
-
-        schema = AresSchemaBuilder.Entry(resolvedType).Build();
         return true;
 
       case AresLangParser.StructTypeRefContext structTypeHint:
@@ -60,7 +70,7 @@ internal static class AresScriptTypeHints
           var structSchema = new AresDataSchema();
           foreach(var field in structTypeHint.structTypeHint().typeHintField())
           {
-            if(!TryParseTypeHint(field.typeHint(), out var fieldSchema))
+            if(!TryParseTypeHint(field.typeHint(), out var fieldSchema, out error))
             {
               return false;
             }
@@ -74,7 +84,7 @@ internal static class AresScriptTypeHints
         }
 
       case AresLangParser.ListTypeRefContext listTypeHint:
-        if(!TryParseTypeHint(listTypeHint.listTypeHint().typeHint(), out var elementSchema))
+        if(!TryParseTypeHint(listTypeHint.listTypeHint().typeHint(), out var elementSchema, out error))
         {
           return false;
         }
@@ -90,14 +100,14 @@ internal static class AresScriptTypeHints
 
   public static SchemaEntry SchemaFromTypeHint(AresLangParser.TypeHintContext? typeHint)
   {
-    return TryParseTypeHint(typeHint, out var schema)
+    return TryParseTypeHint(typeHint, out var schema, out _)
       ? schema
[…]
