### 2026-03-13 507e639
Fixed dummy value factory not making quantities
