+++ AresScript/DummyValueFactory.cs
@@ -60,6 +60,8 @@ public static class DummyValueFactory
         return AresValueHelper.CreateUnit();
       case AresDataType.Function:
         return AresValueHelper.CreateFunction(string.Empty);
+      case AresDataType.Quantity:
+        return AresValueHelper.CreateQuantity(0, schema.QuantitySchema.QuantityType, schema.QuantitySchema.BoundsUnit);
       case AresDataType.UnspecifiedType:
       case AresDataType.Any:
         return new AresValue();
