### 2026-03-13 3426e47
Updated the spool to be an ARES device package
