+++ ChemyxPumpPlugin/Responses/Parsers/ChemyxParsers.cs
@@ -1,0 +1,173 @@
+using Ares.Toolkit.Serial.Commands;
+using ChemyxPumpPlugin.Enums;
+using ChemyxPumpPlugin.Responses;
+using System.Globalization;
+using System.Text;
+using System.Text.RegularExpressions;
+
+namespace ChemyxPumpPlugin.Responses.Parsers;
+
+public abstract class ChemyxResponseParserBase<T>(string originalCommand) : SerialResponseParser<T> where T : SerialResponse
+{
+  protected readonly string OriginalCommand = originalCommand;
+
+  public override bool TryParseResponse(byte[] buffer, out T? response, out ArraySegment<byte>? dataToRemove)
+  {
+    response = null;
+    dataToRemove = null;
+
+    if(buffer == null || buffer.Length == 0) return false;
+
+    var terminatorIdx = Array.IndexOf(buffer, (byte)'>');
+    if(terminatorIdx < 0) return false;
+
+    var utfProxy = Encoding.UTF8.GetString(buffer);
+    var useLast = utfProxy.EndsWith('>');
+    var commandResponses = utfProxy.Split('>', StringSplitOptions.None);
+    if(!useLast) commandResponses = commandResponses.SkipLast(1).ToArray();
+
+    var startIdx = 0;
+    for(var i = 0; i < commandResponses.Length; i++)
+    {
+      var cmdResponse = commandResponses[i];
+      var lines = cmdResponse.Split(["\r\n", "\n", "\r"], StringSplitOptions.RemoveEmptyEntries);
+      var commandEcho = lines.FirstOrDefault() ?? string.Empty;
+      
+      if(commandEcho.Trim() != OriginalCommand.Trim())
+      {
+        startIdx += cmdResponse.Length + 1;
+        continue;
+      }
+
+      var payload = lines.Skip(1).ToArray();
+      response = CreateResponse(commandEcho, payload, cmdResponse);
+      dataToRemove = new ArraySegment<byte>(buffer, startIdx, cmdResponse.Length + 1);
+      return true;
+    }
+
+    return false;
+  }
+
+  protected abstract T CreateResponse(string echo, string[] lines, string raw);
+}
+
+public class ChemyxResponseParser(string originalCommand) : ChemyxResponseParserBase<ChemyxResponse>(originalCommand)
+{
+  protected override ChemyxResponse CreateResponse(string echo, string[] lines, string raw) => new(echo, lines, raw);
+}
+
+public class NumericResponseParser(string originalCommand) : ChemyxResponseParserBase<NumericResponse>(originalCommand)
+{
+  protected override NumericResponse CreateResponse(string echo, string[] lines, string raw)
+  {
+    double? parsed = null;
+    var line = lines.FirstOrDefault();
+    if(!string.IsNullOrWhiteSpace(line))
+    {
+      var lastPart = line.Split(' ', StringSplitOptions.RemoveEmptyEntries).LastOrDefault();
+      if(double.TryParse(lastPart, NumberStyles.Float, CultureInfo.InvariantCulture, out var value))
+        parsed = value;
+    }
+    return new NumericResponse(echo, lines, raw, parsed);
+  }
+}
+
+public class PumpStatusResponseParser(string originalCommand) : ChemyxResponseParserBase<PumpStatusResponse>(originalCommand)
+{
+  protected override PumpStatusResponse CreateResponse(string echo, string[] lines, string raw)
+  {
+    PumpStatus? status = null;
+    var line = lines.FirstOrDefault();
+    if(!string.IsNullOrWhiteSpace(line) && int.TryParse(line.Trim(), out var s))
+      status = (PumpStatus)s;
[…]
