+++ ChemyxPumpPlugin/PumpStatus.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace ChemyxPumpPlugin;
-
-public enum PumpStatus
-{
-  Stopped = 0,
-  Running = 1,
-  Paused = 2
-}
