+++ ChemyxPumpPlugin/Commands/ChemyxPumpCommand.cs
@@ -1,18 +1,0 @@
-﻿namespace ChemyxPumpPlugin;
-
-public enum ChemyxPumpCommand
-{
-  StartPump,
-  StopPump,
-  PausePump,
-  PumpStatus,
-  SetDiameter,
-  DispensedVolume,
-  ElapsedTime,
-  SetVolume,
-  ReadLimitParameter,
-  SetRate,
-  SetTime,
-  SetDelay,
-  SetUnits
-}
