+++ ChemyxPumpPlugin/Commands/Responses/SinglePumpParameters.cs
@@ -1,3 +1,0 @@
-﻿namespace ChemyxPumpPlugin.Commands.Responses;
-
-public record SinglePumpParameters(PumpUnits Units, double Diameter, double Rate, TimeSpan Time, double Volume, TimeSpan Delay);

[…]
