### 2026-03-14 e56af50
Validate the quantity creation to make sure quantity type matches unit.
