### 2026-03-14 c73dd49
Ensure each thrown interpreter exception has line and column numbers
