+++ AresScript/Interpreters/AresBaseInterpreter.cs
@@ -207,8 +207,8 @@ public class AresBaseInterpreter : AresLangBaseVisitor<Task<AresValue>>
         {
           throw new AresInterpreterException(
             "Provided index expression was not a string.",
-            context.Start.Line,
-            context.Start.Column
+            indexContext.expression().Start.Line,
+            indexContext.expression().Start.Column
           );
         }
 
@@ -221,8 +221,8 @@ public class AresBaseInterpreter : AresLangBaseVisitor<Task<AresValue>>
       {
         throw new AresInterpreterException(
           "Provided index expression was not a number.",
-          context.Start.Line,
-          context.Start.Column
+          indexContext.expression().Start.Line,
+          indexContext.expression().Start.Column
         );
       }
 
@@ -469,8 +469,8 @@ public class AresBaseInterpreter : AresLangBaseVisitor<Task<AresValue>>
     {
       throw new AresInterpreterException(
         "Provided index expression was not a number.",
-        context.Start.Line,
-        context.Start.Column
+        context.expression().Start.Line,
+        context.expression().Start.Column
       );
     }
 
@@ -661,7 +661,7 @@ public class AresBaseInterpreter : AresLangBaseVisitor<Task<AresValue>>
     {
       AresLangParser.LambdaSingleParamContext singleParam => singleParam.expression(),
       AresLangParser.LambdaParamListContext paramList => paramList.expression(),
-      _ => throw new AresInterpreterException("Invalid lambda expression.")
+      _ => throw new AresInterpreterException("Invalid lambda expression.", context.Start.Line, context.Start.Column)
     };
 
     var closure = Environment.GetAllUserVariableSymbols()
@@ -730,8 +730,8 @@ public class AresBaseInterpreter : AresLangBaseVisitor<Task<AresValue>>
       {
         throw new AresInterpreterException(
           "Provided index expression was not a string.",
-          context.Start.Line,
-          context.Start.Column
+          context.expression(1).Start.Line,
+          context.expression(1).Start.Column
         );
       }
 
@@ -744,8 +744,8 @@ public class AresBaseInterpreter : AresLangBaseVisitor<Task<AresValue>>
     {
       throw new AresInterpreterException(
         "Provided index expression was not a number.",
-        context.Start.Line,
-        context.Start.Column
+        context.expression(1).Start.Line,
+        context.expression(1).Start.Column
       );
     }
 
@@ -835,7 +835,7 @@ public class AresBaseInterpreter : AresLangBaseVisitor<Task<AresValue>>
     var callee = memberResolution.Callee ?? await Visit(ctx.expression());
 
     if(callee.FunctionValue is null)
-      throw new AresInterpreterException("Attempted to call a non-function");
+      throw new AresInterpreterException("Attempted to call a non-function", ctx.Start.Line, ctx.Start.Column);
 
     var id = callee.FunctionValue.FunctionId;
 
@@ -843,7 +843,7 @@ public class AresBaseInterpreter : AresLangBaseVisitor<Task<AresValue>>
     {
       if(keywordArgs.Count > 0)
       {
-        throw new AresInterpreterException($"Runtime function '{id}' does not support keyword arguments");
+        throw new AresInterpreterException($"Runtime function '{id}' does not support keyword arguments", ctx.Start.Line, ctx.Start.Column);
       }
 
       await CheckExecutionControlAsync();
@@ -857,7 +857,7 @@ public class AresBaseInterpreter : AresLangBaseVisitor<Task<AresValue>>
     {
       if(!Environment.TryGetUserLambda(id, out var userLambda))
       {
[…]
