### 2026-03-14 0b880bd
Fixed some more of the line/column numbers on interpreter exceptions
