+++ AresScript/Interpreters/AresValidationInterpreter.cs
@@ -662,8 +662,8 @@ public sealed class AresValidationInterpreter : AresLangBaseVisitor<Task>
       {
         throw new AresInterpreterException(
           $"Unknown function '{unresolvedMemberCtx.ID().GetText()}' on '{unresolvedMemberCtx.expression().GetText()}'.",
-          unresolvedMemberCtx.Stop.Line,
-          unresolvedMemberCtx.Stop.Column
+          unresolvedMemberCtx.ID().Symbol.Line,
+          unresolvedMemberCtx.ID().Symbol.Column
         );
       }
       throw new AresInterpreterException(
@@ -682,7 +682,12 @@ public sealed class AresValidationInterpreter : AresLangBaseVisitor<Task>
     {
       if(keywordArgs.Count > 0)
       {
-        throw new AresInterpreterException($"Runtime function '{functionId}' does not support keyword arguments", ctx.Start.Line, ctx.Start.Column);
+        var firstKeywordArg = keywordArgs.Values.First();
+        throw new AresInterpreterException(
+          $"Runtime function '{functionId}' does not support keyword arguments",
+          firstKeywordArg.Start.Line,
+          firstKeywordArg.Start.Column
+        );
       }
 
       ValidateSystemFunctionArgs(systemFn, positionalArgs, keywordArgs, ctx);
@@ -694,10 +699,11 @@ public sealed class AresValidationInterpreter : AresLangBaseVisitor<Task>
     {
       if(positionalArgs.Count > userFn.ParameterNames.Count)
       {
+        var extraArgument = positionalArgs[userFn.ParameterNames.Count];
         throw new AresInterpreterException(
           $"Function '{functionId}' expected {userFn.ParameterNames.Count} arguments but got {positionalArgs.Count}",
-          ctx.Start.Line,
-          ctx.Start.Column
+          extraArgument.Start.Line,
+          extraArgument.Start.Column
         );
       }
 
@@ -706,12 +712,22 @@ public sealed class AresValidationInterpreter : AresLangBaseVisitor<Task>
         var index = FindParameterIndex(userFn.ParameterNames, name);
         if(index < 0)
         {
-          throw new AresInterpreterException($"Function '{functionId}' got an unexpected keyword argument '{name}'", ctx.Start.Line, ctx.Start.Column);
+          var keywordArgument = keywordArgs[name];
+          throw new AresInterpreterException(
+            $"Function '{functionId}' got an unexpected keyword argument '{name}'",
+            keywordArgument.Start.Line,
+            keywordArgument.Start.Column
+          );
         }
 
         if(index < positionalArgs.Count)
         {
-          throw new AresInterpreterException($"Function '{functionId}' got multiple values for argument '{name}'", ctx.Start.Line, ctx.Start.Column);
+          var keywordArgument = keywordArgs[name];
+          throw new AresInterpreterException(
+            $"Function '{functionId}' got multiple values for argument '{name}'",
+            keywordArgument.Start.Line,
+            keywordArgument.Start.Column
+          );
         }
       }
 
@@ -737,10 +753,11 @@ public sealed class AresValidationInterpreter : AresLangBaseVisitor<Task>
     {
       if(positionalArgs.Count > lambda.Parameters.Count)
       {
+        var extraArgument = positionalArgs[lambda.Parameters.Count];
         throw new AresInterpreterException(
           $"Function '{functionId}' expected {lambda.Parameters.Count} arguments but got {positionalArgs.Count}",
-          ctx.argList().Start.Line,
-          ctx.argList().Start.Column
+          extraArgument.Start.Line,
+          extraArgument.Start.Column
         );
       }
 
[…]
