+++ AresScript/Interpreters/AresBaseInterpreter.cs
@@ -164,8 +164,8 @@ public class AresBaseInterpreter : AresLangBaseVisitor<Task<AresValue>>
     {
       throw new AresInterpreterException(
         "Assert condition must be boolean.",
-        context.Start.Line,
-        context.Start.Column
+        assertContext.expression(0).Start.Line,
+        assertContext.expression(0).Start.Column
       );
     }
 
@@ -778,7 +778,9 @@ public class AresBaseInterpreter : AresLangBaseVisitor<Task<AresValue>>
     await CheckExecutionControlAsync();
 
     var positionalArgs = new List<AresValue>();
+    var positionalArgExpressions = new List<AresLangParser.ExpressionContext>();
     var keywordArgs = new Dictionary<string, AresValue>(StringComparer.Ordinal);
+    var keywordArgContexts = new Dictionary<string, AresLangParser.KeywordArgContext>(StringComparer.Ordinal);
     var seenKeywordArg = false;
 
     var argContexts = ctx.argList()?.argument() ?? Enumerable.Empty<AresLangParser.ArgumentContext>();
@@ -798,6 +800,7 @@ public class AresBaseInterpreter : AresLangBaseVisitor<Task<AresValue>>
               );
             }
 
+            positionalArgExpressions.Add(positionalArg.expression());
             positionalArgs.Add(await Visit(positionalArg.expression()));
             break;
           }
@@ -815,6 +818,7 @@ public class AresBaseInterpreter : AresLangBaseVisitor<Task<AresValue>>
             }
 
             keywordArgs[name] = await Visit(keywordArg.expression());
+            keywordArgContexts[name] = keywordArg;
             break;
           }
         default:
@@ -826,7 +830,7 @@ public class AresBaseInterpreter : AresLangBaseVisitor<Task<AresValue>>
       }
     }
 
-    var memberResolution = await TryResolveMemberCallee(ctx, ctx.expression(), positionalArgs, keywordArgs);
+    var memberResolution = await TryResolveMemberCallee(ctx, ctx.expression(), positionalArgs, keywordArgs, keywordArgContexts);
     if(memberResolution.ExtensionHandled)
     {
       return memberResolution.ExtensionResult!;
@@ -843,7 +847,12 @@ public class AresBaseInterpreter : AresLangBaseVisitor<Task<AresValue>>
     {
       if(keywordArgs.Count > 0)
       {
-        throw new AresInterpreterException($"Runtime function '{id}' does not support keyword arguments", ctx.Start.Line, ctx.Start.Column);
+        var firstKeywordArg = keywordArgContexts.Values.First();
+        throw new AresInterpreterException(
+          $"Runtime function '{id}' does not support keyword arguments",
+          firstKeywordArg.Start.Line,
+          firstKeywordArg.Start.Column
+        );
       }
 
       await CheckExecutionControlAsync();
@@ -862,7 +871,7 @@ public class AresBaseInterpreter : AresLangBaseVisitor<Task<AresValue>>
 
       return await ExecuteFunctionInvocationAsync(ctx, id, id, AresFunctionInvocationKind.Lambda, async () =>
       {
-        return await ExecuteLambda(userLambda, id, positionalArgs, keywordArgs);
+        return await ExecuteLambda(ctx, userLambda, id, positionalArgs, keywordArgs, keywordArgContexts);
       });
     }
 
@@ -876,10 +885,11 @@ public class AresBaseInterpreter : AresLangBaseVisitor<Task<AresValue>>
     {
       if(positionalArgs.Count > userFn.ParameterNames.Count)
       {
+        var extraArgument = positionalArgExpressions[userFn.ParameterNames.Count];
         throw new AresInterpreterException(
           $"Function '{id}' expected {userFn.ParameterNames.Count} arguments but got {positionalArgs.Count}",
[…]
