### 2026-03-14 e936fe8
Fixed Assert not being validated for bools
