### 2026-03-15 5724e1b
Extra script validation
