+++ AresScript/Interpreters/AresValidationInterpreter.cs
@@ -387,6 +387,16 @@ public sealed class AresValidationInterpreter : AresLangBaseVisitor<Task>
     }
 
     await Visit(expression);
+    var iterableSchema = _typeInference.Visit(expression);
+    if(!IsIterableSchema(iterableSchema) && _mode == ValidationMode.Strict)
+    {
+      throw new AresInterpreterException(
+        $"Value is not iterable: {iterableSchema.Type}.",
+        expression.Start.Line,
+        expression.Start.Column
+      );
+    }
+
     _environment.EnterScope();
     try
     {
@@ -492,6 +502,18 @@ public sealed class AresValidationInterpreter : AresLangBaseVisitor<Task>
     try
     {
       await Visit(block);
+
+      var unitSchema = AresSchemaBuilder.Entry(AresDataType.Unit).Build();
+      if(_mode == ValidationMode.Strict
+        && !AresScriptTypeHints.IsCompatibleWithTypeHint(unitSchema, declaredReturnSchema)
+        && !AlwaysReturns(block))
+      {
+        throw new AresInterpreterException(
+          $"Function '{functionId}' may complete without returning a value of type {declaredReturnSchema.Stringify()}.",
+          functionNameToken.Symbol.Line,
+          functionNameToken.Symbol.Column
+        );
+      }
     }
     finally
     {
@@ -597,6 +619,50 @@ public sealed class AresValidationInterpreter : AresLangBaseVisitor<Task>
       context.Stop.Line,
       context.Stop.Column
     );
+  }
+
+  public override async Task VisitIndexAccess([NotNull] AresLangParser.IndexAccessContext context)
+  {
+    var receiver = context.expression(0);
+    var index = context.expression(1);
+    await Visit(receiver);
+    await Visit(index);
+
+    var receiverSchema = _typeInference.Visit(receiver);
+    if(receiverSchema.Type is not AresDataType.Any and not AresDataType.UnspecifiedType
+      && !IsIndexableSchema(receiverSchema))
+    {
+      throw new AresInterpreterException(
+        "Cannot access index of a value that is not of list or struct type.",
+        receiver.Start.Line,
+        receiver.Start.Column
+      );
+    }
+
+    var indexSchema = _typeInference.Visit(index);
+    if(receiverSchema.Type == AresDataType.Struct)
+    {
+      if(indexSchema.Type is not AresDataType.Any and not AresDataType.UnspecifiedType and not AresDataType.String)
+      {
+        throw new AresInterpreterException(
+          "Provided index expression was not a string.",
+          index.Start.Line,
+          index.Start.Column
+        );
+      }
+
+      return;
+    }
+
+    if(receiverSchema.Type is not AresDataType.Any and not AresDataType.UnspecifiedType
+      && indexSchema.Type is not AresDataType.Any and not AresDataType.UnspecifiedType and not AresDataType.Number)
+    {
+      throw new AresInterpreterException(
+        "Provided index expression was not a number.",
+        index.Start.Line,
+        index.Start.Column
+      );
+    }
   }
 
   public override async Task VisitFunctionCall(AresLangParser.FunctionCallContext ctx)
@@ -817,6 +883,58 @@ public sealed class AresValidationInterpreter : AresLangBaseVisitor<Task>
     {
       throw new AresInterpreterException($"Function '{functionId}' not found.", ctx.Start.Line, ctx.Start.Column);
     }
+  }
+
+  public override async Task VisitUnaryMinus([NotNull] AresLangParser.UnaryMinusContext context)
+  {
+    await Visit(context.expression());
[…]
