### 2026-03-15 b51279a
Moved the semantic token provider to the scripting project.
