### 2026-03-16 f27f263
Support of arithmetic operations between quantity values
