### 2026-03-16 ce5b40f
Separate string formatter for schema. Avoids using the Stringify method provided by the datamodel which is meant for user readability and not script creation. Also committed a file I forgetti.
