### 2026-03-16 27bb065
Enhanced the support of typed schemas in the script builder.
