+++ AresScript/DummyValueFactory.cs
@@ -61,7 +61,10 @@ public static class DummyValueFactory
       case AresDataType.Function:
         return AresValueHelper.CreateFunction(string.Empty);
       case AresDataType.Quantity:
-        return AresValueHelper.CreateQuantity(0, schema.QuantitySchema.QuantityType, schema.QuantitySchema.BoundsUnit);
+        return AresValueHelper.CreateQuantity(
+          0,
+          schema.QuantitySchema?.QuantityType ?? QuantityType.Unspecified,
+          schema.QuantitySchema?.BoundsUnit ?? string.Empty);
       case AresDataType.UnspecifiedType:
       case AresDataType.Any:
         return new AresValue();
