### 2026-03-16 17507e1
Fixed bad parsing of syringe pump states
