### 2026-03-16 3411c0f
Updated the Chemyx plugin to include all the commands again
