+++ ChemyxPumpPlugin/ChemyxPump.cs
@@ -1,5 +1,6 @@
 using Ares.Datamodel;
 using Ares.Datamodel.Device;
+using Ares.Datamodel.Extensions;
 using Ares.Datamodel.Factories;
 using Ares.Device;
 using Ares.Toolkit.Serial;
@@ -7,6 +8,8 @@ using ChemyxPumpPlugin.Commands;
 using ChemyxPumpPlugin.Enums;
 using ChemyxPumpPlugin.Responses;
 using ChemyxPumpPlugin.Simulation;
+using Google.Protobuf.WellKnownTypes;
+using ReactiveUI;
 using System.IO.Ports;
 using System.Reactive.Linq;
 using System.Reactive.Subjects;
@@ -127,10 +130,33 @@ public class ChemyxPump : AresDevice, IChemyxPump
   {
     return Task.FromResult(new List<DeviceCommandDescriptor>
     {
-      new() { Name = ChemyxPumpCommand.Start.ToString(), Description = "Starts the pump." },
-      new() { Name = ChemyxPumpCommand.Stop.ToString(), Description = "Stops the pump." },
-      new() { Name = ChemyxPumpCommand.Pause.ToString(), Description = "Pauses the pump." },
-      new() {
+      new()
+      {
+        Name = ChemyxPumpCommand.Start.ToString(),
+        Description = "Starts the pump.",
+        InputSchema = AresSchemaBuilder.Empty()
+          .AddEntry("PumpIndex", AresSchemaBuilder.NumberEntry().Build())
+          .AddEntry("Mode", AresSchemaBuilder.NumberEntry().AsOptional().Build())
+          .Build()
+      },
+      new()
+      {
+        Name = ChemyxPumpCommand.Stop.ToString(),
+        Description = "Stops the pump.",
+        InputSchema = AresSchemaBuilder.Empty()
+          .AddEntry("PumpIndex", AresSchemaBuilder.NumberEntry().Build())
+          .Build()
+      },
+      new()
+      {
+        Name = ChemyxPumpCommand.Pause.ToString(),
+        Description = "Pauses the pump.",
+        InputSchema = AresSchemaBuilder.Empty()
+          .AddEntry("PumpIndex", AresSchemaBuilder.NumberEntry().Build())
+          .Build()
+      },
+      new()
+      {
         Name = ChemyxPumpCommand.SetRate.ToString(),
         Description = "Sets the flow rate.",
         InputSchema = AresSchemaBuilder.Empty()
@@ -138,38 +164,133 @@ public class ChemyxPump : AresDevice, IChemyxPump
           .AddEntry("Rate", AresSchemaBuilder.NumberEntry().Build())
           .Build()
       },
-      new() {
+      new()
+      {
         Name = ChemyxPumpCommand.SetVolume.ToString(),
         Description = "Sets the target volume.",
         InputSchema = AresSchemaBuilder.Empty()
           .AddEntry("PumpIndex", AresSchemaBuilder.NumberEntry().Build())
           .AddEntry("Volume", AresSchemaBuilder.NumberEntry().Build())
           .Build()
+      },
+      new()
+      {
+        Name = ChemyxPumpCommand.SetDiameter.ToString(),
+        Description = "Sets the inner diameter of the syringe.",
+        InputSchema = AresSchemaBuilder.Empty()
+          .AddEntry("PumpIndex", AresSchemaBuilder.NumberEntry().Build())
+          .AddEntry("Diameter", AresSchemaBuilder.NumberEntry().Build())
+          .Build(),
+        OutputSchema = AresSchemaBuilder.NumberEntry().Build()
+      },
+      new()
+      {
+        Name = ChemyxPumpCommand.SetUnits.ToString(),
[…]
