### 2026-03-16 ad1e520
Fixed up some of the UI components to be a bit prettier.
