+++ HerkulexDRS/Enums/ServoCommand.cs
@@ -1,8 +1,0 @@
-namespace HerkulexDRS.Enums;
-public enum ServoCommand
-{
-  GoUp,
-  GoDown,
-  Reset,
-  GetPosition
-}
