+++ HerkulexDRS/Responses/CommandResponse.cs
@@ -1,13 +1,0 @@
-using Ares.Toolkit.Serial.Commands;
-
-namespace HerkulexDRS.Responses;
-
-public abstract class CommandResponse : SerialResponse
-{
-  public CommandResponse(int servoId)
-  {
-    ServoId = servoId;
-  }
-
-  public int ServoId { get; set; }
-}
