+++ TicStepperController/ITicStepperController.cs
@@ -1,5 +1,0 @@
-using Ares.Device;
-
-namespace TicStepperController;
-
-public interface ITicStepperController : IAresDevice, IAsyncDisposable { }
