+++ ValveController/ValveControllerDevice.cs
@@ -1,253 +1,0 @@
-using Ares.Datamodel;
-using Ares.Datamodel.Device;
-using Ares.Datamodel.Extensions;
-using Ares.Datamodel.Factories;
-using Ares.Device;
-using Ares.Toolkit.Serial;
-using ValveController.Commands;
-using ValveController.Commands.RelayOne;
-using ValveController.Commands.RelayTwo;
-using ValveController.Commands.Responses;
-using ValveController.Connection;
-using ValveController.Enums;
-using ValveController.Interfaces;
-using ValveController.Simulation;
-using System.Reactive.Linq;
-using System.Reactive.Subjects;
-
-namespace ValveController;
-
-public class ValveControllerDevice : AresDevice, IValveController
-{
-  private readonly BehaviorSubject<AresStruct> _stateSubject = new(new AresStruct());
-  private readonly IValveControllerConnection _connection;
-
-  public ValveControllerDevice(DeviceConnectionInfo connectionInfo) : base(connectionInfo)
-  {
-    var serialInfo = connectionInfo.SerialConnectionInfo;
-    
-    if (connectionInfo.Simulated)
-    {
-      _connection = new SimValveControllerConnection(serialInfo.PortName);
-    }
-    else
-    {
-      _connection = new ValveControllerConnection(serialInfo.PortName);
-    }
-  }
-
-  public async Task<RelayStatusResponse> GetRelayStatus()
-  {
-    await _connection.Send(new EnterCommandModeCommand());
-    var response = await _connection.Send(new GetRelayStatusCommand());
-    
-    if (response != null)
-    {
-      RelayOneEngaged = response.RelayOneOn;
-      RelayTwoEngaged = response.RelayTwoOn;
-      UpdateState();
-    }
-    else
-    {
-      throw new Exception("No Status Response from Relay Board!");
-    }
-
-    return response;
-  }
-
-  public override async Task<bool> Activate(CancellationToken ct)
-  {
-    try
-    {
-      await EnableRelays();
-      await GetRelayStatus();
-      UpdateState();
-      Status = new DeviceOperationalStatus { OperationalState = OperationalState.Active, Message = "Valve Controller is active" };
-      return true;
-    }
-    catch (Exception e)
-    {
-      Status = new DeviceOperationalStatus { OperationalState = OperationalState.Error, Message = $"Failed to activate: {e.Message}" };
-      return false;
-    }
-  }
-
-  public override async Task EnterSafeMode(CancellationToken ct)
-  {
-    await DisengageRelayOne();
-    await DisengageRelayTwo();
-  }
-
-  public async Task EngageRelayOne()
-  {
-    await _connection.Send(new EnterCommandModeCommand());
-    await _connection.Send(new EngageRelayOneCommand());
-    RelayOneEngaged = true;
-    UpdateState();
-  }
-
-  public async Task EngageRelayTwo()
-  {
-    await _connection.Send(new EnterCommandModeCommand());
-    await _connection.Send(new EngageRelayTwoCommand());
-    RelayTwoEngaged = true;
-    UpdateState();
-  }
-
-  public async Task DisengageRelayOne()
-  {
-    await _connection.Send(new EnterCommandModeCommand());
-    await _connection.Send(new DisengageRelayOneCommand());
-    RelayOneEngaged = false;
-    UpdateState();
-  }
-
-  public async Task DisengageRelayTwo()
-  {
-    await _connection.Send(new EnterCommandModeCommand());
-    await _connection.Send(new DisengageRelayTwoCommand());
-    RelayTwoEngaged = false;
-    UpdateState();
-  }
-
-  public async Task EnableRelays()
-  {
-    await _connection.Send(new EnterCommandModeCommand());
-    await _connection.Send(new EnableAllDevicesCommand());
-  }
-
[…]
