+++ ValveController/Commands/RelayOne/DisengageRelayOneCommand.cs
@@ -1,10 +1,0 @@
-using Ares.Toolkit.Serial.Commands;
-
-namespace ValveController.Commands.RelayOne;
-public class DisengageRelayOneCommand : SerialCommand
-{
-  protected override byte[] Serialize()
-  {
-    return new byte[] { 0 };
-  }
-}
