### 2026-03-17 0b72b90
Removed the old Tube Furnace implementation
