+++ LindbergFurnace/TubeFurnace.cs
@@ -1,145 +1,0 @@
-﻿using Ares.Datamodel;
-using Ares.Device;
-using Ares.Device.Serial;
-using LindbergFurnace.Commands;
-using LindbergFurnace.Commands.Requests;
-using System.Globalization;
-using System.Reactive.Linq;
-using System.Reactive.Subjects;
-using System.Reactive.Threading.Tasks;
-using TubeFurnace.Messaging;
-using UnitsNet;
-
-namespace LindbergFurnace;
-
-public class TubeFurnace : SerialDevice<ITubeFurnaceConnection>, ITubeFurnace
-{
-  private readonly ISubject<TubeFurnaceState> _statePublisher = new BehaviorSubject<TubeFurnaceState>(new TubeFurnaceState());
-  private readonly BehaviorSubject<AresStruct> _stateSubject = new(new AresStruct());
-
-
-  public TubeFurnace(string name, int address, ITubeFurnaceConnection connection) : base(name, connection)
-  {
-    InternalStateStream = _statePublisher.AsObservable();
-    var initialState =
-      new TubeFurnaceState
-      {
-        Id = UniqueId,
-        AssumedAddress = address,
-      };
-    _statePublisher.OnNext(initialState);
-  }
-
-  public async Task<int> GetSetpoint()
-  {
-    var currentState = await InternalStateStream.Take(1);
-
-    var request = new ReadMultipleRegistersRequest(currentState.AssumedAddress, Register.SP1, 1);
-    var response = await Connection.Send(request);
-    var setpointData = response.RegisterContents.First();
-    var setpointAsciiHEx = setpointData.Select(b => (char)b).ToArray();
-    var setpointInt = int.Parse(setpointAsciiHEx, NumberStyles.HexNumber);
-    var setpoint = Temperature.FromDegreesCelsius(setpointInt);
-
-    currentState.Setpoint = setpoint.DegreesCelsius;
-    _statePublisher.OnNext(currentState);
-
-    return setpointInt;
-  }
-
-  public async Task<int> GetCurrentTemperature()
-  {
-    var request = new ReadMultipleRegistersRequest(1, Register.PV, 1);
-    var response = await Connection.Send(request);
-    var temperatureData = response.RegisterContents.First();
-    var temperatureAsciiHex = temperatureData.Select(b => (char)b).ToArray();
-    var temperatureInt = int.Parse(temperatureAsciiHex, NumberStyles.HexNumber);
-    var temperature = Temperature.FromDegreesCelsius(temperatureInt);
-
-    var currentState = await InternalStateStream.Take(1);
-    currentState.CurrentTemperature = temperature.DegreesCelsius;
-    _statePublisher.OnNext(currentState);
-
-    return temperatureInt;
-  }
-
-  public async Task<int> GetCurrentAddress()
-  {
-    var currentState = await InternalStateStream.Take(1);
-    if(currentState is not null)
-      return currentState.AssumedAddress;
-
-    return -1;
-  }
-
-  public async Task SetAndWaitForSetpoint(Temperature targetTemperature, double delta, double timeout, CancellationToken ct = default)
-  {
-    await SetSetpoint(targetTemperature);
-
-    var task = InternalStateStream
-      .Where(state => Math.Abs(targetTemperature.DegreesCelsius - state.CurrentTemperature) <= delta)
-      .FirstAsync()
-      .ToTask(ct);
-
-    if(timeout == -1)
-    {
-      await task;
-    }
-
-    else
-    {
-      try
-      {
-        var timespan = TimeSpan.FromSeconds(timeout);
-        await task.WaitAsync(timespan, ct);
-      }
-      catch (TimeoutException)
-      {
[…]
