+++ LindbergFurnace/Commands/TubeFurnaceCommand.cs
@@ -1,9 +1,0 @@
-﻿namespace LindbergFurnace.Commands;
-
-public enum TubeFurnaceCommand
-{
-  GetSetpoint,
-  SetSetpoint,
-  GetCurrentTemperature,
-  SetAndWaitForSetpoint
-}

[…]
