### 2026-03-17 afc6b7e
Removed linberg funrance from solution file. Fixed the bad looking transition for devices loading in.
