### 2026-03-17 edeee77
Updated the service starter to run things in parallel when possible, speeding up startup times a bit
