### 2026-03-17 a0ff971
Updated service starter to be a background process. ARES now has a splash screen that appears during the loading process. Devices seem to be having a difficult time displaying on the dashboard properly though, not sure what's going on there yet.
