### 2026-03-17 0669952
Cleaned up the index.razor and it's companion JS logic
