### 2026-03-17 2d697af
Fixed the gridstack resize issue
