### 2026-03-17 5e1eefd
Squashed the migrations into one large migration file
