+++ AresService.Migrations.SqlServer/Migrations/20260225203920_ArchitectureV2_ModularDeviceRework_AresIdentityContext.cs
@@ -1,22 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore.Migrations;
-
-#nullable disable
-
-namespace AresService.Migrations.SqlServer.Migrations.AresIdentity
-{
-    /// <inheritdoc />
-    public partial class ArchitectureV2_ModularDeviceRework_AresIdentityContext : Migration
-    {
-        /// <inheritdoc />
-        protected override void Up(MigrationBuilder migrationBuilder)
-        {
-
-        }
-
-        /// <inheritdoc />
-        protected override void Down(MigrationBuilder migrationBuilder)
-        {
-
-        }
-    }
-}
