+++ AresService.Migrations.Postgres/Migrations/AresDbContextModelSnapshot.cs
@@ -17,7 +17,7 @@ namespace AresService.Migrations.Postgres.Migrations
         {
 #pragma warning disable 612, 618
             modelBuilder
-                .HasAnnotation("ProductVersion", "10.0.1")
+                .HasAnnotation("ProductVersion", "10.0.4")
                 .HasAnnotation("Relational:MaxIdentifierLength", 63);
 
             NpgsqlModelBuilderExtensions.UseIdentityByDefaultColumns(modelBuilder);
@@ -341,6 +341,9 @@ namespace AresService.Migrations.Postgres.Migrations
 
                     b.Property<Guid?>("StepExecutionSummaryUniqueId")
                         .HasColumnType("uuid");
+
+                    b.Property<string>("TemplateId")
+                        .HasColumnType("text");
 
                     b.HasKey("UniqueId");
 
