+++ AresService.Migrations.SqlServer/Migrations/20260310145639_AddedConnectionInfo_AresDbContext.cs
@@ -1,39 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore.Migrations;
-
-#nullable disable
-
-namespace AresService.Migrations.SqlServer.Migrations.AresDb
-{
-    /// <inheritdoc />
-    public partial class AddedConnectionInfo_AresDbContext : Migration
-    {
-        /// <inheritdoc />
-        protected override void Up(MigrationBuilder migrationBuilder)
-        {
-            migrationBuilder.AddColumn<int>(
-                name: "BaudRate",
-                table: "SerialConnection",
-                type: "int",
-                nullable: false,
-                defaultValue: 0);
-
-            migrationBuilder.AddColumn<string>(
-                name: "SerialId",
-                table: "SerialConnection",
-                type: "nvarchar(max)",
-                nullable: true);
-        }
-
-        /// <inheritdoc />
-        protected override void Down(MigrationBuilder migrationBuilder)
-        {
-            migrationBuilder.DropColumn(
-                name: "BaudRate",
-                table: "SerialConnection");
-
-            migrationBuilder.DropColumn(
-                name: "SerialId",
-                table: "SerialConnection");
-        }
-    }
-}
