+++ Ares.Core/Device/Plugins/Drivers/Loading/AresDriverLoadContext.cs
@@ -14,18 +14,13 @@ public class AresDriverLoadContext : AssemblyLoadContext
 
   protected override Assembly? Load(AssemblyName assemblyName)
   {
-    // 1. "PROPER" WAY: Ask the Host first.
-    // If the Host has it (System.IO.Ports, Protobuf, etc.), use the Host's version.
     try
     {
-      // This prevents the "Illegal State" by ensuring we never load 
-      // a second copy of a shared system assembly.
       var hostAssembly = Default.LoadFromAssemblyName(assemblyName);
       if(hostAssembly != null) return hostAssembly;
     }
     catch { /* Host doesn't have it, continue to local resolver */ }
 
-    // 2. Fallback to the plugin's 'bin' folder for unique hardware DLLs.
     string? assemblyPath = _resolver.ResolveAssemblyToPath(assemblyName);
     if(assemblyPath != null)
     {
