### 2026-03-19 1045ea8
Some device state logging updates
