+++ Ares.Core/Device/State/Logging/AresDeviceStateLogger.cs
@@ -1,0 +1,162 @@
+﻿using Ares.Core.EntityConfigurations.Extensions;
+using Ares.Datamodel;
+using Ares.Datamodel.Device;
+using Ares.Device;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
+using System.Reactive;
+using System.Reactive.Disposables;
+using System.Reactive.Linq;
+
+namespace Ares.Core.Device.State.Logging;
+
+public class AresDeviceStateLogger : IDeviceStateLogger
+{
+  private readonly IDbContextFactory<CoreDatabaseContext> _dbContextFactory;
+  private readonly ILogger<AresDeviceStateLogger> _logger;
+  private readonly IAresDevice _device;
+  private IDisposable _stateWatcher = Disposable.Empty;
+  private readonly Dictionary<string, double> _lastDeltaValues = new(StringComparer.OrdinalIgnoreCase);
+  private Dictionary<string, double> _eligibleDeltas = [];
+  
+  public AresDeviceStateLogger(IDbContextFactory<CoreDatabaseContext> dbContextFactory, IAresDevice device, ILogger<AresDeviceStateLogger> logger)
+  {
+    _dbContextFactory = dbContextFactory;
+    _logger = logger;
+    _device = device;
+  }
+
+  public string DeviceId => _device.UniqueId;
+
+  public DeviceLoggingSettings Settings { get; private set; } = new DeviceLoggingSettings { LoggingType = DeviceLoggingSettings.Types.LoggingType.None };
+
+  public Task Start(DeviceLoggingSettings? settings = null)
+  {
+    Settings = settings ?? Settings;
+
+    _eligibleDeltas = Settings.Deltas
+      .Where(d => d.Value > 0)
+      .ToDictionary();
+
+    var stream = _device.StateStream;
+    if(Settings.LoggingType == DeviceLoggingSettings.Types.LoggingType.Interval)
+    {
+      var timer = Observable.Interval(
+        Settings.IntervalMs > 0 ? TimeSpan.FromMilliseconds(Settings.IntervalMs) : TimeSpan.FromMilliseconds(1));
+      _stateWatcher = timer
+        .WithLatestFrom(stream, (_, state) => state)
+        .SelectMany(meme => Observable.FromAsync(() => UpdateState(meme)))
+        .OnErrorResumeNext(Observable.Empty<Unit>())
+        .Subscribe();
+    }
+    else if(Settings.LoggingType == DeviceLoggingSettings.Types.LoggingType.OnChange)
+    {
+      if(Settings.IntervalMs > 0)
+      {
+        stream = stream.Sample(TimeSpan.FromMilliseconds(Settings.IntervalMs));
+      }
+
+      _stateWatcher = stream
+        .Where(state => ShouldEmitByDeltas(state))
+        .SelectMany(meme => Observable.FromAsync(() => UpdateState(meme)))
+        .OnErrorResumeNext(Observable.Empty<Unit>())
+        .Subscribe();
+    }
+
+    return Task.CompletedTask;
+  }
+
+  public Task Stop()
+  {
+    _stateWatcher.Dispose();
+
+    return Task.CompletedTask;
+  }
+
+  public async Task UpdateSettings(DeviceLoggingSettings? settings)
+  {
+    await Stop();
+    await Start(settings);
+  }
+
+  private bool ShouldEmitByDeltas(AresStruct? state)
+  {
+    // If no state or no configured/eligible deltas, allow emission (fallback to original behavior)
+    if(state is null || _eligibleDeltas.Count == 0)
+    {
+      return true;
+    }
+
+    bool anyExceeded = false;
+
+    foreach(var d in _eligibleDeltas)
+    {
+      if(!TryGetDouble(state, d.Key, out var current))
+      {
[…]
