### 2026-03-19 cc6c153
Fixed some problems with the remote devices
