+++ UI/Features/Devices/IAresDeviceViewModelFactory.cs
@@ -1,9 +1,10 @@
 ﻿using Ares.Device;
 using Ares.Toolkit.Device.UI;
+using UI.Features.Devices.Plugin;
 
-namespace UI.Features.Devices.Plugin.Factories;
+namespace UI.Features.Devices;
 
-public interface IPluginViewModelFactory
+public interface IAresDeviceViewModelFactory
 {
   Task<PluginDeviceSettingsListViewModel?> CreateListViewModelAsync(string driverName);
   IDeviceUnitControlViewModel CreateUnitControlViewModel(IAresDevice device);

[…]
