### 2026-03-19 506456a
Flattened the plugin .ares folders to have no bin. Fixed incorrect view model factory assignment
