+++ AresScript/AresScriptTypeHints.cs
@@ -14,10 +14,10 @@ internal static class AresScriptTypeHints
     .Select(type => type.ToString())
     .ToArray();
 
-  public static bool TryParseTypeHint(AresLangParser.TypeHintContext? typeHint, out SchemaEntry schema)
+  public static bool TryParseTypeHint(AresLangParser.TypeHintContext? typeHint, out AresValueSchema schema)
     => TryParseTypeHint(typeHint, out schema, out _);
 
-  public static bool TryParseTypeHint(AresLangParser.TypeHintContext? typeHint, out SchemaEntry schema, out string? error)
+  public static bool TryParseTypeHint(AresLangParser.TypeHintContext? typeHint, out AresValueSchema schema, out string? error)
   {
     error = null;
     schema = AresSchemaBuilder.Entry(AresDataType.Any).Build();
@@ -37,7 +37,7 @@ internal static class AresScriptTypeHints
 
       case AresLangParser.StructTypeRefContext structTypeHint:
         {
-          var structSchema = new AresDataSchema();
+          var structSchema = new AresStructSchema();
           foreach(var field in structTypeHint.structTypeHint().typeHintField())
           {
             if(!TryParseTypeHint(field.typeHint(), out var fieldSchema, out error))
@@ -68,14 +68,14 @@ internal static class AresScriptTypeHints
     }
   }
 
-  public static SchemaEntry SchemaFromTypeHint(AresLangParser.TypeHintContext? typeHint)
+  public static AresValueSchema SchemaFromTypeHint(AresLangParser.TypeHintContext? typeHint)
   {
     return TryParseTypeHint(typeHint, out var schema, out _)
       ? schema
       : AresSchemaBuilder.Entry(AresDataType.Any).Build();
   }
 
-  public static SchemaEntry SchemaFromTypeHint(string? typeHint)
+  public static AresValueSchema SchemaFromTypeHint(string? typeHint)
   {
     if(string.IsNullOrWhiteSpace(typeHint))
     {
@@ -97,14 +97,14 @@ internal static class AresScriptTypeHints
     return SchemaFromTypeHint(functionDecl.functionDeclaration().parameterList()?.parameter().FirstOrDefault()?.typeHint());
   }
 
-  public static bool IsCompatibleWithTypeHint(SchemaEntry actual, SchemaEntry expected)
+  public static bool IsCompatibleWithTypeHint(AresValueSchema actual, AresValueSchema expected)
   {
     return IsCompatible(expected, actual);
   }
 
-  public static bool IsCompatibleWithTypeHint(AresValue actual, SchemaEntry expected)
+  public static bool IsCompatibleWithTypeHint(AresValue actual, AresValueSchema expected)
   {
-    if(!IsCompatibleWithTypeHint(actual.ToSchemaEntry(), expected))
+    if(!IsCompatibleWithTypeHint(actual.ToAresValueSchema(), expected))
     {
       return false;
     }
@@ -122,17 +122,17 @@ internal static class AresScriptTypeHints
     return true;
   }
 
-  public static bool IsCompatibleWithTypeHint(SchemaEntry actual, AresDataType expectedType)
+  public static bool IsCompatibleWithTypeHint(AresValueSchema actual, AresDataType expectedType)
   {
     return IsCompatibleWithTypeHint(actual, AresSchemaBuilder.Entry(expectedType).Build());
   }
 
   public static bool IsCompatibleWithTypeHint(AresValue actual, AresDataType expectedType)
   {
-    return IsCompatibleWithTypeHint(actual.ToSchemaEntry(), expectedType);
+    return IsCompatibleWithTypeHint(actual.ToAresValueSchema(), expectedType);
   }
 
-  private static bool IsCompatible(SchemaEntry expected, SchemaEntry actual)
[…]
