### 2026-03-20 217b3f0
Fixed quantity values not getting properly validated due to missing unit.
