### 2026-03-20 6aa9ed2
Potentially fixed the issue with calling device commands with args in script.
