+++ Ares.Core/Scripting/DeviceSymbolProvider.cs
@@ -44,7 +44,7 @@ public class DeviceSymbolProvider(IAresDeviceProvider deviceProvider) : ISymbolP
                 $"Function '{functionId}' expected at most {descriptor.InputSchema?.Fields.Count ?? 0} arguments but got {args.Count}.");
             }
 
-            var result = await device.ExecuteCommand(commandName, args.Select(arg => new Datamodel.Device.DeviceCommandArgument { ArgName = Guid.NewGuid().ToString(), ArgValue = arg }).ToList(), token.CancellationToken);
+            var result = await device.ExecuteCommand(commandName, args.Select((arg, idx) => new Datamodel.Device.DeviceCommandArgument { ArgName = (inputSchema.Fields.ElementAtOrDefault(idx).Key ?? ""), ArgValue = arg }).ToList(), token.CancellationToken);
             if(!result.Success)
             {
               throw new InvalidOperationException(
