### 2026-03-20 c9f5778
Fixed script not calling device commands with spaces.
