+++ .github/workflows/dotnet-release-preview.yaml
@@ -41,7 +41,7 @@ jobs:
         dotnet-version: '10.x'
         
     - name: Setup Node.js Environment
-      uses: actions/setup-node@v6.2.0
+      uses: actions/setup-node@v6.3.0
       with:
         node-version: 24
 
