### 2026-03-23 23e9ebf
Modified release pipeline to account for service project deletion
