### 2026-03-23 a234896
Trying to include device plugins in build process
