### 2026-03-23 dfc861b
Adding some debugging help to the release pipe
