### 2026-03-23 634d74f
Set working directory before building drivers
