### 2026-03-23 cc64b23
Fix unit test failing due to wrong context being created.
