### 2026-03-23 a9de80b
Updated the release and pre-release pipelines to clone the latest device drivers alongside the ARES build.
