### 2026-03-24 b081fe3
Couple of fixes to the migration files to ensure clean database interactions. Updated the missing driver system to remove and log the removal of null driver id's that can be caused by migrations. Added script for creating a local release package for testing.
