### 2026-03-26 3ec7a36
This is actually a visualization branch now. Working on the menu for adding charts. Device section looks decent, but probably going to work on adding sections for planners, analyzers and eventually dedicated visualization services.
