+++ UI/ServiceCollectionExtensions.cs
@@ -22,6 +22,7 @@ using UI.Features.Auth;
 using UI.Features.CampaignEdit;
 using UI.Features.CampaignEdit.Factories;
 using UI.Features.CampaignEdit.ViewModels;
+using UI.Features.Components;
 using UI.Features.Devices;
 using UI.Features.Devices.Plugin;
 using UI.Features.Devices.Remote.Factory;
@@ -31,6 +32,7 @@ using UI.Features.DeviceStateLogging.Settings;
 using UI.Features.Notifications;
 using UI.Features.Planning.Settings;
 using UI.Features.ServerHealth;
+using UI.Features.Visualization.ViewModels;
 using UI.Infrastructure.Auth;
 using UI.Infrastructure.Devices;
 using UI.Infrastructure.Dialog;
@@ -111,6 +113,8 @@ internal static class ServiceCollectionExtensions
     services.AddScoped<ExecutionHistoryViewModel>();
     services.AddScoped<ExecutionViewModel>();
     services.AddScoped<ScriptPlaygroundViewModel>();
+    services.AddScoped<VisualizationViewModel>();
+    services.AddScoped<VisualizationSidebarViewModel>();
 
     //Device Settings List View Models
     services.AddTransient<DeviceStatesViewModel>();
