### 2026-03-27 9dd7ffb
Fixed the analyzer and planner pages loading poorly, often getting stuck at loading spinner
