### 2026-03-27 76f7935
Working towards visualization functionality
