+++ UI/ServiceCollectionExtensions.cs
@@ -22,7 +22,6 @@ using UI.Features.Auth;
 using UI.Features.CampaignEdit;
 using UI.Features.CampaignEdit.Factories;
 using UI.Features.CampaignEdit.ViewModels;
-using UI.Features.Components;
 using UI.Features.Devices;
 using UI.Features.Devices.Plugin;
 using UI.Features.Devices.Remote.Factory;
