+++ UI/Features/Visualization/ChartStyleEnum.cs
@@ -1,0 +1,11 @@
+﻿namespace UI.Features.Visualization;
+
+public enum ChartStyle
+{
+  Line,
+  Spline,
+  Area,
+  Column,
+  Gauge,
+  TextIndicator
+}
