+++ Ares.Core/Execution/Executors/CampaignExecutor.cs
@@ -206,7 +206,7 @@ public class CampaignExecutor : ICampaignExecutor
 
       if(experimentExecutorResult.ErrorString is not null)
       {
-        _logger.LogError("Experiment Executor Generation Failed! Reason: {error-message}", experimentExecutorResult.ErrorString);
+        _logger.LogError($"Experiment Executor Generation Failed! Reason: {experimentExecutorResult.ErrorString}");
         await _notifier.Notify("Experiment Executor Generation Failure", experimentExecutorResult.ErrorString, NotificationSeverityEnum.Error);
         executionSuccess = false;
         break;
@@ -404,7 +404,7 @@ public class CampaignExecutor : ICampaignExecutor
     var result = new ExperimentExecutorResult();
     var experimentTemplate = template.CloneWithNewIds();
 
-    _logger.LogDebug("Going to try and generate an experiment executor for {TemplateName}.", template.Name);
+    _logger.LogDebug($"Going to try and generate an experiment executor for {template.Name}.");
     if(!experimentTemplate.IsResolved())
     {
       _logger.LogTrace("Experiment was not resolved");
@@ -413,7 +413,7 @@ public class CampaignExecutor : ICampaignExecutor
         Status.PlannerState = PlannerState.PlanningInProgress;
         _executionStatusSubject.OnNext(Status);
         _executionReporter.Report(Status);
-        _logger.LogTrace("Analyses count is {count} and replan rate {rate}", analyses.Count(), ReplanRate);
+        _logger.LogTrace($"Analyses count is {analyses.Count()} and replan rate {ReplanRate}");
         var metadata = new RequestMetadata 
         { 
           CampaignId = Template.UniqueId, 
