### 2026-03-27 460838d
Fixed the ARES core device being missing, fixed a bug that caused the await user confirmation to break some things.
