### 2026-03-27 606b178
Excluded the core device from UI generation
