+++ .github/workflows/dotnet-ci.yaml
@@ -27,7 +27,7 @@ jobs:
         dotnet-version: '10.x'
 
     - name: Setup Node.js Environment
-      uses: actions/setup-node@v6.2.0
+      uses: actions/setup-node@v6
       with:
         node-version: 24
 
