+++ .github/workflows/dotnet-release-preview.yaml
@@ -30,24 +30,24 @@ jobs:
 
     steps:
     - name: Checkout Code
-      uses: actions/checkout@v4
+      uses: actions/checkout@v6
       with:
         ref: ${{ github.event.release.tag_name }}
 
     # Clone the drivers repo into 'DriversRepo'
     - name: Checkout Device Drivers
-      uses: actions/checkout@v4
+      uses: actions/checkout@v6
       with:
         repository: 'AFRL-ARES/Ares.Device.Drivers'
         path: 'DriversRepo'
 
     - name: Set up .NET 10.x
-      uses: actions/setup-dotnet@v4
+      uses: actions/setup-dotnet@v5
       with:
         dotnet-version: '10.x'
         
     - name: Setup Node.js Environment
-      uses: actions/setup-node@v4
+      uses: actions/setup-node@v6
       with:
         node-version: 24
 
