### 2026-03-31 730a2fd
Added some additional plumbing for creating visualization pieces
