+++ Ares.Core/CoreDatabaseContext.cs
@@ -5,6 +5,7 @@ using Ares.Datamodel.Analyzing;
 using Ares.Datamodel.Device;
 using Ares.Datamodel.Planning;
 using Ares.Datamodel.Templates;
+using Ares.Datamodel.Visualizing.Local;
 using Ares.Services;
 using Google.Protobuf.WellKnownTypes;
 using Microsoft.EntityFrameworkCore;
@@ -38,6 +39,7 @@ public class CoreDatabaseContext : DbContext
   public DbSet<Parameter> Parameters => Set<Parameter>();
   public DbSet<DeviceLoggingSettings> DeviceLoggingSettings => Set<DeviceLoggingSettings>();
   public DbSet<DeviceState> DeviceStates => Set<DeviceState>();
+  public DbSet<DeviceVisualizationConfig> DeviceVisualizationConfigs => Set<DeviceVisualizationConfig>();
 
   protected override void OnModelCreating(ModelBuilder modelBuilder)
   {
