### 2026-03-31 8ae62bd
Working towards working visualization of device state data
