### 2026-03-31 c8ef0ea
Fixed the title of the graphs to be more helpful
