### 2026-03-31 94081b3
Added support for extracting the scalar out of a quantity value.
