### 2026-03-31 b0e5e5c
Added an "as(...)" extension to quantity values. That should allow for seamless conversion between compatible units if you needed to get the scalar values in a different unit for some reason.
