### 2026-03-31 bd74f41
Charts are working for the Alicat. Though the longer it runs the slower the charts seem to update, so there's some efficiency improvements to be made
