+++ UI/Features/Visualization/ViewModels/VisualizationViewModel.cs
@@ -1,5 +1,7 @@
 ﻿using Ares.Core.Device.Providers;
+using Ares.Core.Visualization.Managers;
 using Ares.Core.Visualization.Providers;
+using Ares.Core.Visualization.Repos;
 using Ares.Datamodel.Visualizing.Local;
 using DynamicData;
 using ReactiveUI;
@@ -13,21 +15,23 @@ namespace Ares.Core.Visualization.ViewModels;
 public partial class VisualizationViewModel : ReactiveObject, IDisposable
 {
   private readonly IDeviceVisualizationConfigProvider _configProvider;
+  private readonly IVisualizationConfigManager _visualizationConfigManager;
   private readonly IAresDeviceProvider _deviceProvider;
   private readonly IDisposable _subscription;
   private readonly ReadOnlyObservableCollection<VisualizationItemViewModel> _visualizationItems;
 
-  public VisualizationViewModel(IDeviceVisualizationConfigProvider configProvider, IAresDeviceProvider deviceProvider)
+  public VisualizationViewModel(IDeviceVisualizationConfigProvider configProvider, IAresDeviceProvider deviceProvider, IVisualizationConfigManager visualizationConfigManager)
   {
     _configProvider = configProvider;
     _deviceProvider = deviceProvider;
+    _visualizationConfigManager = visualizationConfigManager;
 
     _subscription = _configProvider.Connect()
         // Transform the raw config into our rich, reactive Item ViewModel
         .Transform(config =>
         {
           var device = _deviceProvider.GetDevice(config.DeviceId);
-          return new VisualizationItemViewModel(config, device);
+          return new VisualizationItemViewModel(config, device, OnChartDeleteRequested);
         })
         .DisposeMany()
         .Bind(out _visualizationItems)
@@ -39,6 +43,9 @@ public partial class VisualizationViewModel : ReactiveObject, IDisposable
     _subscription?.Dispose();
     GC.SuppressFinalize(this);
   }
+
+  private void OnChartDeleteRequested(string uniqueId)
+    => _visualizationConfigManager.Remove(uniqueId);
 
   public ReadOnlyObservableCollection<VisualizationItemViewModel> VisualizationItems => _visualizationItems;
 }
