### 2026-03-31 19527a8
Properly resolve the function call values during validation. Specifically when the function is a member access.
