### 2026-04-01 155d3fb
Updates to fix the poor performance of the charts. They no longer freeze after running for awhile, updates seem stable
