### 2026-04-01 ee10b5b
Moved settings to a dialog. Added some additional settings options
