### 2026-04-01 f36d5a2
Additional updates. Adding more user customization to chart display
