### 2026-04-01 31b4ccf
Added the ability to select marker visibility
