### 2026-04-01 74b68d3
Fixed a bug that prevented changes from being saved properly to the database
