### 2026-04-01 7e8cda7
Charts are now keeping their position listed in their configs in order to support loading up in the correct place and in the correct size the user set them too, making this page much more user friendly
