### 2026-04-06 590a55f
Removed extra stuff from the base environment builder into its own class.
