+++ Ares.Core.Tests/EnvironmentSymbolWriterTests.cs
@@ -1,0 +1,278 @@
+using Ares.Core.Scripting;
+using Ares.Datamodel;
+using Ares.Datamodel.Extensions;
+using Ares.Datamodel.Scripting;
+using AresScript;
+using AresScript.Environment;
+using AresScript.Symbols;
+
+namespace Ares.Core.Tests;
+
+[TestFixture]
+public class EnvironmentSymbolWriterTests
+{
+  [Test]
+  public void AddSystemFunction_Adds_Root_Function()
+  {
+    var env = new AresScriptEnvironment();
+    var function = CreateFunction("math::sum", "sum");
+
+    env.AddSystemFunction(function);
+
+    using(Assert.EnterMultipleScope())
+    {
+      Assert.That(env.TryGetSystemFunction(function.Id, out var storedFunction), Is.True);
+      Assert.That(storedFunction, Is.EqualTo(function));
+      Assert.That(env.TryGetSystemValueSymbol("sum", out var value), Is.True);
+      Assert.That(value?.SymbolKind, Is.EqualTo(SymbolKind.Function));
+      Assert.That(value?.Value.FunctionValue.FunctionId, Is.EqualTo(function.Id));
+    }
+  }
+
+  [Test]
+  public void AddSystemFunction_Adds_Nested_Function_From_ParentName()
+  {
+    var env = new AresScriptEnvironment();
+    var function = CreateFunction("devices::scope::run", "run", parentName: "devices.scope");
+
+    env.AddSystemFunction(function);
+
+    using(Assert.EnterMultipleScope())
+    {
+      Assert.That(env.TryGetSystemValueSymbol("devices", out var devices), Is.True);
+      Assert.That(devices?.StructFields?["scope"].StructFields?["run"].Value.FunctionValue.FunctionId, Is.EqualTo(function.Id));
+    }
+  }
+
+  [Test]
+  public void AddSystemFunction_Adds_Nested_Function_From_Namespace()
+  {
+    var env = new AresScriptEnvironment();
+    var function = CreateFunction("math::angle::from", "from", @namespace: "math.angle");
+
+    env.AddSystemFunction(function);
+
+    using(Assert.EnterMultipleScope())
+    {
+      Assert.That(env.TryGetSystemValueSymbol("math", out var math), Is.True);
+      Assert.That(math?.StructFields?["angle"].StructFields?["from"].Value.FunctionValue.FunctionId, Is.EqualTo(function.Id));
+    }
+  }
+
+  [Test]
+  public void AddSystemValue_Adds_Root_Value()
+  {
+    var env = new AresScriptEnvironment();
+    var value = new AresScriptValueSymbol(
+      "answer",
+      AresValueHelper.CreateNumber(42),
+      IsReadOnly: true,
+      SymbolKind: SymbolKind.Variable,
+      Detail: "the answer",
+      Documentation: "doc",
+      IsUserDefined: false);
+
+    env.AddSystemValue(value);
+
+    Assert.That(env.TryGetSystemValueSymbol("answer", out var stored), Is.True);
+    using(Assert.EnterMultipleScope())
+    {
+      Assert.That(stored?.Value.NumberValue, Is.EqualTo(42));
+      Assert.That(stored?.Detail, Is.EqualTo("the answer"));
+      Assert.That(stored?.Documentation, Is.EqualTo("doc"));
+      Assert.That(stored?.IsReadOnly, Is.True);
+      Assert.That(stored?.IsUserDefined, Is.False);
+    }
+  }
+
+  [Test]
+  public void AddSystemValue_Adds_Nested_Value()
+  {
+    var env = new AresScriptEnvironment();
+    var value = new AresScriptValueSymbol(
+      "mode",
+      AresValueHelper.CreateString("auto"),
[…]
