+++ AresScript/Environment/EnvironmentSymbolWriter.cs
@@ -1,0 +1,293 @@
+using Ares.Datamodel;
+using Ares.Datamodel.Scripting;
+using AresScript.Symbols;
+
+namespace AresScript.Environment;
+
+internal static class EnvironmentSymbolWriter
+{
+  public static void AddSystemSymbol(AresScriptEnvironment environment, IScriptSymbol symbol)
+  {
+    switch(symbol)
+    {
+      case AresSystemFunctionSymbol function:
+        AddSystemFunction(environment, function);
+        return;
+      case IValueSymbol value:
+        AddSystemValue(environment, value);
+        return;
+      default:
+        throw new InvalidOperationException($"Unsupported system symbol type: {symbol.GetType().Name}");
+    }
+  }
+
+  public static void AddSystemFunction(AresScriptEnvironment environment, AresSystemFunctionSymbol function)
+  {
+    if(environment.TryGetSystemFunction(function.Id, out _))
+    {
+      throw new InvalidOperationException($"System function '{function.Id}' already exists.");
+    }
+
+    var fieldName = string.IsNullOrWhiteSpace(function.Name) ? function.Id : function.Name;
+    if(!string.IsNullOrWhiteSpace(fieldName))
+    {
+      var parentPath = !string.IsNullOrWhiteSpace(function.ParentName)
+        ? function.ParentName
+        : function.Namespace;
+
+      var projectedFunction = AresSystemValue.Function(function) with
+      {
+        Name = fieldName,
+        ParentName = NormalizePath(parentPath),
+        IsReadOnly = true,
+        SymbolKind = SymbolKind.Function,
+        Detail = function.Detail,
+        Documentation = function.Documentation
+      };
+
+      AddProjectedValue(environment, parentPath, fieldName, projectedFunction);
+    }
+
+    environment.AssignSystemFunctions([function]);
+  }
+
+  public static void AddSystemValue(AresScriptEnvironment environment, IValueSymbol valueSymbol)
+  {
+    if(string.IsNullOrWhiteSpace(valueSymbol.Name))
+    {
+      return;
+    }
+
+    AddProjectedValue(environment, valueSymbol.ParentName, valueSymbol.Name, NormalizeSystemValue(valueSymbol));
+  }
+
+  // Map a flat symbol path like devices.pump.run into the nested system-value tree.
+  private static void AddProjectedValue(
+    AresScriptEnvironment environment,
+    string? parentPath,
+    string fieldName,
+    AresSystemValue value)
+  {
+    var pathSegments = string.IsNullOrWhiteSpace(parentPath) ? [] : SplitPath(parentPath);
+    if(pathSegments.Length == 0)
+    {
+      AssignOrMergeRootValue(environment, fieldName, value);
+      return;
+    }
+
+    var target = GetOrCreateStructAtPath(environment, pathSegments);
+    AssignOrMergeValue(target, fieldName, value, BuildPath(BuildPath(pathSegments), fieldName));
+  }
+
+  private static IDictionary<string, AresSystemValue> GetOrCreateStructAtPath(
+    AresScriptEnvironment environment,
+    IReadOnlyList<string> pathSegments)
+  {
+    IDictionary<string, AresSystemValue>? current = null;
+    string? iterParentPath = null;
+
+    for(var index = 0; index < pathSegments.Count; index++)
+    {
+      var segment = pathSegments[index];
+      var currentPath = iterParentPath is null ? segment : $"{iterParentPath}.{segment}";
+
+      if(current is null)
+      {
+        if(!environment.TryGetSystemValueSymbol(segment, out var existingRoot))
+        {
+          var newRoot = CreateStructPlaceholder(segment, iterParentPath);
[…]
