### 2026-04-07 3a9c67d
Small cleanup work on the visualization UI files. Starting to integrate necessary backend support for visualization of planner and analyzer features
