### 2026-04-07 cbeedbc
Cleaned up analyzers logic a bit and added the creation and logging of analyzer transactions. Starting to play with getting those updates on the UI now
