+++ Ares.Core/Analyzing/AnalysisHelper.cs
@@ -1,10 +1,12 @@
-﻿using Ares.Core.Notifications;
-using Ares.Datamodel;
+﻿using Ares.Datamodel;
 using Ares.Datamodel.Analyzing;
+using Ares.Datamodel.Analyzing.Remote;
 using Ares.Datamodel.Extensions;
 using Ares.Datamodel.Templates;
 using Google.Protobuf.Collections;
+using Google.Protobuf.WellKnownTypes;
 using Grpc.Core;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 
 namespace Ares.Core.Analyzing;
@@ -12,26 +14,45 @@ namespace Ares.Core.Analyzing;
 public class AnalysisHelper
 {
   readonly IAnalyzerRepo _analyzerRepo;
-    private readonly ILogger<AnalysisHelper> _logger;
+  private readonly ILogger<AnalysisHelper> _logger;
+  private readonly IDbContextFactory<CoreDatabaseContext> _dbContextFactory;
 
-    public AnalysisHelper(IAnalyzerRepo analyzerRepo, ILogger<AnalysisHelper> logger)
-    {
-      _analyzerRepo = analyzerRepo;
-      _logger = logger;
-    }
+  public AnalysisHelper(IAnalyzerRepo analyzerRepo, ILogger<AnalysisHelper> logger, IDbContextFactory<CoreDatabaseContext> dbContextFactory)
+  {
+    _analyzerRepo = analyzerRepo;
+    _logger = logger;
+    _dbContextFactory = dbContextFactory;
+  }
 
   public async Task<Analysis> Analyze(ExperimentTemplate template, ExperimentExecutionSummary experimentSummary, ExperimentExecutionSummary startupSummary, RequestMetadata metadata, CancellationToken cancellationToken)
   {
     try
     {
+      var transaction = new AnalyzerTransaction();
       var analyzer = GetAnalyzer(template.AnalyzerId);
       _logger.LogInformation("Analysis requested using analyzer {AnalyzerName}", analyzer.Name);
 
       var combinedResult = AresStructHelper.AppendStruct(experimentSummary.ExperimentOverview.Result, startupSummary.ExperimentOverview.Result);
       var analyzerInputs = ExperimentOutputToAnalyzerInputs(combinedResult, template.AnalyzerMaps);
 
-      // TODO: Maybe add support for settings overrides if needed
-      var analysis = await analyzer.Analyze(analyzerInputs, metadata, cancellationToken);
+      var analysisRequest = new AnalysisRequest 
+      { 
+        Inputs = analyzerInputs, 
+        Metadata = metadata, 
+        Settings = analyzer.Settings 
+      };
+
+      //Populate Transaction Info
+      transaction.UniqueId = Guid.NewGuid().ToString();
+      transaction.AnalyzerName = analyzer.Name;
+      transaction.AnalyzerType = analyzer.Type;
+      transaction.AnalysisRequest = analysisRequest;
+      transaction.TimeRequestSent = DateTime.UtcNow.ToTimestamp();
+
+      var analysis = await analyzer.Analyze(analysisRequest, cancellationToken);
+      transaction.TimeResponseReceived = DateTime.UtcNow.ToTimestamp();
+      transaction.AnalysisResponse = analysis;
+      
       experimentSummary.ExperimentOverview.AnalysisOverview = new AnalysisOverview
       {
         UniqueId = Guid.NewGuid().ToString(),
@@ -41,6 +62,7 @@ public class AnalysisHelper
       };
 
       _logger.LogInformation("Analysis completed {Result}", analysis.Result);
+      await LogAnalyzerTransaction(transaction);
[…]
