### 2026-04-07 e19f0bf
Actually have a nice little chart on the execution page now! It looks quite nice. Handles multiple planning values well, and theoretically should be an easy update when we upgrade analyzers to return more than a simple score.
