+++ UI/Features/Execution/ExecutionViewModel.cs
@@ -14,6 +14,7 @@ using UI.Application.Notifications;
 using UI.Domain.Experiments;
 using Ares.Core.Analyzing;
 using System.Reactive.Linq;
+using Ares.Core.Planning;
 
 namespace UI.Features.Execution;
 
@@ -24,23 +25,31 @@ public partial class ExecutionViewModel : ReactiveObject, INotifyPropertyChanged
   private readonly IAnalyzerTransactionProvider _analyzerTransactionProvider;
   public readonly ObservableCollection<CampaignTemplateSummary> CampaignTemplateSummaries = [];
   private readonly INotificationReceivingService _notificationService;
+  private readonly IPlannerServiceRepo _plannerServiceRepo;
+  private readonly IPlannerTransactionProvider _plannerTransactionProvider;
 
   public ExecutionViewModel(AutomationService automationClient,
     IConfiguration configuration,
     INotificationReceivingService notificationService,
     AnalyzerService analyzerService,
-    IAnalyzerTransactionProvider analysisTransactionProvider)
+    IAnalyzerTransactionProvider analysisTransactionProvider,
+    IPlannerServiceRepo plannerServiceRepo,
+    IPlannerTransactionProvider plannerTransactionProvider)
   {
     _automationClient = automationClient;
     _notificationService = notificationService;
     _analyzerService = analyzerService;
     _analyzerTransactionProvider = analysisTransactionProvider;
+    _plannerServiceRepo = plannerServiceRepo;
+    _plannerTransactionProvider = plannerTransactionProvider;
     PlannerAdapterInfos = [];
+    AnalyzerMetrics = [];
+    PlannerMetricsMap = [];
 
     this.WhenAnyValue(x => x.CurrentPlannerState)
       .Subscribe(newState =>
       {
-        Console.WriteLine($"Planning state has changed to: {newState}");
+        _ = UpdatePlannerTransactions();
       });
 
     this.WhenAnyValue(x => x.CurrentAnalysisState)
@@ -245,29 +254,113 @@ public partial class ExecutionViewModel : ReactiveObject, INotifyPropertyChanged
 
   private async Task UpdateAnalysisTransactions()
   {
-    if(CampaignTemplate is not null)
-    {
-      try
-      {
-        var bleh = new AnalyzerTransactionRequestFilter { AnalyzerId = CampaignTemplate.ExperimentTemplate.AnalyzerId, Start = CurrentCampaignStartTime?.ToTimestamp(), End = DateTime.UtcNow.ToTimestamp() };
-        var blah = await _analyzerTransactionProvider.GetAnalyzerTransactionsAsync(bleh);
-      }
+    if(CampaignTemplate is null || CurrentAnalysisState != AnalysisState.AnalysisComplete)
+      return;
 
-      catch(Exception e)
-      {
-        Console.WriteLine("lol oops");
-      }
-    }
+    var filter = new AnalyzerTransactionRequestFilter 
+    { 
+      AnalyzerId = CampaignTemplate.ExperimentTemplate.AnalyzerId, 
+      Start = CurrentCampaignStartTime.ToTimestamp(), 
+      End = DateTime.UtcNow.ToTimestamp() 
+    };
+
+    var analyzerTransactions = await _analyzerTransactionProvider.GetAnalyzerTransactionsAsync(filter);
+    var newestTransaction = analyzerTransactions.LastOrDefault();
+
+    if(newestTransaction is null)
+      return;
+
+    OnAnalyzerTransactionReceived(newestTransaction, analyzerTransactions.Count());
   }
 
   private async Task UpdatePlannerTransactions()
   {
-    if(CampaignTemplate is not null)
[…]
