### 2026-04-10 e15ed70
Removed step templates from the experiment template
