+++ Ares.Core/EntityConfigurations/Execution/Experiments/ExperimentTemplateEntityConfiguration.cs
@@ -10,12 +10,5 @@ internal class ExperimentTemplateEntityConfiguration : AresEntityTypeBaseConfigu
   {
     base.Configure(builder);
     builder.ToTable("ExperimentTemplates");
-
-    builder.HasMany(experimentTemplate => experimentTemplate.StepTemplates)
-      .WithOne()
-      .OnDelete(DeleteBehavior.Cascade);
-
-    builder.Navigation(experimentTemplate => experimentTemplate.StepTemplates)
-      .AutoInclude();
   }
 }
