+++ Ares.Core/EntityConfigurations/Execution/Steps/StepTemplateEntityConfiguration.cs
@@ -1,20 +1,0 @@
-﻿using Ares.Datamodel.Templates;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-
-namespace Ares.Core.EntityConfigurations.Execution.Steps;
-
-internal class StepTemplateEntityConfiguration : AresEntityTypeBaseConfiguration<StepTemplate>
-{
-  public override void Configure(EntityTypeBuilder<StepTemplate> builder)
-  {
-    base.Configure(builder);
-    builder.ToTable("StepTemplates");
-    builder.HasMany(stepTemplate => stepTemplate.CommandTemplates)
-      .WithOne()
-      .IsRequired();
-
-    builder.Navigation(stepTemplate => stepTemplate.CommandTemplates)
-      .AutoInclude();
-  }
-}
