+++ Ares.Core/Execution/Executors/IExecutor.cs
@@ -1,13 +1,0 @@
-﻿using Ares.Core.Execution.ControlTokens;
-using Google.Protobuf;
-
-namespace Ares.Core.Execution.Executors;
-
-public interface IExecutor<TResult, out TStatus>
-  where TResult : IMessage
-  where TStatus : IMessage
-{
-  IObservable<TStatus> ExperimentStatusObservable { get; }
-  TStatus Status { get; }
-  Task<TResult> Execute(ExecutionControlToken executionToken);
-}
