### 2026-04-13 1a8daa1
Replaced the execution token with the script one. Started getting rid of unnecessary executors.
