+++ Ares.Core/Execution/Executors/CampaignExecutor.cs
@@ -4,8 +4,6 @@ using Ares.Core.Analyzing;
 using Ares.Core.AresEnvironment;
 using Ares.Core.Device.State.Logging;
 using Ares.Core.Exceptions;
-using Ares.Core.Execution.ControlTokens;
-using Ares.Core.Execution.Executors.Composers;
 using Ares.Core.Execution.Extensions;
 using Ares.Core.Execution.StopConditions;
 using Ares.Core.Notifications;
@@ -14,6 +12,7 @@ using Ares.Core.Planning;
 using Ares.Datamodel;
 using Ares.Datamodel.Analyzing;
 using Ares.Datamodel.Templates;
+using AresScript;
 using Google.Protobuf.WellKnownTypes;
 using Microsoft.Extensions.Logging;
 
@@ -23,7 +22,6 @@ public class CampaignExecutor : ICampaignExecutor
 {
   private readonly IExecutionReporter _executionReporter;
   private readonly ISubject<CampaignExecutionStatus> _executionStatusSubject;
-  private readonly ICommandComposer<ExperimentTemplate, ExperimentExecutor> _experimentComposer;
   private readonly IPlanningHelper _planningHelper;
   private readonly IEnumerable<IExecutionSummaryHandler> _summaryHandlers;
   private readonly INotifier _notifier;
@@ -34,7 +32,7 @@ public class CampaignExecutor : ICampaignExecutor
   readonly AnalysisRepo _analysisRepo;
   readonly IAnalyzerRepo _analyzerRepo;
 
-  internal CampaignExecutor(ICommandComposer<ExperimentTemplate, ExperimentExecutor> experimentComposer,
+  internal CampaignExecutor(
     IPlanningHelper planningHelper,
     IExecutionReporter executionReporter,
     AnalysisHelper analysisHelper,
@@ -52,7 +50,6 @@ public class CampaignExecutor : ICampaignExecutor
     _analysisHelper = analysisHelper;
     _variableManager = variableManager;
     _stateLoggerManager = stateLoggerManager;
-    _experimentComposer = experimentComposer;
     _planningHelper = planningHelper;
     _executionReporter = executionReporter;
     _summaryHandlers = resultHandlers;
@@ -70,7 +67,7 @@ public class CampaignExecutor : ICampaignExecutor
     ExperimentStatusObservable = _executionStatusSubject.AsObservable();
   }
 
-  public async Task<CampaignExecutionSummary> Execute(ExecutionControlToken token)
+  public async Task<CampaignExecutionSummary> Execute(ScriptExecutionControlToken token)
   {
     // Make sure we log all the device state changes during execution.
     await _stateLoggerManager.EnableOverrideAsync(new Datamodel.Device.DeviceLoggingSettings
@@ -151,7 +148,7 @@ public class CampaignExecutor : ICampaignExecutor
     return campaignPath;
   }
 
-  private void ResetStatus(ExecutionControlToken token)
+  private void ResetStatus(ScriptExecutionControlToken token)
   {
     Status = new CampaignExecutionStatus
     {
@@ -170,7 +167,7 @@ public class CampaignExecutor : ICampaignExecutor
     _logger.LogInformation("Campaign started named {CampaignName}", Template.Name);
   }
 
-  private async Task<(bool Success, ExperimentExecutionSummary Summary)> ExecuteStartup(List<Analysis> analyses, List<ExperimentExecutionSummary> experimentSummaries, ExecutionControlToken token)
+  private async Task<(bool Success, ExperimentExecutionSummary Summary)> ExecuteStartup(List<Analysis> analyses, List<ExperimentExecutionSummary> experimentSummaries, ScriptExecutionControlToken token)
   {
     _logger.LogDebug("Startup template not null, will run startup script");
[…]
