### 2026-04-13 69c12ca
Gutted and simplified the execution view by breaking it up into multiple smaller components and moved a lot of the old logic into the view model. Added some live charts for device data, currently not being saved in the database.
