+++ UI/Features/Execution/ExecutionViewModel.cs
@@ -1,20 +1,25 @@
+using Ares.Core.Analyzing;
+using Ares.Core.Device.Providers;
+using Ares.Core.Execution;
+using Ares.Core.Grpc.Services;
+using Ares.Core.Planning;
 using Ares.Datamodel;
 using Ares.Datamodel.Analyzing;
 using Ares.Datamodel.Planning;
 using Ares.Datamodel.Templates;
+using Ares.Datamodel.Visualizing.Local;
 using Ares.Services;
-using Ares.Core.Grpc.Services;
 using DynamicData;
 using Google.Protobuf.WellKnownTypes;
 using ReactiveUI;
 using ReactiveUI.SourceGenerators;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
-using UI.Application.Notifications;
-using UI.Domain.Experiments;
-using Ares.Core.Analyzing;
 using System.Reactive.Linq;
-using Ares.Core.Planning;
+using UI.Application.Notifications;
+using UI.Domain.Execution;
+using UI.Domain.Experiments;
+using UI.Features.Visualization.ViewModels;
 
 namespace UI.Features.Execution;
 
@@ -27,6 +32,12 @@ public partial class ExecutionViewModel : ReactiveObject, INotifyPropertyChanged
   private readonly INotificationReceivingService _notificationService;
   private readonly IPlannerServiceRepo _plannerServiceRepo;
   private readonly IPlannerTransactionProvider _plannerTransactionProvider;
+  private readonly IExecutionReportStore _executionReportStore;
+  private readonly IAresDeviceProvider _deviceProvider;
+  public event Action? StateChanged;
+
+  private IDisposable? _experimentSubscription;
+  private IDisposable? _campaignStateSubscription;
 
   public ExecutionViewModel(AutomationService automationClient,
     IConfiguration configuration,
@@ -34,7 +45,9 @@ public partial class ExecutionViewModel : ReactiveObject, INotifyPropertyChanged
     AnalyzerService analyzerService,
     IAnalyzerTransactionProvider analysisTransactionProvider,
     IPlannerServiceRepo plannerServiceRepo,
-    IPlannerTransactionProvider plannerTransactionProvider)
+    IPlannerTransactionProvider plannerTransactionProvider,
+    IExecutionReportStore executionReportStore,
+    IAresDeviceProvider deviceProvider)
   {
     _automationClient = automationClient;
     _notificationService = notificationService;
@@ -42,9 +55,13 @@ public partial class ExecutionViewModel : ReactiveObject, INotifyPropertyChanged
     _analyzerTransactionProvider = analysisTransactionProvider;
     _plannerServiceRepo = plannerServiceRepo;
     _plannerTransactionProvider = plannerTransactionProvider;
+    _executionReportStore = executionReportStore;
+    _deviceProvider = deviceProvider;
+
     PlannerAdapterInfos = [];
     AnalyzerMetrics = [];
     PlannerMetricsMap = [];
+    ExperimentExecutionStatuses = [];
 
     this.WhenAnyValue(x => x.CurrentPlannerState)
       .Subscribe(newState =>
@@ -140,6 +157,37 @@ public partial class ExecutionViewModel : ReactiveObject, INotifyPropertyChanged
     await _automationClient.SetReplanRate(new ReplanRate { ReplanRate_ = DesiredReplanRate }, null);
     var replanRateResponse = await GetCurrentReplanRate();
     DesiredReplanRate = replanRateResponse.ReplanRate;
+  }
+
+  public async Task StartCampaign()
+  {
+    var executionEligibility = await _automationClient.CheckExecutionEligibility(new Empty(), null);
+
+    if(!executionEligibility.IsEligible)
+    {
[…]
