### 2026-04-13 6512ca3
Fixed some visual bugs on the execution page and moved the tracking of execution start time to the execution manager on the backend as that feels more correct architecture wise
