### 2026-04-13 d734c2d
Seperating out the analysis and planning charts. Got the custom tooltips working to see non-normalized data.
