### 2026-04-13 61749bc
Removed plain invocation emitters for functions. Gutted result generator. Added experiment specific function symbol creator. Reworked experiment and campaign executors slightly to use the script. More removals of step and command concepts
