### 2026-04-14 #10054 @espottesmith
Hey @arthursloan, my name is Evan, and I'll be the managing Topic Editor for your JOSS submission.

To start us off, could you suggest some potential reviewers?
