### 2026-04-14 #10054 @espottesmith
Also, heads up that the heading of your "Software Design" section in your manuscript is currently misspelled, which is why our bot didn't catch its existence.
