### 2026-04-15 3b14324
Device plugins now take in a logger from ARES in their constructor
