### 2026-04-15 70ce7e6
Fixed an issue where we weren't properly displaying ID assignment for serial devices
