### 2026-04-15 abcd11e
Don't allow users to add devices of different types onto the same serial port (i.e. if Alicat MFC's are already registered on COM1, don't allow a user to then register a tube furnace on the same port)
