### 2026-04-15 1f7b617
Some result generator cleanup.
