### 2026-04-15 #10054 @arthursloan
Hi @espottesmith, we recommend Prof. Raymundo Arróyave [rarroyave@tamu.edu](mailto:rarroyave@tamu.edu) and Willi Gottstein [willi.gottstein@utoronto.ca](mailto:willi.gottstein@utoronto.ca) as potential reviewers.

We're aware of that typo (and a few others) but didn't want to push a fix as we weren't sure if would interfere with the review pipeline. If it won't cause issues we can get those fixed with a push ASAP.
