+++ Ares.Core/EntityConfigurations/Execution/Steps/StepExecutionStatusEntityConfiguration.cs
@@ -1,18 +1,0 @@
-﻿using Ares.Datamodel;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-
-namespace Ares.Core.EntityConfigurations.Execution.Steps;
-
-internal class StepExecutionStatusEntityConfiguration : AresEntityTypeBaseConfiguration<StepExecutionStatus>
-{
-  public override void Configure(EntityTypeBuilder<StepExecutionStatus> builder)
-  {
-    base.Configure(builder);
-    builder.ToTable("StepExecutionStatuses");
-
-    builder.HasMany(status => status.CommandExecutionStatuses)
-      .WithOne()
-      .OnDelete(DeleteBehavior.Cascade);
-  }
-}
