+++ AresScript/Environment/PredefinedScope.cs
@@ -1,5 +1,8 @@
 ﻿namespace AresScript.Environment;
 
+/// <summary>
+/// A few enums to reduce the reliance on string matching
+/// </summary>
 public enum PredefinedScope
 {
   Global,
