### 2026-04-16 3a86ecb
Sila servers are now detectable by ARES
