+++ AresService.Migrations.Postgres/Migrations/AresDbContextModelSnapshot.cs
@@ -656,6 +656,42 @@ namespace AresService.Migrations.Postgres.Migrations
                     b.ToTable("SerialConnection");
                 });
 
+            modelBuilder.Entity("Ares.Datamodel.Device.SilaDeviceConfig", b =>
+                {
+                    b.Property<Guid>("UniqueId")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("uuid");
+
+                    b.Property<DateTime>("CreationTime")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("timestamp with time zone")
+                        .HasDefaultValueSql("NOW()");
+
+                    b.Property<string>("Description")
+                        .HasColumnType("text");
+
+                    b.Property<DateTime>("LastModified")
+                        .ValueGeneratedOnAddOrUpdate()
+                        .HasColumnType("timestamp with time zone")
+                        .HasDefaultValueSql("NOW()");
+
+                    b.Property<string>("ServerName")
+                        .HasColumnType("text");
+
+                    b.Property<string>("Type")
+                        .HasColumnType("text");
+
+                    b.Property<string>("VendorUri")
+                        .HasColumnType("text");
+
+                    b.Property<string>("Version")
+                        .HasColumnType("text");
+
+                    b.HasKey("UniqueId");
+
+                    b.ToTable("SilaConfigs");
+                });
+
             modelBuilder.Entity("Ares.Datamodel.ExecutionInfo", b =>
                 {
                     b.Property<Guid>("UniqueId")
