### 2026-04-16 473d6d0
Sila GUI improvements, updates to the manager to reflect said GUI improvements
