+++ Ares.Core.Tests/SilaDataConverterTests.cs
@@ -1,0 +1,357 @@
+using Ares.Core.Device.Sila;
+using Ares.Datamodel;
+using Ares.Datamodel.Extensions;
+using Ares.Datamodel.Factories;
+using Google.Protobuf;
+using Tecan.Sila2;
+using Tecan.Sila2.DynamicClient;
+
+namespace Ares.Core.Tests;
+
+[TestFixture]
+public class SilaDataConverterTests
+{
+  [Test]
+  public void ToSilaDataType_NumberSchemaWithRange_UsesConstrainedReal()
+  {
+    var schema = AresSchemaBuilder.Entry(AresDataType.Number)
+      .WithNumberRange(0, 5.5)
+      .Build();
+
+    var silaType = SilaDataConverter.ToSilaDataType(schema);
+
+    Assert.That(silaType.Item, Is.TypeOf<ConstrainedType>());
+    var constrained = (ConstrainedType)silaType.Item;
+    using(Assert.EnterMultipleScope())
+    {
+      Assert.That(constrained.DataType.Item, Is.EqualTo(BasicType.Real));
+      Assert.That(constrained.Constraints.MinimalInclusive, Is.EqualTo("0"));
+      Assert.That(constrained.Constraints.MaximalInclusive, Is.EqualTo("5.5"));
+    }
+  }
+
+  [Test]
+  public void ToAresValueSchema_SilaConstrainedStringList_PreservesChoices()
+  {
+    var silaType = new DataTypeType
+    {
+      Item = new ListType
+      {
+        DataType = new DataTypeType
+        {
+          Item = new ConstrainedType
+          {
+            DataType = new DataTypeType { Item = BasicType.String },
+            Constraints = new Constraints
+            {
+              Set = ["A", "B"]
+            }
+          }
+        }
+      }
+    };
+
+    var schema = SilaDataConverter.ToAresValueSchema(silaType);
+
+    using(Assert.EnterMultipleScope())
+    {
+      Assert.That(schema.Type, Is.EqualTo(AresDataType.StringArray));
+      Assert.That(schema.AvailableChoicesCase, Is.EqualTo(AresValueSchema.AvailableChoicesOneofCase.StringChoices));
+      Assert.That(schema.StringChoices.Strings, Is.EqualTo(new[] { "A", "B" }));
+    }
+  }
+
+  [Test]
+  public void ToSilaProperty_MixedAresList_UsesAnyTypedElements()
+  {
+    var value = AresValueHelper.CreateList(
+    [
+      AresValueHelper.CreateString("alpha"),
+      AresValueHelper.CreateNumber(42)
+    ]);
+
+    var property = SilaDataConverter.ToSilaProperty("Items", value);
+
+    Assert.That(property.Type.Item, Is.TypeOf<ListType>());
+    var listType = (ListType)property.Type.Item;
+    Assert.That(listType.DataType.Item, Is.EqualTo(BasicType.Any));
+
+    var items = ((IEnumerable<object>)property.Value!).ToList();
+    Assert.That(items, Has.Count.EqualTo(2));
+    using(Assert.EnterMultipleScope())
+    {
+      Assert.That(items[0], Is.TypeOf<DynamicObjectProperty>());
+      Assert.That(((DynamicObjectProperty)items[0]).Type.Item, Is.EqualTo(BasicType.String));
+      Assert.That(((DynamicObjectProperty)items[1]).Type.Item, Is.EqualTo(BasicType.Real));
+    }
+  }
+
+  [Test]
+  public void ToAresValue_UnmarkedQuantityShapedStructure_RemainsStruct()
+  {
+    var property = new DynamicObjectProperty(
+      "Quantity",
+      "Quantity",
+      null,
+      new DataTypeType
+      {
+        Item = new StructureType
+        {
[…]
