### 2026-04-16 675ec2a
Built ARES device descriptors from SILA features.
