+++ Ares.Core.Tests/SilaDeviceTests.cs
@@ -1,0 +1,251 @@
+using Ares.Core.Device.Sila;
+using Ares.Datamodel;
+using Ares.Datamodel.Device;
+using Ares.Device;
+using Moq;
+using Tecan.Sila2;
+using Tecan.Sila2.Client;
+
+namespace Ares.Core.Tests;
+
+[TestFixture]
+public class SilaDeviceTests
+{
+  [Test]
+  public async Task GetCommandDescriptorsAsync_BuildsDescriptorForScalarCommandWithoutResponse()
+  {
+    var serverData = CreateServerData(
+    [
+      CreateFeature(
+        "PumpControl",
+        CreateCommand(
+          "Start",
+          parameters:
+          [
+            CreateElement("Rate", BasicType.Real, description: "Pump rate")
+          ]))
+    ]);
+
+    var device = CreateDevice(serverData);
+
+    var descriptors = await device.GetCommandDescriptorsAsync();
+
+    Assert.That(descriptors, Has.Count.EqualTo(1));
+    var descriptor = descriptors[0];
+    using(Assert.EnterMultipleScope())
+    {
+      Assert.That(descriptor.Name, Is.EqualTo("PumpControl.Start"));
+      Assert.That(descriptor.Description, Is.Empty);
+      Assert.That(descriptor.InputSchema.Fields["Rate"].Type, Is.EqualTo(AresDataType.Number));
+      Assert.That(descriptor.InputSchema.Fields["Rate"].Description, Is.EqualTo("Pump rate"));
+      Assert.That(descriptor.OutputSchema, Is.Null);
+    }
+  }
+
+  [Test]
+  public async Task GetCommandDescriptorsAsync_UsesFeatureQualifiedNamesAcrossFeatures()
+  {
+    var serverData = CreateServerData(
+    [
+      CreateFeature("PumpA", CreateCommand("Start")),
+      CreateFeature("PumpB", CreateCommand("Start"))
+    ]);
+
+    var device = CreateDevice(serverData);
+
+    var descriptors = await device.GetCommandDescriptorsAsync();
+
+    Assert.That(descriptors.Select(d => d.Name), Is.EquivalentTo(new[] { "PumpA.Start", "PumpB.Start" }));
+  }
+
+  [Test]
+  public async Task GetCommandDescriptorsAsync_PreservesConstrainedInputSchema()
+  {
+    var constrainedString = new DataTypeType
+    {
+      Item = new ConstrainedType
+      {
+        DataType = new DataTypeType { Item = BasicType.String },
+        Constraints = new Constraints
+        {
+          Set = ["Fast", "Slow"]
+        }
+      }
+    };
+
+    var serverData = CreateServerData(
+    [
+      CreateFeature(
+        "PumpControl",
+        CreateCommand(
+          "Configure",
+          parameters:
+          [
+            CreateElement("Mode", constrainedString, "Operating mode")
+          ]))
+    ]);
+
+    var device = CreateDevice(serverData);
+
+    var descriptors = await device.GetCommandDescriptorsAsync();
+    var modeSchema = descriptors[0].InputSchema.Fields["Mode"];
+
+    using(Assert.EnterMultipleScope())
+    {
+      Assert.That(modeSchema.Type, Is.EqualTo(AresDataType.String));
+      Assert.That(modeSchema.Description, Is.EqualTo("Operating mode"));
+      Assert.That(modeSchema.StringChoices.Strings, Is.EqualTo(new[] { "Fast", "Slow" }));
+    }
+  }
+
+  [Test]
+  public async Task GetCommandDescriptorsAsync_BuildsStructOutputForSingleResponse()
+  {
+    var serverData = CreateServerData(
+    [
+      CreateFeature(
+        "PumpControl",
+        CreateCommand(
+          "ReadStatus",
+          responses:
+          [
[…]
