### 2026-04-16 020dae1
Implemented SILA device command execution. Added a bunch of tests.
