+++ Ares.Core.Tests/SilaDeviceTests.cs
@@ -1,10 +1,14 @@
 using Ares.Core.Device.Sila;
 using Ares.Datamodel;
 using Ares.Datamodel.Device;
+using Ares.Datamodel.Extensions;
 using Ares.Device;
 using Moq;
+using System.Threading.Channels;
 using Tecan.Sila2;
 using Tecan.Sila2.Client;
+using Tecan.Sila2.Discovery;
+using Tecan.Sila2.DynamicClient;
 
 namespace Ares.Core.Tests;
 
@@ -188,7 +192,281 @@ public class SilaDeviceTests
     Assert.That(descriptors.Select(d => d.Name), Is.EqualTo(new[] { "PumpControl.Start" }));
   }
 
-  private static SilaDevice CreateDevice(ServerData serverData)
+  [Test]
+  public async Task ExecuteCommand_UsesExactFeatureQualifiedCommandName()
+  {
+    var featureA = CreateFeature(
+      "PumpA",
+      CreateCommand("Start", responses: [], observable: FeatureCommandObservable.No));
+    var featureB = CreateFeature(
+      "PumpB",
+      CreateCommand("Start", responses: [], observable: FeatureCommandObservable.No));
+
+    var channelMock = CreateChannelMock();
+    channelMock
+      .Setup(c => c.ExecuteUnobservableCommandAsync(
+        It.Is<string>(serviceName => serviceName == $"{featureB.Namespace}.{featureB.Identifier}"),
+        It.Is<string>(commandName => commandName == "Start"),
+        It.IsAny<DynamicRequest>(),
+        It.IsAny<IClientCallInfo>(),
+        It.IsAny<ByteSerializer<DynamicRequest>>(),
+        It.IsAny<ByteSerializer<DynamicObjectProperty>>()))
+      .ReturnsAsync(CreateResponseProperty((FeatureCommand)featureB.Items[0]));
+
+    var device = CreateDevice(CreateServerData([featureA, featureB], channelMock.Object), CreateConfiguredClient());
+
+    var result = await device.ExecuteCommand("PumpB.Start", [], CancellationToken.None);
+
+    using(Assert.EnterMultipleScope())
+    {
+      Assert.That(result.Success, Is.True);
+      Assert.That(result.Result.KindCase, Is.EqualTo(AresValue.KindOneofCase.UnitValue));
+    }
+  }
+
+  [Test]
+  public async Task ExecuteCommand_UnobservableCommand_MapsRequestAndResponse()
+  {
+    var command = CreateCommand(
+      "Dispense",
+      parameters:
+      [
+        CreateElement("Rate", BasicType.Real)
+      ],
+      responses:
+      [
+        CreateElement("Status", BasicType.String)
+      ],
+      observable: FeatureCommandObservable.No);
+
+    var channelMock = CreateChannelMock();
+    channelMock
+      .Setup(c => c.ExecuteUnobservableCommandAsync(
+        It.IsAny<string>(),
+        It.IsAny<string>(),
+        It.Is<DynamicRequest>(request => RequestContains(request, "Rate", value => value != null && value.GetType() == typeof(double) && (double)value == 12.5)),
+        It.IsAny<IClientCallInfo>(),
+        It.IsAny<ByteSerializer<DynamicRequest>>(),
+        It.IsAny<ByteSerializer<DynamicObjectProperty>>()))
+      .ReturnsAsync(CreateResponseProperty(command, (command.Response[0], "Ready")));
+
+    var device = CreateDevice(
+      CreateServerData([CreateFeature("PumpControl", command)], channelMock.Object),
+      CreateConfiguredClient());
+
+    var result = await device.ExecuteCommand(
+      "PumpControl.Dispense",
+      [
[…]
