+++ Ares.Core/Device/Sila/SilaDevice.cs
@@ -1,5 +1,6 @@
 using Ares.Datamodel;
 using Ares.Datamodel.Device;
+using Ares.Datamodel.Extensions;
 using Ares.Datamodel.Factories;
 using Ares.Device;
 using System.Reactive.Linq;
@@ -74,9 +75,56 @@ public sealed class SilaDevice : AresDevice, IAsyncDisposable
     return Task.CompletedTask;
   }
 
-  public override Task<CommandResult> ExecuteCommand(string command, List<DeviceCommandArgument> arguments, CancellationToken token)
+  public override async Task<CommandResult> ExecuteCommand(string command, List<DeviceCommandArgument> arguments, CancellationToken token)
   {
-    throw new NotImplementedException();
+    token.ThrowIfCancellationRequested();
+
+    if(_silaClient.ExecutionManager is null)
+      return CreateFailedResult("SiLA client is not initialized.");
+
+    if(!TryResolveCommand(command, out var feature, out var featureCommand))
+      return CreateFailedResult($"Unknown SiLA command '{command}'.");
+
+    var context = new FeatureContext(feature, _serverData, _silaClient.ExecutionManager);
+
+    try
+    {
+      if(featureCommand.Observable == FeatureCommandObservable.Yes)
+      {
+        if(featureCommand.IntermediateResponse?.Length > 0)
+        {
+          var observableClient = new IntermediateObservableCommandClient(featureCommand, context);
+          var request = BuildRequest(observableClient.CreateRequest(), featureCommand, arguments);
+          var observableCommand = observableClient.Invoke(request);
+
+          using var cancellationRegistration = token.Register(observableCommand.Cancel);
+          // TODO: Surface observable state and intermediate responses once ARES exposes progress/intermediate result APIs.
+          var response = await observableCommand.Response.WaitAsync(token);
+          return CreateSuccessResult(featureCommand, response);
+        }
+
+        var trackedObservableClient = new ObservableCommandClient(featureCommand, context);
+        var trackedRequest = BuildRequest(trackedObservableClient.CreateRequest(), featureCommand, arguments);
+        var trackedObservableCommand = trackedObservableClient.Invoke(trackedRequest);
+
+        using var trackedCancellationRegistration = token.Register(trackedObservableCommand.Cancel);
+        var trackedResponse = await trackedObservableCommand.Response.WaitAsync(token);
+        return CreateSuccessResult(featureCommand, trackedResponse);
+      }
+
+      var nonObservableClient = new NonObservableCommandClient(featureCommand, context);
+      var nonObservableRequest = BuildRequest(nonObservableClient.CreateRequest(), featureCommand, arguments);
+      var nonObservableResponse = await nonObservableClient.InvokeAsync(nonObservableRequest).WaitAsync(token);
+      return CreateSuccessResult(featureCommand, nonObservableResponse);
+    }
+    catch(OperationCanceledException)
+    {
+      throw;
+    }
+    catch(Exception ex)
+    {
+      return CreateFailedResult(ex.Message);
+    }
   }
 
   public override Task<AresStruct> GetSettings()
@@ -115,11 +163,87 @@ public sealed class SilaDevice : AresDevice, IAsyncDisposable
   {
     return new DeviceCommandDescriptor
     {
-      Name = $"{feature.Identifier}.{command.Identifier}",
+      Name = GetCommandDescriptorName(feature, command),
       Description = command.Description ?? string.Empty,
       InputSchema = BuildStructSchema(command.Parameter),
       OutputSchema = BuildOutputSchema(command.Response)
     };
+  }
+
+  private static DynamicRequest BuildRequest(
+    DynamicRequest request,
[…]
