### 2026-04-17 36b68ba
Add a really quick and dirty manual option for adding a SiLA device as we discovered the auto discover feature isn't always reliable
