### 2026-04-17 e1961ee
Added lil null check for discovered SiLA servers. Removed test code
