### 2026-04-20 90fa445
Fixed bug that prevented changing initial value after first assignment
