### 2026-04-21 99c4b2e
Updated the validator to search for nested keys within struct return values.
