### 2026-04-21 2d16f8e
Fixed notification not being sent via proper service causing it to not be tracked in the event viewer
