### 2026-04-21 f555bd5
Added some additional deadlines to client calls that would freeze up the UI if there was a strange crash in a remote device (such as a serial port connection not being closed properly causing a hang). Added some additional logging to hopefully provide better clues if these issues ever popup again.
