+++ UI/ServiceStarter.cs
@@ -28,6 +28,7 @@ public class ServiceStarter : BackgroundService
   private readonly StateLoggerManager _stateLoggerManager;
   private readonly IConfiguration _configuration;
   private readonly StartupStateTracker _tracker;
+  private readonly ILogger<ServiceStarter> _logger;
 
   private readonly string _dataPath;
   private readonly string _resultsPath;
@@ -49,7 +50,8 @@ public class ServiceStarter : BackgroundService
     IDeviceControlViewModelRepo deviceControlViewModelRepo,
     DeviceAdapterManager deviceAdapterManager,
     StateLoggerManager stateLoggerManager,
-    StartupStateTracker tracker)
+    StartupStateTracker tracker,
+    ILogger<ServiceStarter> logger)
   {
     _notificationReceivingService = notificationReceivingService;
     _deviceControlViewModelRepo = deviceControlViewModelRepo;
@@ -71,6 +73,7 @@ public class ServiceStarter : BackgroundService
     _templatesPath = Path.Combine(_dataPath, AppSettings.TemplatesFolder);
     _devicesPath = Path.Combine(_dataPath, AppSettings.DevicesFolder);
     _pluginsPath = PluginPathResolver.Resolve(_configuration.Get<AppSettings>());
+    _logger = logger;
   }
 
   protected override async Task ExecuteAsync(CancellationToken cancellationToken)
@@ -89,6 +92,7 @@ public class ServiceStarter : BackgroundService
       //If we overwrite it too early, our archive wipes out the references to the deleted drivers making
       //it impossible to migrate devices to updated drivers.
       await _driverDbManager.RefreshDriverArchive();
+      _logger.LogInformation("Finished loading the local startup track");
     }, cancellationToken);
 
     var infraTrack = EnsureDataPathsExist();
@@ -96,11 +100,11 @@ public class ServiceStarter : BackgroundService
     var remoteTrack = Task.WhenAll(
       _plannerManager.LoadPlanners(),
       _analyzerManager.LoadAnalyzers(),
-      _remoteDeviceManager.LoadDevices()
-    );
+      _remoteDeviceManager.LoadDevices());
 
     await Task.WhenAll(localTrack, infraTrack, remoteTrack);
     await Task.Delay(TimeSpan.FromSeconds(6));
+    _logger.LogInformation("Successfully Finished the Loading Sequences");
     _tracker.MarkAsReady();
   }
 
