### 2026-04-21 36b0114
Fixed a bug that caused strange behavior when changing from a deleted device command
