### 2026-04-23 #10461 @editorialbot
[REVIEW]: ARES OS 2.0: An Orchestration Software Suite for Autonomous Experimentation Systems and Self-Driving Labs
**Submitting author:** @arthursloan (<a href="http://orcid.org/0000-0002-7066-1678">Arthur Sloan</a>)
**Repository:** https://github.com/AFRL-ARES/ARES
**Branch with paper.md** (empty if default branch): JOSS-Submission
**Version:** 1.0.0
