### 2026-04-28 91188b9
Added the new error states, as well as an enum to define the behaviors available for the user to select from.
