+++ UI/Features/Settings/SystemSettingsViewModel.cs
@@ -1,0 +1,5 @@
+﻿namespace UI.Features.Settings;
+
+public class SystemSettingsViewModel
+{
+}
