### 2026-04-28 2c625ac
Support for dual device visualization
