+++ UI/Features/Visualization/ViewModels/VisualizationItemViewModel.cs
@@ -4,6 +4,8 @@ using Ares.Datamodel.Visualizing.Local;
 using Ares.Device;
 using ReactiveUI;
 using ReactiveUI.SourceGenerators;
+using System.Reactive.Disposables;
+using System.Reactive.Disposables.Fluent;
 using System.Reactive.Linq;
 using UI.Features.Visualization.Models;
 
@@ -11,21 +13,24 @@ namespace UI.Features.Visualization.ViewModels;
 
 public partial class VisualizationItemViewModel : ReactiveObject, IDisposable
 {
-  private IDisposable? _streamSubscription;
-  private readonly IAresDevice _device;
+  private readonly CompositeDisposable _streamSubscription = new();
+  private readonly IEnumerable<IAresDevice> _devices;
   private DeviceVisualizationConfig _config;
   private readonly Action<string> _onDeleteRequested;
   private readonly object _bufferLock = new object();
-  private readonly List<ChartDataPoint> _internalBuffer = new(50);
+
+  // NEW: Changed from a single list to a Dictionary mapped by "DeviceName - Path"
+  private readonly Dictionary<string, List<ChartDataPoint>> _internalBuffers = new();
+
   private readonly Action<string, DeviceVisualizationConfig> _onUpdateRequested;
 
-  public VisualizationItemViewModel(DeviceVisualizationConfig config, 
-    IAresDevice device, 
-    Action<string> onDeleteRequested, 
+  public VisualizationItemViewModel(DeviceVisualizationConfig config,
+    IEnumerable<IAresDevice> devices,
+    Action<string> onDeleteRequested,
     Action<string, DeviceVisualizationConfig> onUpdateRequested)
   {
     _config = config;
-    _device = device;
+    _devices = devices;
     _onDeleteRequested = onDeleteRequested;
     _onUpdateRequested = onUpdateRequested;
 
@@ -34,9 +39,16 @@ public partial class VisualizationItemViewModel : ReactiveObject, IDisposable
     NumberOfDisplayPoints = config.NumberDisplayPoints;
     DisplayLabels = config.ShowDataLabels;
     DisplayMarkers = config.ShowMarkers;
-    DataPoints = [];
+
+    // Initialize the new dictionary
+    DataPoints = new Dictionary<string, IList<ChartDataPoint>>();
+
     UniqueId = config.UniqueId;
-    Title = $"{device.Name} : {config.Path.Path}";
+    if(config.Paths.Count == 1 && devices.Count() == 1)
+      Title = $"{devices.First().Name} : {config.Paths.First().Path}";
+    else
+      Title = $"Multi-Data Display Chart";
+
     Style = config.Style;
 
     GridX = config.GridX;
@@ -49,40 +61,68 @@ public partial class VisualizationItemViewModel : ReactiveObject, IDisposable
 
   public void StartStreamSubscription()
   {
-    _streamSubscription?.Dispose();
-    Console.WriteLine($"Disposing the old stream, creating a new one with polling frequency of {PollingFrequencyMs}");
-    _streamSubscription = _device.StateStream
+    _streamSubscription.Clear();
+
+    lock(_bufferLock)
+    {
+      _internalBuffers.Clear();
+    }
+
+    foreach(var device in _devices)
+    {
+      if(device is null)
+        continue;
+
+      var matchingPaths = _config.Paths.Where(p => p.AssociatedDeviceName == device.UniqueId).ToList();
+      if(!matchingPaths.Any()) continue;
+
+      device.StateStream
         .Sample(TimeSpan.FromMilliseconds(PollingFrequencyMs))
-        .Do(state => ProcessNewState(state, _config.Path))
+        .Do(state => ProcessNewState(state, matchingPaths, device.Name))
         .Subscribe(
             onNext: _ =>
             {
[…]
