### 2026-04-28 ef43521
Added ability to name charts.
