### 2026-04-29 4c3679f
Error handling updates
