+++ Ares.Core/EntityConfigurations/Shared/ErrorHandlingEntityConfiguration.cs
@@ -1,0 +1,7 @@
+﻿using Ares.Datamodel;
+
+namespace Ares.Core.EntityConfigurations.Shared;
+
+internal class ErrorHandlingEntityConfiguration : AresEntityTypeBaseConfiguration<Limits>
+{
+}
