### 2026-04-29 d0c036c
Updated charts to react properly when devices are removed. If a device is removed all charts relying solely on it's data are removed, all charts incoprorating multiple data flows with some not being related to that device are recreated to not include that information.
