+++ Ares.Core/Visualization/Managers/IVisualizationConfigManager.cs
@@ -5,7 +5,7 @@ namespace Ares.Core.Visualization.Managers;
 
 public interface IVisualizationConfigManager
 {
-  Task LoadConfigs();
+  Task Initialize();
   Task AddDeviceVisualization(List<VisualizationPath> paths, ChartStyle style);
   Task Remove(string configId);
   Task UpdateDeviceVisualization(string configId, DeviceVisualizationConfig config);
