### 2026-04-29 a096707
Added a button to clear selected data paths.
