### 2026-05-01 04a1466
Fixed some bad database context logic in the config manager
