### 2026-05-01 e75f1ab
Additional driver management logic cleanup
