### 2026-05-04 ccc9bc4
Updated device toolkit version. Added some additional logging output to device creation process.
