+++ Ares.Core/EntityConfigurations/Helpers/VisualizationSerializeHelper.cs
@@ -2,7 +2,6 @@
 using Google.Protobuf.Collections;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using Microsoft.EntityFrameworkCore.Metadata.Internal;
 using System.Text.Json;
 
 namespace Ares.Core.EntityConfigurations.Helpers;
@@ -16,14 +15,5 @@ public static class VisualizationSerializeHelper
       v => JsonSerializer.Serialize(v, settings),
       v => JsonSerializer.Deserialize<RepeatedField<VisualizationPath>>(v, settings) ?? new RepeatedField<VisualizationPath>())
     .HasColumnType(SerializerSettingsHelper.DetermineColumnType());
-  }
-
-  public static PropertyBuilder<RepeatedField<string>> HasDeviceIds(this PropertyBuilder<RepeatedField<string>> value)
-  {
-    var settings = SerializerSettingsHelper.CreateCustomSerializationSettings();
-    return value.HasConversion(
-      v => JsonSerializer.Serialize(v, settings),
-      v => JsonSerializer.Deserialize<RepeatedField<string>>(v, settings) ?? new RepeatedField<string>())
-      .HasColumnType(SerializerSettingsHelper.DetermineColumnType());
   }
 }
