### 2026-05-05 0bae333
Fixed device creation (likely an artifact from the merge). Fixed manual planning, made it accessible in the new execution page.
