### 2026-05-05 a39f152
Created newest migration to capture changes supporting visualization updates
