### 2026-05-06 d04c7df
Initial model for data explorer
