### 2026-05-08 6b320db
Made things more async when acquiring symbols. This should fix some stuck device command gettings.
