### 2026-05-08 987563e
Made things more async when acquiring symbols. This should fix some stuck device command gettings. (#152)
