+++ Ares.Core.Grpc/Services/AresScriptingService.cs
@@ -33,7 +33,7 @@ public partial class AresScriptingService : Ares.Services.AresScriptingService.A
     [EnumeratorCancellation] CancellationToken cancellationToken = default)
   {
     var channel = Channel.CreateBounded<CoreScriptExecutionEvent>(new BoundedChannelOptions(100));
-    var env = _environmentBuilder.Build();
+    var env = await _environmentBuilder.Build();
     var runner = new ScriptRunner(env);
     var subscriptions = new CompositeDisposable
     {
@@ -86,7 +86,7 @@ public partial class AresScriptingService : Ares.Services.AresScriptingService.A
   public override async Task ExecuteScript(ScriptExecutionRequest request, IServerStreamWriter<GrpcScriptExecutionEvent> responseStream, ServerCallContext context)
   {
     var channel = Channel.CreateBounded<GrpcScriptExecutionEvent>(new BoundedChannelOptions(100));
-    var env = _environmentBuilder.Build();
+    var env = await _environmentBuilder.Build();
     var runner = new ScriptRunner(env);
     var subscriptions = new CompositeDisposable
     {
@@ -145,7 +145,7 @@ public partial class AresScriptingService : Ares.Services.AresScriptingService.A
 
   public override async Task<CompletionResponse> GetCompletions(CompletionRequest request, ServerCallContext context)
   {
-    var environment = _environmentBuilder.Build();
+    var environment = await _environmentBuilder.Build();
     var items = await AresScriptAnalysis.BuildCompletionsAsync(
       environment,
       request.Script,
@@ -158,7 +158,7 @@ public partial class AresScriptingService : Ares.Services.AresScriptingService.A
 
   public override async Task<ValidateScriptResponse> ValidateScript(ValidateScriptRequest request, ServerCallContext context)
   {
-    var environment = _environmentBuilder.Build();
+    var environment = await _environmentBuilder.Build();
     var diagnostics = await AresScriptAnalysis.ValidateScriptAsync(
       request.Script,
       environment,
@@ -171,7 +171,7 @@ public partial class AresScriptingService : Ares.Services.AresScriptingService.A
 
   public override async Task<ScriptSummaryResponse> GetScriptSummary(ScriptSummaryRequest request, ServerCallContext context)
   {
-    var environment = _environmentBuilder.Build();
+    var environment = await _environmentBuilder.Build();
     var (steps, diagnostics) = await AresScriptAnalysis.BuildScriptSummaryAsync(
       request.Script,
       environment,
@@ -187,7 +187,7 @@ public partial class AresScriptingService : Ares.Services.AresScriptingService.A
 
   public override async Task<SymbolMetadataResponse> GetSymbolMetadata(SymbolMetadataRequest request, ServerCallContext context)
   {
-    var environment = _environmentBuilder.Build();
+    var environment = await _environmentBuilder.Build();
     var metadata = await AresScriptAnalysis.BuildSymbolMetadataAsync(
       environment,
       request.Script,
