### 2026-05-11 2116a04
Added a dataset provider for device states based on an input filter.
