+++ Ares.Core.Tests/DataManagement/DataMappers/DeviceStateDatasetGeneratorTests.cs
@@ -1,0 +1,191 @@
+using Ares.Core.DataManagement.DataMappers;
+using Ares.Core.Device.State.Export.StateGetters;
+using Ares.Datamodel;
+using Ares.Datamodel.Device;
+using Ares.Datamodel.Extensions;
+using Google.Protobuf.WellKnownTypes;
+using Moq;
+
+namespace Ares.Core.Tests.DataManagement.DataMappers;
+
+internal class DeviceStateDatasetGeneratorTests
+{
+  [Test]
+  public async Task GenerateAsync_CreatesOneDatasetPerDevice()
+  {
+    var filter = new DeviceStateRequestFilter();
+    var stateGetter = CreateStateGetter(filter, new Dictionary<string, IEnumerable<DeviceState>>
+    {
+      ["Device A"] = [CreateState(DateTime.UnixEpoch, ("Temperature", AresValueHelper.CreateNumber(1)))],
+      ["Device B"] = [CreateState(DateTime.UnixEpoch.AddSeconds(1), ("Enabled", AresValueHelper.CreateBool(true)))]
+    });
+
+    var generator = new DeviceStateDatasetGenerator(stateGetter.Object);
+    var datasets = await generator.GenerateAsync(filter);
+
+    Assert.That(datasets.Select(dataset => dataset.Name), Is.EquivalentTo(["Device A", "Device B"]));
+  }
+
+  [Test]
+  public async Task GenerateAsync_SortsRowsAndUsesTimestampValues()
+  {
+    var filter = new DeviceStateRequestFilter();
+    var firstTimestamp = DateTime.UnixEpoch.AddSeconds(1);
+    var secondTimestamp = DateTime.UnixEpoch.AddSeconds(2);
+    var stateGetter = CreateStateGetter(filter, new Dictionary<string, IEnumerable<DeviceState>>
+    {
+      ["Device A"] =
+      [
+        CreateState(secondTimestamp, ("Temperature", AresValueHelper.CreateNumber(2))),
+        CreateState(firstTimestamp, ("Temperature", AresValueHelper.CreateNumber(1)))
+      ]
+    });
+
+    var generator = new DeviceStateDatasetGenerator(stateGetter.Object);
+    var dataset = (await generator.GenerateAsync(filter)).Single();
+
+    using(Assert.EnterMultipleScope())
+    {
+      Assert.That(dataset.Columns[0].Name, Is.EqualTo("Timestamp"));
+      Assert.That(dataset.Columns[0].Schema.Type, Is.EqualTo(AresDataType.Timestamp));
+      Assert.That(dataset.Rows[0].Data.Fields["Timestamp"].KindCase, Is.EqualTo(AresValue.KindOneofCase.TimestampValue));
+      Assert.That(dataset.Rows[0].Data.Fields["Timestamp"].TimestampValue, Is.EqualTo(Timestamp.FromDateTime(firstTimestamp)));
+      Assert.That(dataset.Rows[1].Data.Fields["Timestamp"].TimestampValue, Is.EqualTo(Timestamp.FromDateTime(secondTimestamp)));
+    }
+  }
+
+  [Test]
+  public async Task GenerateAsync_UsesUnionOfStateFieldsForTypedColumns()
+  {
+    var filter = new DeviceStateRequestFilter();
+    var measuredAt = Timestamp.FromDateTime(DateTime.UnixEpoch.AddSeconds(10));
+    var stateGetter = CreateStateGetter(filter, new Dictionary<string, IEnumerable<DeviceState>>
+    {
+      ["Device A"] =
+      [
+        CreateState(
+          DateTime.UnixEpoch,
+          ("Temperature", AresValueHelper.CreateNumber(12.3)),
+          ("Name", AresValueHelper.CreateString("alpha")),
[…]
