### 2026-05-11 fc1233f
Added campaign dataset generator. Changed the device state getter to return arrays.
