+++ Ares.Core/DataManagement/DataMappers/CampaignDatasetGenerator.cs
@@ -1,0 +1,167 @@
+using Ares.Datamodel;
+using Ares.Datamodel.Extensions;
+using Ares.Datamodel.Templates;
+using Microsoft.EntityFrameworkCore;
+
+namespace Ares.Core.DataManagement.DataMappers;
+
+internal class CampaignDatasetGenerator(IDbContextFactory<CoreDatabaseContext> _dbContextFactory)
+{
+  private const string ExperimentNumberColumnName = "Experiment Number";
+  private const string TimeStartedColumnName = "Time Started";
+  private const string TimeFinishedColumnName = "Time Finished";
+  private const string AnalysisResultColumnName = "Analysis Result";
+  private const string ParameterColumnPrefix = "Parameter.";
+
+  public async ValueTask<AresDataset[]> GenerateAsync(string summaryId, CancellationToken cancellationToken = default)
+  {
+    cancellationToken.ThrowIfCancellationRequested();
+
+    await using var ctx = await _dbContextFactory.CreateDbContextAsync(cancellationToken);
+    var summary = await ctx.CampaignExecutionSummaries.FirstOrDefaultAsync(ces => ces.UniqueId == summaryId, cancellationToken);
+    if(summary is null)
+      return [];
+
+    var experiments = summary.ExperimentSummaries
+      .OrderBy(experiment => experiment.ExecutionInfo?.TimeStarted)
+      .ToArray();
+
+    var dataset = new AresDataset
+    {
+      Name = GetDatasetName(summary, summaryId)
+    };
+
+    dataset.Columns.AddRange(CreateFixedColumns());
+    dataset.Columns.AddRange(CreateDynamicColumns(experiments, cancellationToken));
+    dataset.Rows.AddRange(experiments.Select((experiment, index) => CreateRow(experiment, index + 1, cancellationToken)));
+
+    return [dataset];
+  }
+
+  private static string GetDatasetName(CampaignExecutionSummary summary, string summaryId)
+  {
+    if(!string.IsNullOrWhiteSpace(summary.CampaignName))
+      return summary.CampaignName;
+
+    if(!string.IsNullOrWhiteSpace(summary.CampaignId))
+      return summary.CampaignId;
+
+    return summaryId;
+  }
+
+  private static IEnumerable<AresDataColumn> CreateFixedColumns()
+  {
+    return
+    [
+      new AresDataColumn
+      {
+        Name = ExperimentNumberColumnName,
+        Schema = new AresValueSchema { Type = AresDataType.Int }
+      },
+      new AresDataColumn
+      {
+        Name = TimeStartedColumnName,
+        Schema = new AresValueSchema { Type = AresDataType.Timestamp }
+      },
+      new AresDataColumn
+      {
+        Name = TimeFinishedColumnName,
+        Schema = new AresValueSchema { Type = AresDataType.Timestamp }
+      },
+      new AresDataColumn
+      {
+        Name = AnalysisResultColumnName,
+        Schema = new AresValueSchema { Type = AresDataType.Number, Optional = true }
+      }
+    ];
+  }
+
+  private static IEnumerable<AresDataColumn> CreateDynamicColumns(IEnumerable<ExperimentExecutionSummary> experiments, CancellationToken cancellationToken)
+  {
+    var columns = new Dictionary<string, AresValueSchema>();
+
+    foreach(var experiment in experiments)
+    {
+      cancellationToken.ThrowIfCancellationRequested();
+
+      var resultFields = experiment.ExperimentOverview?.Result?.Fields.OrderBy(field => field.Key)
+        ?? Enumerable.Empty<KeyValuePair<string, AresValue>>();
[…]
