### 2026-05-12 454cc83
Settings menu now has a functionable selection process for handling errors
