+++ Ares.Core/Settings/ISystemSettingsManager.cs
@@ -1,0 +1,10 @@
+﻿using Ares.Datamodel;
+
+namespace Ares.Core.Settings;
+
+public interface ISystemSettingsManager
+{
+  public Task UpdateErrorHandlingSettings(List<DeviceErrorHandlingConfig> configs);
+
+  public Task<IEnumerable<DeviceErrorHandlingConfig>> GetCurrentErrorHandlingSettings();
+}
