### 2026-05-12 f49f299
Working towards integrating the new settings into the loop. First with retying commands, which is done at the step execution level when applicable. Next is going to be dealing with things like replanning and safe mode entry.
