+++ Ares.Core/Settings/ISystemSettingsManager.cs
@@ -7,4 +7,6 @@ public interface ISystemSettingsManager
   public Task UpdateErrorHandlingSettings(List<DeviceErrorHandlingConfig> configs);
 
   public Task<IEnumerable<DeviceErrorHandlingConfig>> GetCurrentErrorHandlingSettings();
+
+  public Task<ErrorHandling> GetErrorHandlingByStatusCode(CommandStatusCode code);
 }
