+++ Ares.Core/Execution/Executors/Composers/StepComposer.cs
@@ -1,5 +1,6 @@
 ﻿using Ares.Core.Device.Repos;
 using Ares.Core.Notifications;
+using Ares.Core.Settings;
 using Ares.Datamodel;
 using Ares.Datamodel.Device;
 using Ares.Datamodel.Templates;
@@ -10,11 +11,12 @@ public class StepComposer : ICommandComposer<StepTemplate, StepExecutor>
 {
   private readonly INotifier _notifier;
   private readonly IAresDeviceRepo _deviceRepo;
-
-  public StepComposer(IAresDeviceRepo deviceRepo, INotifier notifier)
+  private readonly ISystemSettingsManager _settingsManager;
+  public StepComposer(IAresDeviceRepo deviceRepo, INotifier notifier, ISystemSettingsManager settingsManager)
   {
     _deviceRepo = deviceRepo;
     _notifier = notifier;
+    _settingsManager = settingsManager;
   }
 
   public StepExecutor Compose(StepTemplate template)
@@ -41,7 +43,7 @@ public class StepComposer : ICommandComposer<StepTemplate, StepExecutor>
               Func<CancellationToken, Task<CommandResult>> internalAction = async (ct)
                 => await device.ExecuteCommand(commandTemplate.Metadata.Name, commandArgs, ct);
 
-              return new CommandExecutor(internalAction, commandTemplate, _notifier);
+              return new CommandExecutor(internalAction, commandTemplate, _notifier, _settingsManager);
             }
 
             throw new InvalidOperationException("I'm not certain what to do here yet :(");
@@ -52,7 +54,7 @@ public class StepComposer : ICommandComposer<StepTemplate, StepExecutor>
 
     return template.IsParallel
       ? new ParallelStepExecutor(template, executables)
-      : new SequentialStepExecutor(template, executables);
+      : new SequentialStepExecutor(template, executables, _settingsManager, _notifier);
   }
 
   public StepExecutor Compose(StepTemplate template, ExperimentExecutionStatus experimentExecutionStatus)
