### 2026-05-12 c9f57c4
Added a basic data explorer for device states.
