+++ Ares.Core/Device/State/Logging/DeviceStateQueryBuilder.cs
@@ -21,7 +21,8 @@ public static class DeviceStateQueryBuilder
     }
     if(!string.IsNullOrEmpty(request.CompletedCampaignId))
     {
-      var completedCampaign = await dbContext.CampaignExecutionSummaries.FirstOrDefaultAsync(result => result.CampaignId == request.CompletedCampaignId);
+      var completedCampaign = await dbContext.CampaignExecutionSummaries.FirstOrDefaultAsync(result =>
+        result.UniqueId == request.CompletedCampaignId || result.CampaignId == request.CompletedCampaignId);
       if(completedCampaign is not null)
         statesQuery = statesQuery.Where(state => state.Timestamp >= completedCampaign.ExecutionInfo.TimeStarted && state.Timestamp <= completedCampaign.ExecutionInfo.TimeFinished);
     }
