### 2026-05-12 e40a9b9
Device dataset generator will not expand sub structs
