+++ Ares.Core/DataManagement/DataMappers/DeviceStateDatasetGenerator.cs
@@ -39,7 +39,7 @@ public class DeviceStateDatasetGenerator(IDeviceStateGetter _deviceStateGetter)
   {
     var columns = states
       .SelectMany(state => state.Data?.Fields ?? [])
-      .Select(field => KeyValuePair.Create(GetColumnName(field.Key), field.Value))
+      .SelectMany(field => FlattenField(field.Key, field.Value))
       .GroupBy(field => field.Key)
       .OrderBy(group => group.Key)
       .Select(group => new AresDataColumn
@@ -110,13 +110,34 @@ public class DeviceStateDatasetGenerator(IDeviceStateGetter _deviceStateGetter)
     foreach(var field in state.Data?.Fields ?? [])
     {
       cancellationToken.ThrowIfCancellationRequested();
-      data.Fields[GetColumnName(field.Key)] = field.Value.Clone();
+      foreach(var flattenedField in FlattenField(field.Key, field.Value))
+      {
+        cancellationToken.ThrowIfCancellationRequested();
+        data.Fields[flattenedField.Key] = flattenedField.Value.Clone();
+      }
     }
 
     return new AresDataRow
     {
       Data = data
     };
+  }
+
+  private static IEnumerable<KeyValuePair<string, AresValue>> FlattenField(string fieldName, AresValue value)
+  {
+    if(value.KindCase != AresValue.KindOneofCase.StructValue)
+    {
+      yield return KeyValuePair.Create(GetColumnName(fieldName), value);
+      yield break;
+    }
+
+    foreach(var childField in value.StructValue.Fields)
+    {
+      foreach(var flattenedChildField in FlattenField($"{fieldName}.{childField.Key}", childField.Value))
+      {
+        yield return flattenedChildField;
+      }
+    }
   }
 
   private static string GetColumnName(string fieldName)
