+++ Ares.Core.Tests/DataManagement/DataMappers/DeviceStateLongDatasetProjectorTests.cs
@@ -1,0 +1,169 @@
+using Ares.Core.DataManagement.DataMappers;
+using Ares.Datamodel;
+using Ares.Datamodel.Extensions;
+using Google.Protobuf.WellKnownTypes;
+
+namespace Ares.Core.Tests.DataManagement.DataMappers;
+
+internal class DeviceStateLongDatasetProjectorTests
+{
+  [Test]
+  public void Project_CombinesMultipleDeviceDatasets()
+  {
+    var timestamp = Timestamp.FromDateTime(DateTime.UnixEpoch);
+    var deviceDatasets = new[]
+    {
+      CreateDataset("Device A", CreateRow(timestamp, ("Temperature", AresValueHelper.CreateNumber(1)))),
+      CreateDataset("Device B", CreateRow(timestamp, ("Pressure", AresValueHelper.CreateNumber(2))))
+    };
+
+    var longDataset = DeviceStateLongDatasetProjector.Project(deviceDatasets);
+
+    using(Assert.EnterMultipleScope())
+    {
+      Assert.That(longDataset.Columns.Select(column => column.Name), Is.EqualTo([
+        "Timestamp",
+        "Device",
+        "Property",
+        "Value",
+        "Unit"
+      ]));
+      Assert.That(longDataset.Rows.Select(row => row.Data.Fields["Device"].StringValue), Is.EqualTo(["Device A", "Device B"]));
+      Assert.That(longDataset.Rows.Select(row => row.Data.Fields["Property"].StringValue), Is.EqualTo(["Temperature", "Pressure"]));
+      Assert.That(longDataset.Rows.Select(row => row.Data.Fields["Value"].NumberValue), Is.EqualTo([1, 2]));
+    }
+  }
+
+  [Test]
+  public void Project_SplitsQuantityValueAndUnit()
+  {
+    var timestamp = Timestamp.FromDateTime(DateTime.UnixEpoch);
+    var deviceDatasets = new[]
+    {
+      CreateDataset("Device A", CreateRow(timestamp, ("Mass", AresValueHelper.CreateQuantity(4.5, QuantityType.Mass, "g"))))
+    };
+
+    var row = DeviceStateLongDatasetProjector.Project(deviceDatasets).Rows.Single();
+
+    using(Assert.EnterMultipleScope())
+    {
+      Assert.That(row.Data.Fields["Value"].KindCase, Is.EqualTo(AresValue.KindOneofCase.NumberValue));
+      Assert.That(row.Data.Fields["Value"].NumberValue, Is.EqualTo(4.5));
+      Assert.That(row.Data.Fields["Unit"].StringValue, Is.EqualTo("g"));
+    }
+  }
+
+  [Test]
+  public void Project_OmitsUnitForNonQuantityValues()
+  {
+    var timestamp = Timestamp.FromDateTime(DateTime.UnixEpoch);
+    var deviceDatasets = new[]
+    {
+      CreateDataset("Device A", CreateRow(timestamp, ("Enabled", AresValueHelper.CreateBool(true))))
+    };
+
+    var row = DeviceStateLongDatasetProjector.Project(deviceDatasets).Rows.Single();
+
+    using(Assert.EnterMultipleScope())
+    {
+      Assert.That(row.Data.Fields["Value"].BoolValue, Is.True);
+      Assert.That(row.Data.Fields.ContainsKey("Unit"), Is.False);
+    }
+  }
+
+  [Test]
+  public void Project_PreservesFlattenedPropertyNames()
+  {
+    var timestamp = Timestamp.FromDateTime(DateTime.UnixEpoch);
+    var deviceDatasets = new[]
+    {
+      CreateDataset("Device A", CreateRow(timestamp, ("Position.Offset.X", AresValueHelper.CreateNumber(1.2))))
+    };
[…]
