### 2026-05-14 e26128b
Prevent UI from freezing when applying a large dataset
