### 2026-05-14 6d6186c
Working on adding some general settings
