+++ Ares.Core/Settings/ISystemSettingsManager.cs
@@ -4,9 +4,10 @@ namespace Ares.Core.Settings;
 
 public interface ISystemSettingsManager
 {
+  public Task Initialize();
   public Task UpdateErrorHandlingSettings(List<DeviceErrorHandlingConfig> configs);
-
   public Task<IEnumerable<DeviceErrorHandlingConfig>> GetCurrentErrorHandlingSettings();
-
   public Task<ErrorHandling> GetErrorHandlingByStatusCode(CommandStatusCode code);
+  public Task<AresGeneralSettingsConfig?> GetAresGeneralSettings();
+  public Task UpdateAresGeneralSettings(AresGeneralSettingsConfig config);
 }
