+++ UI/Features/Settings/SystemSettingsViewModel.cs
@@ -1,10 +1,11 @@
 ﻿using Ares.Core.Settings;
 using Ares.Datamodel;
 using ReactiveUI;
+using ReactiveUI.SourceGenerators;
 
 namespace UI.Features.Settings;
 
-public class SystemSettingsViewModel : ReactiveObject
+public partial class SystemSettingsViewModel : ReactiveObject
 {
   private readonly ISystemSettingsManager _settingsManager;
 
@@ -24,13 +25,28 @@ public class SystemSettingsViewModel : ReactiveObject
     ).ToList();
 
     await _settingsManager.UpdateErrorHandlingSettings(configs);
+    await _settingsManager.UpdateAresGeneralSettings(new AresGeneralSettingsConfig
+    {
+      ExperimentRetryLimit = ExperimentRetryLimit,
+      RetryCooldown = ExperimentRetryCooldown,
+      CommandLatency = CommandLatency
+    });
   }
 
   public async Task GetUpdatedSettings()
   {
-    var newSettings = await _settingsManager.GetCurrentErrorHandlingSettings();
-    foreach(var setting in newSettings)
+    var newErrorHandlingSettings = await _settingsManager.GetCurrentErrorHandlingSettings();
+    var newGeneralSettings = await _settingsManager.GetAresGeneralSettings();
+
+    foreach(var setting in newErrorHandlingSettings)
       CurrentErrorHandlingSettings[setting.Code] = setting.Handling;
+
+    if(newGeneralSettings is not null)
+    {
+      ExperimentRetryCooldown = newGeneralSettings.RetryCooldown;
+      ExperimentRetryLimit = newGeneralSettings.ExperimentRetryLimit;
+      CommandLatency = newGeneralSettings.CommandLatency;
+    }
   }
 
   private void Initialize()
@@ -42,6 +58,14 @@ public class SystemSettingsViewModel : ReactiveObject
     }
   }
 
-
   public Dictionary<CommandStatusCode, ErrorHandling> CurrentErrorHandlingSettings { get; set; }
+
+  [Reactive]
+  public partial int ExperimentRetryLimit { get; set; }
+
+  [Reactive]
+  public partial int ExperimentRetryCooldown { get; set; }
+
+  [Reactive]
+  public partial int CommandLatency { get; set; }
 }
