+++ Ares.Core/Execution/Executors/CampaignExecutor.cs
@@ -7,10 +7,12 @@ using Ares.Core.Exceptions;
 using Ares.Core.Execution.ControlTokens;
 using Ares.Core.Execution.Executors.Composers;
 using Ares.Core.Execution.Extensions;
+using Ares.Core.Execution.Safety;
 using Ares.Core.Execution.StopConditions;
 using Ares.Core.Notifications;
 using Ares.Core.Output;
 using Ares.Core.Planning;
+using Ares.Core.Settings;
 using Ares.Datamodel;
 using Ares.Datamodel.Analyzing;
 using Ares.Datamodel.Templates;
@@ -33,6 +35,8 @@ public class CampaignExecutor : ICampaignExecutor
   readonly AnalysisHelper _analysisHelper;
   readonly AnalysisRepo _analysisRepo;
   readonly IAnalyzerRepo _analyzerRepo;
+  readonly ISystemSettingsManager _settingsManager;
+  readonly IExecutionSafetyManager _executionSafetyManager;
 
   internal CampaignExecutor(ICommandComposer<ExperimentTemplate, ExperimentExecutor> experimentComposer,
     IPlanningHelper planningHelper,
@@ -45,7 +49,9 @@ public class CampaignExecutor : ICampaignExecutor
     IAnalyzerRepo analyzerRepo,
     ILogger<CampaignExecutor> logger,
     AresVariableManager variableManager,
-    StateLoggerManager stateLoggerManager)
+    StateLoggerManager stateLoggerManager,
+    ISystemSettingsManager settingsManager,
+    IExecutionSafetyManager executionSafetyManager)
   {
     _analyzerRepo = analyzerRepo;
     _analysisRepo = analysisRepo;
@@ -59,6 +65,8 @@ public class CampaignExecutor : ICampaignExecutor
     _logger = logger;
     Template = template;
     _notifier = notifier;
+    _settingsManager = settingsManager;
+    _executionSafetyManager = executionSafetyManager;
 
     Status = new CampaignExecutionStatus
     {
@@ -192,12 +200,198 @@ public class CampaignExecutor : ICampaignExecutor
 
   private async Task<bool> ExecuteExperimentLoop(string campaignPath, List<Analysis> analyses, List<ExperimentExecutionSummary> experimentSummaries, ExecutionControlToken token, bool executionSuccess, ExperimentExecutionSummary startupSummary)
   {
+    var campaignFailedTitle = "Campaign Execution Failed!";
+    var currentState = ExecutionState.InitializeExperiment;
+    executionSuccess = true;
+    var experimentCount = 0;
+    var currentExperimentFolder = "";
+    var currentExperimentPath = "";
+
+    ExperimentExecutorResult currentExecutorResult = new();
+    ExperimentExecutionSummary? currentSummary = null;
+    ExperimentTemplate? currentExperimentTemplate = null;
+
+    while(!ShouldStop() && !token.IsCancelled && executionSuccess != false)
+    {
+      switch(currentState)
+      {
+        case ExecutionState.InitializeExperiment:
+          experimentCount++;
+          currentExperimentFolder = $"Experiment_{++experimentCount}";
+          currentExperimentPath = CampaignOutputHelper.CreateExperimentSubFolder(campaignPath, currentExperimentFolder);
+          currentExperimentTemplate = Template.ExperimentTemplate.CloneWithNewIds();
+          currentExperimentTemplate.Name = Template.Name;
+
+          AresEnvironment.AresEnvironment.SetInternalVariable(InternalVariableType.CurrentExperimentNumber, experimentCount.ToString());
+          _logger.LogInformation("ARES is initializing experiment {exp_number}, and has created the associated directories.", experimentCount);
+          currentState = ExecutionState.Planning;
+          break;
+
[…]
