+++ Ares.Core/Execution/Safety/ExecutionSafetyManager.cs
@@ -1,5 +1,4 @@
 ﻿using Ares.Core.Device.Repos;
-using Microsoft.Build.Framework;
 using Microsoft.Extensions.Logging;
 
 namespace Ares.Core.Execution.Safety;
