### 2026-05-14 8460a5e
Some cleanup and additional UI for general settings
