### 2026-05-14 10f41d3
Added the wide display mode. Fixed the interval filter.
