### 2026-05-14 8cc1d74
Most error handling cases are now working. Only outlier left is the "prompt user" case which will need some additional UI work done.
