+++ Ares.Core/Execution/Executors/CampaignExecutor.cs
@@ -1,6 +1,4 @@
-﻿using System.Reactive.Linq;
-using System.Reactive.Subjects;
-using Ares.Core.Analyzing;
+﻿using Ares.Core.Analyzing;
 using Ares.Core.AresEnvironment;
 using Ares.Core.Device.State.Logging;
 using Ares.Core.Exceptions;
@@ -18,6 +16,8 @@ using Ares.Datamodel.Analyzing;
 using Ares.Datamodel.Templates;
 using Google.Protobuf.WellKnownTypes;
 using Microsoft.Extensions.Logging;
+using System.Reactive.Linq;
+using System.Reactive.Subjects;
 
 namespace Ares.Core.Execution.Executors;
 
@@ -37,6 +37,9 @@ public class CampaignExecutor : ICampaignExecutor
   readonly IAnalyzerRepo _analyzerRepo;
   readonly ISystemSettingsManager _settingsManager;
   readonly IExecutionSafetyManager _executionSafetyManager;
+
+  private readonly string _campaignFailedTitle = "Campaign Execution Failed!";
+  private AresGeneralSettingsConfig? _generalSettingsConfig;
 
   internal CampaignExecutor(ICommandComposer<ExperimentTemplate, ExperimentExecutor> experimentComposer,
     IPlanningHelper planningHelper,
@@ -116,8 +119,6 @@ public class CampaignExecutor : ICampaignExecutor
       executionSuccess = await ExecuteExperimentLoop(campaignPath, analyses, experimentSummaries, token, executionSuccess, startupResult.Summary);
 
       _logger.LogDebug("The campaign loop is now finished. Moving onto the closeout script.");
-
-
       var closeoutResult = await ExecuteCloseout(analyses, experimentSummaries, token, executionSuccess);
       closeoutSummary = closeoutResult.Summary;
 
@@ -156,6 +157,7 @@ public class CampaignExecutor : ICampaignExecutor
     _logger.LogInformation("Analyzer selected {AnalyzerName}", analyzer?.Name ?? "NO ANALYZER");
     await CampaignOutputHelper.OutputVersionFile(campaignPath, Template, analyzer);
 
+    _generalSettingsConfig = await _settingsManager.GetAresGeneralSettings();
     return campaignPath;
   }
 
@@ -200,18 +202,20 @@ public class CampaignExecutor : ICampaignExecutor
 
   private async Task<bool> ExecuteExperimentLoop(string campaignPath, List<Analysis> analyses, List<ExperimentExecutionSummary> experimentSummaries, ExecutionControlToken token, bool executionSuccess, ExperimentExecutionSummary startupSummary)
   {
-    var campaignFailedTitle = "Campaign Execution Failed!";
     var currentState = ExecutionState.InitializeExperiment;
     executionSuccess = true;
     var experimentCount = 0;
     var currentExperimentFolder = "";
     var currentExperimentPath = "";
+    var failedExperimentRetryCount = 0;
+    var failedExperimentRetryLimit = _generalSettingsConfig?.ExperimentRetryLimit ?? 1;
+    var experimentRetryCooldown = _generalSettingsConfig?.RetryCooldown ?? 0;
 
     ExperimentExecutorResult currentExecutorResult = new();
     ExperimentExecutionSummary? currentSummary = null;
     ExperimentTemplate? currentExperimentTemplate = null;
 
-    while(!ShouldStop() && !token.IsCancelled && executionSuccess != false)
+    while(!ShouldStop() && !token.IsCancelled && executionSuccess)
     {
       switch(currentState)
       {
@@ -232,7 +236,7 @@ public class CampaignExecutor : ICampaignExecutor
           {
[…]
