### 2026-05-15 1f417d7
Basic implementation of the user decision request for when commands fail in an experiment
