+++ Ares.Core/Execution/Executors/CampaignExecutor.cs
@@ -40,6 +40,11 @@ public class CampaignExecutor : ICampaignExecutor
 
   private readonly string _campaignFailedTitle = "Campaign Execution Failed!";
   private AresGeneralSettingsConfig? _generalSettingsConfig;
+  private ExperimentExecutorResult _currentExecutorResult = new();
+  private ExperimentExecutionSummary? _currentSummary = null;
+  private ExperimentTemplate? _currentExperimentTemplate = null;
+  private int _experimentCount = 0;
+  private TaskCompletionSource<ErrorHandling>? _userDecisionSource;
 
   internal CampaignExecutor(ICommandComposer<ExperimentTemplate, ExperimentExecutor> experimentComposer,
     IPlanningHelper planningHelper,
@@ -158,6 +163,7 @@ public class CampaignExecutor : ICampaignExecutor
     await CampaignOutputHelper.OutputVersionFile(campaignPath, Template, analyzer);
 
     _generalSettingsConfig = await _settingsManager.GetAresGeneralSettings();
+    _experimentCount = 0;
     return campaignPath;
   }
 
@@ -204,35 +210,30 @@ public class CampaignExecutor : ICampaignExecutor
   {
     var currentState = ExecutionState.InitializeExperiment;
     executionSuccess = true;
-    var experimentCount = 0;
     var currentExperimentFolder = "";
     var currentExperimentPath = "";
     var failedExperimentRetryCount = 0;
     var failedExperimentRetryLimit = _generalSettingsConfig?.ExperimentRetryLimit ?? 1;
     var experimentRetryCooldown = _generalSettingsConfig?.RetryCooldown ?? 0;
 
-    ExperimentExecutorResult currentExecutorResult = new();
-    ExperimentExecutionSummary? currentSummary = null;
-    ExperimentTemplate? currentExperimentTemplate = null;
-
     while(!ShouldStop() && !token.IsCancelled && executionSuccess)
     {
       switch(currentState)
       {
         case ExecutionState.InitializeExperiment:
-          experimentCount++;
-          currentExperimentFolder = $"Experiment_{++experimentCount}";
+          _experimentCount++;
+          currentExperimentFolder = $"Experiment_{++_experimentCount}";
           currentExperimentPath = CampaignOutputHelper.CreateExperimentSubFolder(campaignPath, currentExperimentFolder);
-          currentExperimentTemplate = Template.ExperimentTemplate.CloneWithNewIds();
-          currentExperimentTemplate.Name = Template.Name;
+          _currentExperimentTemplate = Template.ExperimentTemplate.CloneWithNewIds();
+          _currentExperimentTemplate.Name = Template.Name;
 
-          AresEnvironment.AresEnvironment.SetInternalVariable(InternalVariableType.CurrentExperimentNumber, experimentCount.ToString());
-          _logger.LogInformation("ARES is initializing experiment {exp_number}, and has created the associated directories.", experimentCount);
+          AresEnvironment.AresEnvironment.SetInternalVariable(InternalVariableType.CurrentExperimentNumber, _experimentCount.ToString());
+          _logger.LogInformation("ARES is initializing experiment {exp_number}, and has created the associated directories.", _experimentCount);
           currentState = ExecutionState.Planning;
           break;
 
         case ExecutionState.GenerateExecutor:
-          if(currentExperimentTemplate is null)
+          if(_currentExperimentTemplate is null)
           {
             var nullTemplateMsg = "ARES attempted to generate an experiment executor, but failed due to a null experiment template! The experiment will be terminated.";
             _logger.LogError(nullTemplateMsg);
@@ -243,8 +244,8 @@ public class CampaignExecutor : ICampaignExecutor
 
           try
           {
-            currentExecutorResult = new ExperimentExecutorResult();
[…]
