### 2026-05-17 da25c17
Made the invalid characters platform agnostic.
