### 2026-05-18 a792cb8
Updates to solidify execution logic in the new "state based" execution method
