+++ UI/Features/Execution/ExecutionViewModel.cs
@@ -228,20 +228,6 @@ public partial class ExecutionViewModel : ReactiveObject, INotifyPropertyChanged
     return Task.CompletedTask;
   }
 
-  public Task RequestUserDecision()
-  {
-    var notification = new AresNotification();
-    notification.NotificationSeverity = Severity.Warning;
-    notification.Title = "User Decision Required to Proceed";
-    notification.Message = $"ARES has encountered an error and requires your decision on how to proceed. Please check the execution screen.";
-    notification.Timestamp = DateTime.UtcNow.ToTimestamp();
-    notification.Loiter = true;
-
-    _notificationService.PushNotification(notification);
-
-    return Task.CompletedTask;
-  }
-
   public async Task AddTag()
   {
     if(NewTagName is not null && AvailableTags.Any(t => t.TagName == NewTagName))
@@ -470,9 +456,6 @@ public partial class ExecutionViewModel : ReactiveObject, INotifyPropertyChanged
     //TODO: FIX THIS!!!
     if(CampaignExecutionState == ExecutionState.AwaitingUser)
       _ = RequestUserConfirmation();
-
-    if(CampaignExecutionState == ExecutionState.WaitingForUserDecision)
-      _ = RequestUserDecision();
   }
 
   public Task UpdateDeviceChartA()
