+++ Ares.Core/Execution/Executors/CampaignExecutor.cs
@@ -271,20 +271,16 @@ public class CampaignExecutor : ICampaignExecutor
             break;
           }
 
-          var experimentSummary = await ExecuteTemplate(_currentExecutorResult.ExperimentExecutor, token);
-          experimentSummary.ResultOutputPath = currentExperimentPath;
+          _currentSummary = await ExecuteTemplate(_currentExecutorResult.ExperimentExecutor, token);
+          _currentSummary.ResultOutputPath = currentExperimentPath;
 
-          var failedCommandSummary = experimentSummary.StepSummaries.SelectMany(s => s.CommandSummaries.Where(c => !c.Result.Success)).FirstOrDefault();
+          var failedCommandSummary = _currentSummary.StepSummaries.SelectMany(s => s.CommandSummaries.Where(c => !c.Result.Success)).FirstOrDefault();
 
           if(failedCommandSummary is not null)
             currentState = await HandleError(failedCommandSummary);
 
           else
-          {
-            await PostExperimentExecution(experimentSummary);
-            experimentSummaries.Add(experimentSummary);
-            currentState = ExecutionState.InitializeExperiment;
-          }
+            currentState = ExecutionState.Analyzing;
 
           break;
 
@@ -304,6 +300,10 @@ public class CampaignExecutor : ICampaignExecutor
             break;
           }
 
+          Status.State = ExecutionState.Replanning;
+          _executionStatusSubject.OnNext(Status);
+          _executionReporter.Report(Status);
+
           var replanMsg = "An ARES experiment failed, but based on your settings ARES will attempt to re-plan the experiment and run it again.";
           _logger.LogInformation(replanMsg);
           await _notifier.Notify("ARES is Replanning Experiment", replanMsg, NotificationSeverityEnum.Info);
@@ -317,11 +317,17 @@ public class CampaignExecutor : ICampaignExecutor
           }
 
           else
-            currentState = ExecutionState.Running;
+            currentState = ExecutionState.GenerateExecutor;
 
           break;
 
         case ExecutionState.Retrying:
+          Status.State = ExecutionState.Retrying;
+          _executionStatusSubject.OnNext(Status);
+          _executionReporter.Report(Status);
+
+          ResetCurrentExperimentStatus();
+
           failedExperimentRetryCount++;
 
           if(failedExperimentRetryCount > failedExperimentRetryLimit)
@@ -340,7 +346,7 @@ public class CampaignExecutor : ICampaignExecutor
           if(experimentRetryCooldown != 0)
             await Task.Delay(TimeSpan.FromSeconds(experimentRetryCooldown));
 
-          currentState = ExecutionState.Running;
+          currentState = ExecutionState.GenerateExecutor;
           break;
 
         case ExecutionState.EnteringSafeMode:
@@ -358,7 +364,11 @@ public class CampaignExecutor : ICampaignExecutor
               var success = await PlanExperiment(analyses, _currentExperimentTemplate, experimentSummaries, token);
 
               if(!success)
+              {
                 _currentExecutorResult?.ErrorString = "Failed to plan! Experiment will be terminated!";
+                currentState = ExecutionState.Failed;
+                break;
+              }
             }
 
             else
@@ -399,9 +409,17 @@ public class CampaignExecutor : ICampaignExecutor
             
             if(!result.Continue) 
               break;
+
+            await PostExperimentExecution(_currentSummary);
+            experimentSummaries.Add(_currentSummary);
+            currentState = ExecutionState.InitializeExperiment;
           }
+
           else
+          {
[…]
