### 2026-05-18 d9b1fc1
Updates to support durations in settings values
