### 2026-05-18 612daa3
Updated entity configurations
