### 2026-05-18 3f1c39f
Updated system settings to handle new duration type gracefully
