### 2026-05-21 48eeb67
Some work towards finalizing SiLA implementation, but properties seems to be broken. Investigating whether this is our doing or potentially an issue with the SiLA library implementation
