+++ Ares.Core/Device/Sila/SilaDevice.cs
@@ -3,10 +3,12 @@ using Ares.Datamodel.Device;
 using Ares.Datamodel.Extensions;
 using Ares.Datamodel.Factories;
 using Ares.Device;
+using System.Collections.Concurrent;
 using System.Reactive.Linq;
 using System.Reactive.Subjects;
 using Tecan.Sila2;
 using Tecan.Sila2.DynamicClient;
+using Grpc.Core;
 
 namespace Ares.Core.Device.Sila;
 
@@ -18,6 +20,8 @@ public sealed class SilaDevice : AresDevice, IAsyncDisposable
   private CancellationTokenSource _stateStreamCts = new();
   private DevicePollingSettings _pollingSettings = new();
   private ServerData _serverData;
+  private readonly ConcurrentDictionary<string, AresValue> _currentPropertyValues = new();
+  private Task? _monitoringTask;
 
   public SilaDevice(ServerData serverData, DeviceConnectionInfo connectionInfo, SilaClient client) : base(connectionInfo)
   {
@@ -28,16 +32,204 @@ public sealed class SilaDevice : AresDevice, IAsyncDisposable
     Description = serverData.Info.Description;
     Version = serverData.Info.Version;
     Type = serverData.Info.Type;
+    Address = serverData.Address;
   }
 
   public override IObservable<AresStruct> StateStream => _stateSubject.AsObservable();
 
-  public override Task<bool> Activate(CancellationToken ct)
+  //public AresStructSchema StateSchema { get; private set; } = new();
+
+  public override async Task<bool> Activate(CancellationToken ct)
   {
     DeviceFeatures = _serverData.Features.ToArray();
+    BuildStateSchema();
 
-    Status = new DeviceOperationalStatus() { OperationalState = OperationalState.Active, Message = $"Connected to SiLA Device {Name}!" };
-    return Task.FromResult(true);
+    Status = new DeviceOperationalStatus() 
+    { 
+      OperationalState = OperationalState.Active, 
+      Message = $"Connected to SiLA Device {Name}!" 
+    };
+
+    await StartPropertyMonitoring();
+    return true;
+  }
+
+  private void BuildStateSchema()
+  {
+    var schema = new AresStructSchema();
+    foreach(var feature in DeviceFeatures)
+    {
+      foreach(var property in (feature.Items ?? []).OfType<FeatureProperty>())
+      {
+        var key = $"{feature.Identifier}.{property.Identifier}";
+        var fieldSchema = property.DataType is not null
+          ? SilaDataConverter.ToAresValueSchema(property.DataType)
+          : AresSchemaBuilder.Entry(AresDataType.Any).Build();
+
+        if(!string.IsNullOrWhiteSpace(property.Description))
+          fieldSchema.Description = property.Description;
+
+        schema.Fields[key] = fieldSchema;
+      }
+    }
+    StateSchema = schema;
+  }
+
+  private async Task StartPropertyMonitoring()
+  {
+    if(_monitoringTask is not null)
+      return;
+
+    _stateStreamCts = new CancellationTokenSource();
+    var token = _stateStreamCts.Token;
+
+    var monitorTasks = new List<Task>();
+
+    foreach(var feature in DeviceFeatures)
+    {
+      if(feature.Identifier.Contains("SiLAService")) 
+        continue;
+
+      var properties = (feature.Items ?? []).OfType<FeatureProperty>();
+      var context = new FeatureContext(feature, _serverData, _silaClient.ExecutionManager);
+
+      foreach(var property in properties)
+      {
+        if(property.Observable == FeaturePropertyObservable.Yes)
+          monitorTasks.Add(MonitorObservableProperty(property, context, token));
+
[…]
