### 2026-05-21 032306e
Changed from an experiment output map for each command to command just outputting a variable.
