+++ UI/Features/CampaignEdit/ViewModels/CommandDesignerViewModel.cs
@@ -61,9 +61,13 @@ public partial class CommandDesignerViewModel : ReactiveObject
   public string? MetadataDeviceName { get; private set; }
   public string? TemplateCommandName => CommandTemplate.Metadata?.Name;
 
-  public bool TemplateOutputProvider => CommandTemplate.UserOutputKeyMap.Any();
+  public bool TemplateOutputProvider => CommandTemplate.HasOutputVarName;
 
   public bool OutputProvider { get; set; }
+
+  public bool HasOutputMetadata => CommandMetadata?.OutputMetadata?.DataSchema is not null;
+
+  public string? OutputVariableName { get; set; }
 
   public IEnumerable<Parameter> Arguments => CommandTemplate.Parameters;
 
@@ -77,8 +81,6 @@ public partial class CommandDesignerViewModel : ReactiveObject
       InitMetadata(value);
     }
   }
-
-  public UserOutputSelection[] OutputKeyMap { get; private set; } = [];
 
   public MetadataPickerViewModel? MetadataPickerViewModel { get; set; }
 
@@ -97,17 +99,10 @@ public partial class CommandDesignerViewModel : ReactiveObject
       
 
     CommandTemplate.Index = Index;
-    CommandTemplate.UserOutputKeyMap.Clear();
-    if(OutputProvider)
+    CommandTemplate.ClearOutputVarName();
+    if(OutputProvider && HasOutputMetadata && !string.IsNullOrWhiteSpace(OutputVariableName))
     {
-      foreach(var selection in OutputKeyMap)
-      {
-        CommandTemplate.UserOutputKeyMap[selection.DeviceOutputName] = selection.CustomName;
-      }
-    }
-    else
-    {
-      CommandTemplate.UserOutputKeyMap.Clear();
+      CommandTemplate.OutputVarName = OutputVariableName.Trim();
     }
 
     return CommandTemplate;
@@ -120,18 +115,8 @@ public partial class CommandDesignerViewModel : ReactiveObject
     ArgumentDesigners = [.. existingParamDesigners];
     MetadataPickerViewModel = _metadataPickerFactory.Create(existingTemplate.Metadata);
 
-    foreach(var kvp in existingTemplate.UserOutputKeyMap)
-    {
-      var existingValue = OutputKeyMap.FirstOrDefault(keyValue => keyValue.DeviceOutputName == kvp.Key);
-      if(existingValue is null)
-      {
-        continue;
-      }
-
-      existingValue.CustomName = kvp.Value;
-    }
-
-    OutputProvider = existingTemplate.UserOutputKeyMap.Any();
+    OutputProvider = existingTemplate.HasOutputVarName;
+    OutputVariableName = existingTemplate.HasOutputVarName ? existingTemplate.OutputVarName : null;
 
     // Revisit this once we have some sort of caching on the UI end.
     // that way we don't have to bother the service every time
@@ -156,51 +141,10 @@ public partial class CommandDesignerViewModel : ReactiveObject
         .ToArray() ?? [];
 
     ArgumentDesigners = newArgumentDesigners;
-    var outputSchema = existingMetadata?.OutputMetadata?.DataSchema;
-
-    if(outputSchema is null)
+    if(existingMetadata?.OutputMetadata?.DataSchema is null)
     {
-      OutputKeyMap = [];
-    }
-    else
-    {
-      // Normalize the schema
-      var availableFields = new Dictionary<string, AresValueSchema>();
-
-      if(outputSchema.Type == AresDataType.Struct && outputSchema.StructSchema is not null)
-      {
-        foreach(var field in outputSchema.StructSchema.Fields)
-        {
-          availableFields[field.Key] = field.Value;
-        }
-      }
[…]
