+++ UI/Features/CampaignEdit/ViewModels/ExperimentDesignerViewModel.cs
@@ -46,9 +46,9 @@ public partial class ExperimentDesignerViewModel : ReactiveObject
   {
     Name = existingTemplate.Name;
     StepDesigners = existingTemplate.StepTemplates.Select(template => _stepDesignerFactory.Create(template)).OrderBy(model => model.Index).ToList();
-    if(existingTemplate.StepTemplates.Select(step => step.CommandTemplates.Select(cmd => cmd.UserOutputKeyMap)).Any())
+    if(existingTemplate.StepTemplates.SelectMany(step => step.CommandTemplates).Any(cmd => cmd.HasOutputVarName))
     {
-      var commandDesigners = StepDesigners.SelectMany(model => model.CommandDesigners).Where(model => model.CommandTemplate.UserOutputKeyMap.Any());
+      var commandDesigners = StepDesigners.SelectMany(model => model.CommandDesigners).Where(model => model.CommandTemplate.HasOutputVarName);
       foreach(var designer in commandDesigners)
       {
         designer.OutputProvider = true;
