+++ UI/Domain/Experiments/ExperimentTemplateExtensions.cs
@@ -18,5 +18,5 @@ internal static class ExperimentTemplateExtensions
   public static CommandTemplate[] GetAllOutputCommands(this ExperimentTemplate template)
     => template.StepTemplates
     .SelectMany(step => step.CommandTemplates)
-    .Where(command => command.UserOutputKeyMap.Any()).ToArray();
+    .Where(command => command.HasOutputVarName).ToArray();
 }

[…]
