### 2026-05-28 fab6a90
Added support for adding output variables to commands and then consuming them in subsequent commands.
