+++ Ares.Core/Execution/Executors/StepExecutor.cs
@@ -7,7 +7,7 @@ namespace Ares.Core.Execution.Executors;
 
 public abstract class StepExecutor : IExecutor<StepExecutionSummary, StepExecutionStatus>
 {
-  public StepExecutor(StepTemplate template, IExecutor<CommandExecutionSummary, CommandExecutionStatus>[] commandExecutors)
+  public StepExecutor(StepTemplate template, CommandExecutor[] commandExecutors)
   {
     Template = template;
     CommandExecutors = commandExecutors;
@@ -33,9 +33,11 @@ public abstract class StepExecutor : IExecutor<StepExecutionSummary, StepExecuti
     ExperimentStatusObservable = commandExecutionObservation;
   }
 
-  public IExecutor<CommandExecutionSummary, CommandExecutionStatus>[] CommandExecutors { get; }
+  public CommandExecutor[] CommandExecutors { get; }
   protected StepTemplate Template { get; }
   public IObservable<StepExecutionStatus> ExperimentStatusObservable { get; }
   public StepExecutionStatus Status { get; }
   public abstract Task<StepExecutionSummary> Execute(ExecutionControlToken token);
+  public virtual Task<StepExecutionSummary> Execute(ExecutionControlToken token, IReadOnlyDictionary<string, AresValue> variableScope)
+    => Execute(token);
 }
