+++ Ares.Core/Execution/Executors/Composers/StepComposer.cs
@@ -2,6 +2,7 @@
 using Ares.Core.Notifications;
 using Ares.Datamodel;
 using Ares.Datamodel.Device;
+using Ares.Datamodel.Extensions;
 using Ares.Datamodel.Templates;
 
 namespace Ares.Core.Execution.Executors.Composers;
@@ -38,7 +39,7 @@ public class StepComposer : ICommandComposer<StepTemplate, StepExecutor>
               Func<CancellationToken, Task<CommandResult>> internalAction = async (ct)
                 => await device.ExecuteCommand(
                   commandTemplate.Metadata.Name,
-                  commandTemplate.Parameters.Select(p => new DeviceCommandArgument() { ArgName = p.Metadata.Name, ArgValue = p.Value }).ToList(),
+                  commandTemplate.Parameters.Select(p => new DeviceCommandArgument() { ArgName = p.Metadata.Name, ArgValue = p.GetValue() }).ToList(),
                   ct);
 
               return new CommandExecutor(internalAction, commandTemplate, _notifier);
