### 2026-05-28 6d6a71f
Improved the Parameter datamodel object. Improved the parameter display.
