### 2026-05-29 10a1002
Changed the name of the parameter partial class in the dm, so updating it here.
