+++ AresService.Migrations.Sqlite/Migrations/AresDbContextModelSnapshot.cs
@@ -1,4 +1,4 @@
-﻿// <auto-generated />
+// <auto-generated />
 using System;
 using AresService.Data;
 using Microsoft.EntityFrameworkCore;
@@ -1381,7 +1381,7 @@ namespace AresService.Migrations.Sqlite.Migrations
                         .HasColumnType("TEXT")
                         .HasDefaultValueSql("DATETIME('now')");
 
-                    b.Property<string>("SourcePersistence")
+                    b.Property<string>("SourceJson")
                         .HasColumnType("TEXT")
                         .HasColumnName("Source");
 
