+++ AresService.Migrations.SqlServer/Migrations/AresDbContextModelSnapshot.cs
@@ -1,4 +1,4 @@
-﻿// <auto-generated />
+// <auto-generated />
 using System;
 using AresService.Data;
 using Microsoft.EntityFrameworkCore;
@@ -1397,7 +1397,7 @@ namespace AresService.Migrations.SqlServer.Migrations
                         .HasColumnType("datetime2")
                         .HasDefaultValueSql("getdate()");
 
-                    b.Property<string>("SourcePersistence")
+                    b.Property<string>("SourceJson")
                         .HasColumnType("nvarchar(max)")
                         .HasColumnName("Source");
 
