+++ AresService.Migrations.Postgres/Migrations/20260528230904_CommandVariableSupport_AresDbContext.Designer.cs
@@ -1,4 +1,4 @@
-﻿// <auto-generated />
+// <auto-generated />
 using System;
 using AresService.Data;
 using Microsoft.EntityFrameworkCore;
@@ -1389,7 +1389,7 @@ namespace AresService.Migrations.Postgres.Migrations.AresDb
                         .HasColumnType("timestamp with time zone")
                         .HasDefaultValueSql("NOW()");
 
-                    b.Property<string>("SourcePersistence")
+                    b.Property<string>("SourceJson")
                         .HasColumnType("jsonb")
                         .HasColumnName("Source");
 
