+++ AresService.Migrations.Sqlite/Migrations/20260528230842_CommandVariableSupport_AresDbContext.Designer.cs
@@ -1,4 +1,4 @@
-﻿// <auto-generated />
+// <auto-generated />
 using System;
 using AresService.Data;
 using Microsoft.EntityFrameworkCore;
@@ -1384,7 +1384,7 @@ namespace AresService.Migrations.Sqlite.Migrations.AresDb
                         .HasColumnType("TEXT")
                         .HasDefaultValueSql("DATETIME('now')");
 
-                    b.Property<string>("SourcePersistence")
+                    b.Property<string>("SourceJson")
                         .HasColumnType("TEXT")
                         .HasColumnName("Source");
 
