+++ AresService.Migrations.SqlServer/Migrations/20260528230925_CommandVariableSupport_AresDbContext.Designer.cs
@@ -1,4 +1,4 @@
-﻿// <auto-generated />
+// <auto-generated />
 using System;
 using AresService.Data;
 using Microsoft.EntityFrameworkCore;
@@ -1400,7 +1400,7 @@ namespace AresService.Migrations.SqlServer.Migrations.AresDb
                         .HasColumnType("datetime2")
                         .HasDefaultValueSql("getdate()");
 
-                    b.Property<string>("SourcePersistence")
+                    b.Property<string>("SourceJson")
                         .HasColumnType("nvarchar(max)")
                         .HasColumnName("Source");
 
