+++ AresService.Migrations.SqlServer/Migrations/AresDbContextModelSnapshot.cs
@@ -1,4 +1,4 @@
-// <auto-generated />
+﻿// <auto-generated />
 using System;
 using AresService.Data;
 using Microsoft.EntityFrameworkCore;
@@ -385,6 +385,9 @@ namespace AresService.Migrations.SqlServer.Migrations
                         .HasColumnType("datetime2")
                         .HasDefaultValueSql("getdate()");
 
+                    b.Property<int>("StatusCode")
+                        .HasColumnType("int");
+
                     b.Property<Guid?>("StepExecutionSummaryUniqueId")
                         .HasColumnType("uniqueidentifier");
 
@@ -428,6 +431,9 @@ namespace AresService.Migrations.SqlServer.Migrations
 
                     b.Property<string>("Result")
                         .HasColumnType("nvarchar(max)");
+
+                    b.Property<int>("StatusCode")
+                        .HasColumnType("int");
 
                     b.Property<bool>("Success")
                         .HasColumnType("bit");
