### 2026-05-29 d32714d
Data explorer (#155)
* Added initial support for the int, float, and timestamp

* Ares script updates for the new data types.

* Added a dataset provider for device states based on an input filter.

* Added campaign dataset generator.
Changed the device state getter to return arrays.

* Added a basic data explorer for device states.

* Device dataset generator will not expand sub structs

* Added a long dataset viewer.

* Page size options to the dataset viewer

* Css tweaks. Removed unused css

* Prevent UI from freezing when applying a large dataset

* Horizontal table scroll. Fixed bug with
[…]
