+++ UI/Features/DeviceStateLogging/DeviceStateFilterViewModel.cs
@@ -1,91 +1,0 @@
-﻿using Ares.Datamodel;
-using Ares.Datamodel.Device;
-using Ares.Services;
-using Ares.Services.Device;
-using Ares.Core.Grpc.Services;
-using Google.Protobuf.WellKnownTypes;
-using ReactiveUI;
-using ReactiveUI.SourceGenerators;
-
-namespace UI.Features.DeviceStateLogging;
-
-public partial class DeviceStateFilterViewModel : ReactiveObject
-{
-  private readonly DevicesService _deviceClient;
-  readonly AutomationService _automationClient;
-
-  public DeviceStateFilterViewModel(DevicesService devicesClient, AutomationService automationClient)
-  {
-    _automationClient = automationClient;
-    _deviceClient = devicesClient;
-    _ = _automationClient.GetAvailableCampaignExecutionSummaries(new Empty(), null)
-      .ContinueWith(task => { if(task.IsCompletedSuccessfully) UpdateCampaigns(task.Result); });
-    _ = _deviceClient.GetAllAvailableDevices(new Empty(), null)
-      .ContinueWith(task => { if(task.IsCompletedSuccessfully) UpdateDevices(task.Result); });
-
-    var currentTime = DateTime.Now;
-    // probably don't need millisecond precision
-    var truncatedTime = new DateTime(currentTime.Ticks - (currentTime.Ticks % TimeSpan.TicksPerMinute));
-    StartTime = truncatedTime - TimeSpan.FromHours(1);
-    EndTime = truncatedTime;
-  }
-
-  private void UpdateCampaigns(AvailableCampaignExecutionSummariesResponse response)
-  {
-    Campaigns = response.AvailableCampaignSummaries.Select(result => new CampaignExecutionSummaryMetadata { SummaryId = result.SummaryId, CampaignName = $"result.CampaignName-{result.CompletionTime}", CompletionTime = result.CompletionTime });
-  }
-
-  private void UpdateDevices(AvailableDevicesResponse response)
-  {
-    AvailableDevices = response.Devices.ToList();
-  }
-
-  [Reactive]
-  public partial IEnumerable<AresDeviceDescription>? AvailableDevices { get; private set; }
-
-  public IEnumerable<AresDeviceDescription>? SelectedDevices { get; set; }
-
-  [Reactive]
-  public partial IEnumerable<CampaignExecutionSummaryMetadata>? Campaigns { get; private set; }
-
-  public async Task UpdateExperiments(string? campaignResultId)
-  {
-    Experiments = null;
-    var campaignResult = await _automationClient.GetCampaignSummary(
-      new CampaignExecutionSummaryRequest { SummaryId = campaignResultId }, null);
-    Experiments = campaignResult.ExperimentSummaries;
-  }
-
-  [Reactive]
-  public partial IEnumerable<ExperimentExecutionSummary>? Experiments { get; private set; }
-
-  public string? SelectedExperimentId { get; set; }
-
-  public DateTime StartTime { get; set; }
-
-  public DateTime EndTime { get; set; }
-
-  public TimeSpan Interval { get; set; }
-
-  public bool UseStartTime { get; set; }
-  public bool UseEndTime { get; set; }
-  public bool UseExperiment { get; set; }
-
-  public DeviceStateRequestFilter GetStateRequestFilter()
-  {
-    var request = new DeviceStateRequestFilter
-    {
-      Start = UseStartTime ? StartTime.ToUniversalTime().ToTimestamp() : default,
-      End = UseEndTime ? EndTime.ToUniversalTime().ToTimestamp() : default,
-      CompletedExperimentId = UseExperiment ? SelectedExperimentId : string.Empty,
[…]

[…]
