### 2026-06-01 7ddad89
Working towards support for batch planning
