+++ Ares.Core/Execution/Executors/ICampaignExecutor.cs
@@ -6,8 +6,8 @@ namespace Ares.Core.Execution.Executors;
 public interface ICampaignExecutor : IExecutor<CampaignExecutionSummary, CampaignExecutionStatus>
 {
   IList<IStopCondition> StopConditions { get; }
-  double ReplanRate { get; set; }
+  int ReplicateRate { get; set; }
+  int BatchPlanningSize { get; set; }
   void UpdateExecutionNotes(string executionNotes);
-
   void UpdateCampaignTags(List<AresCampaignTag> campaignTags);
 }
