+++ Ares.Core.Grpc/Services/AutomationService.cs
@@ -561,7 +561,7 @@ public class AutomationService : AresAutomation.AresAutomationBase
       return [];
 
     var usedPlanners = _activeCampaignTemplateStore.CampaignTemplate.ExperimentTemplate.GetAllPlannedParameters()
-      .Select(p => p.GetPlanningMetadata()?.PlannerName)
+      .Select(p => p.GetPlanningMetadata()?.PlannerName ?? "")
       .Where(name => !string.IsNullOrWhiteSpace(name))
       .Select(_plannerServiceRepo.GetPlannerByName)
       .Where(p => p is not null)
