### 2026-06-03 1bb6a7f
Updated to use datamodel v0.25.0 and utilized the new helper methods to make our analysis mapping process more robust.
