+++ Ares.Core/Execution/Executors/CommandExecutor.cs
@@ -63,8 +63,8 @@ public class CommandExecutor : IExecutor<CommandExecutionSummary, CommandExecuti
     var timeStarted = DateTime.UtcNow;
     var execInfo = new ExecutionInfo { TimeStarted = DateTime.UtcNow.ToTimestamp() };
     var variableResolutionError = CommandVariableResolver.ResolveParameters(Template.Parameters, variableScope);
-    var result = variableResolutionError is null
-      ? await InternalExecute(token.CancellationToken)
+    var result = variableResolutionError is null 
+      ? await InternalExecute(token.CancellationToken) 
       : new CommandResult { Success = false, Error = variableResolutionError };
 
     execInfo.TimeFinished = DateTime.UtcNow.ToTimestamp();
