### 2026-06-03 427f971
Fixed missing planning metadata assignment
