### 2026-06-03 5a9eebc
Removed completed TODO comment
