### 2026-06-03 12a31b5
Variable input (#160)
* Changed from an experiment output map for each command to command just outputting a variable.

* Added support for adding output variables to commands and then consuming them in subsequent commands.

* Changed the way datamodel presents the parameter and its source.

* Improved the Parameter datamodel object. Improved the parameter display.

* Migrations

* Changed the name of the parameter partial class in the dm, so updating it here.

* Done redid them there migrashuns as they was broken.

* DM Update

* Fixed build errors

* Fixed compiler warnings. Package updates.

[…]
