+++ UI/Domain/Experiments/ExperimentTemplateExtensions.cs
@@ -1,3 +1,4 @@
+using Ares.Datamodel.Extensions;
 using Ares.Datamodel.Templates;
 
 namespace UI.Domain.Experiments;
@@ -10,13 +11,13 @@ internal static class ExperimentTemplateExtensions
       .SelectMany(commandTemplate => commandTemplate.Parameters);
 
   public static IEnumerable<Parameter> GetAllPlannedParameters(this ExperimentTemplate template)
-    => template.GetAllParameters().Where(parameter => parameter.Planned);
+    => template.GetAllParameters().Where(parameter => parameter.IsPlanned());
 
   public static bool IsResolved(this ExperimentTemplate template)
-    => template.GetAllParameters().All(parameter => parameter.Value is not null);
+    => template.GetAllParameters().All(parameter => parameter.GetValue() is not null);
 
   public static CommandTemplate[] GetAllOutputCommands(this ExperimentTemplate template)
     => template.StepTemplates
     .SelectMany(step => step.CommandTemplates)
-    .Where(command => command.UserOutputKeyMap.Any()).ToArray();
+    .Where(command => command.HasOutputVarName).ToArray();
 }
