+++ AresService.Migrations.Sqlite/Migrations/20260529032528_CommandVariableSupport_AresIdentityContext.cs
@@ -1,0 +1,22 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace AresService.Migrations.Sqlite.Migrations.AresIdentity
+{
+    /// <inheritdoc />
+    public partial class CommandVariableSupport_AresIdentityContext : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+
+        }
+    }
+}
