+++ Ares.Core.Tests/Execution/CommandVariableResolutionTests.cs
@@ -3,6 +3,7 @@ using Ares.Core.Execution.ControlTokens;
 using Ares.Core.Execution.Executors;
 using Ares.Core.Execution.Executors.Composers;
 using Ares.Core.Notifications;
+using Ares.Core.Settings;
 using Ares.Datamodel;
 using Ares.Datamodel.Device;
 using Ares.Datamodel.Templates;
@@ -15,6 +16,14 @@ namespace Ares.Core.Tests.Execution;
 
 internal class CommandVariableResolutionTests
 {
+  private ISystemSettingsManager _systemSettingsManager;
+  
+  [OneTimeSetUp]
+  public void OneTimeSetUp()
+  {
+    _systemSettingsManager = new Mock<SystemSettingsManager>().Object;
+  }
+
   [Test]
   public async Task SequentialStepExecutor_ResolvesVariableParameterFromEarlierCommandOutput()
   {
@@ -26,7 +35,7 @@ internal class CommandVariableResolutionTests
     stepTemplate.CommandTemplates.Add(CreateSourceCommand());
     stepTemplate.CommandTemplates.Add(CreateConsumerCommand("sourceResult"));
 
-    var stepComposer = new StepComposer(deviceRepo, new Mock<INotifier>().Object);
+    var stepComposer = new StepComposer(deviceRepo, new Mock<INotifier>().Object, _systemSettingsManager);
     var stepExecutor = stepComposer.Compose(stepTemplate);
 
     using var tokenSource = new ExecutionControlTokenSource();
@@ -52,7 +61,8 @@ internal class CommandVariableResolutionTests
         return Task.FromResult(new CommandResult { Success = true });
       },
       template,
-      new Mock<INotifier>().Object);
+      new Mock<INotifier>().Object,
+      _systemSettingsManager);
 
     using var tokenSource = new ExecutionControlTokenSource();
     var summary = await executor.Execute(tokenSource.Token);
