### 2026-06-04 271774b
Fixed some bugs in the retry command logic. Added notifications to confirm settings were saved
