### 2026-06-04 0564bc1
Updates to make the notification stream a bit more stable and thread safe. Also prompt user for decision when receiving the default (unset) status code from a command failure.
