### 2026-06-04 6531f6c
Report status consistently
