+++ AresService.Migrations.SqlServer/Migrations/20260603203704_error-handling-updates_AresDbContext.cs
@@ -1,0 +1,57 @@
+﻿using System;
+using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace AresService.Migrations.SqlServer.Migrations.AresDb
+{
+    /// <inheritdoc />
+    public partial class errorhandlingupdates_AresDbContext : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.CreateTable(
+                name: "AresGeneralSettingsConfig",
+                columns: table => new
+                {
+                    UniqueId = table.Column<Guid>(type: "uniqueidentifier", nullable: false),
+                    ExperimentRetryLimit = table.Column<int>(type: "int", nullable: false),
+                    CommandRetryLimit = table.Column<int>(type: "int", nullable: false),
+                    RetryCooldown = table.Column<string>(type: "nvarchar(max)", nullable: true),
+                    CommandLatency = table.Column<string>(type: "nvarchar(max)", nullable: true),
+                    CreationTime = table.Column<DateTime>(type: "datetime2", nullable: false, defaultValueSql: "getdate()"),
+                    LastModified = table.Column<DateTime>(type: "datetime2", nullable: false, defaultValueSql: "getdate()")
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_AresGeneralSettingsConfig", x => x.UniqueId);
+                });
+
+            migrationBuilder.CreateTable(
+                name: "ErrorHandlingConfigs",
+                columns: table => new
+                {
+                    UniqueId = table.Column<Guid>(type: "uniqueidentifier", nullable: false),
+                    Code = table.Column<int>(type: "int", nullable: false),
+                    Handling = table.Column<int>(type: "int", nullable: false),
+                    CreationTime = table.Column<DateTime>(type: "datetime2", nullable: false, defaultValueSql: "getdate()"),
+                    LastModified = table.Column<DateTime>(type: "datetime2", nullable: false, defaultValueSql: "getdate()")
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_ErrorHandlingConfigs", x => x.UniqueId);
+                });
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropTable(
+                name: "AresGeneralSettingsConfig");
+
+            migrationBuilder.DropTable(
+                name: "ErrorHandlingConfigs");
+        }
+    }
+}
