### 2026-06-04 cb18639
Fixed duplicate entity config. Fixed visual bug in execution screen.
