### 2026-06-04 4c522e9
Added a campaign data explorer for tabulating individual experiments.
