+++ Ares.Core/DataManagement/DataMappers/CampaignDatasetGenerator.cs
@@ -5,13 +5,19 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Ares.Core.DataManagement.DataMappers;
 
-internal class CampaignDatasetGenerator(IDbContextFactory<CoreDatabaseContext> _dbContextFactory)
+public class CampaignDatasetGenerator(IDbContextFactory<CoreDatabaseContext> _dbContextFactory)
 {
   private const string ExperimentNumberColumnName = "Experiment Number";
+  private const string ExperimentExecutionIdColumnName = "Experiment Execution ID";
+  private const string ExperimentIdColumnName = "Experiment ID";
+  private const string ExperimentTemplateColumnName = "Experiment Template";
   private const string TimeStartedColumnName = "Time Started";
   private const string TimeFinishedColumnName = "Time Finished";
+  private const string DurationSecondsColumnName = "Duration Seconds";
   private const string AnalysisResultColumnName = "Analysis Result";
-  private const string ParameterColumnPrefix = "Parameter.";
+  private const string ResultOutputPathColumnName = "Result Output Path";
+  private const string InputColumnPrefix = "Input.";
+  private const string OutputColumnPrefix = "Output.";
 
   public async ValueTask<AresDataset[]> GenerateAsync(string summaryId, CancellationToken cancellationToken = default)
   {
@@ -28,7 +34,7 @@ internal class CampaignDatasetGenerator(IDbContextFactory<CoreDatabaseContext> _
 
     var dataset = new AresDataset
     {
-      Name = GetDatasetName(summary, summaryId)
+      Name = "Experiments"
     };
 
     dataset.Columns.AddRange(CreateFixedColumns());
@@ -38,42 +44,29 @@ internal class CampaignDatasetGenerator(IDbContextFactory<CoreDatabaseContext> _
     return [dataset];
   }
 
-  private static string GetDatasetName(CampaignExecutionSummary summary, string summaryId)
-  {
-    if(!string.IsNullOrWhiteSpace(summary.CampaignName))
-      return summary.CampaignName;
-
-    if(!string.IsNullOrWhiteSpace(summary.CampaignId))
-      return summary.CampaignId;
-
-    return summaryId;
-  }
-
   private static IEnumerable<AresDataColumn> CreateFixedColumns()
   {
     return
     [
-      new AresDataColumn
-      {
-        Name = ExperimentNumberColumnName,
-        Schema = new AresValueSchema { Type = AresDataType.Int }
-      },
-      new AresDataColumn
-      {
-        Name = TimeStartedColumnName,
-        Schema = new AresValueSchema { Type = AresDataType.Timestamp }
-      },
-      new AresDataColumn
-      {
-        Name = TimeFinishedColumnName,
-        Schema = new AresValueSchema { Type = AresDataType.Timestamp }
-      },
-      new AresDataColumn
-      {
-        Name = AnalysisResultColumnName,
-        Schema = new AresValueSchema { Type = AresDataType.Number, Optional = true }
-      }
+      CreateColumn(ExperimentNumberColumnName, AresDataType.Int),
+      CreateColumn(ExperimentExecutionIdColumnName, AresDataType.String, optional: true),
+      CreateColumn(ExperimentIdColumnName, AresDataType.String, optional: true),
+      CreateColumn(ExperimentTemplateColumnName, AresDataType.String, optional: true),
+      CreateColumn(TimeStartedColumnName, AresDataType.Timestamp, optional: true),
+      CreateColumn(TimeFinishedColumnName, AresDataType.Timestamp, optional: true),
[…]
