+++ Ares.Core.Tests/DataManagement/DataMappers/CampaignDatasetGeneratorTests.cs
@@ -39,25 +39,33 @@ internal class CampaignDatasetGeneratorTests
 
     using(Assert.EnterMultipleScope())
     {
-      Assert.That(dataset.Name, Is.EqualTo("Campaign A"));
-      Assert.That(dataset.Columns.Take(4).Select(column => column.Name), Is.EqualTo([
+      Assert.That(dataset.Name, Is.EqualTo("Experiments"));
+      Assert.That(dataset.Columns.Take(9).Select(column => column.Name), Is.EqualTo([
         "Experiment Number",
+        "Experiment Execution ID",
+        "Experiment ID",
+        "Experiment Template",
         "Time Started",
         "Time Finished",
-        "Analysis Result"
+        "Duration Seconds",
+        "Analysis Result",
+        "Result Output Path"
       ]));
       Assert.That(ColumnSchema(dataset, "Experiment Number").Type, Is.EqualTo(AresDataType.Int));
       Assert.That(ColumnSchema(dataset, "Time Started").Type, Is.EqualTo(AresDataType.Timestamp));
       Assert.That(ColumnSchema(dataset, "Time Finished").Type, Is.EqualTo(AresDataType.Timestamp));
+      Assert.That(ColumnSchema(dataset, "Duration Seconds").Type, Is.EqualTo(AresDataType.Number));
       Assert.That(ColumnSchema(dataset, "Analysis Result").Type, Is.EqualTo(AresDataType.Number));
       Assert.That(ColumnSchema(dataset, "Analysis Result").Optional, Is.True);
       Assert.That(dataset.Rows[0].Data.Fields["Experiment Number"].IntValue, Is.EqualTo(1));
       Assert.That(dataset.Rows[0].Data.Fields["Time Started"].TimestampValue, Is.EqualTo(Timestamp.FromDateTime(firstStart)));
       Assert.That(dataset.Rows[0].Data.Fields["Time Finished"].TimestampValue, Is.EqualTo(Timestamp.FromDateTime(firstEnd)));
+      Assert.That(dataset.Rows[0].Data.Fields["Duration Seconds"].NumberValue, Is.EqualTo(1));
       Assert.That(dataset.Rows[0].Data.Fields["Analysis Result"].NumberValue, Is.EqualTo(1.5));
       Assert.That(dataset.Rows[1].Data.Fields["Experiment Number"].IntValue, Is.EqualTo(2));
       Assert.That(dataset.Rows[1].Data.Fields["Time Started"].TimestampValue, Is.EqualTo(Timestamp.FromDateTime(secondStart)));
       Assert.That(dataset.Rows[1].Data.Fields["Time Finished"].TimestampValue, Is.EqualTo(Timestamp.FromDateTime(secondEnd)));
+      Assert.That(dataset.Rows[1].Data.Fields["Duration Seconds"].NumberValue, Is.EqualTo(1));
       Assert.That(dataset.Rows[1].Data.Fields["Analysis Result"].NumberValue, Is.EqualTo(2.5));
     }
   }
@@ -84,17 +92,99 @@ internal class CampaignDatasetGeneratorTests
 
     using(Assert.EnterMultipleScope())
     {
-      Assert.That(ColumnSchema(dataset, "Yield").Type, Is.EqualTo(AresDataType.Number));
-      Assert.That(ColumnSchema(dataset, "Yield").Optional, Is.True);
-      Assert.That(ColumnSchema(dataset, "Comment").Type, Is.EqualTo(AresDataType.String));
-      Assert.That(ColumnSchema(dataset, "Comment").Optional, Is.True);
[…]
