### 2026-06-04 d95b1e6
Added command view to the campaign data explorer
