+++ Ares.Core.Tests/DataManagement/DataMappers/CampaignDatasetGeneratorTests.cs
@@ -35,21 +35,18 @@ internal class CampaignDatasetGeneratorTests
       CreateExperiment(firstStart, firstEnd, analysisResult: 1.5));
     var generator = CreateGenerator(summary);
 
-    var dataset = (await generator.GenerateAsync(summaryId)).Single();
+    var dataset = GetDataset(await generator.GenerateAsync(summaryId), "Experiments");
 
     using(Assert.EnterMultipleScope())
     {
       Assert.That(dataset.Name, Is.EqualTo("Experiments"));
-      Assert.That(dataset.Columns.Take(9).Select(column => column.Name), Is.EqualTo([
+      Assert.That(dataset.Columns.Take(6).Select(column => column.Name), Is.EqualTo([
         "Experiment Number",
-        "Experiment Execution ID",
-        "Experiment ID",
         "Experiment Template",
         "Time Started",
         "Time Finished",
         "Duration Seconds",
-        "Analysis Result",
-        "Result Output Path"
+        "Analysis Result"
       ]));
       Assert.That(ColumnSchema(dataset, "Experiment Number").Type, Is.EqualTo(AresDataType.Int));
       Assert.That(ColumnSchema(dataset, "Time Started").Type, Is.EqualTo(AresDataType.Timestamp));
@@ -87,7 +84,7 @@ internal class CampaignDatasetGeneratorTests
       ]));
     var generator = CreateGenerator(summary);
 
-    var dataset = (await generator.GenerateAsync(summaryId)).Single();
+    var dataset = GetDataset(await generator.GenerateAsync(summaryId), "Experiments");
     original.StringValue = "after";
 
     using(Assert.EnterMultipleScope())
@@ -106,24 +103,16 @@ internal class CampaignDatasetGeneratorTests
   }
 
   [Test]
-  public async Task GenerateAsync_IncludesExperimentIdentityTemplateAndOutputPath()
+  public async Task GenerateAsync_IncludesExperimentTemplate()
   {
     var summaryId = Guid.NewGuid().ToString();
     var experiment = CreateExperiment(DateTime.UnixEpoch, DateTime.UnixEpoch.AddSeconds(1));
-    experiment.ExperimentId = "experiment-id";
     experiment.ExperimentOverview.Template = new ExperimentTemplate { Name = "Template A" };
-    experiment.ResultOutputPath = "results/experiment.json";
     var generator = CreateGenerator(CreateCampaignSummary(summaryId, "Campaign A", experiment));
 
-    var row = (await generator.GenerateAsync(summaryId)).Single().Rows.Single();
+    var row = GetDataset(await generator.GenerateAsync(summaryId), "Experiments").Rows.Single();
 
-    using(Assert.EnterMultipleScope())
-    {
-      Assert.That(row.Data.Fields["Experiment Execution ID"].StringValue, Is.EqualTo(experiment.UniqueId));
-      Assert.That(row.Data.Fields["Experiment ID"].StringValue, Is.EqualTo("experiment-id"));
-      Assert.That(row.Data.Fields["Experiment Template"].StringValue, Is.EqualTo("Template A"));
-      Assert.That(row.Data.Fields["Result Output Path"].StringValue, Is.EqualTo("results/experiment.json"));
-    }
+    Assert.That(row.Data.Fields["Experiment Template"].StringValue, Is.EqualTo("Template A"));
   }
 
   [Test]
@@ -144,7 +133,7 @@ internal class CampaignDatasetGeneratorTests
[…]
