+++ Ares.Core/DataManagement/DataMappers/CampaignDatasetGenerator.cs
@@ -8,16 +8,21 @@ namespace Ares.Core.DataManagement.DataMappers;
 public class CampaignDatasetGenerator(IDbContextFactory<CoreDatabaseContext> _dbContextFactory)
 {
   private const string ExperimentNumberColumnName = "Experiment Number";
-  private const string ExperimentExecutionIdColumnName = "Experiment Execution ID";
-  private const string ExperimentIdColumnName = "Experiment ID";
   private const string ExperimentTemplateColumnName = "Experiment Template";
+  private const string StepNumberColumnName = "Step Number";
+  private const string CommandNumberColumnName = "Command Number";
+  private const string CommandNameColumnName = "Command Name";
+  private const string CommandDescriptionColumnName = "Command Description";
   private const string TimeStartedColumnName = "Time Started";
   private const string TimeFinishedColumnName = "Time Finished";
   private const string DurationSecondsColumnName = "Duration Seconds";
   private const string AnalysisResultColumnName = "Analysis Result";
-  private const string ResultOutputPathColumnName = "Result Output Path";
+  private const string StatusColumnName = "Status";
+  private const string SuccessColumnName = "Success";
+  private const string ErrorColumnName = "Error";
   private const string InputColumnPrefix = "Input.";
   private const string OutputColumnPrefix = "Output.";
+  private const string OutputColumnName = "Output";
 
   public async ValueTask<AresDataset[]> GenerateAsync(string summaryId, CancellationToken cancellationToken = default)
   {
@@ -32,44 +37,71 @@ public class CampaignDatasetGenerator(IDbContextFactory<CoreDatabaseContext> _db
       .OrderBy(experiment => experiment.ExecutionInfo?.TimeStarted)
       .ToArray();
 
+    return [
+      CreateExperimentsDataset(experiments, cancellationToken),
+      CreateCommandsDataset(experiments, cancellationToken)
+    ];
+  }
+
+  private static AresDataset CreateExperimentsDataset(ExperimentExecutionSummary[] experiments, CancellationToken cancellationToken)
+  {
     var dataset = new AresDataset
     {
       Name = "Experiments"
     };
 
-    dataset.Columns.AddRange(CreateFixedColumns());
-    dataset.Columns.AddRange(CreateDynamicColumns(experiments, cancellationToken));
-    dataset.Rows.AddRange(experiments.Select((experiment, index) => CreateRow(experiment, index + 1, cancellationToken)));
-
-    return [dataset];
+    dataset.Columns.AddRange(CreateExperimentColumns(experiments, cancellationToken));
+    dataset.Rows.AddRange(experiments.Select((experiment, index) => CreateExperimentRow(experiment, index + 1, cancellationToken)));
+    return dataset;
   }
 
-  private static IEnumerable<AresDataColumn> CreateFixedColumns()
+  private static AresDataset CreateCommandsDataset(ExperimentExecutionSummary[] experiments, CancellationToken cancellationToken)
+  {
+    var commandRecords = CreateCommandRecords(experiments, cancellationToken).ToArray();
+    var dataset = new AresDataset
+    {
+      Name = "Commands"
+    };
+
+    dataset.Columns.AddRange(CreateCommandColumns(commandRecords, cancellationToken));
+    dataset.Rows.AddRange(commandRecords.Select(record => CreateCommandRow(record, cancellationToken)));
+    return dataset;
+  }
+
[…]
