### 2026-06-08 30da787
Analyzer and planner transaction views
