+++ Ares.Core/DataManagement/DataMappers/CampaignDatasetGenerator.cs
@@ -1,6 +1,10 @@
 using Ares.Datamodel;
+using Ares.Datamodel.Analyzing;
+using Ares.Datamodel.Analyzing.Remote;
 using Ares.Datamodel.Extensions;
+using Ares.Datamodel.Planning;
 using Ares.Datamodel.Templates;
+using Google.Protobuf.WellKnownTypes;
 using Microsoft.EntityFrameworkCore;
 
 namespace Ares.Core.DataManagement.DataMappers;
@@ -20,6 +24,17 @@ public class CampaignDatasetGenerator(IDbContextFactory<CoreDatabaseContext> _db
   private const string StatusColumnName = "Status";
   private const string SuccessColumnName = "Success";
   private const string ErrorColumnName = "Error";
+  private const string PlannerNameColumnName = "Planner Name";
+  private const string PlannerTypeColumnName = "Planner Type";
+  private const string PlannerVersionColumnName = "Planner Version";
+  private const string AnalyzerNameColumnName = "Analyzer Name";
+  private const string AnalyzerTypeColumnName = "Analyzer Type";
+  private const string AnalyzerVersionColumnName = "Analyzer Version";
+  private const string TimeRequestSentColumnName = "Time Request Sent";
+  private const string TimeResponseReceivedColumnName = "Time Response Received";
+  private const string OutcomeColumnName = "Outcome";
+  private const string AnalysisResultsColumnName = "Analysis Results";
+  private const string ResultColumnName = "Result";
   private const string InputColumnPrefix = "Input.";
   private const string OutputColumnPrefix = "Output.";
   private const string OutputColumnName = "Output";
@@ -36,10 +51,43 @@ public class CampaignDatasetGenerator(IDbContextFactory<CoreDatabaseContext> _db
     var experiments = summary.ExperimentSummaries
       .OrderBy(experiment => experiment.ExecutionInfo?.TimeStarted)
       .ToArray();
+    var experimentNumbers = experiments
+      .Select((experiment, index) => new { experiment.ExperimentId, ExperimentNumber = index + 1 })
+      .Where(item => !string.IsNullOrWhiteSpace(item.ExperimentId))
+      .GroupBy(item => item.ExperimentId)
+      .ToDictionary(group => group.Key, group => group.First().ExperimentNumber);
+    var plannerRecords = new List<PlannerRecord>();
+    var analyzerRecords = new List<AnalyzerRecord>();
+    var campaignStart = summary.ExecutionInfo?.TimeStarted;
+    var campaignEnd = summary.ExecutionInfo?.TimeFinished;
+    if(campaignStart is not null && campaignEnd is not null)
+    {
+      var plannerTransactions = await ctx.PlannerTransactions
+        .Where(transaction => transaction.TimeRequestSent >= campaignStart && transaction.TimeResponseReceived <= campaignEnd)
+        .ToListAsync(cancellationToken);
+      foreach(var transaction in plannerTransactions)
+      {
+        cancellationToken.ThrowIfCancellationRequested();
+        if(TryCreatePlannerRecord(transaction, summary, experimentNumbers, out var record))
+          plannerRecords.Add(record);
+      }
+
+      var analyzerTransactions = await ctx.AnalyzerTransactions
+        .Where(transaction => transaction.TimeRequestSent >= campaignStart && transaction.TimeResponseReceived <= campaignEnd)
+        .ToListAsync(cancellationToken);
+      foreach(var transaction in analyzerTransactions)
+      {
+        cancellationToken.ThrowIfCancellationRequested();
[…]
