+++ Ares.Core.Tests/DataManagement/DataMappers/CampaignDatasetGeneratorTests.cs
@@ -1,6 +1,9 @@
 using Ares.Core.DataManagement.DataMappers;
 using Ares.Datamodel;
+using Ares.Datamodel.Analyzing;
+using Ares.Datamodel.Analyzing.Remote;
 using Ares.Datamodel.Extensions;
+using Ares.Datamodel.Planning;
 using Ares.Datamodel.Templates;
 using Google.Protobuf.WellKnownTypes;
 using Microsoft.EntityFrameworkCore;
@@ -177,7 +180,7 @@ internal class CampaignDatasetGeneratorTests
   }
 
   [Test]
-  public async Task GenerateAsync_ReturnsExperimentsAndCommandsDatasets()
+  public async Task GenerateAsync_ReturnsAllCampaignDatasets()
   {
     var summaryId = Guid.NewGuid().ToString();
     var generator = CreateGenerator(CreateCampaignSummary(
@@ -187,7 +190,17 @@ internal class CampaignDatasetGeneratorTests
 
     var datasets = await generator.GenerateAsync(summaryId);
 
-    Assert.That(datasets.Select(dataset => dataset.Name), Is.EqualTo(["Experiments", "Commands"]));
+    using(Assert.EnterMultipleScope())
+    {
+      Assert.That(datasets.Select(dataset => dataset.Name), Is.EqualTo([
+        "Experiments",
+        "Commands",
+        "Planner Transactions",
+        "Analyzer Transactions"
+      ]));
+      Assert.That(GetDataset(datasets, "Planner Transactions").Rows, Is.Empty);
+      Assert.That(GetDataset(datasets, "Analyzer Transactions").Rows, Is.Empty);
+    }
   }
 
   [Test]
@@ -390,6 +403,142 @@ internal class CampaignDatasetGeneratorTests
   }
 
   [Test]
+  public async Task GenerateAsync_CreatesPlannerTransactionRows()
+  {
+    var summaryId = Guid.NewGuid().ToString();
+    var experiment = CreateExperiment(DateTime.UnixEpoch, DateTime.UnixEpoch.AddSeconds(5));
+    var nestedOutput = AresValueHelper.CreateStruct();
+    nestedOutput.StructValue.Fields["Value"] = AresValueHelper.CreateNumber(12.5);
+    var transaction = CreatePlannerTransaction(
+      experiment.ExperimentId,
+      DateTime.UnixEpoch.AddSeconds(1),
+      DateTime.UnixEpoch.AddSeconds(3),
+      ("Temperature", nestedOutput));
+    transaction.PlanningRequest.AnalysisResults.AddRange([1.5, 2.5]);
+    transaction.PlanningResponse.PlanningOutcome = Outcome.Warning;
+    transaction.PlanningResponse.ErrorString = "planner warning";
+    var generator = CreateGenerator(CreateCampaignSummary(summaryId, "Campaign A", experiment), [transaction]);
+
+    var dataset = GetDataset(await generator.GenerateAsync(summaryId), "Planner Transactions");
+    var row = dataset.Rows.Single();
+
+    using(Assert.EnterMultipleScope())
+    {
+      Assert.That(dataset.Columns.Take(11).Select(column => column.Name), Is.EqualTo([
+        "Experiment Number",
+        "Planner Name",
+        "Planner Type",
+        "Planner Version",
+        "Time Request Sent",
+        "Time Response Received",
+        "Duration Seconds",
+        "Outcome",
+        "Error",
+        "Analysis Results",
+        "Output.Temperature.Value"
+      ]));
+      Assert.That(row.Data.Fields["Experiment Number"].IntValue, Is.EqualTo(1));
[…]
