### 2026-06-08 5dc5afc
Command exporter now has inputs in addition to outputs.
