+++ Ares.Core.Tests/DataManagement/DataMappers/CampaignDatasetGeneratorTests.cs
@@ -350,6 +350,107 @@ internal class CampaignDatasetGeneratorTests
   }
 
   [Test]
+  public async Task GenerateAsync_IncludesResolvedCommandInputsBeforeOutputs()
+  {
+    var summaryId = Guid.NewGuid().ToString();
+    var structValue = AresValueHelper.CreateStruct();
+    structValue.StructValue.Fields["Temperature"] = AresValueHelper.CreateNumber(22.5);
+    var literal = CreateParameter("Literal", AresValueHelper.CreateString("literal"));
+    var planned = CreateParameter("Planned", AresValueHelper.CreateNumber(1));
+    planned.SetPlannedSource(new ParameterMetadata { Name = "Planned" });
+    planned.SetResolvedValue(AresValueHelper.CreateNumber(2));
+    var environment = CreateParameter("Environment", AresValueHelper.CreateString("initial"));
+    environment.SetEnvironmentSource(VariableType.VarUnspecified, "environment");
+    environment.SetResolvedValue(structValue);
+    var variable = CreateParameter("Variable", AresValueHelper.CreateString("initial"));
+    variable.SetCommandVariableSource("previous");
+    variable.SetResolvedValue(AresValueHelper.CreateList([AresValueHelper.CreateNumber(3), AresValueHelper.CreateNumber(4)]));
+    var template = CreateCommandTemplate("command-template", literal, planned, environment, variable);
+    var command = CreateCommand(
+      DateTime.UnixEpoch,
+      DateTime.UnixEpoch.AddSeconds(1),
+      result: new CommandResult { Result = AresValueHelper.CreateNumber(10) },
+      templateId: template.UniqueId);
+    var experiment = CreateExperiment(
+      DateTime.UnixEpoch,
+      DateTime.UnixEpoch.AddSeconds(2),
+      steps: [CreateStep("step", DateTime.UnixEpoch, DateTime.UnixEpoch.AddSeconds(2), command)]);
+    SetCommandTemplates(experiment, template);
+    var generator = CreateGenerator(CreateCampaignSummary(summaryId, "Campaign A", experiment));
+
+    var dataset = GetDataset(await generator.GenerateAsync(summaryId), "Commands");
+    var columns = dataset.Columns.Select(column => column.Name).ToArray();
+    var row = dataset.Rows.Single();
+
+    using(Assert.EnterMultipleScope())
+    {
+      Assert.That(columns, Does.Contain("Input.Literal"));
+      Assert.That(columns, Does.Contain("Input.Planned"));
+      Assert.That(columns, Does.Contain("Input.Environment.Temperature"));
+      Assert.That(columns, Does.Contain("Input.Variable"));
+      Assert.That(Array.IndexOf(columns, "Input.Literal"), Is.LessThan(Array.IndexOf(columns, "Output")));
+      Assert.That(row.Data.Fields["Input.Literal"].StringValue, Is.EqualTo("literal"));
+      Assert.That(row.Data.Fields["Input.Planned"].NumberValue, Is.EqualTo(2));
+      Assert.That(row.Data.Fields["Input.Environment.Temperature"].NumberValue, Is.EqualTo(22.5));
+      Assert.That(row.Data.Fields["Input.Variable"].ListValue.Values.Select(value => value.NumberValue), Is.EqualTo([3, 4]));
+      Assert.That(row.Data.Fields["Output"].NumberValue, Is.EqualTo(10));
+    }
+  }
+
+  [Test]
+  public async Task GenerateAsync_MatchesCommandInputsByTemplateId()
+  {
+    var summaryId = Guid.NewGuid().ToString();
[…]
