+++ Ares.Core/Execution/Executors/CampaignExecutor.cs
@@ -1,21 +1,23 @@
-﻿using System.Reactive.Linq;
-using System.Reactive.Subjects;
-using Ares.Core.Analyzing;
+﻿using Ares.Core.Analyzing;
 using Ares.Core.AresEnvironment;
 using Ares.Core.Device.State.Logging;
 using Ares.Core.Exceptions;
 using Ares.Core.Execution.ControlTokens;
 using Ares.Core.Execution.Executors.Composers;
 using Ares.Core.Execution.Extensions;
+using Ares.Core.Execution.Safety;
 using Ares.Core.Execution.StopConditions;
 using Ares.Core.Notifications;
 using Ares.Core.Output;
 using Ares.Core.Planning;
+using Ares.Core.Settings;
 using Ares.Datamodel;
 using Ares.Datamodel.Analyzing;
 using Ares.Datamodel.Templates;
 using Google.Protobuf.WellKnownTypes;
 using Microsoft.Extensions.Logging;
+using System.Reactive.Linq;
+using System.Reactive.Subjects;
 
 namespace Ares.Core.Execution.Executors;
 
@@ -33,6 +35,16 @@ public class CampaignExecutor : ICampaignExecutor
   readonly AnalysisHelper _analysisHelper;
   readonly AnalysisRepo _analysisRepo;
   readonly IAnalyzerRepo _analyzerRepo;
+  readonly ISystemSettingsManager _settingsManager;
+  readonly IExecutionSafetyManager _executionSafetyManager;
+
+  private readonly string _campaignFailedTitle = "Campaign Execution Failed!";
+  private AresGeneralSettingsConfig? _generalSettingsConfig;
+  private ExperimentExecutorResult _currentExecutorResult = new();
+  private ExperimentExecutionSummary? _currentSummary = null;
+  private ExperimentTemplate? _currentExperimentTemplate = null;
+  private int _experimentCount = 0;
+  private TaskCompletionSource<ErrorHandling>? _userDecisionSource;
 
   internal CampaignExecutor(ICommandComposer<ExperimentTemplate, ExperimentExecutor> experimentComposer,
     IPlanningHelper planningHelper,
@@ -45,7 +57,9 @@ public class CampaignExecutor : ICampaignExecutor
     IAnalyzerRepo analyzerRepo,
     ILogger<CampaignExecutor> logger,
     AresVariableManager variableManager,
-    StateLoggerManager stateLoggerManager)
+    StateLoggerManager stateLoggerManager,
+    ISystemSettingsManager settingsManager,
+    IExecutionSafetyManager executionSafetyManager)
   {
     _analyzerRepo = analyzerRepo;
     _analysisRepo = analysisRepo;
@@ -59,6 +73,8 @@ public class CampaignExecutor : ICampaignExecutor
     _logger = logger;
     Template = template;
     _notifier = notifier;
+    _settingsManager = settingsManager;
+    _executionSafetyManager = executionSafetyManager;
 
     Status = new CampaignExecutionStatus
     {
@@ -108,8 +124,6 @@ public class CampaignExecutor : ICampaignExecutor
       executionSuccess = await ExecuteExperimentLoop(campaignPath, analyses, experimentSummaries, token, executionSuccess, startupResult.Summary);
 
       _logger.LogDebug("The campaign loop is now finished. Moving onto the closeout script.");
-
-
       var closeoutResult = await ExecuteCloseout(analyses, experimentSummaries, token, executionSuccess);
       closeoutSummary = closeoutResult.Summary;
 
@@ -148,6 +162,8 @@ public class CampaignExecutor : ICampaignExecutor
[…]
