### 2026-06-10 66cdacf
Provide status code to plan requests
