### 2026-06-10 3130e4e
Campaign data export (#168)
* Added a campaign data explorer for tabulating individual experiments.

* Added campaign notes and tag support for display and filtering.

* Added command view to the campaign data explorer

* Analyzer and planner transaction views

* Command exporter now has inputs in addition to outputs.
