+++ Ares.Core.Tests/DataManagement/DataMappers/CampaignDatasetGeneratorTests.cs
@@ -1,6 +1,9 @@
 using Ares.Core.DataManagement.DataMappers;
 using Ares.Datamodel;
+using Ares.Datamodel.Analyzing;
+using Ares.Datamodel.Analyzing.Remote;
 using Ares.Datamodel.Extensions;
+using Ares.Datamodel.Planning;
 using Ares.Datamodel.Templates;
 using Google.Protobuf.WellKnownTypes;
 using Microsoft.EntityFrameworkCore;
@@ -35,29 +38,34 @@ internal class CampaignDatasetGeneratorTests
       CreateExperiment(firstStart, firstEnd, analysisResult: 1.5));
     var generator = CreateGenerator(summary);
 
-    var dataset = (await generator.GenerateAsync(summaryId)).Single();
+    var dataset = GetDataset(await generator.GenerateAsync(summaryId), "Experiments");
 
     using(Assert.EnterMultipleScope())
     {
-      Assert.That(dataset.Name, Is.EqualTo("Campaign A"));
-      Assert.That(dataset.Columns.Take(4).Select(column => column.Name), Is.EqualTo([
+      Assert.That(dataset.Name, Is.EqualTo("Experiments"));
+      Assert.That(dataset.Columns.Take(6).Select(column => column.Name), Is.EqualTo([
         "Experiment Number",
+        "Experiment Template",
         "Time Started",
         "Time Finished",
+        "Duration Seconds",
         "Analysis Result"
       ]));
       Assert.That(ColumnSchema(dataset, "Experiment Number").Type, Is.EqualTo(AresDataType.Int));
       Assert.That(ColumnSchema(dataset, "Time Started").Type, Is.EqualTo(AresDataType.Timestamp));
       Assert.That(ColumnSchema(dataset, "Time Finished").Type, Is.EqualTo(AresDataType.Timestamp));
+      Assert.That(ColumnSchema(dataset, "Duration Seconds").Type, Is.EqualTo(AresDataType.Number));
       Assert.That(ColumnSchema(dataset, "Analysis Result").Type, Is.EqualTo(AresDataType.Number));
       Assert.That(ColumnSchema(dataset, "Analysis Result").Optional, Is.True);
       Assert.That(dataset.Rows[0].Data.Fields["Experiment Number"].IntValue, Is.EqualTo(1));
       Assert.That(dataset.Rows[0].Data.Fields["Time Started"].TimestampValue, Is.EqualTo(Timestamp.FromDateTime(firstStart)));
       Assert.That(dataset.Rows[0].Data.Fields["Time Finished"].TimestampValue, Is.EqualTo(Timestamp.FromDateTime(firstEnd)));
+      Assert.That(dataset.Rows[0].Data.Fields["Duration Seconds"].NumberValue, Is.EqualTo(1));
       Assert.That(dataset.Rows[0].Data.Fields["Analysis Result"].NumberValue, Is.EqualTo(1.5));
       Assert.That(dataset.Rows[1].Data.Fields["Experiment Number"].IntValue, Is.EqualTo(2));
       Assert.That(dataset.Rows[1].Data.Fields["Time Started"].TimestampValue, Is.EqualTo(Timestamp.FromDateTime(secondStart)));
       Assert.That(dataset.Rows[1].Data.Fields["Time Finished"].TimestampValue, Is.EqualTo(Timestamp.FromDateTime(secondEnd)));
+      Assert.That(dataset.Rows[1].Data.Fields["Duration Seconds"].NumberValue, Is.EqualTo(1));
       Assert.That(dataset.Rows[1].Data.Fields["Analysis Result"].NumberValue, Is.EqualTo(2.5));
     }
   }
@@ -79,22 +87,390 @@ internal class CampaignDatasetGeneratorTests
[…]
