+++ Ares.Core/Execution/Executors/ExecutorSummaryHelpers.cs
@@ -58,7 +58,7 @@ internal static class ExecutorSummaryHelpers
       CommandDescription = template.Metadata.Description,
       CommandName = template.Metadata.Name,
       StatusCode = deviceResult?.StatusCode ?? CommandStatusCode.StatusUnspecified
-     };
+    };
 
     if(template.HasOutputVarName)
       commandExecutionSummary.VarName = template.OutputVarName;
