### 2026-06-11 ad58390
Fixed averaging command
