+++ Ares.Core/CoreDevice/AresCoreDevice.cs
@@ -1,6 +1,7 @@
 ﻿using Ares.Core.EntityConfigurations.Extensions;
 using Ares.Datamodel;
 using Ares.Datamodel.Device;
+using Ares.Datamodel.Extensions;
 using Ares.Datamodel.Factories;
 using Ares.Device;
 using System.Reactive.Linq;
@@ -126,19 +127,19 @@ public class AresCoreDevice : AresDevice
         { 
           case AresValue.KindOneofCase.NumberArrayValue:
             var average = dataToBeAveraged.NumberArrayValue.Numbers.Average();
-            result.Result.FloatValue = average;
+            result.Result = AresValueHelper.CreateFloat(average);
             result.Success = true;
             break;
 
           case AresValue.KindOneofCase.FloatArrayValue:
             var floatAverage = dataToBeAveraged.FloatArrayValue.Floats.Average();
-            result.Result.FloatValue = floatAverage;
+            result.Result = AresValueHelper.CreateFloat(floatAverage);
             result.Success = true;
             break;
 
           case AresValue.KindOneofCase.IntArrayValue:
             var intAverage = dataToBeAveraged.IntArrayValue.Ints.Average();
-            result.Result.FloatValue = intAverage;
+            result.Result = AresValueHelper.CreateFloat(intAverage);
             result.Success = true;
             break;
 
