### 2026-06-11 60ee973
Fixed the timestamp core command
