### 2026-06-11 50a0a97
Fixed manual planner not re-seeding itself before the start of a campaign
