### 2026-06-13 08cb7c9
Refactored campaign executor and fixed the bug where it failed to properly finish last experiment in a set.
