+++ Ares.Core/Execution/Executors/CampaignExecutor.cs
@@ -23,6 +23,27 @@ namespace Ares.Core.Execution.Executors;
 
 public class CampaignExecutor : ICampaignExecutor
 {
+  private enum ExperimentPhase
+  {
+    Initialize,
+    Plan,
+    Compose,
+    Execute,
+    Analyze,
+    Retry,
+    Replan,
+    Complete,
+    Failed,
+    Canceled
+  }
+
+  private enum ExperimentLoopOutcome
+  {
+    Succeeded,
+    Failed,
+    Canceled
+  }
+
   private readonly IExecutionReporter _executionReporter;
   private readonly ISubject<CampaignExecutionStatus> _executionStatusSubject;
   private readonly ICommandComposer<ExperimentTemplate, ExperimentExecutor> _experimentComposer;
@@ -111,9 +132,6 @@ public class CampaignExecutor : ICampaignExecutor
       ResetStatus(token);
       await NotifyCampaignStart();
 
-      var executionSuccess = true;
-
-
       var startupResult = await ExecuteStartup(analyses, experimentSummaries, token);
       if(!startupResult.Success)
       {
@@ -124,7 +142,8 @@ public class CampaignExecutor : ICampaignExecutor
       startupSummary = startupResult.Summary;
       
 
-      executionSuccess = await ExecuteExperimentLoop(campaignPath, analyses, experimentSummaries, token, executionSuccess, startupResult.Summary);
+      var loopOutcome = await ExecuteExperimentLoop(campaignPath, analyses, experimentSummaries, token, startupResult.Summary);
+      var executionSuccess = loopOutcome == ExperimentLoopOutcome.Succeeded;
 
       _logger.LogDebug("The campaign loop is now finished. Moving onto the closeout script.");
       var closeoutResult = await ExecuteCloseout(analyses, experimentSummaries, token, executionSuccess);
@@ -179,8 +198,7 @@ public class CampaignExecutor : ICampaignExecutor
     };
 
     _analysisRepo.ClearAnalyses();
-    Status.State = token.IsPaused ? ExecutionState.Paused : ExecutionState.Running;
-    _executionReporter.Report(Status);
+    ReportCampaignStatus(token.IsPaused ? ExecutionState.Paused : ExecutionState.Running);
   }
 
   private async Task NotifyCampaignStart()
@@ -209,291 +227,300 @@ public class CampaignExecutor : ICampaignExecutor
     return (true, startupSummary);
   }
 
-  private async Task<bool> ExecuteExperimentLoop(string campaignPath, List<Analysis> analyses, List<ExperimentExecutionSummary> experimentSummaries, ExecutionControlToken token, bool executionSuccess, ExperimentExecutionSummary startupSummary)
+  private async Task<ExperimentLoopOutcome> ExecuteExperimentLoop(string campaignPath, List<Analysis> analyses, List<ExperimentExecutionSummary> experimentSummaries, ExecutionControlToken token, ExperimentExecutionSummary startupSummary)
   {
-    var currentState = ExecutionState.InitializeExperiment;
-    executionSuccess = true;
-    var currentExperimentFolder = "";
+    var currentPhase = ExperimentPhase.Initialize;
     var currentExperimentPath = "";
     var failedExperimentRetryCount = 0;
     var failedExperimentRetryLimit = _generalSettingsConfig?.ExperimentRetryLimit ?? 1;
     var experimentRetryCooldown = _generalSettingsConfig?.RetryCooldown ?? new Duration();
-    var shouldStop = false;
 
-    while(!shouldStop && !token.IsCancelled && executionSuccess)
+    try
     {
-      _executionStatusSubject.OnNext(Status);
-      _executionReporter.Report(Status);
-
-      switch(currentState)
[…]
