### 2026-06-15 6a35f87
Added a switch to the device logging menu that controls whether logging is enabled or disabled
