### 2026-06-15 9472f4d
Migration updates to support new logging setting
