### 2026-06-15 a83197c
Vibecoded an initial pass at an execution page rework.
