+++ UI/Features/Execution/ExecutionViewModel.cs
@@ -84,6 +84,9 @@ public partial class ExecutionViewModel : ReactiveObject, INotifyPropertyChanged
 
   public async Task SelectCampaignTemplate(object? templateSummary)
   {
+    if(CampaignActive)
+      return;
+
     if(templateSummary is null || templateSummary is not CampaignTemplateSummary campaignTemplateSummary)
       return;
 
@@ -98,7 +101,8 @@ public partial class ExecutionViewModel : ReactiveObject, INotifyPropertyChanged
     else
       SelectedTemplateSummary = null;
 
-    DisplayExecutionSummary = false;
+    SelectedExecutionTabIndex = 0;
+    await RefreshCampaignSetup();
   }
 
   public async Task UpdateCurrentTemplate()
@@ -108,6 +112,7 @@ public partial class ExecutionViewModel : ReactiveObject, INotifyPropertyChanged
     if(CampaignTemplate is null)
       return;
 
+    AnalyzerInfo = null;
     PlannerAdapterInfos = CampaignTemplate.ExperimentTemplate.GetAllPlannedParameters()
     .Select(parameter => parameter.GetPlanningMetadata())
     .Select(metadata => CampaignTemplate.PlannerAllocations
@@ -131,6 +136,9 @@ public partial class ExecutionViewModel : ReactiveObject, INotifyPropertyChanged
     await _automationClient.SetAnalysisResultStopCondition(
       new AnalysisResultCondition { DesiredResult = DesiredResult, Leeway = DesiredLeeway }, null);
     CurrentStopCondition = await GetCurrentStopCondition();
+    ActiveStopConditionMode = ExecutionStopConditionMode.AnalyzerResult;
+    await RefreshExecutionEligibility();
+    StateChanged?.Invoke();
   }
 
   public Task<ExperimentStopConditionResponse> GetCurrentStopCondition()
@@ -153,6 +161,9 @@ public partial class ExecutionViewModel : ReactiveObject, INotifyPropertyChanged
   {
     await _automationClient.SetNumExperimentsStopCondition(new NumExperimentsCondition { NumExperiments = ExperimentsToRun }, null);
     CurrentStopCondition = await GetCurrentStopCondition();
+    ActiveStopConditionMode = ExecutionStopConditionMode.NumExperiments;
+    await RefreshExecutionEligibility();
+    StateChanged?.Invoke();
   }
 
   public async Task SetReplanRate()
@@ -164,7 +175,13 @@ public partial class ExecutionViewModel : ReactiveObject, INotifyPropertyChanged
 
   public async Task StartCampaign()
   {
+    await ApplyActiveStopCondition();
+
+    if(!CampaignActive)
+      await SetReplanRate();
+
     var executionEligibility = await _automationClient.CheckExecutionEligibility(new Empty(), null);
+    LastExecutionEligibility = executionEligibility;
 
     if(!executionEligibility.IsEligible)
     {
@@ -189,8 +206,11 @@ public partial class ExecutionViewModel : ReactiveObject, INotifyPropertyChanged
     await _automationClient.StartExecution(request, null);
     PlannerMetricsMap.Clear();
     AnalyzerMetrics.Clear();
-    DisplayExecutionSummary = true;
+    SelectedExecutionTabIndex = 1;
   }
+
+  public Task StartConfiguredCampaign()
+    => StartCampaign();
 
   public Task StopCampaign()
     => _automationClient.StopExecution(new Empty(), null);
@@ -203,6 +223,20 @@ public partial class ExecutionViewModel : ReactiveObject, INotifyPropertyChanged
 
   public Task SubmitUserDecision(ErrorHandling decision)
[…]
