### 2026-06-16 bf97790
Additional rework changes
