### 2026-06-16 ccd1ac8
Fixing some more items up, moving new page towards full functionality
