### 2026-06-16 e54babd
Updates to fix a handful of miscellaneous problems. Device command statuses are coming across correctly. We now have a "time elapsed" display in the execution page.
