### 2026-06-17 f63a882
Fixed a bug where logging preferences weren't being updated in the database properly
