### 2026-06-17 52d390d
Bug fixes (#188)
* Fixed averaging command

* Fixed the timestamp core command

* Fixed blank UI values when hardcoding inputs to comands

* Fixed manual planner not re-seeding itself before the start of a campaign

* Made it so the execution page reloads the latest version of the template on intialization. Ensured if the template was deleted since last opened the page doesn't crash. Ensured the user cannot delete templates that are currently actively running.

* Added campaign, experiment, and step info to the device state export lines.

* Added an output variable name to the dataset generato
[…]
