+++ AresService.Migrations.Sqlite/Migrations/20260615170447_logging-settings-update_AresDbContext.cs
@@ -1,0 +1,29 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace AresService.Migrations.Sqlite.Migrations.AresDb
+{
+    /// <inheritdoc />
+    public partial class loggingsettingsupdate_AresDbContext : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<bool>(
+                name: "LoggingEnabled",
+                table: "DeviceLoggingSettings",
+                type: "INTEGER",
+                nullable: false,
+                defaultValue: false);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "LoggingEnabled",
+                table: "DeviceLoggingSettings");
+        }
+    }
+}
