+++ Ares.Core/Execution/ExecutionManager.cs
@@ -140,7 +140,7 @@ public class ExecutionManager : IExecutionManager
     return true;
   }
 
-  public void UpdateReplanRate(int newRate)
+  public void UpdateReplicateRate(int newRate)
   {
     ReplanRate = newRate;
   }
