### 2026-06-17 824d734
Working on campaign executor
