+++ AresService.Migrations.Sqlite/Migrations/20260617194346_FixCorruptedParameterSources_AresDbContext.cs
@@ -1,0 +1,22 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace AresService.Migrations.Sqlite.Migrations.AresDb
+{
+    /// <inheritdoc />
+    public partial class FixCorruptedParameterSources_AresDbContext : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+          migrationBuilder.Sql("UPDATE Parameters SET Source = '{}' WHERE Source = '0'");
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+
+        }
+    }
+}
