### 2026-06-18 570624c
Fixed logging overrides breaking the device state logging.
