+++ Ares.Core.Tests/Device/State/Logging/StateLoggerManagerTests.cs
@@ -1,0 +1,215 @@
+using Ares.Core.Device.Providers;
+using Ares.Core.Device.State.Logging;
+using Ares.Core.Tests.Data.Device;
+using Ares.Datamodel.Device;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
+using Moq;
+
+namespace Ares.Core.Tests.Device.State.Logging;
+
+internal class StateLoggerManagerTests
+{
+  private DbContextOptions<CoreDatabaseContext> _dbOptions;
+  private Mock<IDbContextFactory<CoreDatabaseContext>> _dbContextFactory;
+  private DeviceStateLoggerRepository _repository;
+  private Mock<IDeviceStateLoggerFactory> _loggerFactory;
+  private StateLoggerManager _manager;
+
+  [SetUp]
+  public void SetUp()
+  {
+    _dbOptions = new DbContextOptionsBuilder<CoreDatabaseContext>()
+      .UseInMemoryDatabase(Guid.NewGuid().ToString())
+      .Options;
+    _dbContextFactory = new Mock<IDbContextFactory<CoreDatabaseContext>>();
+    _dbContextFactory
+      .Setup(factory => factory.CreateDbContext())
+      .Returns(() => new CoreDatabaseContext(_dbOptions));
+    _repository = new DeviceStateLoggerRepository();
+    _loggerFactory = new Mock<IDeviceStateLoggerFactory>();
+    _manager = new StateLoggerManager(
+      _repository,
+      _loggerFactory.Object,
+      Mock.Of<ILogger<StateLoggerManager>>(),
+      _dbContextFactory.Object,
+      Mock.Of<IAresDeviceProvider>());
+  }
+
+  [Test]
+  public async Task EnableOverrideAsync_WithLoggingEnabled_AppliesEnabledClone()
+  {
+    var logger = AddLogger("device-1");
+    var settings = CreateSettings(loggingEnabled: false);
+    DeviceLoggingSettings appliedSettings = null;
+    logger
+      .Setup(stateLogger => stateLogger.UpdateSettings(It.IsAny<DeviceLoggingSettings>()))
+      .Callback<DeviceLoggingSettings>(value => appliedSettings = value)
+      .Returns(Task.CompletedTask);
+
+    await _manager.EnableOverrideAsync(settings, true);
+
+    Assert.Multiple(() =>
+    {
+      Assert.That(appliedSettings, Is.Not.Null);
+      Assert.That(appliedSettings!.LoggingEnabled, Is.True);
+      Assert.That(appliedSettings.LoggingType, Is.EqualTo(settings.LoggingType));
+      Assert.That(settings.LoggingEnabled, Is.False);
+      Assert.That(appliedSettings, Is.Not.SameAs(settings));
+    });
+  }
+
+  [Test]
+  public async Task EnableOverrideAsync_WithLoggingDisabled_AppliesDisabledSettings()
+  {
+    var logger = AddLogger("device-1");
+    DeviceLoggingSettings appliedSettings = null;
+    logger
+      .Setup(stateLogger => stateLogger.UpdateSettings(It.IsAny<DeviceLoggingSettings>()))
+      .Callback<DeviceLoggingSettings>(value => appliedSettings = value)
+      .Returns(Task.CompletedTask);
+
+    await _manager.EnableOverrideAsync(CreateSettings(loggingEnabled: true), false);
+
+    Assert.That(appliedSettings, Is.Not.Null);
+    Assert.That(appliedSettings!.LoggingEnabled, Is.False);
+  }
+
+  [Test]
+  public async Task DisableOverrideAsync_RestoresPersistedSettings()
+  {
+    var logger = AddLogger("device-1");
[…]
