### 2026-06-18 211288f
More logging bugfixes that optimizer thought were pertinent.
