+++ Ares.Core.Tests/Device/State/Logging/StateLoggerManagerTests.cs
@@ -54,6 +54,7 @@ internal class StateLoggerManagerTests
       Assert.That(appliedSettings, Is.Not.Null);
       Assert.That(appliedSettings!.LoggingEnabled, Is.True);
       Assert.That(appliedSettings.LoggingType, Is.EqualTo(settings.LoggingType));
+      Assert.That(appliedSettings.DeviceId, Is.EqualTo("device-1"));
       Assert.That(settings.LoggingEnabled, Is.False);
       Assert.That(appliedSettings, Is.Not.SameAs(settings));
     });
@@ -89,6 +90,28 @@ internal class StateLoggerManagerTests
       stateLogger => stateLogger.UpdateSettings(It.Is<DeviceLoggingSettings>(
         settings => settings.DeviceId == "device-1" && !settings.LoggingEnabled)),
       Times.Once);
+  }
+
+  [Test]
+  public async Task DisableOverrideAsync_WithoutPersistedSettings_DisablesLogger()
+  {
+    var logger = AddLogger("device-1");
+    DeviceLoggingSettings restoredSettings = null;
+    logger
+      .Setup(stateLogger => stateLogger.UpdateSettings(It.IsAny<DeviceLoggingSettings>()))
+      .Callback<DeviceLoggingSettings>(value => restoredSettings = value)
+      .Returns(Task.CompletedTask);
+
+    await _manager.EnableOverrideAsync(CreateSettings(loggingEnabled: true), true);
+    await _manager.DisableOverrideAsync();
+
+    Assert.Multiple(() =>
+    {
+      Assert.That(restoredSettings, Is.Not.Null);
+      Assert.That(restoredSettings.DeviceId, Is.EqualTo("device-1"));
+      Assert.That(restoredSettings.LoggingEnabled, Is.False);
+      Assert.That(restoredSettings.LoggingType, Is.EqualTo(DeviceLoggingSettings.Types.LoggingType.None));
+    });
   }
 
   [Test]
@@ -136,6 +159,7 @@ internal class StateLoggerManagerTests
     {
       Assert.That(startedSettings, Is.Not.Null);
       Assert.That(startedSettings!.LoggingEnabled, Is.True);
+      Assert.That(startedSettings.DeviceId, Is.EqualTo(device.UniqueId));
       Assert.That(startedSettings, Is.Not.SameAs(suppliedSettings));
     });
   }
@@ -172,7 +196,85 @@ internal class StateLoggerManagerTests
       Assert.That(startedSettings.LoggingType, Is.EqualTo(persistedSettings.LoggingType));
       Assert.That(startedSettings.IntervalMs, Is.EqualTo(persistedSettings.IntervalMs));
     });
-    newLogger.Verify(logger => logger.UpdateSettings(updatedSettings), Times.Once);
+    newLogger.Verify(
+      logger => logger.UpdateSettings(It.Is<DeviceLoggingSettings>(
+        settings => settings.DeviceId == updatedSettings.DeviceId
+          && settings.LoggingEnabled == updatedSettings.LoggingEnabled
+          && settings.LoggingType == updatedSettings.LoggingType
+          && settings.IntervalMs == updatedSettings.IntervalMs)),
+      Times.Once);
+  }
+
+  [Test]
+  public async Task RemoveLogger_PreservesSettingsByDefault()
+  {
+    AddLogger("device-1");
+    await SaveSettings(CreateSettings("device-1", loggingEnabled: true));
+
+    await _manager.RemoveLogger("device-1");
+
+    Assert.That(await _manager.GetDatabaseLoggerSettings("device-1"), Is.Not.Null);
+  }
+
+  [Test]
[…]
