+++ Ares.Core/Device/State/Logging/StateLoggerManager.cs
@@ -90,7 +90,9 @@ public class StateLoggerManager
 
       using var ctx = _dbContextFactory.CreateDbContext();
       var existingSettings = await ctx.DeviceLoggingSettings.FirstOrDefaultAsync(s => s.DeviceId == device.UniqueId);
-      var settings = _overrideActive ? _overrideSettings?.Clone() : existingSettings;
+      var settings = _overrideActive
+        ? CreateOverrideSettings(device.UniqueId)
+        : existingSettings ?? CreateDisabledSettings(device.UniqueId);
 
       await logger.Start(settings);
       _stateLoggerRepository.Add(device.UniqueId, logger);
@@ -101,7 +103,7 @@ public class StateLoggerManager
     }
   }
 
-  public async Task RemoveLogger(string deviceId)
+  public async Task RemoveLogger(string deviceId, bool removeSettings = false)
   {
     await _overrideLock.WaitAsync();
 
@@ -113,14 +115,19 @@ public class StateLoggerManager
         _stateLoggerRepository.Remove(deviceId);
       }
 
-      using var ctx = _dbContextFactory.CreateDbContext();
-      var existingSettings = await ctx.DeviceLoggingSettings.FirstOrDefaultAsync(s => s.DeviceId == deviceId);
-      if(existingSettings is not null)
+      if(removeSettings)
       {
-        ctx.DeviceLoggingSettings.Remove(existingSettings);
-      }
+        using var ctx = _dbContextFactory.CreateDbContext();
+        var existingSettings = await ctx.DeviceLoggingSettings.FirstOrDefaultAsync(s => s.DeviceId == deviceId);
+        if(existingSettings is not null)
+        {
+          ctx.DeviceLoggingSettings.Remove(existingSettings);
+          await ctx.SaveChangesAsync();
+        }
 
-      await ctx.SaveChangesAsync();
+        _logger.LogInformation("Removed logger and settings for device id {DeviceId}", deviceId);
+        return;
+      }
 
       _logger.LogInformation("Removed logger for device id {DeviceId}", deviceId);
     }
@@ -132,33 +139,33 @@ public class StateLoggerManager
 
   public async Task UpdateLogger(string deviceId, DeviceLoggingSettings settings)
   {
-    if(_stateLoggerRepository.TryGetValue(deviceId, out var logger))
+    await _overrideLock.WaitAsync();
+
+    try
     {
-      await _overrideLock.WaitAsync();
-
-
-      try
+      if(!_stateLoggerRepository.TryGetValue(deviceId, out var logger))
       {
-        await UpdateDatabase(deviceId, settings);
-
-        if(!_overrideActive)
-        {
-          await logger.UpdateSettings(settings);
-          _logger.LogInformation("Updated logger for device id {DeviceId}", deviceId);
-        }
+        throw new KeyNotFoundException($"No logger found for device {deviceId}");
       }
-      catch(Exception e)
+
+      var deviceSettings = settings.Clone();
+      deviceSettings.DeviceId = deviceId;
+      await UpdateDatabase(deviceId, deviceSettings);
+
+      if(!_overrideActive)
       {
-        _logger.LogError("Error updating device state logger for device {DeviceId}: {Exception}", deviceId, e);
-      }
-      finally
-      {
-        _overrideLock.Release();
+        await logger.UpdateSettings(deviceSettings);
+        _logger.LogInformation("Updated logger for device id {DeviceId}", deviceId);
       }
     }
-    else
+    catch(Exception e)
     {
-      throw new KeyNotFoundException($"No logger found for device {deviceId}");
+      _logger.LogError("Error updating device state logger for device {DeviceId}: {Exception}", deviceId, e);
+      throw;
+    }
+    finally
+    {
+      _overrideLock.Release();
     }
   }
[…]
