### 2026-06-18 a4af650
Device state logging fixins (#190)
* Fixed logging overrides breaking the device state logging.

* More logging bugfixes that optimizer thought were pertinent.
