+++ Ares.Core/Device/State/Logging/StateLoggerManager.cs
@@ -20,6 +20,7 @@ public class StateLoggerManager
   private readonly IAresDeviceProvider _deviceProvider;
   private readonly CompositeDisposable _cleanup = new();
   private bool _overrideActive;
+  private DeviceLoggingSettings? _overrideSettings;
   private readonly SemaphoreSlim _overrideLock = new(1, 1);
 
   public StateLoggerManager(IDeviceStateLoggerRepository stateLoggerRepository,
@@ -68,74 +69,103 @@ public class StateLoggerManager
 
   public async Task SetupLogger(IAresDevice device)
   {
-    // Stop and remove any existing logger for the device
-    if(_stateLoggerRepository.TryGetValue(device.UniqueId, out var existingLogger))
+    await _overrideLock.WaitAsync();
+
+    try
     {
-      await existingLogger.Stop();
-      _stateLoggerRepository.Remove(device.UniqueId);
-    }
+      // Stop and remove any existing logger for the device
+      if(_stateLoggerRepository.TryGetValue(device.UniqueId, out var existingLogger))
+      {
+        await existingLogger.Stop();
+        _stateLoggerRepository.Remove(device.UniqueId);
+      }
 
-    if(_deviceLoggerFactory is null)
+      if(_deviceLoggerFactory is null)
+      {
+        _logger.LogError("No suitable logger factory found for device type {DeviceType}", device.GetType().Name);
+        return;
+      }
+
+      var logger = _deviceLoggerFactory.Create(device);
+
+      using var ctx = _dbContextFactory.CreateDbContext();
+      var existingSettings = await ctx.DeviceLoggingSettings.FirstOrDefaultAsync(s => s.DeviceId == device.UniqueId);
+      var settings = _overrideActive
+        ? CreateOverrideSettings(device.UniqueId)
+        : existingSettings ?? CreateDisabledSettings(device.UniqueId);
+
+      await logger.Start(settings);
+      _stateLoggerRepository.Add(device.UniqueId, logger);
+    }
+    finally
     {
-      _logger.LogError("No suitable logger factory found for device type {DeviceType}", device.GetType().Name);
-      return;
+      _overrideLock.Release();
     }
-
-    var logger = _deviceLoggerFactory.Create(device);
-
-    using var ctx = _dbContextFactory.CreateDbContext();
-    var existingSettings = await ctx.DeviceLoggingSettings.FirstOrDefaultAsync(s => s.DeviceId == device.UniqueId);
-
-    await logger.Start(existingSettings);
-    _stateLoggerRepository.Add(device.UniqueId, logger);
   }
 
-  public async Task RemoveLogger(string deviceId)
+  public async Task RemoveLogger(string deviceId, bool removeSettings = false)
   {
-    if(_stateLoggerRepository.TryGetValue(deviceId, out var logger))
-    {
-      await logger.Stop();
-      _stateLoggerRepository.Remove(deviceId);
-    }
+    await _overrideLock.WaitAsync();
 
-    using var ctx = _dbContextFactory.CreateDbContext();
-    var existingSettings = await ctx.DeviceLoggingSettings.FirstOrDefaultAsync(s => s.DeviceId == deviceId);
-    if(existingSettings is not null)
+    try
     {
-      ctx.DeviceLoggingSettings.Remove(existingSettings);
-    }
+      if(_stateLoggerRepository.TryGetValue(deviceId, out var logger))
+      {
+        await logger.Stop();
+        _stateLoggerRepository.Remove(deviceId);
+      }
 
-    _logger.LogInformation("Removed logger for device id {DeviceId}", deviceId);
+      if(removeSettings)
+      {
[…]
