### 2026-06-22 0c13fb3
Cleanup to the SiLA implementation
