+++ Ares.Core/Device/Sila/SilaDevice.cs
@@ -8,26 +8,32 @@ using System.Reactive.Linq;
 using System.Reactive.Subjects;
 using Tecan.Sila2;
 using Tecan.Sila2.DynamicClient;
-using Grpc.Core;
 
 namespace Ares.Core.Device.Sila;
 
 public sealed class SilaDevice : AresDevice, IAsyncDisposable
 {
   private readonly SilaClient _silaClient;
-  private DeviceCommandDescriptor[] _commands = [];
+  private readonly ServerData _serverData;
   private readonly BehaviorSubject<AresStruct> _stateSubject = new(new AresStruct());
+  private readonly ConcurrentDictionary<string, AresValue> _currentPropertyValues = new();
+
+  private DeviceCommandDescriptor[] _commands = [];
   private CancellationTokenSource _stateStreamCts = new();
   private DevicePollingSettings _pollingSettings = new();
-  private ServerData _serverData;
-  private readonly ConcurrentDictionary<string, AresValue> _currentPropertyValues = new();
   private Task? _monitoringTask;
+
+  public string Address { get; private set; }
+  private IEnumerable<Feature> DeviceFeatures { get; set; } = [];
+
+  public override IObservable<AresStruct> StateStream => _stateSubject.AsObservable();
 
   public SilaDevice(ServerData serverData, DeviceConnectionInfo connectionInfo, SilaClient client) : base(connectionInfo)
   {
-    _serverData = serverData;
-    DeviceFeatures = _serverData.Features;
-    _silaClient = client;
+    _serverData = serverData ?? throw new ArgumentNullException(nameof(serverData));
+    _silaClient = client ?? throw new ArgumentNullException(nameof(client));
+
+    DeviceFeatures = _serverData.Features.ToArray();
 
     Description = serverData.Info.Description;
     Version = serverData.Info.Version;
@@ -35,19 +41,14 @@ public sealed class SilaDevice : AresDevice, IAsyncDisposable
     Address = serverData.Address;
   }
 
-  public override IObservable<AresStruct> StateStream => _stateSubject.AsObservable();
-
-  //public AresStructSchema StateSchema { get; private set; } = new();
-
   public override async Task<bool> Activate(CancellationToken ct)
   {
-    DeviceFeatures = _serverData.Features.ToArray();
     BuildStateSchema();
 
-    Status = new DeviceOperationalStatus() 
-    { 
-      OperationalState = OperationalState.Active, 
-      Message = $"Connected to SiLA Device {Name}!" 
+    Status = new DeviceOperationalStatus()
+    {
+      OperationalState = OperationalState.Active,
+      Message = $"Connected to SiLA Device {Name}!"
     };
 
     await StartPropertyMonitoring();
@@ -63,8 +64,8 @@ public sealed class SilaDevice : AresDevice, IAsyncDisposable
       {
         var key = $"{feature.Identifier}.{property.Identifier}";
         var fieldSchema = property.DataType is not null
-          ? SilaDataConverter.ToAresValueSchema(property.DataType)
-          : AresSchemaBuilder.Entry(AresDataType.Any).Build();
+            ? SilaDataConverter.ToAresValueSchema(property.DataType)
+            : AresSchemaBuilder.Entry(AresDataType.Any).Build();
 
         if(!string.IsNullOrWhiteSpace(property.Description))
           fieldSchema.Description = property.Description;
@@ -87,7 +88,7 @@ public sealed class SilaDevice : AresDevice, IAsyncDisposable
 
     foreach(var feature in DeviceFeatures)
     {
-      if(feature.Identifier.Contains("SiLAService")) 
+      if(feature.Identifier.Contains("SiLAService"))
         continue;
 
       var properties = (feature.Items ?? []).OfType<FeatureProperty>();
[…]
