+++ UI/ServiceCollectionExtensions.cs
@@ -27,7 +27,6 @@ using UI.Features.Devices;
 using UI.Features.Devices.Plugin;
 using UI.Features.Devices.Remote.Factory;
 using UI.Features.Devices.Sila;
-using UI.Features.DeviceStateExport;
 using UI.Features.DeviceStateLogging;
 using UI.Features.DeviceStateLogging.Settings;
 using UI.Features.Notifications;
