### 2026-06-22 7d521f7
Refactored the planning helper class to move towards supporting batch planning and to make things more readable. Also enabled more efficient requesting of plans by firing all requests at once rather than one at a time.
