+++ Ares.Core/Planning/PlanningHelper.cs
@@ -8,6 +8,7 @@ using Ares.Datamodel.Templates;
 using Google.Protobuf.WellKnownTypes;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
+using System.Collections.Concurrent;
 
 namespace Ares.Core.Planning;
 
@@ -38,101 +39,227 @@ public class PlanningHelper : IPlanningHelper
     CancellationToken cancellationToken)
   {
     var parameterArray = parameters.ToArray();
+    var plannerToMetadataMaps = MapParameterMetadataToPlanners(plannerAllocations);
+
+    if(plannerToMetadataMaps is null)
+      return false;
+
+    var planGroup = plannerToMetadataMaps.GroupBy(pair => pair.Planner);
+    var seedAnalysesArr = seedAnalyses.ToArray();
+
+    var planningTasks = planGroup.Select(async grouping =>
+    {
+      var planner = grouping.Key;
+
+      var planQueue = PlannerQueueDictionary.GetOrAdd(planner.UniqueId, _ => new ConcurrentQueue<Plan>());
+
+      if(planQueue.TryDequeue(out var plan))
+      {
+        ResolveParametersFromPlan(plan, parameterArray);
+        return true;
+      }
+
+      return await RequestNewPlans(planner, grouping, seedExperiments, seedAnalysesArr, statusCode, metadata, planQueue, parameterArray, cancellationToken);
+    });
+
+    var results = await Task.WhenAll(planningTasks);
+    return results.All(success => success);
+  }
+
+  /// <summary>
+  /// Requests a fresh set of plans from a planner service, resolves using the first one received and queues any additional.
+  /// </summary>
+  /// <param name="planner">The planner service to request plans from</param>
+  /// <param name="grouping">The grouping of planner and parameter metadata</param>
+  /// <param name="seedExperiments">Seed experiments for the planning request</param>
+  /// <param name="seedAnalysesArr">Seed analyses for the planning request</param>
+  /// <param name="statusCode">The status code of the plan</param>
+  /// <param name="metadata">Request metadata</param>
+  /// <param name="planQueue"></param>
+  /// <param name="parameterArray"></param>
+  /// <param name="cancellationToken"></param>
+  /// <returns></returns>
+  private async Task<bool> RequestNewPlans(IPlannerService planner, 
+    IGrouping<IPlannerService, (IPlannerService, ParameterMetadata)> grouping,
+    IEnumerable<ExperimentOverview> seedExperiments,
+    Analysis[] seedAnalysesArr,
+    PlanStatusCode statusCode,
+    RequestMetadata metadata,
+    ConcurrentQueue<Plan> planQueue,
+    Parameter[] parameterArray,
+    CancellationToken cancellationToken)
+  {
+    var planTransaction = new PlannerTransaction()
+    {
+      UniqueId = Guid.NewGuid().ToString(),
+      PlannerName = planner.Name,
+      PlannerType = planner.Type,
+      PlannerVersion = planner.Version,
+      PlannerId = planner.UniqueId
+    };
+
+    try
+    {
+      var plannableParameters = grouping.Select(pair => pair.Item2).ToArray();
+      planTransaction.TimeRequestSent = DateTime.UtcNow.ToTimestamp();
+
+      //Create the plan request. Store it in the transaction.
+      var planRequest = CreatePlanningRequest(plannableParameters, seedExperiments, seedAnalysesArr, statusCode, metadata);
+      planTransaction.PlanningRequest = planRequest;
+
+      var planResponse = await planner.Plan(planRequest, cancellationToken);
+      planTransaction.TimeResponseReceived = DateTime.UtcNow.ToTimestamp();
+      planTransaction.PlanningResponse = planResponse;
+
+      if(planResponse.Plans.Any())
+      {
[…]
