+++ Ares.Core.Tests/Execution/CampaignExecutorTests.cs
@@ -59,6 +59,7 @@ internal class CampaignExecutorTests
         It.IsAny<IEnumerable<Parameter>>(),
         It.IsAny<IEnumerable<Ares.Datamodel.Analyzing.Analysis>>(),
         It.IsAny<IEnumerable<ExperimentOverview>>(),
+        It.IsAny<int>(),
         It.IsAny<CancellationToken>()))
       .ReturnsAsync(true);
 
@@ -218,6 +219,7 @@ internal class CampaignExecutorTests
       It.IsAny<IEnumerable<Parameter>>(),
       It.IsAny<IEnumerable<Ares.Datamodel.Analyzing.Analysis>>(),
       It.IsAny<IEnumerable<ExperimentOverview>>(),
+      It.IsAny<int>(),
       It.IsAny<CancellationToken>()), Times.Once);
   }
 
