### 2026-06-23 e1c357b
Additional changes to support batch planning in ARES
