+++ Ares.Core.Tests/Execution/CampaignExecutorTests.cs
@@ -60,7 +60,7 @@ internal class CampaignExecutorTests
         It.IsAny<IEnumerable<Ares.Datamodel.Analyzing.Analysis>>(),
         It.IsAny<IEnumerable<ExperimentOverview>>(),
         It.IsAny<int>(),
-        It.IsAny<PlanStatusCode>(),
+        It.IsAny<List<PlanStatusCode>>(),
         It.IsAny<CancellationToken>()))
       .ReturnsAsync(true);
 
@@ -221,7 +221,7 @@ internal class CampaignExecutorTests
       It.IsAny<IEnumerable<Ares.Datamodel.Analyzing.Analysis>>(),
       It.IsAny<IEnumerable<ExperimentOverview>>(),
       It.IsAny<int>(),
-      It.IsAny<PlanStatusCode>(),
+      It.IsAny<List<PlanStatusCode>>(),
       It.IsAny<CancellationToken>()), Times.Once);
   }
 
