### 2026-06-23 eb38a04
Fix a bug that causes the UI to hang when pressing stop during a remote device command
