### 2026-06-24 b62f9b5
Updated datamodel version and added support for campaign start time
