### 2026-06-25 462570f
Added entity configs to the custom commands
