### 2026-06-25 d2bcd0f
Fix little script bug where comment right after function definition was marked as an error.
