### 2026-06-25 7a5cc17
Script editor is now just responsible just for editing the script body. The function definition is rendered separately.
