### 2026-06-25 38305d4
Added another card for diagnostics feedback of the command.
