### 2026-06-26 d3747c9
Updated ARES to support the new analysis capabilities. Still some work to do
