### 2026-06-26 3626024
Added persistence and the migrations for the custom commands
