### 2026-06-26 a2fadd4
Fixed some indexing issues
