### 2026-06-29 76768d5
Fixed the display charts for analysis to support displaying multiple objectives
