### 2026-06-29 d3b3eef
Batch planning (#193)
* Added the new error states, as well as an enum to define the behaviors available for the user to select from.

* Error handling updates

* Settings menu now has a functionable selection process for handling errors

* Working towards integrating the new settings into the loop. First with retying commands, which is done at the step execution level when applicable. Next is going to be dealing with things like replanning and safe mode entry.

* Spacing change

* Working on adding some general settings

* Forgot some additional changes

* Some cleanup and additional UI for g
[…]
