+++ Ares.Core/Execution/ExecutionManager.cs
@@ -55,7 +55,7 @@ public class ExecutionManager : IExecutionManager
     var startConditions = await Task.WhenAll(startConditionTasks);
     return startConditions.All(condition => condition?.Success ?? true);
   }
-  public int ReplanRate { get; private set; } = 1;
+
   public async Task Start(string executionNotes, List<AresCampaignTag> campaignTags)
   {
     var err = await CheckCampaignStartPrerequisites();
@@ -73,7 +73,8 @@ public class ExecutionManager : IExecutionManager
       executor.UpdateCampaignTags(campaignTags);
 
     executor.StopConditions.AddRange(CampaignStopConditions);
-    executor.ReplanRate = ReplanRate;
+    executor.ReplicateRate = ReplicateRate;
+    executor.BatchPlanningSize = PlanningBatchSize;
     _executionControlTokenSource = new ExecutionControlTokenSource();
     CampaignExecutionSummary campaignExecutionSummary;
     ExecutionStartTime = DateTime.UtcNow;
@@ -111,7 +112,7 @@ public class ExecutionManager : IExecutionManager
   public async Task<string> CheckCampaignStartPrerequisites()
   {
     if(_activeCampaignTemplateStore.CampaignTemplate is null)
-      return "CampaignTemplate was not assigned to the active template store.";
+      return "Campaign Template was not assigned to the active template store.";
 
     if(!CampaignStopConditions.Any())
       return "The Campaign has no stop conditions, please set a stop condition before starting campaign.";
@@ -142,7 +143,12 @@ public class ExecutionManager : IExecutionManager
 
   public void UpdateReplicateRate(int newRate)
   {
-    ReplanRate = newRate;
+    ReplicateRate = newRate;
+  }
+
+  public void UpdateBatchPlanningSize(int newBatchSize)
+  {
+    PlanningBatchSize = newBatchSize;
   }
 
   public void SubmitUserDecision(ErrorHandling decision)
@@ -171,8 +177,11 @@ public class ExecutionManager : IExecutionManager
     {
       throw;
     }
-
   }
 
   public DateTime? ExecutionStartTime { get; set; }
+
+  public int ReplicateRate { get; private set; } = 1;
+
+  public int PlanningBatchSize { get; private set; } = 1;
 }
