+++ Ares.Core/Planning/IPlanningHelper.cs
@@ -23,6 +23,8 @@ public interface IPlanningHelper
     IEnumerable<Parameter> parameters,
     IEnumerable<Analysis> seedAnalyses,
     IEnumerable<ExperimentOverview> seedExperiments,
+    int batchSize,
+    List<PlanStatusCode> codes,
     CancellationToken cancellationToken);
 
   /// <summary>
