+++ Ares.Core/Execution/Executors/CampaignExecutor.cs
@@ -3,6 +3,7 @@ using Ares.Core.AresEnvironment;
 using Ares.Core.Device.State.Logging;
 using Ares.Core.Exceptions;
 using Ares.Core.Execution.ControlTokens;
+using Ares.Core.Execution.Enums;
 using Ares.Core.Execution.Executors.Composers;
 using Ares.Core.Execution.Extensions;
 using Ares.Core.Execution.Safety;
@@ -13,6 +14,7 @@ using Ares.Core.Planning;
 using Ares.Core.Settings;
 using Ares.Datamodel;
 using Ares.Datamodel.Analyzing;
+using Ares.Datamodel.Planning;
 using Ares.Datamodel.Templates;
 using Google.Protobuf.WellKnownTypes;
 using Microsoft.Extensions.Logging;
@@ -22,28 +24,7 @@ using System.Reactive.Subjects;
 namespace Ares.Core.Execution.Executors;
 
 public class CampaignExecutor : ICampaignExecutor
-{
-  private enum ExperimentPhase
-  {
-    Initialize,
-    Plan,
-    Compose,
-    Execute,
-    Analyze,
-    Retry,
-    Replan,
-    Complete,
-    Failed,
-    Canceled
-  }
-
-  private enum ExperimentLoopOutcome
-  {
-    Succeeded,
-    Failed,
-    Canceled
-  }
-
+{ 
   private readonly IExecutionReporter _executionReporter;
   private readonly ISubject<CampaignExecutionStatus> _executionStatusSubject;
   private readonly ICommandComposer<ExperimentTemplate, ExperimentExecutor> _experimentComposer;
@@ -66,6 +47,7 @@ public class CampaignExecutor : ICampaignExecutor
   private ExperimentTemplate? _currentExperimentTemplate = null;
   private int _experimentCount = 0;
   private TaskCompletionSource<ErrorHandling>? _userDecisionSource;
+  private List<PlanStatusCode> _latestPlanStatusCodes = new List<PlanStatusCode>();
 
   internal CampaignExecutor(ICommandComposer<ExperimentTemplate, ExperimentExecutor> experimentComposer,
     IPlanningHelper planningHelper,
@@ -226,8 +208,13 @@ public class CampaignExecutor : ICampaignExecutor
     return (true, startupSummary);
   }
 
-  private async Task<ExperimentLoopOutcome> ExecuteExperimentLoop(string campaignPath, List<Analysis> analyses, List<ExperimentExecutionSummary> experimentSummaries, ExecutionControlToken token, ExperimentExecutionSummary startupSummary)
+  private async Task<ExperimentLoopOutcome> ExecuteExperimentLoop(string campaignPath, 
+    List<Analysis> analyses, 
+    List<ExperimentExecutionSummary> experimentSummaries, 
+    ExecutionControlToken token, 
+    ExperimentExecutionSummary startupSummary)
   {
+    _latestPlanStatusCodes = new List<PlanStatusCode>();
     var currentPhase = ExperimentPhase.Initialize;
     var currentExperimentPath = "";
     var failedExperimentRetryCount = 0;
@@ -317,7 +304,7 @@ public class CampaignExecutor : ICampaignExecutor
     if(!_currentExperimentTemplate.IsResolved())
     {
       _logger.LogTrace("Experiment has not been resolved, ARES will now begin the planning process.");
-      if(analyses.Count % ReplanRate == 0)
+      if(analyses.Count % ReplicateRate == 0)
       {
         if(!await PlanExperiment(analyses, _currentExperimentTemplate, experimentSummaries, token))
         {
@@ -397,9 +384,17 @@ public class CampaignExecutor : ICampaignExecutor
[…]
