### 2026-06-30 c8f1e90
Playing around with the idea of reworking some of the execution process, inspired by the need to make the system level actions (sleep, wait for user, get user input, etc.) work better within the ARES architecture.
