+++ Ares.Core/Execution/System/SystemCommandParameters.cs
@@ -1,0 +1,11 @@
+﻿namespace Ares.Core.Execution.System;
+
+public enum SystemCommandParameters
+{
+  Duration,
+  NumericData,
+  IntData,
+  FloatData,
+  UserInput,
+  UserPrompt
+}
