+++ Ares.Core/Execution/Interaction/InteractionType.cs
@@ -1,0 +1,7 @@
+﻿namespace Ares.Core.Execution.Interaction;
+
+public enum InteractionType
+{
+  Confirmation,
+  DataInput
+}
