+++ Ares.Core/Execution/Interaction/InteractionRequest.cs
@@ -1,0 +1,7 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Ares.Core.Execution.Interaction;
+
+public record InteractionRequest(string RequestId, InteractionType Type, string Prompt);

[…]
