### 2026-07-02 c55f93b
Custom command execution device skeleton
