+++ Ares.Core/CustomCommands/ICustomCommandPersistenceService.cs
@@ -1,4 +1,4 @@
-using CustomCommandModel = Ares.Datamodel.Automation.CustomCommand;
+using Ares.Datamodel.Automation;
 
 namespace Ares.Core.CustomCommands;
 
@@ -6,9 +6,11 @@ public interface ICustomCommandPersistenceService
 {
   Task<IReadOnlyList<CustomCommandSummary>> GetSummariesAsync();
 
-  Task<CustomCommandModel?> GetAsync(Guid id);
+  Task<IReadOnlyList<CustomCommand>> GetCommandsAsync();
 
-  Task<Guid> SaveAsync(Guid? id, CustomCommandModel command);
+  Task<CustomCommand?> GetAsync(Guid id);
+
+  Task<Guid> SaveAsync(Guid? id, CustomCommand command);
 
   Task DeleteAsync(Guid id);
 }
