### 2026-07-09 a8b593f
Fixed campaign execution screen not properly displaying startup and closeout in tree view.
