### 2026-07-09 4cea149
Added support for multiple command types. The types included so far are system command, device command, and custom command.
