+++ Ares.Core/CustomCommands/CustomCommandPersistenceService.cs
@@ -1,6 +1,10 @@
+using Ares.Datamodel.Automation;
 using Ares.Datamodel.Extensions;
+using Google.Protobuf;
 using Microsoft.EntityFrameworkCore;
+using System.Security.Cryptography;
 using CustomCommandModel = Ares.Datamodel.Automation.CustomCommand;
+using CustomCommandVersionModel = Ares.Datamodel.Automation.CustomCommandVersion;
 
 namespace Ares.Core.CustomCommands;
 
@@ -11,68 +15,64 @@ internal sealed class CustomCommandPersistenceService(IDbContextFactory<CoreData
 
   public async Task<IReadOnlyList<CustomCommandSummary>> GetSummariesAsync()
   {
-    await using var context = await dbContextFactory.CreateDbContextAsync();
-    var rows = await context.CustomCommands
-      .AsNoTracking()
-      .Include(command => command.InputParameters)
+    var commands = await GetCurrentVersionsAsync();
+    return commands
+      .Select(command => new CustomCommandSummary(
+        ParseEntityId(command.CustomCommandId),
+        string.IsNullOrWhiteSpace(command.Name) ? "(Unnamed command)" : command.Name,
+        command.Description,
+        BuildInputSummary(command),
+        BuildOutputSummary(command)))
       .OrderBy(command => command.Name)
-      .Select(command => new
-      {
-        Id = EF.Property<string>(command, UniqueIdPropertyName),
-        Command = command
-      })
-      .ToListAsync();
-
-    return rows
-      .Select(row => new CustomCommandSummary(
-        ParseEntityId(row.Id),
-        string.IsNullOrWhiteSpace(row.Command.Name) ? "(Unnamed command)" : row.Command.Name,
-        row.Command.Description,
-        BuildInputSummary(row.Command),
-        BuildOutputSummary(row.Command)))
       .ToArray();
   }
 
-  public async Task<CustomCommandModel?> GetAsync(Guid id)
+  public async Task<CustomCommandVersionModel?> GetAsync(Guid id)
   {
     await using var context = await dbContextFactory.CreateDbContextAsync();
-    return await context.CustomCommands
+    var currentVersionId = await context.CustomCommands
       .AsNoTracking()
-      .Include(command => command.InputParameters)
-      .FirstOrDefaultAsync(command => EF.Property<string>(command, UniqueIdPropertyName) == id.ToString());
+      .Where(command => EF.Property<string>(command, UniqueIdPropertyName) == id.ToString())
+      .Select(command => command.CurrentVersionId)
+      .FirstOrDefaultAsync();
+
+    return string.IsNullOrWhiteSpace(currentVersionId)
+      ? null
+      : await GetVersionAsync(context, currentVersionId);
   }
 
-  public async Task<Guid> SaveAsync(Guid? id, CustomCommandModel command)
+  public async Task<Guid> SaveAsync(Guid? id, CustomCommandVersionModel command)
   {
     await using var context = await dbContextFactory.CreateDbContextAsync();
     var commandId = id ?? Guid.NewGuid();
     var existingCommand = id is null
       ? null
       : await context.CustomCommands
-        .Include(command => command.InputParameters)
         .FirstOrDefaultAsync(command => EF.Property<string>(command, UniqueIdPropertyName) == id.Value.ToString());
+
+    var nextVersionNumber = existingCommand is null
+      ? 1
+      : await context.CustomCommandVersions
+        .Where(version => version.CustomCommandId == commandId.ToString())
+        .Select(version => (long?)version.VersionNumber)
+        .MaxAsync() ?? 0;
+
+    if(existingCommand is not null)
+      nextVersionNumber++;
+
+    var version = CreateVersion(commandId, nextVersionNumber, command);
[…]
