### 2026-07-09 c0af44d
Gutted the ares core device and moved its functionality to a system operation executor.
