+++ Ares.Core/CoreDevice/AresCoreDeviceCommandParameter.cs
@@ -1,8 +1,0 @@
-﻿namespace Ares.Core.CoreDevice;
-public enum AresCoreDeviceCommandParameter
-{
-  Duration,
-  NumericData,
-  IntData,
-  FloatData
-}
