### 2026-07-10 9a2b78f
Added an executor for the custom command.
