### 2026-07-10 81b597b
Add support for actually creating the different available command template types.
