+++ UI/Features/CampaignEdit/ViewModels/CommandDesignerViewModel.cs
@@ -1,4 +1,8 @@
-﻿using Ares.Core.Grpc.Services;
+using Ares.Core.CustomCommands;
+using Ares.Core.Execution.Executors;
+using Ares.Core.Grpc.Services;
+using Ares.Datamodel;
+using Ares.Datamodel.Automation;
 using Ares.Datamodel.Extensions;
 using Ares.Datamodel.Templates;
 using Ares.Services.Device;
@@ -13,61 +17,136 @@ public partial class CommandDesignerViewModel : ReactiveObject
   private readonly CommandParameterDesignerFactory _commandParameterDesignerFactory;
   private readonly MetadataPickerFactory _metadataPickerFactory;
   private readonly DevicesService _devicesClient;
+  private readonly ICustomCommandPersistenceService _customCommandPersistenceService;
+  private readonly Task _initializationTask;
   private CommandMetadata? _commandMetadata;
-  private CommandTemplate _commandTemplate = null!;
+  private CustomCommandVersion? _selectedCustomCommand;
   private CommandOutputVariableReference[] _availableVariableReferences = [];
 
   public CommandDesignerViewModel(
     CommandTemplate existingTemplate,
     CommandParameterDesignerFactory commandParameterDesignerFactory,
     MetadataPickerFactory metadataPickerFactory,
-    DevicesService devicesClient)
+    DevicesService devicesClient,
+    ICustomCommandPersistenceService customCommandPersistenceService)
   {
     ArgumentDesigners = [];
+    AvailableCustomCommands = [];
     _commandParameterDesignerFactory = commandParameterDesignerFactory;
     _metadataPickerFactory = metadataPickerFactory;
     _devicesClient = devicesClient;
-
+    _customCommandPersistenceService = customCommandPersistenceService;
     CommandTemplate = existingTemplate;
+    _initializationTask = InitializeAsync(existingTemplate);
   }
 
-  public CommandDesignerViewModel(CommandParameterDesignerFactory commandParameterDesignerFactory, MetadataPickerFactory metadataPickerFactory, DevicesService devicesClient)
+  public CommandDesignerViewModel(
+    CommandParameterDesignerFactory commandParameterDesignerFactory,
+    MetadataPickerFactory metadataPickerFactory,
+    DevicesService devicesClient,
+    ICustomCommandPersistenceService customCommandPersistenceService)
+    : this(
+      new CommandTemplate
+      {
+        UniqueId = Guid.NewGuid().ToString(),
+        DeviceCommand = new DeviceCommand()
+      },
+      commandParameterDesignerFactory,
+      metadataPickerFactory,
+      devicesClient,
+      customCommandPersistenceService)
   {
-    ArgumentDesigners = [];
-    _commandParameterDesignerFactory = commandParameterDesignerFactory;
-    _metadataPickerFactory = metadataPickerFactory;
-    _devicesClient = devicesClient;
-
-    CommandTemplate = new CommandTemplate
-    {
-      UniqueId = Guid.NewGuid().ToString(),
-      DeviceCommand = new DeviceCommand()
-    };
   }
 
-  public CommandTemplate CommandTemplate
-  {
-    get => _commandTemplate;
-
-    set
-    {
-      _commandTemplate = value;
-      CommandMetadata = value.DeviceCommand?.Metadata;
-      InitTemplate(value);
-    }
-  }
+  public CommandTemplate CommandTemplate { get; }
 
   public int Index { get; set; }
 
-  public string? TemplateDeviceName { get; private set; }
[…]
