### 2026-07-11 081a28b
Forgot to commit the system operations catalog
