### 2026-07-11 b135a7f
Added migrations for the custom command data stuff
