+++ AresService.Migrations.Postgres/Migrations/AresDbContextModelSnapshot.cs
@@ -321,22 +321,13 @@ namespace AresService.Migrations.Postgres.Migrations
                         .HasColumnType("timestamp with time zone")
                         .HasDefaultValueSql("NOW()");
 
-                    b.Property<string>("Description")
+                    b.Property<string>("CurrentVersionId")
                         .HasColumnType("text");
 
                     b.Property<DateTime>("LastModified")
                         .ValueGeneratedOnAddOrUpdate()
                         .HasColumnType("timestamp with time zone")
                         .HasDefaultValueSql("NOW()");
-
-                    b.Property<string>("Name")
-                        .HasColumnType("text");
-
-                    b.Property<string>("OutputSchema")
-                        .HasColumnType("text");
-
-                    b.Property<string>("ScriptBody")
-                        .HasColumnType("text");
 
                     b.HasKey("UniqueId");
 
@@ -354,7 +345,7 @@ namespace AresService.Migrations.Postgres.Migrations
                         .HasColumnType("timestamp with time zone")
                         .HasDefaultValueSql("NOW()");
 
-                    b.Property<Guid>("CustomCommandId")
+                    b.Property<Guid>("CustomCommandVersionId")
                         .HasColumnType("uuid");
 
                     b.Property<DateTime>("LastModified")
@@ -370,9 +361,51 @@ namespace AresService.Migrations.Postgres.Migrations
 
                     b.HasKey("UniqueId");
 
-                    b.HasIndex("CustomCommandId");
+                    b.HasIndex("CustomCommandVersionId");
 
                     b.ToTable("CustomCommandParameters", (string)null);
+                });
+
+            modelBuilder.Entity("Ares.Datamodel.Automation.CustomCommandVersion", b =>
+                {
+                    b.Property<Guid>("UniqueId")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("uuid");
+
+                    b.Property<DateTime>("CreationTime")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("timestamp with time zone")
+                        .HasDefaultValueSql("NOW()");
+
+                    b.Property<Guid>("CustomCommandId")
+                        .HasColumnType("uuid");
+
+                    b.Property<string>("Description")
+                        .HasColumnType("text");
+
+                    b.Property<DateTime>("LastModified")
+                        .ValueGeneratedOnAddOrUpdate()
+                        .HasColumnType("timestamp with time zone")
+                        .HasDefaultValueSql("NOW()");
+
+                    b.Property<string>("Name")
+                        .HasColumnType("text");
+
+                    b.Property<string>("OutputSchema")
+                        .HasColumnType("text");
+
+                    b.Property<string>("ScriptBody")
+                        .HasColumnType("text");
+
+                    b.Property<long>("VersionNumber")
+                        .HasColumnType("bigint");
+
+                    b.HasKey("UniqueId");
+
+                    b.HasIndex("CustomCommandId", "VersionNumber")
+                        .IsUnique();
+
+                    b.ToTable("CustomCommandVersions", (string)null);
                 });
 
             modelBuilder.Entity("Ares.Datamodel.CampaignExecutionSummary", b =>
@@ -1361,9 +1394,6 @@ namespace AresService.Migrations.Postgres.Migrations
                         .ValueGeneratedOnAdd()
                         .HasColumnType("uuid");
 
-                    b.Property<Guid>("CommandTemplateId")
-                        .HasColumnType("uuid");
-
                     b.Property<DateTime>("CreationTime")
                         .ValueGeneratedOnAdd()
[…]
