### 2026-07-11 2356017
Fixed custom commands not having proper output schema for later reference.
