### 2026-07-11 b1ad7b9
Fixed funny script translation issues when it comes to custom command creation.
