+++ Ares.Core/Execution/Executors/ICommandDisplayNameResolver.cs
@@ -1,0 +1,10 @@
+using Ares.Datamodel.Templates;
+
+namespace Ares.Core.Execution.Executors;
+
+public interface ICommandDisplayNameResolver
+{
+  Task RefreshAsync();
+
+  string Resolve(CommandTemplate template);
+}

[…]
