### 2026-07-11 ef4d5b5
Made campaign templates be stored and retrieved from the database.
