+++ Ares.Core.Grpc/Services/AutomationService.cs
@@ -1,13 +1,13 @@
 using System.Threading;
 using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using System.Reactive.Linq;
 using System.Reactive.Threading.Tasks;
 using System.Text.Json;
 using System.Threading.Tasks;
 using Ares.Core.Analyzing;
+using Ares.Core.Campaigns;
 using Ares.Core.EntityConfigurations.Helpers;
 using Ares.Core.Execution;
 using Ares.Core.Execution.StartConditions;
@@ -40,6 +40,7 @@ public class AutomationService : AresAutomation.AresAutomationBase
   private readonly IPlannerServiceRepo _plannerServiceRepo;
   private readonly IPlannerTransactionProvider _plannerTransactionProvider;
   private readonly IAnalyzerTransactionProvider _analyzerTransactionProvider;
+  private readonly ICampaignTemplatePersistenceService _campaignTemplatePersistenceService;
 
   public AutomationService(IDbContextFactory<CoreDatabaseContext> coreContextFactory,
     IExecutionManager executionManager,
@@ -50,7 +51,8 @@ public class AutomationService : AresAutomation.AresAutomationBase
     IDesiredAnalysisResultFactory desiredAnalysisResultFactory,
     IPlannerServiceRepo plannerServiceRepo,
     IPlannerTransactionProvider plannerTransactionProvider,
-    IAnalyzerTransactionProvider analyzerTransactionProvider)
+    IAnalyzerTransactionProvider analyzerTransactionProvider,
+    ICampaignTemplatePersistenceService campaignTemplatePersistenceService)
   {
     _desiredAnalysisResultFactory = desiredAnalysisResultFactory;
     _coreContextFactory = coreContextFactory;
@@ -63,6 +65,7 @@ public class AutomationService : AresAutomation.AresAutomationBase
     _plannerServiceRepo = plannerServiceRepo;
     _plannerTransactionProvider = plannerTransactionProvider;
     _analyzerTransactionProvider = analyzerTransactionProvider;
+    _campaignTemplatePersistenceService = campaignTemplatePersistenceService;
   }
 
   public override async Task<ProjectsResponse> GetAllProjects(Empty request, ServerCallContext? context)
@@ -77,83 +80,35 @@ public class AutomationService : AresAutomation.AresAutomationBase
   public override async Task<GetAllCampaignsResponse> GetAllCampaigns(GetAllCampaignsRequest request, ServerCallContext? context)
   {
     var campaignResponse = new GetAllCampaignsResponse();
-    foreach(var file in Directory.EnumerateFiles(AresConfig.TemplatePath, "*.json"))
-    {
-      try
-      {
-        var contents = await File.ReadAllTextAsync(file);
-        var campaignTemplate = JsonSerializer.Deserialize<CampaignTemplate>(contents, _serializerSettings);
-        if(campaignTemplate is not null)
-        {
-          var summary = new CampaignTemplateSummary { CampaignName  = campaignTemplate.Name, UniqueId = campaignTemplate.UniqueId };
-          campaignResponse.Campaigns.Add(summary);
-        }
-          
-
-        else
-          throw new Exception("Deserialization of campaign template failed");
-      }
-
-      catch(Exception ex)
-      {
-        HandleNotification("Error Loading Campaign Template", $"{file} - {ex.Message}", NotificationSeverityEnum.Error);
-      }
-    }
+    var campaigns = await _campaignTemplatePersistenceService.GetSummariesAsync(context?.CancellationToken ?? CancellationToken.None);
+    campaignResponse.Campaigns.AddRange(campaigns);
 
     return campaignResponse;
   }
 
-  public override Task<BoolValue> CampaignExists(CampaignRequest request, ServerCallContext? context)
[…]

[…]
