### 2026-07-11 daa781c
Campaign template import/export support. Added legacy transformer to support the pre command template rework template imports.
