+++ Ares.Core.Tests/Campaigns/CampaignTemplateTransferServiceTests.cs
@@ -1,0 +1,266 @@
+using Ares.Core.Campaigns;
+using Ares.Core.CustomCommands;
+using Ares.Core.Device.Repos;
+using Ares.Core.EntityConfigurations.Helpers;
+using Ares.Datamodel.Planning;
+using Ares.Datamodel.Templates;
+using Microsoft.Data.Sqlite;
+using Microsoft.EntityFrameworkCore;
+using Moq;
+using System.Text.Json;
+
+namespace Ares.Core.Tests.Campaigns;
+
+internal class CampaignTemplateTransferServiceTests
+{
+  private SqliteConnection _connection;
+  private IDbContextFactory<CoreDatabaseContext> _contextFactory;
+  private CampaignTemplatePersistenceService _persistenceService;
+  private CampaignTemplateTransferService _transferService;
+
+  [SetUp]
+  public async Task SetUp()
+  {
+    DatabaseRuntimeEnvironment.DatabaseProvider = "Sqlite";
+    _connection = new SqliteConnection("Data Source=:memory:");
+    await _connection.OpenAsync();
+    var options = new DbContextOptionsBuilder<CoreDatabaseContext>().UseSqlite(_connection).Options;
+    _contextFactory = new TestContextFactory(options);
+    await using var context = await _contextFactory.CreateDbContextAsync();
+    await context.Database.EnsureCreatedAsync();
+    _persistenceService = new CampaignTemplatePersistenceService(_contextFactory);
+    var customCommands = new Mock<ICustomCommandPersistenceService>();
+    customCommands.Setup(service => service.GetCommandsAsync()).ReturnsAsync([]);
+    var deviceRepo = new Mock<IAresDeviceRepo>();
+    deviceRepo.Setup(repo => repo.GetAll()).Returns([]);
+    _transferService = new CampaignTemplateTransferService(
+      _persistenceService,
+      customCommands.Object,
+      deviceRepo.Object,
+      _contextFactory);
+  }
+
+  [TearDown]
+  public async Task TearDown()
+    => await _connection.DisposeAsync();
+
+  [Test]
+  public async Task ExportAndImport_RoundTripsCurrentCampaignWithFreshGraphIds()
+  {
+    var original = CreateCampaign("Round Trip");
+    await _persistenceService.AddAsync(original);
+    var export = await _transferService.ExportAsync(original.UniqueId);
+
+    var result = await _transferService.ImportAsync(export!.Json);
+    var importedCommands = result.Template.ExperimentTemplate.StepTemplates.Single().CommandTemplates.OrderBy(command => command.Index).ToArray();
+    var originalCommands = original.ExperimentTemplate.StepTemplates.Single().CommandTemplates.OrderBy(command => command.Index).ToArray();
+
+    using(Assert.EnterMultipleScope())
+    {
+      Assert.That(export.Template.UniqueId, Is.EqualTo(original.UniqueId));
+      Assert.That(export.SuggestedFileName, Is.EqualTo("Round Trip.json"));
+      Assert.That(result.Template.UniqueId, Is.Not.EqualTo(original.UniqueId));
+      Assert.That(result.Template.Name, Is.EqualTo("Round Trip (Imported)"));
+      Assert.That(importedCommands[0].UniqueId, Is.Not.EqualTo(originalCommands[0].UniqueId));
+      Assert.That(importedCommands[0].CommandTypeCase, Is.EqualTo(CommandTemplate.CommandTypeOneofCase.DeviceCommand));
+      Assert.That(importedCommands[1].CommandTypeCase, Is.EqualTo(CommandTemplate.CommandTypeOneofCase.SystemCommand));
[…]
