### 2026-07-11 5fe5a07
Fixed campaign edit bug.
