### 2026-07-12 d3c2cfb
Fixed the analyzer validator not validating the commands if they were not device commands.
