+++ Ares.Core/Validation/Validators/GoodAnalyzerValidator.cs
@@ -1,4 +1,5 @@
 ﻿using Ares.Core.Analyzing;
+using Ares.Core.Execution.Executors;
 using Ares.Core.Execution.Extensions;
 using Ares.Datamodel;
 using Ares.Datamodel.Connection;
@@ -9,8 +10,15 @@ namespace Ares.Core.Validation.Validators;
 public static class GoodAnalyzerValidator
 {
   public static async Task<ValidationResult> Validate(ExperimentTemplate experimentTemplate, ExperimentTemplate? startupTemplate, IAnalyzerRepo analyzerRepo)
+    => await Validate(experimentTemplate, startupTemplate, analyzerRepo, new Dictionary<string, AresValueSchema>(StringComparer.OrdinalIgnoreCase));
+
+  public static async Task<ValidationResult> Validate(
+    ExperimentTemplate experimentTemplate,
+    ExperimentTemplate? startupTemplate,
+    IAnalyzerRepo analyzerRepo,
+    IReadOnlyDictionary<string, AresValueSchema> customCommandOutputSchemas)
   {
-    if(experimentTemplate.AnalyzerId is null)
+    if(!experimentTemplate.HasAnalyzerId || string.IsNullOrWhiteSpace(experimentTemplate.AnalyzerId))
       return new ValidationResult(true);
 
     var analyzer = analyzerRepo.GetAnalyzerById(experimentTemplate.AnalyzerId);
@@ -41,52 +49,120 @@ public static class GoodAnalyzerValidator
     }
 
     var inputSchema = new AresStructSchema();
+    var mappingErrors = new List<string>();
+    var customSchemas = ToCaseInsensitiveLookup(customCommandOutputSchemas);
 
     foreach(var map in experimentTemplate.AnalyzerMaps)
     {
-      var splitVarName = map.Value.Split('.');
-      if(!splitVarName.Any())
+      var splitVarName = map.Value.Split('.', StringSplitOptions.RemoveEmptyEntries);
+      if(splitVarName.Length == 0)
+      {
+        mappingErrors.Add($"Analyzer input '{map.Key}' does not reference a command output.");
         continue;
+      }
 
       var matchingCommand = outputCommands.FirstOrDefault(cmd => cmd.HasOutputVarName && cmd.OutputVarName.Equals(splitVarName.First()));
 
       if(matchingCommand is null)
+      {
+        mappingErrors.Add($"Analyzer input '{map.Key}' references '{map.Value}', but output variable '{splitVarName[0]}' is unavailable.");
         continue;
+      }
 
-      var cmdSchema = matchingCommand.DeviceCommand?.Metadata?.OutputMetadata?.DataSchema;
+      var cmdSchema = ResolveOutputSchema(matchingCommand, customSchemas, out var schemaError);
       if(cmdSchema is null)
+      {
+        mappingErrors.Add($"Analyzer input '{map.Key}' references '{map.Value}', but {schemaError}");
         continue;
+      }
 
       var matchingSchema = FindNestedSchema(cmdSchema, splitVarName.Skip(1).ToArray());
-      if(matchingSchema is not null)
-        inputSchema.Fields[map.Key] = matchingSchema.Clone();
+      if(matchingSchema is null)
+      {
+        mappingErrors.Add($"Analyzer input '{map.Key}' references '{map.Value}', but that output member is unavailable.");
+        continue;
+      }
+      if(matchingSchema.Type is AresDataType.Unit or AresDataType.UnspecifiedType)
+      {
+        mappingErrors.Add($"Analyzer input '{map.Key}' references '{map.Value}', but that output member does not have a usable schema.");
+        continue;
+      }
+
+      inputSchema.Fields[map.Key] = matchingSchema.Clone();
     }
 
     var result = await analyzer.ValidateInputs(inputSchema);
[…]
