+++ Ares.Core.Tests/Validation/GoodAnalyzerValidatorTests.cs
@@ -1,8 +1,12 @@
 using Ares.Core.Analyzing;
+using Ares.Core.CustomCommands;
+using Ares.Core.Execution.Executors;
+using Ares.Core.Validation.Campaign;
 using Ares.Core.Validation.Validators;
 using Ares.Datamodel;
 using Ares.Datamodel.Analyzing;
 using Ares.Datamodel.Analyzing.Remote;
+using Ares.Datamodel.Automation;
 using Ares.Datamodel.Connection;
 using Ares.Datamodel.Templates;
 using Moq;
@@ -51,6 +55,186 @@ internal class GoodAnalyzerValidatorTests
 
     Assert.That(result.Success, Is.False);
     Assert.That(result.Messages, Does.Contain("Missing AnalyzerInput"));
+    Assert.That(result.Messages, Has.Some.Contains("output member is unavailable"));
+  }
+
+  [Test]
+  public async Task Validate_MapsSystemCommandOutputSchema()
+  {
+    var command = new CommandTemplate
+    {
+      OutputVarName = "timestamp",
+      SystemCommand = new SystemCommand { Operation = SystemOperation.GetTimestamp }
+    };
+    var experimentTemplate = CreateExperimentTemplate(command, "timestamp");
+    var analyzer = CreateAnalyzer(
+      inputSchema => inputSchema.Fields.TryGetValue("AnalyzerInput", out var input)
+        && input.Type == AresDataType.Timestamp,
+      AresDataType.Timestamp);
+
+    var result = await GoodAnalyzerValidator.Validate(
+      experimentTemplate,
+      null,
+      CreateAnalyzerRepo(analyzer.Object).Object);
+
+    Assert.That(result.Success, Is.True);
+  }
+
+  [Test]
+  public async Task Validate_MapsCustomCommandScalarOutputByStableIdCaseInsensitively()
+  {
+    const string savedId = "A2A403BA-1284-4E1F-8C67-082AF05E879B";
+    var experimentTemplate = CreateExperimentTemplate(CreateCustomCommand(savedId), "result");
+    var analyzer = CreateAnalyzer(inputSchema =>
+      inputSchema.Fields.TryGetValue("AnalyzerInput", out var input)
+      && input.Type == AresDataType.Number);
+    var schemas = new Dictionary<string, AresValueSchema>
+    {
+      [savedId.ToLowerInvariant()] = new() { Type = AresDataType.Number }
+    };
+
+    var result = await GoodAnalyzerValidator.Validate(
+      experimentTemplate,
+      null,
+      CreateAnalyzerRepo(analyzer.Object).Object,
+      schemas);
+
+    Assert.That(result.Success, Is.True);
+  }
+
+  [Test]
+  public async Task Validate_MapsCustomCommandNestedStructOutput()
+  {
+    const string commandId = "custom-command-id";
+    var experimentTemplate = CreateExperimentTemplate(CreateCustomCommand(commandId), "result.outer.inner");
+    var analyzer = CreateAnalyzer(inputSchema =>
+      inputSchema.Fields.TryGetValue("AnalyzerInput", out var input)
+      && input.Type == AresDataType.Number);
+
+    var result = await GoodAnalyzerValidator.Validate(
+      experimentTemplate,
+      null,
+      CreateAnalyzerRepo(analyzer.Object).Object,
+      new Dictionary<string, AresValueSchema> { [commandId] = CreateNestedOutputSchema() });
+
+    Assert.That(result.Success, Is.True);
+  }
+
+  [Test]
+  public async Task Validate_FailsClearlyWhenCustomCommandIsMissing()
+  {
+    var experimentTemplate = CreateExperimentTemplate(CreateCustomCommand("deleted-command"), "result");
+    var analyzer = CreateAnalyzer(_ => false);
+
[…]
