### 2026-07-13 9746354
Small bug fix to ensure we ignore "Add" button clicks for campaign tags when the text box is empty.
