+++ Ares.Core/DataManagement/DataMappers/CampaignDatasetGenerator.cs
@@ -6,6 +6,7 @@ using Ares.Datamodel.Planning;
 using Ares.Datamodel.Templates;
 using Google.Protobuf.WellKnownTypes;
 using Microsoft.EntityFrameworkCore;
+using static System.Runtime.InteropServices.JavaScript.JSType;
 
 namespace Ares.Core.DataManagement.DataMappers;
 
@@ -99,7 +100,8 @@ public class CampaignDatasetGenerator(IDbContextFactory<CoreDatabaseContext> _db
       Name = "Experiments"
     };
 
-    dataset.Columns.AddRange(CreateExperimentColumns(experiments, cancellationToken));
+    var createdColumns = CreateExperimentColumns(experiments, cancellationToken);
+    dataset.Columns.AddRange(createdColumns);
     dataset.Rows.AddRange(experiments.Select((experiment, index) => CreateExperimentRow(experiment, index + 1, cancellationToken)));
     return dataset;
   }
@@ -143,16 +145,19 @@ public class CampaignDatasetGenerator(IDbContextFactory<CoreDatabaseContext> _db
 
   private static IEnumerable<AresDataColumn> CreateExperimentColumns(IEnumerable<ExperimentExecutionSummary> experiments, CancellationToken cancellationToken)
   {
-    return
-    [
+    var columns = new List<AresDataColumn>
+    {
       CreateColumn(ExperimentNumberColumnName, AresDataType.Int),
       CreateColumn(ExperimentTemplateColumnName, AresDataType.String, optional: true),
       CreateColumn(TimeStartedColumnName, AresDataType.Timestamp, optional: true),
       CreateColumn(TimeFinishedColumnName, AresDataType.Timestamp, optional: true),
       CreateColumn(DurationSecondsColumnName, AresDataType.Number, optional: true),
-      CreateColumn(AnalysisResultColumnName, AresDataType.Number, optional: true),
-      .. CreateExperimentDynamicColumns(experiments, cancellationToken)
-    ];
+      CreateColumn(AnalysisResultColumnName, AresDataType.Number, optional: true)
+    };
+
+    var dynamicColumns = CreateExperimentDynamicColumns(experiments, cancellationToken);
+    columns.AddRange(dynamicColumns);
+    return columns;
   }
 
   private static IEnumerable<AresDataColumn> CreateCommandColumns(IEnumerable<CommandRecord> commandRecords, CancellationToken cancellationToken)
@@ -237,19 +242,19 @@ public class CampaignDatasetGenerator(IDbContextFactory<CoreDatabaseContext> _db
         }
       }
 
-      var parameters = experiment.ExperimentOverview?.Parameters.OrderBy(GetParameterColumnName)
-        ?? Enumerable.Empty<Parameter>();
-      foreach(var parameter in parameters)
+      var cleanedParameters = CleanParametersOfDuplicateNames(experiment, cancellationToken);
+
+      foreach(var item in cleanedParameters)
       {
         cancellationToken.ThrowIfCancellationRequested();
-        var parameterValue = parameter.GetValue();
+        var parameterValue = item.Item1.GetValue();
+
         if(parameterValue is null)
           continue;
 
-        foreach(var flattenedField in AresValueFlattener.Flatten(GetParameterColumnName(parameter), parameterValue))
-        {
+        // Use item.UniqueName here instead of calling GetParameterColumnName again
+        foreach(var flattenedField in AresValueFlattener.Flatten(item.Item2, parameterValue))
           TryAddDynamicColumn(columns, flattenedField.Key, flattenedField.Value);
-        }
       }
     }
 
[…]
