### 2026-07-15 e0fe19c
Added stop condition for planner lead experimenting.
