### 2026-07-15 13117cb
update readme to point to doc site. (#205)
Co-authored-by: Jalen <jshell@dcscorp.com>
