### 2026-07-16 d122cc0
Finished up a method in the execution view model
