### 2026-07-17 31216f5
Planner lead campaigns (#206)
* Added stop condition for planner lead experimenting.

* Updated datamodel version

* Finished up a method in the execution view model
