+++ Ares.Core/Execution/StopConditions/PlannerLead/IPlannerLeadStopConditionFactory.cs
@@ -1,0 +1,6 @@
+﻿namespace Ares.Core.Execution.StopConditions.PlannerLead;
+
+public interface IPlannerLeadStopConditionFactory
+{
+  PlannerLeadStopCondition Create();
+}
