+++ .github/workflows/dotnet-release.yaml
@@ -30,23 +30,23 @@ jobs:
 
     steps:
     - name: Checkout Code
-      uses: actions/checkout@v4
+      uses: actions/checkout@v7
       with:
         ref: ${{ github.event.release.tag_name }}
       
     - name: Checkout Device Drivers
-      uses: actions/checkout@v4
+      uses: actions/checkout@v7
       with:
         repository: 'AFRL-ARES/Ares.Device.Drivers'
         path: 'DriversRepo'
 
     - name: Set up .NET 10.x
-      uses: actions/setup-dotnet@v4
+      uses: actions/setup-dotnet@v6
       with:
         dotnet-version: '10.x'
         
     - name: Setup Node.js Environment
-      uses: actions/setup-node@v4
+      uses: actions/setup-node@v7
       with:
         node-version: 24
 
@@ -91,7 +91,7 @@ jobs:
         zip -r ../${{ env.ZIP_FILE_NAME }} .
     
     - name: Upload Published Binaries to Release
-      uses: softprops/action-gh-release@v2
+      uses: softprops/action-gh-release@v3
       with:
         files: ${{ env.ZIP_FILE_NAME }}
       env:
