### 2026-07-20 97cbfcb
Updated datamodel version, fixed existing build errors
