### 2026-07-20 679b612
Custom commands (#201)
* Initial command builder placeholder

* Added a schema designer and integrated it into the custom command builder

* Added entity configs to the custom commands

* Slight visual tweaks

* Replace script placeholder with actual editor

* Fix little script bug where comment right after function definition was marked as an error.

* Script editor is now just responsible just for editing the script body. The function definition is rendered separately.

* Added another card for diagnostics feedback of the command.

* Css update for schema designer and command builder

* Prev
[…]
