### 2026-07-22 f571141
Updated tests and fixed build errors associated with the new changes
