+++ Ares.Core/DataManagement/DataMappers/CampaignDatasetGenerator.cs
@@ -1,12 +1,10 @@
 using Ares.Datamodel;
 using Ares.Datamodel.Analyzing;
-using Ares.Datamodel.Analyzing.Remote;
 using Ares.Datamodel.Extensions;
 using Ares.Datamodel.Planning;
 using Ares.Datamodel.Templates;
 using Google.Protobuf.WellKnownTypes;
 using Microsoft.EntityFrameworkCore;
-using static System.Runtime.InteropServices.JavaScript.JSType;
 
 namespace Ares.Core.DataManagement.DataMappers;
 
@@ -22,10 +20,10 @@ public class CampaignDatasetGenerator(IDbContextFactory<CoreDatabaseContext> _db
   private const string TimeStartedColumnName = "Time Started";
   private const string TimeFinishedColumnName = "Time Finished";
   private const string DurationSecondsColumnName = "Duration Seconds";
-  private const string AnalysisResultColumnName = "Analysis Result";
   private const string StatusColumnName = "Status";
   private const string SuccessColumnName = "Success";
   private const string ErrorColumnName = "Error";
+  private const string PlanNumberColumnName = "Plan Number";
   private const string PlannerNameColumnName = "Planner Name";
   private const string PlannerTypeColumnName = "Planner Type";
   private const string PlannerVersionColumnName = "Planner Version";
@@ -35,7 +33,7 @@ public class CampaignDatasetGenerator(IDbContextFactory<CoreDatabaseContext> _db
   private const string TimeRequestSentColumnName = "Time Request Sent";
   private const string TimeResponseReceivedColumnName = "Time Response Received";
   private const string OutcomeColumnName = "Outcome";
-  private const string AnalysisResultsColumnName = "Analysis Results";
+  private const string ObjectiveStatusColumnName = "Objective Status";
   private const string ResultColumnName = "Result";
   private const string InputColumnPrefix = "Input.";
   private const string OutputColumnPrefix = "Output.";
@@ -137,7 +135,8 @@ public class CampaignDatasetGenerator(IDbContextFactory<CoreDatabaseContext> _db
     };
 
     dataset.Columns.AddRange(CreatePlannerTransactionColumns(records, cancellationToken));
-    dataset.Rows.AddRange(records.Select(record => CreatePlannerTransactionRow(record, cancellationToken)));
+    dataset.Rows.AddRange(records.SelectMany(record => CreatePlannerTransactionRows(record, cancellationToken)));
+
     return dataset;
   }
 
@@ -161,8 +160,7 @@ public class CampaignDatasetGenerator(IDbContextFactory<CoreDatabaseContext> _db
       CreateColumn(ExperimentTemplateColumnName, AresDataType.String, optional: true),
       CreateColumn(TimeStartedColumnName, AresDataType.Timestamp, optional: true),
       CreateColumn(TimeFinishedColumnName, AresDataType.Timestamp, optional: true),
-      CreateColumn(DurationSecondsColumnName, AresDataType.Number, optional: true),
-      CreateColumn(AnalysisResultColumnName, AresDataType.Number, optional: true)
+      CreateColumn(DurationSecondsColumnName, AresDataType.Number, optional: true)
     };
 
     var dynamicColumns = CreateExperimentDynamicColumns(experiments, cancellationToken);
@@ -196,13 +194,13 @@ public class CampaignDatasetGenerator(IDbContextFactory<CoreDatabaseContext> _db
     return
     [
       CreateColumn(ExperimentNumberColumnName, AresDataType.Int),
+      CreateColumn(PlanNumberColumnName, AresDataType.Int, optional: true),
[…]
