### 2026-07-22 04fa0b7
Fixed duplicate named fields not being properly ordered
