### 2026-07-22 2ba16a6
Fixed some problems causing unit test failures
