### 2026-07-22 ce4466b
Working towards full functionality. restoring an old field for backwards compatability
