### 2026-07-23 4dd99be
More work. Might change directions on this.
