### 2026-07-23 2ed0ad9
Fixed a bug in which analyzer maps would not save, preventing proper campaign building from taking place.
