+++ AresService.Migrations.Postgres/Migrations/AresDbContextModelSnapshot.cs
@@ -17,7 +17,7 @@ namespace AresService.Migrations.Postgres.Migrations
         {
 #pragma warning disable 612, 618
             modelBuilder
-                .HasAnnotation("ProductVersion", "10.0.8")
+                .HasAnnotation("ProductVersion", "10.0.10")
                 .HasAnnotation("Relational:MaxIdentifierLength", 63);
 
             NpgsqlModelBuilderExtensions.UseIdentityByDefaultColumns(modelBuilder);
@@ -1493,6 +1493,9 @@ namespace AresService.Migrations.Postgres.Migrations
                         .HasColumnType("uuid");
 
                     b.Property<string>("AnalyzerId")
+                        .HasColumnType("text");
+
+                    b.Property<string>("AnalyzerMaps")
                         .HasColumnType("text");
 
                     b.Property<Guid?>("CampaignCloseoutId")
