### 2026-07-24 f82cd40
Working towards fixing device state data outputs
