+++ Ares.Core/Device/State/Export/ExportStreamProviders/CombinedDeviceStateExportStreamProvider.cs
@@ -1,8 +1,8 @@
 ﻿using System.Globalization;
+using Ares.Core.DataManagement.DataMappers;
+using Ares.Datamodel;
 using Ares.Datamodel.Device;
-using CsvHelper;
 using CsvHelper.Configuration;
-using Google.Protobuf.Collections;
 
 namespace Ares.Core.Device.State.Export.ExportStreamProviders;
 
@@ -11,16 +11,16 @@ namespace Ares.Core.Device.State.Export.ExportStreamProviders;
 /// </summary>
 public class CombinedDeviceStateExportStreamProvider : IDeviceStateExportStreamProvider
 {
-  readonly IEnumerable<IDeviceStateDataProvider> _dataProviders;
+  readonly DeviceStateDatasetGenerator _stateGetter;
 
   public string Name => "Single File CSV Exporter";
 
-  public CombinedDeviceStateExportStreamProvider(IEnumerable<IDeviceStateDataProvider> dataProviders)
+  public CombinedDeviceStateExportStreamProvider(DeviceStateDatasetGenerator dataProviders)
   {
-    _dataProviders = dataProviders;
+    _stateGetter = dataProviders;
   }
 
-  public async Task<ExportStateStream> Export(DeviceStateRequestFilter request)
+  public async Task<MemoryStream> Export(DeviceStateRequestFilter request)
   {
     var config = new CsvConfiguration(CultureInfo.InvariantCulture)
     {
@@ -29,102 +29,13 @@ public class CombinedDeviceStateExportStreamProvider : IDeviceStateExportStreamP
 
     var exportData = await GetStateExportData(request);
 
-    var stream = new MemoryStream();
-    var writer = new StreamWriter(stream);
-    using(var csv = new CsvWriter(writer, config, true))
-    {
-      csv.Context.TypeConverterCache.AddConverter<RepeatedField<string>>(new StringCollectionConverter());
-      WriteHeader(csv, exportData);
-
-      var lines = exportData.SelectMany(d => d.ExportLines);
-      var timeOrderedLines = lines.OrderBy(l => l.Timestamp);
-      //The local time offset won't change in between records 
-      var localTimeOffset = DateTimeOffset.Now.Offset.ToString();
-
-      if(!timeOrderedLines.Any())
-        return new ExportStateStream(stream, "csv");
-
-      var interval = request.Interval?.ToTimeSpan() ?? default;
-      if(interval.TotalMilliseconds < 1)
-      {
-        foreach(var line in timeOrderedLines)
-        {
-          WriteRecords(csv, exportData, line.Timestamp, localTimeOffset);
-          await csv.NextRecordAsync();
-        }
-      }
-      else
-      {
-        var startTime = request.Start?.ToDateTime() ?? timeOrderedLines.First().Timestamp;
-        var endTime = request.End?.ToDateTime() ?? timeOrderedLines.Last().Timestamp;
-        for(var i = startTime; i <= endTime; i += interval)
-        {
-          WriteRecords(csv, exportData, i, localTimeOffset);
-          await csv.NextRecordAsync();
-        }
-      }
-      await csv.FlushAsync();
-
-    }
-    stream.Seek(0, SeekOrigin.Begin);
-
-    return new ExportStateStream(stream, "csv");
+    var stream = AresDatasetExporter.ExportZip(exportData);
+    return stream;
   }
 
-  private void WriteRecords(CsvWriter csvWriter, IEnumerable<SingleDeviceStateExportData> exportDataCollection, DateTime timestamp, string localTimeOffset)
+  private async Task<AresDataset[]> GetStateExportData(DeviceStateRequestFilter request)
   {
-    csvWriter.WriteField(timestamp.ToString("yyyy-MM-ddTHH:mm:ss.fff"));
-    csvWriter.WriteField(localTimeOffset);
[…]
