### 2026-07-24 db31265
Reducing old state data code, moving towards standard functionality
