+++ Ares.Core/Device/State/Export/ExportStreamProviders/IDeviceStateExportStreamProvider.cs
@@ -17,10 +17,4 @@ public interface IDeviceStateExportStreamProvider
   ///         If start/end times are not given, the first and last timestamped state will be used</param>
   /// <returns>A stream that can be written/downloaded/etc. Not guaranteed to be at position 0.</returns>
   public Task<MemoryStream> Export(DeviceStateRequestFilter request);
-
-
-  /// <summary>
-  /// The name of the exporter that can be presented to the user if they had multiple exporters to pick from
-  /// </summary>
-  public string Name { get; }
 }
