### 2026-07-24 fd1856c
Stanford bug fixes (#212)
* Working towards fixing device state data outputs

* Reducing old state data code, moving towards standard functionality

* Removed unnecessary remote device state schema assignment
