### 2026-08-05 57e8ef3
Updated ARES to check for datamodel compatability with external services.
