+++ AresService.Migrations.Sqlite/Migrations/20260803174831_GeneralSettingsUpdates_AresDbContext.Designer.cs
@@ -1,0 +1,2277 @@
+﻿// <auto-generated />
+using System;
+using AresService.Data;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Migrations;
+using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
+
+#nullable disable
+
+namespace AresService.Migrations.Sqlite.Migrations.AresDb
+{
+    [DbContext(typeof(AresDbContext))]
+    [Migration("20260803174831_GeneralSettingsUpdates_AresDbContext")]
+    partial class GeneralSettingsUpdates_AresDbContext
+    {
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
+        {
+#pragma warning disable 612, 618
+            modelBuilder.HasAnnotation("ProductVersion", "10.0.10");
+
+            modelBuilder.Entity("Ares.Datamodel.AnalysisOverview", b =>
+                {
+                    b.Property<Guid>("UniqueId")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("TEXT");
+
+                    b.Property<string>("AnalyzerInfo")
+                        .HasColumnType("TEXT");
+
+                    b.Property<DateTime>("CreationTime")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("TEXT")
+                        .HasDefaultValueSql("DATETIME('now')");
+
+                    b.Property<Guid?>("ExperimentOverviewId")
+                        .HasColumnType("TEXT");
+
+                    b.Property<DateTime>("LastModified")
+                        .ValueGeneratedOnAddOrUpdate()
+                        .HasColumnType("TEXT")
+                        .HasDefaultValueSql("DATETIME('now')");
+
+                    b.Property<double>("Result")
+                        .HasColumnType("REAL");
+
+                    b.HasKey("UniqueId");
+
+                    b.HasIndex("ExperimentOverviewId")
+                        .IsUnique();
+
+                    b.ToTable("AnalysisOverview");
+                });
+
+            modelBuilder.Entity("Ares.Datamodel.Analyzing.Analysis", b =>
+                {
+                    b.Property<Guid>("UniqueId")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("TEXT");
+
+                    b.Property<int>("AnalysisOutcome")
+                        .HasColumnType("INTEGER");
+
+                    b.Property<DateTime>("CreationTime")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("TEXT")
+                        .HasDefaultValueSql("DATETIME('now')");
+
+                    b.Property<string>("ErrorString")
+                        .HasColumnType("TEXT");
+
+                    b.Property<DateTime>("LastModified")
+                        .ValueGeneratedOnAddOrUpdate()
+                        .HasColumnType("TEXT")
+                        .HasDefaultValueSql("DATETIME('now')");
+
+                    b.Property<float>("Result")
+                        .HasColumnType("REAL");
+
+                    b.HasKey("UniqueId");
+
+                    b.ToTable("Analyses", (string)null);
+                });
+
+            modelBuilder.Entity("Ares.Datamodel.Analyzing.AnalyzerCapabilities", b =>
+                {
+                    b.Property<Guid>("UniqueId")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("TEXT");
+
+                    b.Property<Guid>("AnalyzerInfoId")
+                        .HasColumnType("TEXT");
+
+                    b.Property<DateTime>("CreationTime")
[…]
