### 2026-08-05 f139f1b
Updated ARES to check for datamodel compatability with external services. (#218)
