### 2026-08-17 #10461 @arthursloan
Hi @espottesmith, so far we've left the `JOSS-Submission` branch untouched since initial submission and made any changes indicated in the reviews elsewhere. With two reviews in should we go ahead and make the changes in the submission repo or will that interfere with the process? Thanks!
