### 2026-08-17 8c54fa8
Built a basic UI addition to handle selecting what objectives are passed to the planner during campaigns. Still need to round out the logic for populating objective info for planners, but currently am successfully seeing analyzers report exactly what objectives they return and then successfully marking what objectives to use for planning and saving that off to experiment templates.
