+++ UI/Features/CampaignEdit/ViewModels/PlanningViewModel.cs
@@ -1,10 +1,11 @@
 using ReactiveUI;
 using System.Collections.ObjectModel;
 using Ares.Datamodel.Templates;
-using Ares.Services;
 using Ares.Datamodel.Planning;
 using Ares.Core.Grpc.Services;
 using UI.Application.Notifications;
+using Ares.Core.Analyzing;
+using Ares.Datamodel;
 
 namespace UI.Features.CampaignEdit.ViewModels;
 
@@ -13,22 +14,39 @@ public class PlanningViewModel : ReactiveObject
   private readonly CampaignTemplate _template;
   private readonly PlannerService _client;
   private readonly IUiNotificationService _notificationService;
+  private readonly IAnalyzer? _selectedAnalyzer;
 
   public PlanningViewModel(CampaignTemplate template, 
     IEnumerable<PlannerServiceInfo> plannerAdapters, 
     PlannerService client,
-    IUiNotificationService notificationService)
+    IUiNotificationService notificationService,
+    IAnalyzer? selectedAnalyzer)
   {
     _template = template;
     _client = client;
     _notificationService = notificationService;
+    _selectedAnalyzer = selectedAnalyzer;
     PlannerAdapters = new ReadOnlyCollection<PlannerServiceInfo>(plannerAdapters.ToList());
     PlannerAllocationEditors = template.PlannableParameters.Select(metadata => new PlannerAllocationEditorViewModel(metadata, template.PlannerAllocations.FirstOrDefault(allocation => allocation.Parameter.Equals(metadata))?.Planner, PlannerAdapters, client, notificationService)).ToArray();
   }
 
-  public IEnumerable<PlannerAllocationEditorViewModel> PlannerAllocationEditors { get; private set; }
+  public async Task UpdateAnalyzerObjectives()
+  {
+    if(_selectedAnalyzer is not null)
+      AvailableObjectives = await _selectedAnalyzer.GetObjectiveOutputs();
 
-  public IEnumerable<PlannerServiceInfo> PlannerAdapters { get; }
+    SelectedObjectives = AvailableObjectives?.Fields.Where(obj => _template.ExperimentTemplate.PlanObjectives.Contains(obj.Key)).ToList() ?? [];
+  }
+
+  public IEnumerable<PlannerAllocationEditorViewModel> PlannerAllocationEditors { get; private set; } = [];
+
+  public IEnumerable<PlannerServiceInfo> PlannerAdapters { get; } = [];
+
+  public AresStructSchema? AvailableObjectives { get; private set; }
+
+  public List<KeyValuePair<string, AresValueSchema>> SelectedObjectives { get; private set; } = [];
+
+  public List<string> IncludedObjectives { get; private set; }
 
   public void Save()
   {
@@ -39,6 +57,8 @@ public class PlanningViewModel : ReactiveObject
       .Select(editor => editor.Save())
       .Where(allocation => allocation is not null)
       .Where(allocation => _template.PlannableParameters.Any(meta => meta.UniqueId == allocation!.Parameter.UniqueId)));
+
+    _template.ExperimentTemplate.PlanObjectives.AddRange(SelectedObjectives.Select(obj => obj.Key));
 
     PlannerAllocationEditors = _template.PlannableParameters
     .Select(metadata => new PlannerAllocationEditorViewModel(metadata, _template.PlannerAllocations
