### 2026-08-17 ebc8ca9
Setting default empty list value to handle warning
