### 2026-08-18 #10461 @espottesmith
You should ideally merge either into `JOSS-Submission` or your main branch so that reviewers can easily see the all of the changes that you've made.
