### 2026-08-18 #10461 @nkleiner
Based on @SissiFeng's feedback it seems the ideal solution would be to merge the commit tagged for release v2.4.0 into the JOSS Submission branch. This would fix both of the blocker issues I believe. Assuming JOSS has no issues with that I am happy to complete that merge.

It's worth clarifying that "ARES OS 2.0" is the name of the software, and is entirely independent of the actual latest release version (in this case v2.4.0). ARES 1.0 was a spiritual predecessor, but realistically was very different software.
