### 2026-08-18 #10461 @espottesmith
Again, you're free to merge your code as you see fit in order to address the reviewers' comments. Please go ahead and merge, if code is ready.

Re: naming, it might be confusing for the paper to say "OS 2.0" and the final version to be "v2.4.0". Perhaps "ARES OS v2" would prevent such confusion?
