### 2026-08-18 #10461 @nkleiner
I've gone ahead and merged develop into the `JOSS-Submission` branch. This means it is now up to date with the latest release version (v2.4.0). I'll discuss with Arthur further regarding naming conventions for the paper.
