### 2026-08-18 d057a09
Updates to support new objectives in planning. Successfully saw a full closed loop run, where the analyzer provided a schema defining it's expected objective outputs as well as the objective outputs themselves, and the planner was able to successfully receive a list of objective outputs.
