+++ Ares.Core.Tests/Execution/CampaignExecutorTests.cs
@@ -56,7 +56,7 @@ internal class CampaignExecutorTests
       .Setup(helper => helper.TryResolveParameters(
         It.IsAny<IEnumerable<PlannerAllocation>>(),
         It.IsAny<RequestMetadata>(),
-        It.IsAny<IEnumerable<Parameter>>(),
+        It.IsAny<ExperimentTemplate>(),
         It.IsAny<IEnumerable<Datamodel.Analyzing.AnalysisResponse>>(),
         It.IsAny<IEnumerable<ExperimentOverview>>(),
         It.IsAny<int>(),
@@ -220,7 +220,7 @@ internal class CampaignExecutorTests
     _planningHelper.Verify(helper => helper.TryResolveParameters(
       It.IsAny<IEnumerable<PlannerAllocation>>(),
       It.IsAny<RequestMetadata>(),
-      It.IsAny<IEnumerable<Parameter>>(),
+      It.IsAny<ExperimentTemplate>(),
       It.IsAny<IEnumerable<Datamodel.Analyzing.AnalysisResponse>>(),
       It.IsAny<IEnumerable<ExperimentOverview>>(),
       It.IsAny<int>(),
