### 2026-08-18 d5a101f
Fixed a problem in the plan request construction. Now when no objective outputs are selected (either because the analyzer doesn't support advertising it's objective outputs or the user simply didn't select any) all objectives are passed to the planner for use.
