### 2026-08-19 b3c4bc1
Fixed a small bug that caused issues displaying analyzer transaction data properly
