### 2026-08-19 e271a44
Fixed a few issues with planners not reconnecting very gracefully. Likely going to do the same for planners, analyzers and devices. Also made the UI for showing the errors much cleaner with an expansion button.
