+++ UI/Features/CampaignEdit/ViewModels/PlannerAllocationEditorViewModel.cs
@@ -12,7 +12,7 @@ public partial class PlannerAllocationEditorViewModel : ReactiveObject
 {
   private readonly PlannerService _plannerClient;
   private readonly IUiNotificationService _notificationService;
-  private PlannerServiceInfo? _selectedAdapter;
+  private PlannerServiceInfo? _selectedService;
 
   public PlannerAllocationEditorViewModel(ParameterMetadata metadata,
     PlannerServiceInfo? plannerInfo,
@@ -27,13 +27,26 @@ public partial class PlannerAllocationEditorViewModel : ReactiveObject
     _plannerClient = plannerClient;
     _notificationService = notificationService;
 
+    // Initialize to empty by default
+    PlannerOptions = Enumerable.Empty<Planner>();
+
     if(plannerInfo is not null)
     {
-      SelectedService = plannerInfo;
-      SelectedPlannerOption = plannerInfo.Capabilities.AvailablePlanners.FirstOrDefault(p => p.PlannerName == metadata.PlannerName);
-    }
+      // Match against the fresh live adapter collection instead of using the stale deserialized snapshot
+      var liveService = plannerArray.FirstOrDefault(p => p.UniqueId == plannerInfo.UniqueId || p.Name == plannerInfo.Name);
 
-    PlannerOptions = Enumerable.Empty<Planner>();
+      SelectedService = liveService ?? plannerInfo;
+
+      // Restore specific planner option selection if specified in metadata
+      if(!string.IsNullOrEmpty(metadata.PlannerName))
+      {
+        var matchedOption = PlannerOptions.FirstOrDefault(p => p.PlannerName == metadata.PlannerName);
+        if(matchedOption != null)
+        {
+          SelectedPlannerOption = matchedOption;
+        }
+      }
+    }
   }
 
   public PlannerAllocation? Save()
@@ -61,22 +74,20 @@ public partial class PlannerAllocationEditorViewModel : ReactiveObject
 
     var updatedInfo = await _plannerClient.GetInfo(new PlannerInfoRequest { PlannerId = SelectedService.UniqueId }, null);
 
-    if(updatedInfo.Info.Name != string.Empty)
+    if(updatedInfo?.Info is not null && !string.IsNullOrEmpty(updatedInfo.Info.Name))
     {
-      PlannerOptions = updatedInfo.Info.Capabilities.AvailablePlanners;
-
-      //Not all adapters will have multiple options, if not auto assign the value
-      if(PlannerOptions.Count() == 1)
-        SelectedPlannerOption = SelectedService.Capabilities.AvailablePlanners.First();  
+      // Update the SelectedService reference with the fresh gRPC response
+      SelectedService = updatedInfo.Info;
     }
-
     else
     {
-      var notification = new UiNotificationMessage();
-      notification.Summary = "Assigned Planner Unavailable!";
-      notification.Detail = $"This template uses a planner that ARES no longer has a connection with. The template won't be usable until this is resolved.";
-      notification.Severity = UiNotificationSeverity.Warning;
-      notification.CloseOnClick = true;
+      var notification = new UiNotificationMessage
+      {
+        Summary = "Assigned Planner Unavailable!",
+        Detail = "This template uses a planner that ARES no longer has a connection with. The template won't be usable until this is resolved.",
+        Severity = UiNotificationSeverity.Warning,
+        CloseOnClick = true
+      };
 
       _notificationService.Notify(notification);
     }
@@ -84,18 +95,13 @@ public partial class PlannerAllocationEditorViewModel : ReactiveObject
 
   public PlannerServiceInfo? SelectedService
   {
-    get => _selectedAdapter;
-
[…]
