### 2026-08-19 234429e
Fixed typos and other errors in paper.md
