+++ paper.md
@@ -2,9 +2,9 @@
 title: 'ARES OS 2.0: An Orchestration Software Suite for Autonomous Experimentation Systems and Self-Driving Labs'
 tags:
  - Automation
- - Autonomoy
+ - Autonomy
  - Self-driving lab
- - Self-driving Laboritories
+ - Self-driving Laboratories
  - Autonomous Experimentation
  
 authors:
@@ -25,7 +25,7 @@ authors:
    affiliation: 5
  - name: Daylond Hooper
    affiliation: 5
- - name: Benji Maruyajma
+ - name: Benji Maruyama
    orcid: 0000-0002-3832-628X
    affiliation: 1
 
@@ -46,35 +46,35 @@ bibliography: paper.bib
 ---
 
 # Summary
-`ARES OS 2.0` (hereinafter `ARES OS`) is an open-source software suite to enable laboratory automation and closed-loop autonomous experimentation. Its function is to orchestrate experimental actions and data handoff between lab equipment, analysis routines, and experimental planning modules through a service-oriented architecture. `ARES OS` is abstracted to apply to general experimental flows common in materials science, chemistry, and biology and related disciplines. The core of `ARES OS` provides central control over all modules, along with the heavy lifting of UI creation, data management, and experimental design tools. `ARES OS` modules communicate with the core software over `protobuf` and `gRPC`, allowing them to be language-agnostic and user-creatable. This allows users to easily implement modules that control experimental hardware, process collected data , or plan experiments to meet their specific research needs. `ARES OS` lowers the barrier to entry for researchers to build their own self-driving labs, allowing them to focus on scientific programming for their use case and reducing the effort and time needed to bring an autonomous experimentation system online. 
+`ARES OS 2.0` (hereinafter `ARES OS`) is an open-source software suite to enable laboratory automation and closed-loop autonomous experimentation. Its function is to orchestrate experimental actions and data handoff between lab equipment, analysis routines, and experimental planning modules through a service-oriented architecture. `ARES OS` is abstracted to apply to general experimental flows common in materials science, chemistry, biology, and related disciplines. The core of `ARES OS` provides central control over all modules, along with the heavy lifting of UI creation, data management, and experimental design tools. `ARES OS` modules communicate with the core software over `protobuf` and `gRPC`, allowing them to be language-agnostic and user-creatable. This allows users to easily implement modules that control experimental hardware, process collected data, or plan experiments to meet their specific research needs. `ARES OS` lowers the barrier to entry for researchers to build their own self-driving labs, allowing them to focus on scientific programming for their use case and reducing the effort and time needed to bring an autonomous experimentation system online. 
 
 # Statement of Need
-Research and technology development in the physical sciences has historically been a slow, expensive, and labor-intensive process. To overcome these issues and accelerate the pace of discovery, researchers across a variety of fields have started a revolution in how science is done: Autonomous Experimentation (AE). AE, also called self-driving labs (SDLs), combines robotic high throughput experimentation (HTE) techniques with in situ and in-line analysis methods, and artificial intelligence/machine learning (AI/ML) planning routines to autonomously plan, execute, and analyze experiments in pursuit of a user defined goal [@stach:2021; @abolhasani:2023], with the objecting of making scientific research faster, better, and cheaper. This process flow is shown in \autoref{fig:fig_1}. AE systems have been demonstrated to provide faster research progress, lower experimental variability, and a reduced number of experiments to reach a goal compared to traditional manual planning and experimentation [@stach:2021]. Our group published the first autonomous experimentation system, ARES, for materials in 2016, and `ARES OS` has been in development since [@nikolaev:2016].
[…]
