### 2026-08-19 #10461 @arthursloan
Just pushed an updated version of paper.md that fixes typos and other minor errors that the reviewers notes and that we've found since the initial submission.
