### 2026-08-21 08f3dad
Analysis improvements (#224)
* Updated ARES to support the new analysis capabilities. Still some work to do

* Fixed the display charts for analysis to support displaying multiple objectives

* Updated tests and fixed build errors associated with the new changes

* Working towards full functionality. restoring an old field for backwards compatability

* More work. Might change directions on this.

* Forgot an entity config

* Saving place

* Built a basic UI addition to handle selecting what objectives are passed to the planner during campaigns. Still need to round out the logic for populating
[…]
