+++ Ares.Core/CoreDatabaseContext.cs
@@ -70,6 +70,9 @@ public class CoreDatabaseContext : DbContext
     // Utilities
     configurationBuilder.Properties<Timestamp>().HaveConversion<AresTimestampConverter>();
 
+    // Ignores
+    configurationBuilder.IgnoreAny<Objective>();
+
     base.ConfigureConventions(configurationBuilder);
   }
 }
