+++ Ares.Core/Analyzing/AnalyzerBase.cs
@@ -69,11 +69,13 @@ public abstract class AnalyzerBase : IAnalyzer
     }
   }
 
-  public abstract Task<Analysis> Analyze(AnalysisRequest request, CancellationToken cancellationToken);
+  public abstract Task<AnalysisResponse> Analyze(AnalysisRequest request, CancellationToken cancellationToken);
 
-  public abstract Task<Analysis> Analyze(AnalysisRequest request, AresStruct settings, CancellationToken cancellationToken);
+  public abstract Task<AnalysisResponse> Analyze(AnalysisRequest request, AresStruct settings, CancellationToken cancellationToken);
 
   public abstract Task<AresStructSchema> GetParameters(CancellationToken cancellationToken);
+
+  public abstract Task<AresStructSchema> GetObjectiveOutputs(CancellationToken cancellationToken);
 
   private static ParameterValidationResult ValidateParameterTypes(KeyValuePair<string, AresValueSchema> analyzerField, AresStructSchema parameters)
   {

[…]
