### 2026-08-25 d6d9c85
Updated migrations for the database, uncommented the database saving logic for the SiLA devices, and updated to the latest Tecan packages.
