+++ AresService.Migrations.SqlServer/Migrations/AresDbContextModelSnapshot.cs
@@ -904,6 +904,45 @@ namespace AresService.Migrations.SqlServer.Migrations
                     b.ToTable("SerialConnection");
                 });
 
+            modelBuilder.Entity("Ares.Datamodel.Device.SilaDeviceConfig", b =>
+                {
+                    b.Property<Guid>("UniqueId")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("uniqueidentifier");
+
+                    b.Property<string>("Address")
+                        .HasColumnType("nvarchar(max)");
+
+                    b.Property<DateTime>("CreationTime")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("datetime2")
+                        .HasDefaultValueSql("getdate()");
+
+                    b.Property<string>("Description")
+                        .HasColumnType("nvarchar(max)");
+
+                    b.Property<DateTime>("LastModified")
+                        .ValueGeneratedOnAddOrUpdate()
+                        .HasColumnType("datetime2")
+                        .HasDefaultValueSql("getdate()");
+
+                    b.Property<string>("ServerName")
+                        .HasColumnType("nvarchar(max)");
+
+                    b.Property<string>("Type")
+                        .HasColumnType("nvarchar(max)");
+
+                    b.Property<string>("VendorUri")
+                        .HasColumnType("nvarchar(max)");
+
+                    b.Property<string>("Version")
+                        .HasColumnType("nvarchar(max)");
+
+                    b.HasKey("UniqueId");
+
+                    b.ToTable("SilaConfigs");
+                });
+
             modelBuilder.Entity("Ares.Datamodel.DeviceErrorHandlingConfig", b =>
                 {
                     b.Property<Guid>("UniqueId")

[…]
