### 2026-08-25 3d7218e
Fixing a potential null reference crash in the beta release
