### 2026-08-31 61a74f5
Sila devices are now loading from the database and display an auto generated GUI that gives the user an immediate understanding of the device.
