+++ Ares.Core/Device/Sila/SilaDeviceManager.cs
@@ -1,11 +1,11 @@
 ﻿using Ares.Core.Device.Repos;
-using Ares.Core.Execution;
+using Ares.Core.Device.State.Logging;
 using Ares.Datamodel;
 using Ares.Datamodel.Device;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
+using System.Net;
 using Tecan.Sila2;
-using Tecan.Sila2.DynamicClient;
-using static System.Runtime.InteropServices.JavaScript.JSType;
 
 namespace Ares.Core.Device.Sila;
 
@@ -14,14 +14,22 @@ public class SilaDeviceManager : ISilaDeviceManager
   private readonly IDbContextFactory<CoreDatabaseContext> _dbContextFactory;
   private readonly SilaClient _silaClient;
   private readonly IAresDeviceRepo _deviceRepo;
+  private readonly List<SilaDeviceMonitor> _monitors;
+  private readonly ILoggerFactory _loggerFactory;
+  private readonly StateLoggerManager _stateLoggerManager;
 
   public SilaDeviceManager(SilaClient client, 
     IAresDeviceRepo deviceRepo, 
-    IDbContextFactory<CoreDatabaseContext> dbContextFactory)
+    IDbContextFactory<CoreDatabaseContext> dbContextFactory,
+    ILoggerFactory loggerFactory,
+    StateLoggerManager stateLoggerManager)
   {
     _silaClient = client;
     _deviceRepo = deviceRepo;
     _dbContextFactory = dbContextFactory;
+    _loggerFactory = loggerFactory;
+    _monitors = new List<SilaDeviceMonitor>();
+    _stateLoggerManager = stateLoggerManager;
   }
 
   public async Task<SilaDevice?> Create(ServerData data)
@@ -88,5 +96,38 @@ public class SilaDeviceManager : ISilaDeviceManager
 
     var silaDevice = new SilaDevice(data, deviceInfo, _silaClient);
     return silaDevice;
+  }
+
+  public async Task LoadSilaDevices()
+  {
+    using var ctx = _dbContextFactory.CreateDbContext();
+    var configs = await ctx.SilaConfigs.ToArrayAsync();
+    var devices = await Task.WhenAll(configs.Select(LoadExistingSilaDevice));
+    var nonNullDevices = devices.OfType<SilaDevice>().ToArray();
+    foreach(var device in nonNullDevices)
+    {
+      _deviceRepo.AddOrUpdate(device);
+      var monitor = new SilaDeviceMonitor(device, _loggerFactory.CreateLogger<SilaDeviceMonitor>());
+      _monitors.Add(monitor);
+
+      await _stateLoggerManager.SetupLogger(device);
+    }
+  }
+
+  public async Task<SilaDevice?> LoadExistingSilaDevice(SilaDeviceConfig deviceConfig)
+  {
+    var parsed = IPEndPoint.TryParse(deviceConfig.Address, out var endPoint);
+
+    if(!parsed || endPoint is null)
+      return null;
+
+    var serverData = _silaClient.TryConnectToServer(endPoint.Address.ToString(), endPoint.Port);
+
+    if(serverData is null)
+      return null;
+
+    var device = ConfigToDevice(deviceConfig, serverData);
+    await device.Activate(CancellationToken.None);
+    return device;
   }
 }
