+++ Ares.Core/Device/Sila/ISilaDeviceManager.cs
@@ -7,4 +7,5 @@ public interface ISilaDeviceManager
   Task<SilaDevice?> Create(ServerData data);
   Task<IEnumerable<ServerData>> UpdateAvailableSilaDevices();
   Task<SilaDevice?> Create(string address, int port);
+  Task LoadSilaDevices();
 }
