+++ UI/Application/Devices/Repos/DeviceControlViewModelRepo.cs
@@ -9,6 +9,7 @@ using System.Reactive.Disposables;
 using System.Reactive.Disposables.Fluent;
 using UI.Features.Devices;
 using UI.Features.Devices.Remote.Factory;
+using UI.Features.Devices.Sila.Factory;
 
 namespace UI.Application.Devices.Repos
 {
@@ -18,18 +19,21 @@ namespace UI.Application.Devices.Repos
     private readonly IDeviceAdapterRepository _deviceAdapterRepo;
     private readonly IAresDeviceViewModelFactory _factory;
     private readonly IRemoteDeviceControlViewModelFactory _remoteVmFactory;
+    private readonly ISilaDeviceControlViewModelFactory _silaVmFactory;
     private readonly SourceCache<IDeviceUnitControlViewModel, string> _viewModelCache = new(vm => vm.DeviceId);
     private readonly CompositeDisposable _cleanup = new();
 
     public DeviceControlViewModelRepo(IAresDeviceProvider deviceProvider,
       IAresDeviceViewModelFactory factory,
       IRemoteDeviceControlViewModelFactory remoteVmFactory,
-      IDeviceAdapterRepository deviceAdapterRepo)
+      IDeviceAdapterRepository deviceAdapterRepo,
+      ISilaDeviceControlViewModelFactory silaVmFactory)
     {
       _deviceProvider = deviceProvider;
       _factory = factory;
       _deviceAdapterRepo = deviceAdapterRepo;
       _remoteVmFactory = remoteVmFactory;
+      _silaVmFactory = silaVmFactory;
     }
 
     public void Initialize()
@@ -37,6 +41,13 @@ namespace UI.Application.Devices.Repos
       _deviceProvider.Connect()
         .Filter(IsPluginDevice)
         .Transform(_factory.CreateUnitControlViewModel)
+        .DisposeMany()
+        .PopulateInto(_viewModelCache)
+        .DisposeWith(_cleanup);
+
+      _deviceProvider.Connect()
+        .Filter(d => d is SilaDevice)
+        .Transform(d => _silaVmFactory.Create((SilaDevice)d))
         .DisposeMany()
         .PopulateInto(_viewModelCache)
         .DisposeWith(_cleanup);
