+++ UI/Features/Devices/Sila/Factory/ISilaDeviceControlViewModelFactory.cs
@@ -1,0 +1,10 @@
+﻿using Ares.Core.Device.Sila;
+using Ares.Toolkit.Device.UI;
+using UI.Application.Devices;
+
+namespace UI.Features.Devices.Sila.Factory;
+
+public interface ISilaDeviceControlViewModelFactory
+{
+  IDeviceUnitControlViewModel Create(SilaDevice silaDevice);
+}

[…]
