### 2026-09-01 65f177d
Some visual changes to the SiLA device UI. Shortened down some of the display names for a better UI experience.
